% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_bayesian_model_parameters}
\alias{add_bayesian_model_parameters}
\alias{add_bayesian_model_parameters,bayesian_alpha_test_strategy-method}
\alias{add_bayesian_model_parameters,ss_backtest_config-method}
\title{Add Bayesian Model Parameters}
\usage{
add_bayesian_model_parameters(
  object,
  user_priors = NULL,
  prior_derivation_control = NULL,
  brms_control = NULL
)

\S4method{add_bayesian_model_parameters}{bayesian_alpha_test_strategy}(
  object,
  user_priors = NULL,
  prior_derivation_control = NULL,
  brms_control = NULL
)

\S4method{add_bayesian_model_parameters}{ss_backtest_config}(
  object,
  user_priors = NULL,
  prior_derivation_control = NULL,
  brms_control = NULL
)
}
\arguments{
\item{object}{The object to which Bayesian model parameters will be added.}

\item{user_priors}{An optional object of class \code{brmsprior}.}

\item{prior_derivation_control}{An optional list containing prior derivation control parameters.}

\item{brms_control}{An optional list of parameters for \code{brms::brm}.}
}
\value{
The updated object with the \code{bayesian_model_parameters} added.
}
\description{
Generic function to add Bayesian model parameters.
}
