% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_concentration_constraint_policy}
\alias{add_concentration_constraint_policy}
\title{Add Concentration Constraint Policy}
\usage{
add_concentration_constraint_policy(
  portfolio_policies_obj,
  benchmark = NULL,
  max_abs_active_individual_weight = NULL,
  max_abs_active_group_weight = NULL
)
}
\arguments{
\item{portfolio_policies_obj}{A \code{portfolio_policies} object to which the concentration constraint policy will be added.}

\item{benchmark}{A character string representing benchmark based on which concentration policy will be applied.}

\item{max_abs_active_individual_weight}{A number indicating the absolute weight differential from benchmark weights.}

\item{max_abs_active_group_weight}{A named vector indicating absolute group (sector) weight differentials in relation to the benchmark.}
}
\value{
The updated \code{portfolio_policies} object with the added liquidity constraint policy.
}
\description{
Adds a concentration constraint policy to a \code{portfolio_policies} object.
}
\examples{
portfolio <- create_portfolio_policies()
portfolio <- add_concentration_constraint_policy(portfolio, benchmark = "IBOV", max_abs_active_individual_weight = 0.05, max_abs_active_group_weight = c(Sector = 0.1, Subsector = 0.5))

}
