% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_cov_est_method}
\alias{add_cov_est_method}
\alias{add_cov_est_method,sb_backtest_config,cov_est_method-method}
\alias{add_cov_est_method,sb_backtest_config,missing-method}
\alias{add_cov_est_method,port_backtest_config,cov_est_method-method}
\alias{add_cov_est_method,port_backtest_config,missing-method}
\title{Add covariance estimation method to a backtest configuration}
\usage{
add_cov_est_method(object, cov_est_method, ...)

\S4method{add_cov_est_method}{sb_backtest_config,cov_est_method}(object, cov_est_method, ...)

\S4method{add_cov_est_method}{sb_backtest_config,missing}(
  object,
  cov_est_method,
  cov_estimation_method = "sample",
  cov_matrix_sample_size = 36,
  active_returns = TRUE,
  cov_matrix_benchmark = NULL,
  ...
)

\S4method{add_cov_est_method}{port_backtest_config,cov_est_method}(object, cov_est_method, ...)

\S4method{add_cov_est_method}{port_backtest_config,missing}(
  object,
  cov_est_method,
  cov_estimation_method = "sample",
  cov_matrix_sample_size = 252,
  active_returns = TRUE,
  cov_matrix_benchmark = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{port_backtest_config}.}

\item{cov_est_method}{An existing object of class \code{cov_est_method}.}

\item{...}{Additional arguments.}

\item{cov_estimation_method}{A character string representing the covariance estimation method. Must be one of \code{"sample"}, \code{"ewma"}, \code{"cc"}, \code{"pca1"}, \code{"pca2"}, \code{"shrink_id"} or \code{"shrink_cc"}.}

\item{cov_matrix_sample_size}{Number of periods to subset return sample when estimating the covariance matrix. A high number provides}

\item{active_returns}{logical. If \code{TRUE}, the covariance matrix is estimated using active returns. If \code{FALSE}, raw returns are used.}

\item{cov_matrix_benchmark}{A character string representing the benchmark for covariance matrix estimation. This is used when \code{cov_estimation_method} is \code{"shrink_id"} or \code{"shrink_cc"}.}
}
\value{
An updated object of class \code{sb_backtest_config} or \code{port_backtest_config} with the \code{cov_est_method} added.
}
\description{
This function allows either directly adding a pre-existing \code{cov_est_method} object or creating one dynamically by passing additional arguments.
When \code{cov_est_method} is not provided, a new one will be created using the values for \code{cov_estimation_method}, \code{cov_matrix_sample_size}, and \code{active_returns}, passed via the \code{...} argument.
}
\section{Functions}{
\itemize{
\item \code{add_cov_est_method(
  object = sb_backtest_config,
  cov_est_method = cov_est_method
)}: Add existing \code{cov_est_method} object to a \code{sb_backtest_config} object.

This method allows to add an already existing \code{cov_est_method} object to an \code{sb_backtest_config}.

\item \code{add_cov_est_method(object = sb_backtest_config, cov_est_method = missing)}: Create a \code{cov_est_method} object to a \code{sb_backtest_config} object.

This method allows to dynamically create a \code{cov_est_method} object and add to \code{sb_backtest_config}.

\item \code{add_cov_est_method(
  object = port_backtest_config,
  cov_est_method = cov_est_method
)}: Add existing \code{cov_est_method} object to a \code{port_backtest_config} object.

This method allows to add an already existing \code{cov_est_method} object to an \code{port_backtest_config}.

\item \code{add_cov_est_method(object = port_backtest_config, cov_est_method = missing)}: Create a \code{cov_est_method} object to a \code{port_backtest_config} object.

This method allows to dynamically create a \code{cov_est_method} object and add to \code{port_backtest_config}.

}}
