% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_hrp_parameters}
\alias{add_hrp_parameters}
\alias{add_hrp_parameters,sb_backtest_config,hrp_parameters-method}
\alias{add_hrp_parameters,sb_backtest_config,missing-method}
\alias{add_hrp_parameters,port_backtest_config,hrp_parameters-method}
\alias{add_hrp_parameters,port_backtest_config,missing-method}
\title{Add hrp_parameters to a backtest config}
\usage{
add_hrp_parameters(object, hrp_params, ...)

\S4method{add_hrp_parameters}{sb_backtest_config,hrp_parameters}(object, hrp_params, level = NULL, ...)

\S4method{add_hrp_parameters}{sb_backtest_config,missing}(
  object,
  hrp_params,
  linkage = "single",
  exp_ret_score_tilt = "none",
  exp_ret_score_tilt_eta = NULL,
  level = NULL,
  ...
)

\S4method{add_hrp_parameters}{port_backtest_config,hrp_parameters}(object, hrp_params, level = NULL, ...)

\S4method{add_hrp_parameters}{port_backtest_config,missing}(
  object,
  hrp_params,
  linkage = "single",
  exp_ret_score_tilt = "none",
  exp_ret_score_tilt_eta = NULL,
  level = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{sb_backtest_config} or \code{port_backtest_config}.}

\item{hrp_params}{An object of class \code{hrp_parameters}, or missing if a new object is to be created.}

\item{...}{Additional arguments used to create a new \code{hrp_parameters} object when \code{hrp_params} is missing.
These arguments must include:
\itemize{
\item \strong{linkage}:  A character indicating the linkage method to use for hierarchical clustering.
}}

\item{linkage}{Character indicating the linkage method to use for hierarchical clustering.}
}
\value{
An updated object of class \code{sb_backtest_config} or \code{port_backtest_config} with
the \code{hrp_parameters} added.
}
\description{
This function allows either directly adding a pre-existing \code{hrp_parameters} object
or creating one dynamically by passing additional arguments.
}
\section{Functions}{
\itemize{
\item \code{add_hrp_parameters(object = sb_backtest_config, hrp_params = hrp_parameters)}: Add an existing \code{hrp_parameters} object to a \code{sb_backtest_config} object.

\item \code{add_hrp_parameters(object = sb_backtest_config, hrp_params = missing)}: Dynamically create a \code{hrp_parameters} object and add it to a \code{sb_backtest_config} object.

\item \code{add_hrp_parameters(object = port_backtest_config, hrp_params = hrp_parameters)}: Add an existing \code{hrp_parameters} object to a \code{port_backtest_config} object.

\item \code{add_hrp_parameters(object = port_backtest_config, hrp_params = missing)}: Dynamically create a \code{hrp_parameters} object and add it to a \code{port_backtest_config} object.

}}
