% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_hyper_grid_domain}
\alias{add_hyper_grid_domain}
\alias{add_hyper_grid_domain,tuning_strategy,hyper_grid_domain-method}
\alias{add_hyper_grid_domain,sb_backtest_config,hyper_grid_domain-method}
\title{Add a \code{hyper_grid_domain} Object}
\usage{
add_hyper_grid_domain(object, hyper_grid_domain)

\S4method{add_hyper_grid_domain}{tuning_strategy,hyper_grid_domain}(object, hyper_grid_domain)

\S4method{add_hyper_grid_domain}{sb_backtest_config,hyper_grid_domain}(object, hyper_grid_domain)
}
\arguments{
\item{object}{An object of class \code{sb_backtest_config}. Must already have a
\code{tuning_strategy} attached (via \code{add_tuning_strategy()}) -- this method writes
into \code{object@tuning_strategy@hyper_grid_domain} and will error on a \code{NULL}/unset strategy.}

\item{hyper_grid_domain}{An object of class \code{hyper_grid_domain}.}
}
\value{
The appropriate object with the added \code{hyper_grid_domain}.
}
\description{
This function adds a \code{hyper_grid_domain} S4 class to a \code{tuning_strategy} or a \code{sb_backtest_config}.
It allows users to add a \code{hyper_grid_domain} already built or extracted from other objects.
}
\section{Functions}{
\itemize{
\item \code{add_hyper_grid_domain(
  object = tuning_strategy,
  hyper_grid_domain = hyper_grid_domain
)}: Add \code{hyper_grid_domain} to \code{tuning_strategy} object

\item \code{add_hyper_grid_domain(
  object = sb_backtest_config,
  hyper_grid_domain = hyper_grid_domain
)}: Add \code{hyper_grid_domain} to \code{sb_backtest_config} object

}}
