% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_hyperparameter,hyperparameter_tuning_strategy-method}
\alias{add_hyperparameter,hyperparameter_tuning_strategy-method}
\title{Add Hyperparameter for \code{hyperparameter_tuning_strategy}}
\usage{
\S4method{add_hyperparameter}{hyperparameter_tuning_strategy}(
  object,
  hyperparameter,
  grid = NULL,
  distribution_choice = NULL,
  pars = NULL,
  bounds = NULL
)
}
\arguments{
\item{hyperparameter}{A vector of characters indicating the name of the hyperparameter to be added.}

\item{grid}{A numeric vector or list of numeric vectors for grid search values (only used for grid_search).}

\item{distribution_choice}{A character vector indicating the distribution to sample from (only used for random_search).}

\item{pars}{A numeric named vector or list of numeric named vectors specifying parameter values (only used for random_search).}

\item{bounds}{A vector of length 2 indicating minimum and maximum bounds for each hyperparameter (only used for bayesian_opt).}
}
\description{
Adds hyperparameters to the \code{hyper_grid_domain} inside the \code{hyperparameter_tuning_strategy} based on the specified tuning method.
}
