% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_keras_layer}
\alias{add_keras_layer}
\alias{add_keras_layer,keras_architecture_parameters-method}
\alias{add_keras_layer,sb_backtest_config-method}
\title{Add Layer to Keras Architecture}
\usage{
add_keras_layer(object, units, activation, batch_norm_option)

\S4method{add_keras_layer}{keras_architecture_parameters}(object, units, activation, batch_norm_option)

\S4method{add_keras_layer}{sb_backtest_config}(object, units, activation, batch_norm_option)
}
\arguments{
\item{object}{An object of class \code{sb_backtest_config}}

\item{units}{A numeric value for the number of units in the new layer.}

\item{activation}{A character string specifying the activation function for the new layer (e.g., "relu").}

\item{batch_norm_option}{A character string indicating whether to apply batch normalization for the new layer (e.g., "yes").}
}
\value{
An updated object of class \code{keras_architecture_parameters}.
}
\description{
Method to add a new layer to the Keras architecture.
}
\section{Functions}{
\itemize{
\item \code{add_keras_layer(keras_architecture_parameters)}: Add a keras layer to an object of class \code{keras_architecture_parameters}

\item \code{add_keras_layer(sb_backtest_config)}: Add a keras layer to an object of class \code{sb_backtest_config}

}}
