% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_ml_backtest_config}
\alias{add_ml_backtest_config}
\title{Add one or more ml_backtest_config objects to an ml_metabacktest_config}
\usage{
add_ml_backtest_config(object, ...)
}
\arguments{
\item{object}{An \code{ml_metabacktest_config} object.}

\item{...}{One or more \code{ml_backtest_config} objects to add.}
}
\value{
An updated \code{ml_metabacktest_config} object with added configurations.
}
\description{
Add one or more ml_backtest_config objects to an ml_metabacktest_config
}
