% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_mmaf_parameters}
\alias{add_mmaf_parameters}
\alias{add_mmaf_parameters,sb_backtest_config,mmaf_parameters-method}
\alias{add_mmaf_parameters,sb_backtest_config,missing-method}
\alias{add_mmaf_parameters,port_backtest_config,mmaf_parameters-method}
\alias{add_mmaf_parameters,port_backtest_config,missing-method}
\title{Add mmaf_parameters to a backtest config}
\usage{
add_mmaf_parameters(object, mmaf_params, ...)

\S4method{add_mmaf_parameters}{sb_backtest_config,mmaf_parameters}(object, mmaf_params, ...)

\S4method{add_mmaf_parameters}{sb_backtest_config,missing}(
  object,
  mmaf_params,
  mmaf_method = "bottom_up",
  top_down_proxy_port_method = if (mmaf_method == "top_down") "ew" else NULL,
  mmaf_group_col,
  micro_port_construction_method,
  macro_port_construction_method,
  ...
)

\S4method{add_mmaf_parameters}{port_backtest_config,mmaf_parameters}(object, mmaf_params, ...)

\S4method{add_mmaf_parameters}{port_backtest_config,missing}(
  object,
  mmaf_params,
  mmaf_method = "bottom_up",
  top_down_proxy_port_method = if (mmaf_method == "top_down") "ew" else NULL,
  mmaf_group_col,
  micro_port_construction_method,
  macro_port_construction_method,
  ...
)
}
\arguments{
\item{object}{An object of class \code{sb_backtest_config} or \code{port_backtest_config}.}

\item{mmaf_params}{An object of class \code{mmaf_parameters}, or missing if a new one is to be created.}

\item{...}{Additional arguments used to create a new \code{mmaf_parameters} object when \code{mmaf_params} is missing.}

\item{mmaf_method}{A character indicating the MMAF method to be used. Must be one of 'top_down' or 'bottom_up'.}

\item{top_down_proxy_port_method}{A character indicating the proxy portfolio method for top-down MMAF.}

\item{mmaf_group_col}{A character string (length 1) with the grouping column name.}

\item{micro_port_construction_method}{A character indicating the micro portfolio construction method.}

\item{macro_port_construction_method}{A character indicating the macro portfolio construction method.}
}
\value{
An updated object of class \code{sb_backtest_config} or \code{port_backtest_config} with
the \code{mmaf_parameters} added.
}
\description{
This function allows either directly adding a pre-existing \code{mmaf_parameters} object
or creating one dynamically by passing additional arguments.
}
\section{Functions}{
\itemize{
\item \code{add_mmaf_parameters(object = sb_backtest_config, mmaf_params = mmaf_parameters)}: Add an existing \code{mmaf_parameters} object to a \code{sb_backtest_config} object.

\item \code{add_mmaf_parameters(object = sb_backtest_config, mmaf_params = missing)}: Dynamically create a \code{mmaf_parameters} object and add it to a \code{sb_backtest_config} object.

\item \code{add_mmaf_parameters(
  object = port_backtest_config,
  mmaf_params = mmaf_parameters
)}: Add an existing \code{mmaf_parameters} object to a \code{port_backtest_config} object.

\item \code{add_mmaf_parameters(object = port_backtest_config, mmaf_params = missing)}: Dynamically create a \code{mmaf_parameters} object and add it to a \code{port_backtest_config} object.

}}
