% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_mvo_parameters}
\alias{add_mvo_parameters}
\alias{add_mvo_parameters,sb_backtest_config,mvo_parameters-method}
\alias{add_mvo_parameters,sb_backtest_config,missing-method}
\alias{add_mvo_parameters,port_backtest_config,mvo_parameters-method}
\alias{add_mvo_parameters,port_backtest_config,missing-method}
\title{Add mvo_parameters to a backtest config}
\usage{
add_mvo_parameters(object, mvo_params, ...)

\S4method{add_mvo_parameters}{sb_backtest_config,mvo_parameters}(object, mvo_params, level = NULL, ...)

\S4method{add_mvo_parameters}{sb_backtest_config,missing}(
  object,
  mvo_params,
  opt_method = "random",
  random_ports_method = "sample",
  n_random_ports = 1000,
  opt_objective = "sharpe",
  ridge_pen = NULL,
  n_resamples = 0,
  exp_ret_score_jitter = 0,
  cov_eigval_jitter = 0,
  level = NULL,
  ...
)

\S4method{add_mvo_parameters}{port_backtest_config,mvo_parameters}(object, mvo_params, level = NULL, ...)

\S4method{add_mvo_parameters}{port_backtest_config,missing}(
  object,
  mvo_params,
  opt_method = "random",
  random_ports_method = "sample",
  n_random_ports = 1000,
  opt_objective = "sharpe",
  ridge_pen = NULL,
  n_resamples = 0,
  exp_ret_score_jitter = 0,
  cov_eigval_jitter = 0,
  level = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{sb_backtest_config} or \code{port_backtest_config}.}

\item{mvo_params}{An object of class \code{mvo_parameters}, or missing if a new object is to be created.}

\item{...}{Additional arguments used to create a new \code{mvo_parameters} object when \code{mvo_params} is missing.
These arguments must include:
\itemize{
\item \strong{opt_method}: A character indicating the optimization method.
The only current available method is 'random'.
\item \strong{random_ports_method}: A character string representing the method to generate random portfolios.
Options are 'sample', 'simplex' or 'grid'.
\item \strong{n_random_ports}: Number of random portfolios to generate.
Only needed when \code{opt_method} is 'random'.
\item \strong{opt_objective}: A character indicating the optimization objective.
Possible options are 'return', 'risk' or 'sharpe'.
}}

\item{level}{A character indicating the level to which the parameters should be applied when using 'mmaf' strategy.}

\item{opt_method}{A character indicating the optimization method.}

\item{random_ports_method}{A character string representing the method to generate random portfolios.}

\item{n_random_ports}{Number of random portfolios to generate.}

\item{opt_objective}{A character indicating the optimization objective.}

\item{ridge_pen}{A numeric value representing the ridge penalty to be used in the optimization.}

\item{n_resamples}{A numeric value indicating the number of bootstrap resamples to perform}

\item{exp_ret_score_jitter}{A numeric value indicating the jitter to be applied to the expected return scores}

\item{cov_eigval_jitter}{A numeric value indicating the jitter to be applied to the covariance matrix eigenvalues}
}
\value{
An updated object of class \code{sb_backtest_config} or \code{port_backtest_config} with
the \code{mvo_parameters} added.
}
\description{
This function allows either directly adding a pre-existing \code{mvo_parameters} object
or creating one dynamically by passing additional arguments.
}
\section{Functions}{
\itemize{
\item \code{add_mvo_parameters(object = sb_backtest_config, mvo_params = mvo_parameters)}: Add existing \code{mvo_parameters} object to a \code{sb_backtest_config} object.

\item \code{add_mvo_parameters(object = sb_backtest_config, mvo_params = missing)}: Dynamically create a \code{mvo_parameters} object and add it to a \code{sb_backtest_config} object.

\item \code{add_mvo_parameters(object = port_backtest_config, mvo_params = mvo_parameters)}: Add existing \code{mvo_parameters} object to a \code{port_backtest_config} object.

\item \code{add_mvo_parameters(object = port_backtest_config, mvo_params = missing)}: Dynamically create a \code{mvo_parameters} object and add it to a \code{port_backtest_config} object.

}}
