% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_rp_parameters}
\alias{add_rp_parameters}
\alias{add_rp_parameters,sb_backtest_config,rp_parameters-method}
\alias{add_rp_parameters,sb_backtest_config,missing-method}
\alias{add_rp_parameters,port_backtest_config,rp_parameters-method}
\alias{add_rp_parameters,port_backtest_config,missing-method}
\title{Add rp_parameters to a backtest config}
\usage{
add_rp_parameters(object, rp_params, ...)

\S4method{add_rp_parameters}{sb_backtest_config,rp_parameters}(object, rp_params, ...)

\S4method{add_rp_parameters}{sb_backtest_config,missing}(object, rp_params, rp_method = "cyclical-spinu", ...)

\S4method{add_rp_parameters}{port_backtest_config,rp_parameters}(object, rp_params, ...)

\S4method{add_rp_parameters}{port_backtest_config,missing}(object, rp_params, rp_method = "cyclical-spinu", ...)
}
\arguments{
\item{object}{An object of class \code{sb_backtest_config} or \code{port_backtest_config}.}

\item{rp_params}{An object of class \code{rp_parameters}, or missing if a new object is to be created.}

\item{...}{Additional arguments used to create a new \code{rp_parameters} object when \code{rp_params} is missing.
These arguments must include:
\itemize{
\item \strong{rp_method}: A character indicating the method to compute the risk-parity solution.
}}
}
\value{
An updated object of class \code{sb_backtest_config} or \code{port_backtest_config} with
the \code{rp_parameters} added.
}
\description{
This function allows either directly adding a pre-existing \code{rp_parameters} object
or creating one dynamically by passing additional arguments.
}
\section{Functions}{
\itemize{
\item \code{add_rp_parameters(object = sb_backtest_config, rp_params = rp_parameters)}: Add an existing \code{rp_parameters} object to a \code{sb_backtest_config} object.

\item \code{add_rp_parameters(object = sb_backtest_config, rp_params = missing)}: Dynamically create a \code{rp_parameters} object and add it to a \code{sb_backtest_config} object.

\item \code{add_rp_parameters(object = port_backtest_config, rp_params = rp_parameters)}: Add an existing \code{rp_parameters} object to a \code{port_backtest_config} object.

\item \code{add_rp_parameters(object = port_backtest_config, rp_params = missing)}: Dynamically create a \code{rp_parameters} object and add it to a \code{port_backtest_config} object.

}}
