% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_sb_backtest_results}
\alias{add_sb_backtest_results}
\title{Add one sb_backtest_results object to a port_backtest_config}
\usage{
add_sb_backtest_results(object, ...)
}
\arguments{
\item{object}{A\code{port_backtest_config} object.}

\item{One}{or more \code{sb_backtest_results} objects to add.}
}
\value{
An updated \code{port_backtest_config} object with added sb_backtest_results.
}
\description{
Add one sb_backtest_results object to a port_backtest_config
}
