% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_signal_selection_policy}
\alias{add_signal_selection_policy}
\title{Add Signal Selection Policy}
\usage{
add_signal_selection_policy(
  port_backtest_config_obj,
  chosen_signals = NULL,
  signal_positions = NULL,
  signal_blending_method = NULL,
  chosen_sb_metric = NULL,
  sb_benchmark_weighting_method = NULL,
  max_abs_active_individual_weight = NULL,
  max_abs_active_group_weight = NULL,
  p_correction_method = "none",
  signal_significance_threshold = 0.05,
  data_availability_cutoff = 60,
  priors_type = NULL,
  priors_informative_data = NULL
)
}
\arguments{
\item{port_backtest_config_obj}{A \code{port_backtest_config} object to which the signal selection policy will be added.}

\item{chosen_signals}{A vector of chosen signals.}

\item{signal_positions}{A data structure indicating positions of signals.}

\item{signal_blending_method}{A method for blending signals.}

\item{chosen_sb_metric}{A metric chosen for signal blending.}

\item{max_abs_active_individual_weight}{A number indicating the absolute weight differential from benchmark weights.}

\item{max_abs_active_group_weight}{A named vector indicating absolute group (sector) weight differentials in relation to the benchmark.}

\item{p_correction_method}{Method for p-value correction.}

\item{signal_significance_threshold}{A threshold for signal significance.}

\item{data_availability_cutoff}{A cutoff for data availability.}

\item{priors_type}{Type of priors to use.}

\item{sb_benchmark_weighting}{A weighting strategy for the benchmark.}

\item{chosen_informative_data}{A data structure with informative data.}
}
\value{
The updated \code{port_backtest_config} object with the added signal selection policy.
}
\description{
Adds a signal selection policy to a \code{port_backtest_config} object.
}
\examples{
portfolio <- create_port_backtest_config()
portfolio <- add_signal_selection_policy(portfolio, signal_blending_method = "method1", ...)

}
