% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_slsaf_parameters}
\alias{add_slsaf_parameters}
\alias{add_slsaf_parameters,port_backtest_config,slsaf_parameters-method}
\alias{add_slsaf_parameters,port_backtest_config,missing-method}
\title{Add slsaf_parameters to a backtest config}
\usage{
add_slsaf_parameters(object, slsaf_params, ...)

\S4method{add_slsaf_parameters}{port_backtest_config,slsaf_parameters}(object, slsaf_params, ...)

\S4method{add_slsaf_parameters}{port_backtest_config,missing}(object, slsaf_params, ...)
}
\arguments{
\item{object}{An object of class \code{port_backtest_config}.}

\item{slsaf_params}{An object of class \code{slsaf_parameters}, or missing if a new one is
to be created.}

\item{...}{Additional arguments passed to \code{\link{create_slsaf_parameters}} when
\code{slsaf_params} is missing.}
}
\value{
An updated \code{port_backtest_config} with the \code{slsaf_parameters} added.
}
\description{
Either add an existing \code{slsaf_parameters} object to a \code{port_backtest_config}, or
create one dynamically from the same arguments as
\code{\link{create_slsaf_parameters}}.
}
\section{Functions}{
\itemize{
\item \code{add_slsaf_parameters(
  object = port_backtest_config,
  slsaf_params = slsaf_parameters
)}: Add an existing \code{slsaf_parameters} object.

\item \code{add_slsaf_parameters(object = port_backtest_config, slsaf_params = missing)}: Dynamically create an \code{slsaf_parameters} object and add it.

}}
