% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_ss_backtest_obj}
\alias{add_ss_backtest_obj}
\alias{add_ss_backtest_obj,sb_backtest_config,ss_backtest_config-method}
\alias{add_ss_backtest_obj,sb_backtest_config,ss_backtest_results-method}
\alias{add_ss_backtest_obj,sb_backtest_config,missing-method}
\title{Add a \code{ss_backtest_config} or a \code{ss_backtest_results} to an existing \code{sb_backtest_config}.}
\usage{
add_ss_backtest_obj(object, ss_backtest_obj, ...)

\S4method{add_ss_backtest_obj}{sb_backtest_config,ss_backtest_config}(object, ss_backtest_obj)

\S4method{add_ss_backtest_obj}{sb_backtest_config,ss_backtest_results}(object, ss_backtest_obj)

\S4method{add_ss_backtest_obj}{sb_backtest_config,missing}(
  object,
  data_availability_cutoff,
  initial_sample_size,
  rebalancing_months,
  active_returns = TRUE,
  split_method = "expanding",
  alpha_test_strategy = NULL,
  config_name = "not_identified"
)
}
\arguments{
\item{object}{A \code{sb_backtest_config} object to which a ss_backtest_obj will be added.}

\item{ss_backtest_obj}{An object of class \code{sb_backtest_config}.}

\item{...}{Additional arguments required to create a new \code{ss_backtest_config}, only needed when ss_backtest_obj is \code{missing}.}

\item{ss_backtest_object}{An object of class \code{ss_backtest_config}, \code{ss_backtest_results} or \code{missing}.
If \code{NULL}, additional parameters must be provided to create a new \code{ss_backtest_config}.}
}
\value{
An updated \code{sb_backtest_config} object with the specified or newly created ss_backtest_obj.

The updated \code{sb_backtest_config} object with the provided \code{ss_backtest_obj}.

The updated \code{sb_backtest_config} object with the provided \code{ss_backtest_obj}.

The updated \code{sb_backtest_config} object with the created \code{ss_backtest_config}.
}
\description{
This generic function adds an existing \code{ss_backtest_config} or a \code{ss_backtest_results}
to an \code{sb_backtest_config} object or creates a new \code{ss_backtest_config} if none is provided (i.e., when ss_backtest_obj is \code{missing}).
}
\section{Functions}{
\itemize{
\item \code{add_ss_backtest_obj(
  object = sb_backtest_config,
  ss_backtest_obj = ss_backtest_config
)}: Add an existing ss_backtest_obj to the \code{sb_backtest_config}.

This method adds a pre-existing \code{ss_backtest_config} to the \code{sb_backtest_config}.
It replaces any existing ss_backtest_obj in \code{sb_backtest_config}.

\item \code{add_ss_backtest_obj(
  object = sb_backtest_config,
  ss_backtest_obj = ss_backtest_results
)}: Add an existing ss_backtest_obj to the \code{sb_backtest_config}.

This method adds a pre-existing \code{ss_backtest_results} to the \code{sb_backtest_config}.
It replaces any existing ss_backtest_obj in \code{sb_backtest_config}.

\item \code{add_ss_backtest_obj(object = sb_backtest_config, ss_backtest_obj = missing)}: Add an existing ss_backtest_obj to the \code{sb_backtest_config}.

This method creates a \code{ss_backtest_config} to add to the \code{sb_backtest_config}.

}}
