% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_turnover_constraint_policy}
\alias{add_turnover_constraint_policy}
\title{Add Turnover Constraint Policy}
\usage{
add_turnover_constraint_policy(portfolio_policies_obj, turnover_rules)
}
\arguments{
\item{portfolio_policies_obj}{A \code{portfolio_policies} object to which the liquidity constraint policy will be added.}

\item{turnover_rule}{A list with the following structure, specifying the turnover rule:
#' \itemize{
\item \strong{liquidity_classification:} One of "micro_caps", "small_caps", "mid_caps", "large_caps" or "mega_caps"
\item \strong{turnover_cap:} A number indicating the turnover_cap
\item \strong{top_stock_quantile_buffer:} A number indicating the minimum signal quantile
}}
}
\value{
The updated \code{portfolio_policies} object with the added liquidity constraint policy.
}
\description{
Adds a turnover constraint policy to a \code{portfolio_policies} object.
}
\examples{
portfolio <- create_portfolio_policies()
portfolio <- add_turnover_constraint_policy(portfolio, turnover_rule = list(liquidity_classification = "micro_caps", turnover_cap = 0.01, top_stock_quantile_buffer = 0.66))

}
