% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{alpha_test_strategy-class}
\alias{alpha_test_strategy-class}
\title{alpha_test_strategy Class}
\description{
The alpha_test_strategy class is designed to specify parameters of hypothesis testing regarding
CAPM alpha under a multiple testing framework, with frequentist and bayesian approaches.
In the latter, the user can change the hierarchical model specification and how priors are going to be set.
}
\section{Slots}{

\describe{
\item{\code{signal_significance_threshold}}{A decimal indicating the hypothesis testing zero-alpha null-hypothesis rejection criteria. If one wants to select all chosen_signals,
provide 1. In any case, a signal being selected demands a significant CAPM alpha.}

\item{\code{p_correction_method}}{The method for p-value correction. Possible options are:
\itemize{
\item{"none"}: No correction.
\item{"bayesian"}: When bayesian is set, a hierarchical mixed-effects bayesian linear model is fitted to the data, using the \code{brms} package,
which is an interface to the \code{Stan} probabilistic programming language.
The user can also choose one of the following frequentist methods, which will control Family-Wise Error Rate (FWER) or the False Discovery Rate (FDR).
FDR is less stringent than FWER.
For FWER, possible options are:
\item{"bonferroni"}: Bonferroni correction, which is dominated by Holm's method.
\item{"holm"}: Holm's (1979) method.
\item{"hochberg"}: Hochberg's (1988) method, valid when hypothesis tests are independent or non-negatively associated. Less powerful than Hommel's (1988) method, but
faster to compute.
\item{"hommel"}: Hommel's (1988) method, also valid when hypothesis tests are independent or non-negatively associated, but is more powerful than Hochberg (1988).
For FDR, possible options are:
\item{"BH" or "fdr"}: Benjamini-Hochberg (1995) procedure.
\item{"BY"}: Benjamini-Yekutieli (2001) procedure.
}}

\item{\code{market_factor_proxy}}{A character string indicating the market factor proxy to be used in the CAPM model.
Should correspond to one of the columns in \code{benchmark_returns_df}.}

\item{\code{bayesian_model_parameters}}{An object of class \code{bayesian_model_parameters}, containing the
parameters needed to build the hierarhicical bayesian model and specify its priors.}
}}

