% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-apply_buffer_rule.R
\name{apply_buffer_rule}
\alias{apply_buffer_rule}
\title{Applies a buffer rule based on the creation of a buffer zone for illiquid stocks}
\usage{
apply_buffer_rule(
  signals_m_d_ref,
  top_assets_quantile_buffer,
  portfolio_weights_m_lstd_ref,
  liquidity_m_d_ref,
  liquidity_floor_cutoffs_list,
  buffer_rule,
  verbose = TRUE
)
}
\arguments{
\item{signals_m_d_ref}{A data frame of selected signals generated by the \code{generate_signal_matrix} function.}

\item{portfolio_weights_m_lstd_ref}{A data frame containing columns for id, tickers, dates and weights from the old portfolio (pre-rebalancing)
All tickers in current stock universe must have a unique correspondence in this data frame.}

\item{liquidity_m_d_ref}{A data frame (similar to \code{features_m_df}) containing columns for id, tickers, dates, and one or more market liquidity measures (e.g., inflation-adjusted mean financial volume).
All tickers in current stock universe must have a unique correspondence in this data frame.}

\item{liquidity_floor_cutoffs_list}{A list of named vectors containing cutoff values for liquidity metrics specified in \code{liquidity_m_d_ref}.
The names should match the metrics and values should be the minimum acceptable values (adjust for inflation)
Elements should be: "micro_caps", "small_caps", "mid_caps" and "large_caps"
Classification should be in ascending order (from lest liquid to most liquid) for all metrics.}

\item{buffer_rule}{Character string specifying the buffer rule target (one of "micro_caps", "small_caps", "mid_caps", "large_caps", "mega_caps").
Stocks that are less liquid than \code{buffer_rule} and that are included in the buffer_zone will be promoted.
Stocks meeting this rule will be included in the \code{filtered_stock_universe} even if they fail the liquidity floor rule.}

\item{top_stocks_quantile_buffer}{Numeric value indicating the quantile to apply the buffer rule.
Stocks in this quantile and pre-rebalancing portfolio will be included in the \code{filtered_stock_universe} if they meet the buffer rule policy.}
}
\description{
This rule is applied to reduce portfolio turnover and transaction costs.
}
