% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{as.data.frame,meta_dataframe-method}
\alias{as.data.frame,meta_dataframe-method}
\title{Coerce a meta_dataframe Object to Data Frame}
\usage{
\S4method{as.data.frame}{meta_dataframe}(x)
}
\arguments{
\item{x}{An object of class \code{meta_dataframe}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{data.frame} containing the contents of the \code{data} slot.
}
\description{
This method extracts the \code{data} slot from a \code{meta_dataframe} object and returns it as a standard \code{data.frame}.
}
