% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{as.list,concentration_constraint_policy-method}
\alias{as.list,concentration_constraint_policy-method}
\title{Turn a concentration_constraint_policy into a list}
\usage{
\S4method{as.list}{concentration_constraint_policy}(x)
}
\arguments{
\item{x}{An S4 object of class \code{concentration_constraint_policy}.}
}
\value{
A named list containing:
\itemize{
\item \code{benchmark}
\item \code{max_abs_active_individual_weight}
\item \code{max_abs_active_group_weight}
}
}
\description{
Turn a concentration_constraint_policy into a list
}
