% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{as.list,liquidity_constraint_policy-method}
\alias{as.list,liquidity_constraint_policy-method}
\title{as.list Method for liquidity_constraint_policy S4 Class}
\usage{
\S4method{as.list}{liquidity_constraint_policy}(x, ...)
}
\arguments{
\item{x}{A liquidity_constraint_policy S4 object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list with elements \code{liquidity_floor_rule} and \code{liquidity_cap_rules}.
}
\description{
Converts a liquidity_constraint_policy S4 object to a list with elements
\code{liquidity_floor_rule} and \code{liquidity_cap_rules}.
}
