% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{as.list,transaction_costs_parameters-method}
\alias{as.list,transaction_costs_parameters-method}
\title{as.list Method for transaction_costs_parameters S4 Class}
\usage{
\S4method{as.list}{transaction_costs_parameters}(x, ...)
}
\arguments{
\item{x}{A transaction_costs_parameters S4 object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list with elements \code{quantile_range_buffer} and \code{turnover_cap_rules}.
}
\description{
Converts a transaction_costs_parameters S4 object to a list with elements
\code{direct_transaction_cost}, \code{alpha}, \code{lambda} and \code{strategy_aum}.
}
