% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{as.list,turnover_constraint_policy-method}
\alias{as.list,turnover_constraint_policy-method}
\title{Convert Turnover Constraint Policy to List}
\usage{
\S4method{as.list}{turnover_constraint_policy}(x)
}
\arguments{
\item{x}{An S4 object of class \code{turnover_constraint_policy}.}
}
\value{
A named list containing:
\itemize{
\item \code{quantile_range_buffer}
\item \code{turnover_cap_rules}
}
}
\description{
Convert Turnover Constraint Policy to List
}
