% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{assets_meta_xts-class}
\alias{assets_meta_xts-class}
\title{An S4 subclass of meta_xts for asset returns with no holes.}
\description{
In addition to the parent slots, it has:
\itemize{
\item \code{assets}: a character vector with names of asset columns.
\item \code{n_assets}: a numeric value equal to the number of asset columns.
}
}
\section{Slots}{

\describe{
\item{\code{assets}}{Character. Names of the columns (assets).}

\item{\code{n_assets}}{Numeric. Number of asset columns.}
}}

