% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{bake.step_impute_sector}
\alias{bake.step_impute_sector}
\title{Apply imputation during baking}
\usage{
\method{bake}{step_impute_sector}(object, new_data, ...)
}
\description{
Apply imputation during baking
}
