% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{bake.step_winsorize}
\alias{bake.step_winsorize}
\title{Apply winsorization during baking}
\usage{
bake.step_winsorize(object, new_data, ...)
}
\description{
Apply winsorization during baking
}
