% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{bayesian_alpha_test_strategy-class}
\alias{bayesian_alpha_test_strategy-class}
\title{bayesian_alpha_test_strategy Class}
\description{
A subclass of alpha_test_strategy for Bayesian methods.
}
\section{Slots}{

\describe{
\item{\code{bayesian_model_parameters}}{Parameters for the hierarchical Bayesian model.}
}}

