% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{bayesian_model_parameters-class}
\alias{bayesian_model_parameters-class}
\title{bayesian_model_parameters Class}
\description{
A class encapsulating parameters necessary to specify the hierarchical Bayesian model and its priors.
}
\section{Slots}{

\describe{
\item{\code{user_priors}}{An object of class \code{brmsprior} with user-defined priors for the hierarchical Bayesian model.
Should be structured according to the \code{model_spec_theme_level}.}

\item{\code{prior_derivation_control}}{A list of additional parameters for deriving priors when \code{priors_type} is \code{"informative_exogenous_dataset"}.
Should include:
\itemize{
\item \code{half_t_df}: Degrees of freedom for the half-t distribution applied to sd priors.
\item \code{lmer_optimizer}: Optimizer to be used in \code{lme4::lmer} for deriving priors.
Options include: \code{"nloptwrap"}, \code{"bobyqa"}, \code{"Nelder_Mead"}, \code{"nlminbwrap"}.
\item \code{lmer_optimization_objective}: Criteria to be optimized in \code{lme4::lmer} for deriving priors.
Options include: \code{likelihood}, \code{REML}.
}}

\item{\code{brms_control}}{A list of additional parameters to be passed to \code{brms::brm} for MCMC sampling, including:
\itemize{
\item \code{chains}: Number of Markov chains to run (default is 4).
\item \code{iter}: Total number of iterations per chain (default is 2000).
\item \code{warmup}: Number of warmup iterations per chain (default is \code{floor(iter / 2)}).
\item \code{thin}: Thinning interval for MCMC sampling (default is 1).
\item \code{seed}: Seed for reproducibility (default is \code{NA} for random seeding).
\item \code{adapt_delta}: Target acceptance probability for the Hamiltonian Monte Carlo sampler (default is 0.99).
}}
}}

