% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{bayesian_opt_strategy-class}
\alias{bayesian_opt_strategy-class}
\title{Bayesian Opt Tuning Strategy}
\description{
A subclass of \code{tuning_strategy} that implements bayesian optimization.
}
\section{Slots}{

\describe{
\item{\code{tuning_method}}{Character string indicating the hyperparameter tuning method ('grid_search', 'random_search', or 'bayesian_opt').}

\item{\code{validation_sample_size}}{Numeric value representing the size of the validation sample. If provided a decimal, it will be set based on proportion of training sample size.}

\item{\code{chosen_eval_metric}}{Character or NULL, specifying the evaluation metric to be optimized.}

\item{\code{hyper_grid_domain}}{An object of class \code{hyper_grid_domain}, representing the hyperparameter domain based on which tuning will be performed.
It contains a list slot \code{hyperparameter_list} with the hyperparameters relevant to the specified machine learning algorithm.
The structure of this list depends on the specified tuning method:
\itemize{
\item \strong{For grid search:} Must be a list of named vectors:
\item \strong{For random search:} Must be a list of named lists, where each named list contains:
\itemize{
\item \code{distribution_choice}: A character string specifying the distribution (one of "normal", "uniform", "lognormal", "constant").
\item \code{pars}: A named numeric vector of parameters corresponding to the chosen distribution.
\item \code{value}: A numeric value (only present if \code{distribution_choice} is "constant").
}
\item \strong{For Bayesian optimization:} Must be a list of named numeric vectors, each of length 2, representing the boundaries for the hyperparameters.
}}

\item{\code{early_stop}}{Sets a halting criteria to prevent overfitting in xgb and nn.}

\item{\code{n_iter}}{For bayesian_opt, it should be the number of times the ml algorithm will be evaluated after initialization.}

\item{\code{acq}}{Acquisition function for Bayesian optimization: "ucb", "ei", or "poi".}

\item{\code{init_points}}{Number of initial random points for Bayesian optimization.}

\item{\code{k_iter}}{Integer that specifies the number of times to sample eval_function at each Epoch during Bayesian optimization.
If the intention is running in parallel, set k_iter to a multiple of the number of cores. Must be lower and preferably a multiple of n_iter.}
}}

