% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-blend_signals.R
\name{blend_signals}
\alias{blend_signals}
\title{Blend signals based on a given blending method}
\usage{
blend_signals(
  current_date,
  signals_m_df,
  target_m_df = NULL,
  signal_selection_policy,
  backtest_returns_df = NULL,
  selected_benchmark_returns_df = NULL,
  signals_groups_m_d_ref,
  priors_m_df_list = NULL,
  covariance_estimation_method = NULL,
  rp_method = "sample",
  n_random_portfolios = 2000,
  mto_port_objective = "IR",
  eligible_signals = NULL,
  ml_walk_forward_validation_results = NULL,
  signal_weights = NULL,
  upper_quantile_winsorization = 0.975,
  lower_quantile_winsorization = 0.025,
  verbose = TRUE
)
}
\arguments{
\item{signal_selection_policy}{A named list containing objects to blend signals. Elements of the list are:
\itemize{
\item \code{chosen_signals} A vector that indicates which characteristics or ml_algorithms should be used as signals
\item \code{signals_positions} A named vector with same length and names as chosen_signals describing whether positions should be taken "long" or "short"
\item \code{benchmark}  A character describing the benchmark to be used to calculate alphas.
\item \code{signal_blending_method} A signal blending (ensemble) method. It can be accept a portfolio_construction_method, such as "EW", "SW", "RP" or "MTO" (note that cap-weighting
does not make sense for blending signals) or a ml_algorithm ("OLS", "GLMNET", "RF", "XGB" or "NN").
\item \code{final_sb_signal}. In case of  \code{signal_blending_method} being a portfolio_construction_method, this variable indicates which backtest summary should be used to blend
signals. Possible options are: "mean_active_return", "IR", "alpha", "AP" or "treynor". For simplicity, AP will be considered as the t-stat in frequentist version.
\item \code{sb_benchmark_weighting}. A character describing how the signal-blend benchmark should be built, something needed when applying constraints on signal weights.
Possible options are "individual" or "theme", in which case the respective dataframe in \code{groups_m_df_list} will be used to classify signals.
If individual, all signals receive the same weight. Otherwise, themes receive the same weight and signals inside the same group receive the same weight.
This helps when one has very unbalanced classes.
\item \code{data_availability_cutoff} Minimum number of periods needed to make decisions based on backtested returns. When a signal's backtest has a length lower than this number,
its active returns will be assigned as 0.
\item \code{signal_significance_threshold} A decimal indicating the hypothesis testing zero-alpha null-hypothesis rejection criteria. If one wants to select all chosen_signals to blend,
provide 1. In any case, a signal being selected demands a significant CAPM alpha.
\item \code{p_correction_method}: A character describing a method to adjust alpha CAPM p-values.
One of "none", holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "bayesian". In case of bayesian, a bayesian hierarhical mixed effects model is fit in order to
perform inference. When selecting "bayesian"the probability of direction will be used to select signals and a bayesian posterior version of the final_sb_signal
will be used to give weights.
In this case, the user needs to provide the following additional elements:
\item \code{chosen_informative_data}: The dataframe in priors_m_d_ref_list that will be used to build informative priors.
\item \code{set_informative_priors_on}: One of all, intercept or none. If all, priors will be built to all parameters of the mixed effects model, including variance parameters. If intercept,
only location parameters will be assigned priors. Parameters not specified to receive informative priors will receive uninformative priors.
\item \code{max_abs_active_individual_weight} The maximum absolute individual active weights.
\item \code{max_abs_active_group_weight}: The maximum absolute group active weight used for creating group constraints in \code{generate_group_constraints}.
}}

\item{covariance_estimation_method}{One of SAM, EWMA, CC (Constant Correlation), PCA1, PCA2, Shrink_ID or Shrink_CC.
If NULL, blending_method can only be EW or IR}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization when creating signals}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization when creating signals}

\item{selected_signals_corrected_positions_m_upd_ref}{A dataframe containing tickers and signals for the current_date}

\item{selected_signals_backtest_returns_upd_ref}{Optional. A backtest_df with active_returns generated by backtest_portfolio function. It should include individual backtests for each signal in signal_df
A object of class "factorverse_backtest" will also carry the portfolio_construction_method and signal_blending_method
Only necessary if signal_blending_method is equal to either RP or MTO.}

\item{priors_m_d_ref}{Optional. A dataframe with alpha priors for signals, containing priors for chosen_signals. Columns should match those in
groups_m_d_ref}

\item{selected_benchmark_returns_upd_ref}{A dataframe containing date and benchmark returns columns}
}
\description{
Blend signals based on a given blending method
}
