% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr.R
\name{cagr}
\alias{cagr}
\title{Flexible function for calculating Compound Annual Growth Rate (CAGR) with handling for various scenarios, including negative values and NAs}
\usage{
cagr(begin, final, period)
}
\arguments{
\item{begin}{Initial Value}

\item{final}{Final Value}

\item{period}{Number of periods}
}
\value{
A numeric value
}
\description{
Flexible function for calculating Compound Annual Growth Rate (CAGR) with handling for various scenarios, including negative values and NAs
}
\examples{

cagr(1,2,1)

}
