% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr_matrix.R
\name{cagr_matrix}
\alias{cagr_matrix}
\title{Calculate Element-Wise Compound Annual Growth Rate (CAGR) for Matrices}
\usage{
cagr_matrix(matrix_begin, matrix_final, period)
}
\arguments{
\item{matrix_begin}{A numeric matrix containing the initial values.}

\item{matrix_final}{A numeric matrix containing the final values.}

\item{period}{The number of periods.}
}
\value{
A numeric matrix of CAGR values for each element.
}
\description{
Calculate Element-Wise Compound Annual Growth Rate (CAGR) for Matrices
}
\examples{
matrix_begin <- matrix(c(100, 200, 300, 400), nrow = 2)
matrix_final <- matrix(c(150, 250, 350, 450), nrow = 2)
period <- 1
cagr_matrix(matrix_begin, matrix_final, period)
}
