% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_eval_metrics.R
\name{calculate_eval_metrics}
\alias{calculate_eval_metrics}
\title{Calculate Evaluation Metrics}
\usage{
calculate_eval_metrics(
  pred,
  target,
  huber_delta = 1,
  quantile_tau = 0.5,
  chosen_eval_metric = "rmse",
  early_stop = NULL,
  best_iteration = NULL,
  return_error = FALSE
)
}
\arguments{
\item{pred}{Numeric vector of predicted values.}

\item{target}{Numeric vector of actual target values.}

\item{huber_delta}{Numeric scalar, delta parameter for Pseudo-Huber loss calculation.}

\item{quantile_tau}{Numeric scalar, quantile parameter for Pinball loss calculation.}

\item{chosen_eval_metric}{Metric to optimize during tuning: "rss", "rmse", "cp", "mae", "mphe", "mpe", "mape", "hr" and "mb"}

\item{early_stop}{If numeric, will include Halting criteria to prevent overfitting in xgb and nn}

\item{best_iteration}{Best iteration according to early stopping criteria implemented}

\item{return_error}{Should residuals be returned?}
}
\value{
A data frame containing the calculated evaluation metrics:
\itemize{
\item \code{rss}: R-squared (coefficient of determination).
\item \code{cp}: Cross-product of predicted and actual values.
\item \code{rmse}: Root Mean Squared Error.
\item \code{mae}: Mean Absolute Error.
\item \code{mphe}: Mean Pseudo-Huber Error.
\item \code{mpe}: Mean Pinball Error.
\item \code{mape}: Mean Absolute Percentage Error.
\item \code{hr}: Hit Rate (percentage of predictions with correct sign).
\item \code{mb}: Mean Bias (mean of prediction errors).
}
}
\description{
Calculate various evaluation metrics for model prediction performance.
}
