% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_port_metrics.R
\name{calculate_port_metrics}
\alias{calculate_port_metrics}
\title{Calculate Consolidated Portfolio and Benchmark Metrics}
\usage{
calculate_port_metrics(port_weights_m_d_ref, custom_stock_metrics_m_d_ref)
}
\arguments{
\item{port_weights_m_d_ref}{A data frame containing portfolio weights with at least the columns:}

\item{custom_stock_metrics_m_d_ref}{A data frame containing custom stock metrics with at least the columns:}
}
\value{
A data frame with consolidated portfolio metrics. If benchmark metrics are calculated,
the data frame will also include columns with a \code{"bench_"} prefix.
}
\description{
This function joins custom stock metrics to portfolio allocation data and computes
consolidated portfolio metrics by calculating the weighted sum of each metric column.
Optionally, if benchmark weights are provided (i.e. if \code{selected_benchmark} is not \code{NULL}),
the function also computes benchmark metrics and merges them with the portfolio metrics.
}
