% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_port_returns.R
\name{calculate_port_returns}
\alias{calculate_port_returns}
\title{Calculate Portfolio Returns}
\usage{
calculate_port_returns(
  clean_fwd_return_1m_m_d_ref,
  fwd_selected_benchmark_return,
  port_weights_m_d_ref,
  total_cost,
  verbose = TRUE
)
}
\arguments{
\item{clean_fwd_return_1m_m_d_ref}{A data frame containing forward stock returns. Must include
a column named \code{id} (or an equivalent identifier) and a column named \code{fwd_return_1m} for returns.}

\item{fwd_selected_benchmark_return}{Numeric. The forward return of the selected benchmark.}

\item{port_weights_m_d_ref}{A data frame containing portfolio weights. Must include a column named \code{id} (or an equivalent identifier).}

\item{total_cost}{Numeric. The total cost associated with the portfolio.}

\item{verbose}{Logical. If \code{TRUE}, messages will be printed about the calculation process.}
}
\value{
A list containing:
\describe{
\item{\code{transactions_m_d_ref}}{Data frame with the updated forward returns after joining with \code{clean_fwd_return_1m_m_d_ref}.}
\item{\code{fwd_raw_return}}{Numeric. The raw weighted forward return.}
\item{\code{fwd_raw_active_return}}{Numeric. The raw active return.}
\item{\code{fwd_net_return}}{Numeric. The net return (subtracting total costs from the raw return).}
\item{\code{fwd_net_active_return}}{Numeric. The net active return (subtracting the benchmark forward return from \code{fwd_net_return}).}
\item{\code{turnover}}{Numeric. The average of the absolute \code{delta} values.}
}
}
\description{
This function calculates the portfolio's forward returns, active returns, net returns, and turnover,
given stock- and benchmark-level forward returns, transactions information, and cost data.
}
\details{
The function first joins \code{transactions_m_d_ref} with \code{clean_fwd_return_1m_m_d_ref} by \code{id}.
Any missing forward returns are replaced with 0.
It then calculates several return metrics:
\itemize{
\item \strong{Raw Return:} Weighted sum of \code{fwd_return_1m}.
\item \strong{Raw Active Return:} Raw Return minus the benchmark forward return.
\item \strong{Net Return:} Raw Return minus the total cost.
\item \strong{Net Active Return:} Net Return minus the benchmark forward return.
\item \strong{Turnover:} Mean absolute value of \code{delta}.
}
}
