% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_portfolio_returns.R
\name{calculate_portfolio_returns}
\alias{calculate_portfolio_returns}
\title{Calculate Returns from a Portfolio and Update Weights}
\usage{
calculate_portfolio_returns(
  current_date,
  is_rebalancing_month,
  stock_universe_m_d_ref,
  fwd_returns_m_d_ref,
  portfolio_weights_m_d_ref,
  portfolio_weights_m_lstd_ref,
  portfolio_returns_df,
  selected_benchmark_returns_df,
  main_liquidity_metric,
  liquidity_m_d_ref,
  volatility_m_d_ref,
  transaction_costs_list,
  verbose
)
}
\arguments{
\item{current_date}{A Date object representing the current date.}

\item{is_rebalancing_month}{A logical indicating whether the current month is a rebalancing month.}

\item{stock_universe_m_d_ref}{A data frame containing stock universe data with columns for tickers and current weights.}

\item{fwd_returns_m_d_ref}{A data frame containing target stock data with columns for tickers and forward returns.}

\item{portfolio_weights_m_d_ref}{A data frame containing current portfolio weights with columns for tickers and weights.}

\item{portfolio_weights_m_lstd_ref}{A data frame containing portfolio weights from the last period, including delisted stocks.}

\item{portfolio_returns_df}{A data frame to which portfolio returns will be added, including raw and net returns.}

\item{main_liquidity_metric}{A string specifying the liquidity metric used to estimate transaction costs. Default is "mean_volfin_3m".}

\item{liquidity_m_d_ref}{A data frame containing liquidity metrics for the stocks.}

\item{volatility_m_d_ref}{A data frame containing volatility metrics for the stocks.}

\item{transaction_costs_list}{A list containing parameters to estimate transaction costs, including:
\itemize{
\item \code{strategy_aum} - The assets under management for the strategy.
\item \code{alpha} - The BARRA model alpha parameter.
\item \code{lambda} - The BARRA model lambda parameter.
\item \code{direct_transaction_cost} - The direct transaction cost per trade.
}}

\item{benchmark_returns_df}{A data frame containing benchmark returns.}
}
\value{
A list with updated portfolio data, including:
\itemize{
\item \code{portfolio_weights_m_d_ref} - Updated portfolio weights.
\item \code{transactions_m_d_ref} - Data frame with rebalancing information including transaction costs.
\item \code{portfolio_returns_df} - Data frame with updated portfolio returns.
\item \code{updated_portfolio_weights} - Updated portfolio weights after applying returns.
}
}
\description{
This function calculates the returns of a portfolio, including transaction costs and updates the portfolio weights based on the latest data.
}
