% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_trade_orders.R
\name{calculate_trade_orders}
\alias{calculate_trade_orders}
\title{Calculate Trade Orders}
\usage{
calculate_trade_orders(
  merged_port_results_list,
  updated_port_weights_m_lstd_ref,
  liquidity_m_d_ref,
  volatility_m_d_ref,
  main_liquidity_metric,
  strategy_aum,
  verbose = TRUE
)
}
\arguments{
\item{merged_port_results_list}{A list of results from merge_and_rescale_weights.}

\item{updated_port_weights_m_lstd_ref}{A data frame containing the updated portfolio weights; must have a column named \code{tickers}.}

\item{liquidity_m_d_ref}{data frame containing liquidity information; must have a column matching \code{main_liquidity_metric}.}

\item{volatility_m_d_ref}{data frame containing volatility information; must contain \code{daily_vol}.}

\item{main_liquidity_metric}{character string naming the column in \code{liquidity_m_d_ref} with liquidity data.}

\item{strategy_aum}{numeric, the AUM used to size trades.}

\item{verbose}{logical. If TRUE, prints IPO/delisting info to console.}
}
\value{
A data frame with columns including \code{tickers}, \code{bop_port_weights}, \code{eop_port_weights},
\code{delta}, \code{order}, \code{relative_order_size}, and any liquidity/volatility columns.
}
\description{
This function calculates the required trades to move from a beginning-of-period (BOP)
portfolio weight to an end-of-period (EOP) portfolio weight, given liquidity and volatility data.
}
