% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_ml_walk_forward_validation_inputs.R
\name{check_inputs_ml_wf_val}
\alias{check_inputs_ml_wf_val}
\title{Perform validation checks on inputs for ML workflow}
\usage{
check_inputs_ml_wf_val(
  features_m_df,
  target_m_df,
  training_sample_size,
  target_fwd_name,
  validation_sample_size,
  rebalancing_months,
  split_method,
  ml_algorithm,
  custom_objective,
  chosen_eval_metric,
  huber_delta,
  quantile_tau,
  hyper_grid_domain_list,
  tuning_method,
  n_iter,
  k_iter,
  acq,
  init_points,
  early_stop,
  keras_architecture_parameters,
  show_plots,
  verbose = FALSE,
  parallel
)
}
\arguments{
\item{features_m_df}{A data frame or matrix containing features data.}

\item{target_m_df}{A data frame or matrix containing target variable data.}

\item{training_sample_size}{Numeric, size of the training sample.}

\item{target_fwd_name}{Character, name of the forward looking target.}

\item{validation_sample_size}{Numeric, size of the validation sample.}

\item{rebalancing_months}{Numeric, number of months for rebalancing.}

\item{split_method}{Character, method of data splitting (currently only "expanding" is supported).}

\item{ml_algorithm}{Character, choice of machine learning algorithm ("ols", "glmnet", "rf", "xgb", "nn").}

\item{custom_objective}{Character, custom objective function for loss (for "xgb" and "nn" algorithms).}

\item{chosen_eval_metric}{Character, chosen evaluation metric ("rmse", "mae", "cp", "rss", "mphe", "mpe", "hr", "mape").}

\item{huber_delta}{Numeric, delta parameter for Huber loss (for "pseudo_huber_error" custom objective).}

\item{quantile_tau}{Numeric, tau parameter for quantile loss (for "quantile_error" custom objective).}

\item{hyper_grid_domain_list}{List, domain list of hyperparameters for tuning.}

\item{tuning_method}{Character, method of hyperparameter tuning ("random_search", "grid_search", "bayesian_opt").}

\item{n_iter}{number of iterations for tuning.}

\item{k_iter}{number of iterations for k-fold cross-validation.}

\item{acq}{Character, acquisition function for Bayesian optimization ("ucb", "ei", "poi").}

\item{init_points}{Numeric, number of initial points for Bayesian optimization.}

\item{early_stop}{Numeric, number of epochs for early stopping (for "xgb" and "nn" algorithms).}

\item{keras_architecture_parameters}{List, containing units (numeric), n_layers (numeric between 1 and 5), activation_function and nn_optimizer ("Adam" or "RMSProp")}

\item{show_plots}{Logical, whether to show diagnostic plots.}

\item{verbose}{Logical, whether to print verbose output.}

\item{parallel}{Logical, whether to use parallel computation.}
}
\value{
NULL. This function is used for validation and does not return a value; it stops on errors.
}
\description{
This function validates and checks various inputs required for a machine learning workflow.
}
\details{
This function performs comprehensive validation checks on various inputs required for a machine learning workflow.
It validates data formats, correctness of hyperparameters, consistency of dates, and other specific requirements for
different machine learning algorithms.
}
\references{
For more information on machine learning algorithms and their usage, refer to appropriate documentation.
}
\keyword{internal}
\keyword{machine-learning}
\keyword{validation}
