% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_inputs_ss_backtest.R
\name{check_inputs_ss_backtest}
\alias{check_inputs_ss_backtest}
\title{Perform validation checks on inputs for Signal Selection (SS) Workflow}
\usage{
check_inputs_ss_backtest(
  rebalancing_months,
  data_availability_cutoff,
  signals_m_df,
  chosen_signals,
  signal_positions,
  backtest_returns_df,
  selected_benchmark_returns_df,
  p_correction_method,
  signal_significance_threshold,
  priors_m_df,
  priors_type,
  signal_themes_m_df
)
}
\arguments{
\item{rebalancing_months}{Months (numeric) when signal selection should be implemented.}

\item{data_availability_cutoff}{The minimum number of non-NA observations required for a backtest to be considered.}

\item{signals_m_df}{A (meta) data frame with columns including "id", "tickers", "dates", and the selected signals.}

\item{chosen_signals}{A vector with user-defined characteristics to be considered.}

\item{signal_positions}{A named vector with same length and names as chosen_signals describing whether positions should be taken "long" or "short".}

\item{backtest_returns_df}{A data frame with a 'dates' column and remaining columns named according to signals in signals_m_df, containing historical backtested returns.}

\item{selected_benchmark_returns_df}{A data frame with a 'dates' column and a column with benchmark returns, named accordingly.}

\item{p_correction_method}{The method for p-value correction. Possible options are:
\itemize{
\item{"none"}: No correction.
\item{"bayesian"}: When bayesian is set, a hierarchical mixed-effects bayesian linear model is fitted to the data, using the \code{brms} package,
which is an interface to the \code{Stan} probabilistic programming language.
The user can also choose one of the following frequentist methods, which will control Family-Wise Error Rate (FWER) or the False Discovery Rate (FDR).
FDR is less stringent than FWER.
For FWER, possible options are:
\item{"bonferroni"}: Bonferroni correction, which is dominated by Holm's method.
\item{"holm"}: Holm's (1979) method.
\item{"hochberg"}: Hochberg's (1988) method, valid when hypothesis tests are independent or non-negatively associated. Less powerful than Hommel's (1988) method, but
faster to compute.
\item{"hommel"}: Hommel's (1988) method, also valid when hypothesis tests are independent or non-negatively associated, but is more powerful than Hochberg (1988).
For FDR, possible options are:
\item{"BH" or "fdr"}: Benjamini-Hochberg (1995) procedure.
\item{"BY"}: Benjamini-Yekutieli (2001) procedure.
}}

\item{signal_significance_threshold}{A decimal indicating the hypothesis testing zero-alpha null-hypothesis rejection criteria. If one wants to select all chosen_signals,
provide 1. In any case, a signal being selected demands a significant CAPM alpha.}

\item{priors_m_df}{A (meta) data frame with columns including "id", "characteristic/signal", "dates", "theme" (used for clustering in hierarchical bayesian model)
and values for alpha (mean and se), beta (mean and se) and sigma, which are used to build priors.}

\item{signal_themes_m_df}{A (meta) data frame with id, tickers ("signals") and dates column contemplating all signals in \code{signals_m_df} and a "theme" column providing group membership for each signal, which is needed
for defining clusters in bayesian hierarchical model.}
}
\description{
This function validates and checks various inputs required for signal selection in the context of run_ss_backtest function.
}
