% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-choose_prior.R
\name{choose_prior}
\alias{choose_prior}
\title{Choose a prior based on the sample fit to a range of distributions}
\usage{
choose_prior(vector, parameter_class)
}
\arguments{
\item{vector}{A numeric vector of data for which the best distribution will be selected.}

\item{parameter_class}{A character string indicating the type of parameterization. Options are "scale" and "location". This determines which distributions will be considered.}
}
\value{
A list containing:
\item{bic}{A matrix of BIC values for each distribution considered, where each row corresponds to a different distribution.}
\item{distribution}{A character string indicating the distribution with the minimum BIC.}
\item{estimated_parameters}{A numeric vector of estimated parameters for the chosen distribution.}
\item{plot}{A ggplot object showing the comparative plot of the data and the chosen distribution.}
}
\description{
This function estimates parameters for several distributions for a given vector and selects the one that best represents the data based on the Bayesian Information Criterion (BIC).
It provides the estimated parameters of the selected distribution and also a comparative plot.
The sample used should represent data from a pilot study based on which an informative prior will be constructed in the context of define_signal_eligibility
}
