% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{combine_benchmark_returns}
\alias{combine_benchmark_returns}
\title{Consolidate benchmark_returns_xts}
\usage{
combine_benchmark_returns(
  meta_benchmark_returns_xts,
  base_benchmark_returns_xts
)
}
\description{
Consolidate benchmark_returns_xts
}
