% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodity_factor_surprise.R
\name{commodity_factor_surprise}
\alias{commodity_factor_surprise}
\title{Macro Factors - Commodities}
\usage{
commodity_factor_surprise(
  segment_classification,
  segments_with_positive_sensibility_to_surprise,
  segments_with_negative_sensibility_to_surprise,
  surprise_matrix
)
}
\arguments{
\item{segment_classification}{A matrix indicating the classification of each observation into different segments.
Rows represent observations, and columns represent segments.}

\item{segments_with_positive_sensibility_to_surprise}{A list of vectors indicating the segments with a positive sensitivity to commodity surprises.}

\item{segments_with_negative_sensibility_to_surprise}{A list of vectors indicating the segments with a negative sensitivity to commodity surprises.}

\item{surprise_matrix}{A matrix containing the surprise values for different commodities and time periods.
Rows represent commodities, and columns represent time periods.}
}
\value{
A matrix representing the impact of commodity factors on different segments.
Rows represent observations, and columns represent time periods.
}
\description{
This function calculates the surprise impact of commodity factors on different segments based on their sensitivities to these factors.
}
