% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sector_map.R
\name{compute_sector_map}
\alias{compute_sector_map}
\alias{compute_sector_map,meta_dataframe,character,list-method}
\title{Compute Sector-Based Column Transformations}
\usage{
compute_sector_map(meta_dataframe, sector_column, mapper, ...)

\S4method{compute_sector_map}{meta_dataframe,character,list}(
  meta_dataframe,
  sector_column,
  mapper,
  feature_name = NULL,
  ...
)
}
\arguments{
\item{meta_dataframe}{A \code{meta_dataframe} object containing the target data.}

\item{sector_column}{A \code{character} specifying the column in \code{meta_dataframe} that contains sector classifications.}

\item{mapper}{A named \code{list} in which names represent sector names and elements are formulas defining how to compute values (e.g., \code{~-Alpha}, \code{~ Beta + Alpha}).}

\item{feature_name}{A \code{character} specifying the name of the new computed column.}
}
\value{
A modified \code{meta_dataframe} with the new column containing computed values.
}
\description{
This function creates a new column in a \code{meta_dataframe} based on sector classification
and applies transformations specified in a mapping list of formulas.
}
\details{
The function:
\itemize{
\item Maps each row in \code{meta_dataframe} to the corresponding transformation based on \code{sector_column}
\item Evaluates the formulas in \code{mapper} within the context of the \code{meta_dataframe}
}
}
