% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sector_wise.R
\name{compute_sector_wise}
\alias{compute_sector_wise}
\title{Compute Sector-Wise Calculation for a Given Signal in a meta_dataframe}
\usage{
compute_sector_wise(
  features_m_df,
  sector_column,
  signal,
  FUN,
  na.rm = TRUE,
  feature_name = NULL,
  min_non_na = 0
)
}
\arguments{
\item{features_m_df}{A \code{meta_dataframe} object.}

\item{sector_column}{A \code{character} specifying the column name representing sector classification in the dataset.}

\item{signal}{A \code{character} specifying the column name on which the function is computed.}

\item{FUN}{A \code{character} specifying the function to apply. Options are "median", "mean", "sd", "skew", "sur", "mean_std".}

\item{na.rm}{A \code{logical} indicating whether to remove NA values (default TRUE).}

\item{feature_name}{A \code{character} specifying the name of the feature to be added to the meta_dataframe. If NULL,
the feature name will be set to "\if{html}{\out{<signal>}}\emph{sector}\if{html}{\out{<FUN>}}". Default is NULL.}

\item{min_non_na}{A \code{numeric} value specifying the minimum number of non-NA values required to compute the metric. Default is 0.}
}
\value{
A \code{meta_dataframe} object with an added column named \code{<signal>_sector_<FUN>} in its
\code{data} slot containing the computed values.
}
\description{
This function computes a calculation for each observation in a \code{meta_dataframe} object by applying a
predefined function (specified by a character) to all observations within the same sector on the same date.
}
\details{
For each row (current observation), the function groups observations by \code{sector_column} and \code{dates},
then applies the specified function to the \code{signal} column.
If no matching observation is found, the resulting value is \code{NA}. The available functions are:
\itemize{
\item \strong{median}: \code{stats::median(x, na.rm = na.rm)}
\item \strong{mean}: \code{stats::mean(x, na.rm = na.rm)}
\item \strong{sd}: \code{stats::sd(x, na.rm = na.rm)}
\item \strong{mean_std}: \code{stats::mean(x, na.rm = na.rm) / stats::sd(x, na.rm = na.rm)}
}
}
