% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_window.R
\name{compute_window,meta_xts,numeric,character,character-method}
\alias{compute_window,meta_xts,numeric,character,character-method}
\title{method for meta_xts}
\usage{
\S4method{compute_window}{meta_xts,numeric,character,character}(
  data,
  period,
  FUN,
  window = "rolling",
  metric,
  benchmark_returns_m_xts = NULL,
  selected_bench = NULL,
  na.rm = TRUE,
  only_unique = FALSE,
  feature_name = NULL,
  min_non_na = 0
)
}
\description{
method for meta_xts
}
