% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{concentration_constraint_policy-class}
\alias{concentration_constraint_policy-class}
\title{Concentration Constraint Policy}
\description{
An S4 class to represent a concentration constraint policy
in portfolio construction.
}
\section{Slots}{

\describe{
\item{\code{benchmark}}{A character vector indicating which benchmark(s) to use.
For stocks, must be a column in benchmarks_m_df.
For signals, must be theme_ss or theme_sb}

\item{\code{max_abs_active_individual_weight}}{A numeric value indicating the
maximum absolute active weight for individual assets.}

\item{\code{max_abs_active_group_weight}}{A \strong{named} numeric vector indicating
maximum absolute group weights in relation to the benchmark.
Names should match columns in a groups_m_df.}
}}

