% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{consolidate_backtest_returns_m_xts}
\alias{consolidate_backtest_returns_m_xts}
\title{Consolidate Backtest Returns (XTS)}
\usage{
consolidate_backtest_returns_m_xts(
  meta_backtest_returns_m_xts,
  base_backtest_returns_m_xts
)
}
\arguments{
\item{meta_backtest_returns_m_xts}{An XTS object representing meta backtest returns.}

\item{base_backtest_returns_m_xts}{An XTS object representing base backtest returns.}
}
\value{
A consolidated XTS object containing the merged data with updated metadata.
If \code{meta_backtest_returns_m_xts} is NULL, NULL is returned.
}
\description{
This function consolidates two XTS objects containing backtest returns by merging them.
If one of the inputs is NULL, the function returns the other non-NULL object.
If both are non-NULL, it merges them using a left join and removes any missing values.
}
\examples{
consolidated <- consolidate_backtest_returns_xts(meta_xts, base_xts)

}
