% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{consolidate_backtest_returns_xts}
\alias{consolidate_backtest_returns_xts}
\title{Consolidate backtest_returns_xts}
\usage{
consolidate_backtest_returns_xts(
  meta_backtest_returns_xts,
  base_backtest_returns_xts
)
}
\description{
Consolidate backtest_returns_xts
}
