% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{consolidate_benchmark_returns_m_xts}
\alias{consolidate_benchmark_returns_m_xts}
\title{Consolidate Benchmark Returns (XTS)}
\usage{
consolidate_benchmark_returns_m_xts(
  meta_benchmark_returns_m_xts,
  base_benchmark_returns_m_xts
)
}
\arguments{
\item{meta_benchmark_returns_m_xts}{An XTS object containing meta benchmark returns.}

\item{base_benchmark_returns_m_xts}{An XTS object containing base benchmark returns.}
}
\value{
A combined XTS object with updated metadata or NULL.
}
\description{
This function consolidates two benchmark return XTS objects.
If one is NULL, it returns the other. If both are present, they are merged.
}
\examples{
result <- consolidate_benchmark_returns_xts(meta_xts, base_xts)
}
