% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{consolidate_benchmark_returns_xts}
\alias{consolidate_benchmark_returns_xts}
\title{Consolidate benchmark_returns_xts}
\usage{
consolidate_benchmark_returns_xts(
  meta_benchmark_returns_m_xts,
  base_benchmark_returns_m_xts
)
}
\description{
Consolidate benchmark_returns_xts
}
