% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{consolidate_generic_meta_dataframes}
\alias{consolidate_generic_meta_dataframes}
\title{Consolidate Generic Meta Dataframes}
\usage{
consolidate_generic_meta_dataframes(
  main_generic_m_df,
  supplemental_generic_m_df,
  type,
  consolidate_name = TRUE,
  require_main = TRUE
)
}
\arguments{
\item{main_generic_m_df}{A meta dataframe object.}

\item{supplemental_generic_m_df}{A base dataframe object.}

\item{type}{A character string indicating the type of data (e.g., "groups").}

\item{consolidate_name}{A logical indicating whether to consolidate the name of the meta dataframe.}

\item{require_main}{A logical indicating whether to return NULL if main_generic_m_df is NULL.}
}
\value{
A consolidated dataframe combining main and new data or the appropriate object depending on require_main
}
\description{
Merges meta and base dataframes, handling cases where either dataframe is NULL.
If both are present, they are combined using a full join and sorted by ID.
}
