% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{consolidate_generic_meta_xts}
\alias{consolidate_generic_meta_xts}
\title{Consolidate Generic Meta XTS}
\usage{
consolidate_generic_meta_xts(
  main_generic_m_xts,
  supplemental_generic_m_xts,
  type = "returns",
  consolidate_name = TRUE,
  require_main = TRUE,
  operation = "merge"
)
}
\arguments{
\item{main_generic_m_xts}{An XTS object representing the most important object.}

\item{supplemental_generic_m_xts}{An XTS object representing supplementary.}

\item{type}{A character string indicating the type of data (e.g., "groups").}

\item{consolidate_name}{A logical indicating whether to consolidate the name of the meta dataframe.}

\item{require_main}{A logical indicating whether to return NULL if main_generic_m_df is NULL.}

\item{operation}{A character indicating whether to 'merge' or 'bind_rows' the data.}
}
\value{
A consolidated xts combining main and new data or the appropriate object depending on require_main
}
\description{
Merges meta and base XTS objects, handling cases where either object is NULL.
If both are present, they are combined using bind_rows.
}
