% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_window.R
\name{count_if}
\alias{count_if}
\title{Count elements that satisfy a condition}
\usage{
count_if(values, count_condition_fun, na.rm = TRUE)
}
\arguments{
\item{values}{A numeric vector of values to evaluate.}

\item{count_condition_fun}{A function that takes a numeric vector and returns a logical vector.
The function should return \code{TRUE} for elements that should be counted.}

\item{na.rm}{Logical. If \code{TRUE}, ignores \code{NA} values when counting. Default is \code{TRUE}.}
}
\value{
An integer representing the count of values satisfying the condition.
}
\description{
The \code{count_if} function counts the number of elements in \code{values}
that satisfy the condition specified in \code{count_condition_fun}.
}
