% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{cov_est_method-class}
\alias{cov_est_method-class}
\title{Define the \code{cov_est_method} S4 Class}
\value{
An S4 object of class \code{cov_est_method}.
}
\description{
S4 class to represent a set of configurations for estimating the covariance matrix.
}
\section{Slots}{

\describe{
\item{\code{cov_estimation_method}}{A character string representing the covariance estimation method. Must be one of 'sample', 'ewma', 'cc', 'pca1', 'pca2', 'shrink_id' or 'shrink_cc'.}

\item{\code{cov_matrix_sample_size}}{Number of periods to subset return sample when estimating the covariance matrix. A high number will provide
higher degrees of freedom, but old returns might not reflect current risk due to parameter shift. A low number will tend to expose estimation
to dimensionality curse.}

\item{\code{active_returns}}{logical. If TRUE, the covariance matrix will be estimated using active returns.
If FALSE, the covariance matrix will be estimated using raw returns.}

\item{\code{cov_matrix_benchmark}}{A character string representing the benchmark from benchmark_returns_xts to be used when estimating the covariance matrix.
Only needed when active_returns is TRUE.}
}}

