% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_alpha_test_strategy}
\alias{create_alpha_test_strategy}
\title{Create an alpha_test_strategy object}
\usage{
create_alpha_test_strategy(
  signal_significance_threshold = 0.05,
  p_correction_method = "none",
  market_factor_proxy,
  bayesian_model_parameters = NULL
)
}
\arguments{
\item{signal_significance_threshold}{A numeric value indicating the hypothesis testing zero-alpha null-hypothesis rejection criteria.
Must be between 0 and 1. Defaults to 0.05.}

\item{p_correction_method}{A character string specifying the p-value correction method.
Options include \code{"none"}, \code{"bonferroni"}, \code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"BH"}, \code{"fdr"}, \code{"BY"}, and \code{"bayesian"}.}

\item{market_factor_proxy}{A character string indicating the market factor proxy to be used in the CAPM model.}

\item{bayesian_model_parameters}{(Optional) An object of class \code{bayesian_model_parameters}.
Required when \code{p_correction_method} is \code{"bayesian"}.}
}
\value{
An object of class \code{alpha_test_strategy}, \code{frequentist_alpha_test_strategy}, or \code{bayesian_alpha_test_strategy}.
}
\description{
A constructor function to create instances of alpha_test_strategy or its subclasses
(frequentist_alpha_test_strategy and bayesian_alpha_test_strategy).
}
