% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_bayesian_model_parameters}
\alias{create_bayesian_model_parameters}
\title{Create Bayesian Model Parameters}
\usage{
create_bayesian_model_parameters(
  user_priors = NULL,
  prior_derivation_control = NULL,
  brms_control = list(chains = 4, iter = 2000, warmup = 1000, thin = 1, seed = NA,
    adapt_delta = 0.8)
)
}
\arguments{
\item{user_priors}{An object of class \code{brmsprior}, or \code{NULL}.
Structured according to the \code{model_spec_theme_level}.}

\item{prior_derivation_control}{A list of additional parameters for deriving priors when \code{priors_type} is
\code{"informative_exogenous_dataset"}. Must be a list with the following elements:
\itemize{
\item \code{half_t_df}: Degrees of freedom for the half-t distribution applied to sd priors.
\item \code{lmer_optimizer}: Optimizer to be used in \code{lme4::lmer} for deriving priors.
\item \code{lmer_optimization_objective}: Criteria to be optimized in \code{lme4::lmer} for deriving priors,
e.g. \code{"likelihood"} or \code{"REML"}.
}}

\item{brms_control}{A list of parameters to be passed to \code{brms::brm} for MCMC sampling, including:
\itemize{
\item \code{chains}: Number of Markov chains (default is 4).
\item \code{iter}: Number of iterations per chain (default is 2000).
\item \code{warmup}: Number of warmup iterations per chain (default is \code{floor(iter / 2)}).
\item \code{thin}: Thinning interval (default is 1).
\item \code{seed}: Seed for reproducibility (default is \code{NA}).
\item \code{adapt_delta}: Target acceptance probability for HMC (default is 0.80).
}}
}
\value{
An S4 object of class \code{bayesian_model_parameters}.
}
\description{
Constructor for an S4 object of class \code{bayesian_model_parameters}.
}
