% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_cap_weighted_portfolio.R
\name{create_cap_weighted_portfolio}
\alias{create_cap_weighted_portfolio}
\title{Create a cap_weighted portfolio}
\usage{
create_cap_weighted_portfolio(
  universe_m_d_ref,
  liquidity_m_d_ref,
  cap_weighting_metric,
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975,
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe with tickers, is_eligible and final_signal columns generated by filter_stock_universe}

\item{liquidity_m_d_ref}{A data frame (similar to \code{features_m_df}) containing columns for id, tickers, dates, and one or more market liquidity measures (e.g., inflation-adjusted mean financial volume).
All tickers in current stock universe must have a unique correspondence in this data frame.}

\item{cap_weighting_metric}{cap weighting metric in liquidity_m_d_ref to be used to weight stocks when portfolio_construction_method is CW}

\item{lower_quantile_winsorization}{lower quantile to winsorize cap_weighting_metric}

\item{upper_quantile_winsorization}{upper quantile to winsorize cap_weighting_metric}

\item{verbose}{If TRUE, will print messages to the console}
}
\description{
Create a cap_weighted portfolio
}
