% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_concentration_constraint_policy}
\alias{create_concentration_constraint_policy}
\title{Create Concentration Constraint Policy}
\usage{
create_concentration_constraint_policy(
  benchmark = character(0),
  max_abs_active_individual_weight = NA_real_,
  max_abs_active_group_weight = numeric(0)
)
}
\arguments{
\item{benchmark}{A character vector (can be empty if no benchmark specified).}

\item{max_abs_active_individual_weight}{A numeric indicating the max
absolute active weight for individual assets.}

\item{max_abs_active_group_weight}{A named numeric vector for group constraints.}
}
\value{
An S4 object of class \code{concentration_constraint_policy}.
}
\description{
Constructor for a \code{concentration_constraint_policy} object.
}
