% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_cov_est_method}
\alias{create_cov_est_method}
\title{Create Covariance Estimation Method}
\usage{
create_cov_est_method(
  cov_estimation_method = "sample",
  cov_matrix_sample_size,
  active_returns = TRUE,
  cov_matrix_benchmark = NULL
)
}
\arguments{
\item{cov_estimation_method}{A character string representing the covariance estimation method. Must be one of 'sample', 'ewma', 'cc', 'pca1', 'pca2', 'shrink_id' or 'shrink_cc'.}

\item{cov_matrix_sample_size}{Number of periods to subset return sample when estimating the covariance matrix. A high number will provide
higher degrees of freedom, but old returns might not reflect current risk due to parameter shift. A low number will tend to expose estimation
to dimensionality curse.}

\item{active_returns}{logical. If TRUE, the covariance matrix will be estimated using active returns. If FALSE, the covariance matrix will be estimated using raw returns.}

\item{cov_matrix_benchmark}{A character string representing the benchmark for covariance matrix estimation. If NULL, the covariance matrix will be estimated using the active returns.}
}
\description{
Constructor for creating an instance of \code{cov_est_method}.
}
