% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_custom_weighted_portfolio.R
\name{create_custom_weighted_portfolio}
\alias{create_custom_weighted_portfolio}
\title{Create a custom_weighted portfolio for signals or stocks}
\usage{
create_custom_weighted_portfolio(
  universe_m_d_ref,
  custom_weights_m_d_ref,
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe with identifiers (tickers or signal column), is_eligible and a weighting column as defined by signal_weighting metric.
This object could either be the result of filter_stock_universe or created in the context of the blend_signals_function}

\item{custom_weights_m_d_ref}{A dataframe with identifiers (tickers or signal column) and custom weights.}
}
\description{
Create a custom_weighted portfolio for signals or stocks
}
