% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_equal_weighted_portfolio.R
\name{create_equal_weighted_portfolio}
\alias{create_equal_weighted_portfolio}
\title{Create a equal_weighted portfolio for signals or stocks}
\usage{
create_equal_weighted_portfolio(universe_m_d_ref)
}
\arguments{
\item{universe_m_d_ref}{A dataframe with identifiers (tickers or signal column), is_eligible and a weighting column as defined by signal_weighting metric.
This object could either be the result of filter_stock_universe or created in the context of the blend_signals_function}
}
\description{
Create a equal_weighted portfolio for signals or stocks
}
