% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_fwd_meta_dataframe}
\alias{create_fwd_meta_dataframe}
\title{Create a Forward MetaDataFrame Object}
\usage{
create_fwd_meta_dataframe(data)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to be converted to a \code{fwd_meta_dataframe}.}
}
\value{
An object of class \code{fwd_meta_dataframe} if the input data frame meets all validation criteria.
The returned object includes metadata such as the number of unique dates, unique tickers, total number
of observations, and the forward-looking target variables.
}
\description{
This function creates an object of class \code{fwd_meta_dataframe} from a provided data frame.
The data frame must meet the same requirements as \code{meta_dataframe}, with additional
constraints on the non-required columns. These columns must follow the format \code{XXXXX_ym},
where \code{XXXXX} represents any string of length 5, and \code{ym} indicates the number of
forward months (01 to 12). The \code{fwd_target} slot will be populated with the names of these
forward-looking columns.
}
\details{
\itemize{
\item The 'id' column should be in the format \code{paste0(tickers, "-", dates)}.
\item The 'dates' column must be of class \code{Date} and in ascending chronological order.
\item Non-required columns must follow the format \code{XXXXX_ym}, where \code{ym} represents the number
of forward months (01 to 12).
}
}
\examples{
# Create a sample data frame
df <- data.frame(
  id = c("A-2024-01-01", "B-2024-02-01"),
  tickers = c("A", "B"),
  dates = as.Date(c("2024-01-01", "2024-02-01")),
  value_01 = c(10, 20),
  value_02 = c(15, 25)
)

# Create a fwd_meta_dataframe object
fwd_meta_df <- create_fwd_meta_dataframe(df)

}
