% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_hyper_grid_domain}
\alias{create_hyper_grid_domain}
\title{Create a \code{hyper_grid_domain} Object}
\usage{
create_hyper_grid_domain(tuning_method, ml_algorithm)
}
\arguments{
\item{tuning_method}{A character string specifying the tuning method to be applied.
Must be one of "grid_search", "random_search", or "bayesian_optimization".}

\item{ml_algorithm}{A character string specifying the machine-learning algorithm to be used.
Must be one of "glmnet", "rf", "xgb", "nn", or "ols".}
}
\value{
An instance of the \code{hyper_grid_domain} S4 class.
}
\description{
This function creates an instance of the \code{hyper_grid_domain} S4 class.
It allows users to specify the tuning method, machine learning algorithm, and relevant hyperparameters.
}
\examples{
# Creating a hyper_grid_domain object for a random forest model
hyper_grid_rf <- create_hyper_grid_domain(
  tuning_method = "grid_search",
  ml_algorithm = "rf"
)


}
