% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_liquidity_constraint_policy}
\alias{create_liquidity_constraint_policy}
\title{Create Liquidity Constraint Policy}
\usage{
create_liquidity_constraint_policy(
  liquidity_floor_rule = NULL,
  liquidity_cap_rules = NULL
)
}
\arguments{
\item{liquidity_floor_rule}{A character string indicating the minimum liquidity classification.}

\item{liquidity_cap_rules}{A named numeric vector indicating liquidity cap rules.}
}
\value{
An S4 object of class \code{liquidity_constraint_policy}.
}
\description{
Constructor for a \code{liquidity_constraint_policy} object.
}
