% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_meta_xts}
\alias{create_meta_xts}
\alias{create_meta_xts,xts-method}
\alias{create_meta_xts,data.frame-method}
\alias{create_meta_xts,meta_dataframe-method}
\title{Create a meta_xts, assets_meta_xts or metrics_meta_xts object.}
\usage{
create_meta_xts(
  data,
  type = c("returns", "metrics"),
  asset_type = "not_identified",
  meta_xts_name = "not_identified",
  metric_name = NULL,
  workflow = NULL,
  source = NULL,
  ...
)

\S4method{create_meta_xts}{xts}(
  data,
  type = c("returns", "metrics"),
  asset_type = "not_identified",
  meta_xts_name = "not_identified",
  metric_name = NULL,
  workflow = NULL,
  source = NULL
)

\S4method{create_meta_xts}{data.frame}(
  data,
  type = c("returns", "metrics"),
  asset_type = "not_identified",
  meta_xts_name = "not_identified",
  metric_name = NULL,
  workflow = NULL,
  source = NULL,
  data_format = c("wide", "long"),
  dates = NULL
)

\S4method{create_meta_xts}{meta_dataframe}(
  data,
  type = c("returns", "metrics"),
  asset_type = "not_identified",
  meta_xts_name = "not_identified",
  metric_name = NULL,
  source = NULL
)
}
\arguments{
\item{data}{An xts object containing your time series data.}

\item{type}{Character. Either \code{"assets"} (no holes allowed) or
\code{"metrics"} (holes allowed). Defaults to \code{c("assets","metrics")}
which means you must pick one.}

\item{asset_type}{A character string indicating the type of asset.}

\item{meta_xts_name}{A character string. Defaults to \code{"default_name"}.}

\item{metric_name}{A character string identifying the name of the metric}

\item{workflow}{An ANY object for workflow. Defaults to \code{NULL}.}

\item{source}{A character vector indicating data origin for each column.
If \code{NULL}, defaults to \code{"not_identified"} repeated for each column.}

\item{...}{Additional arguments passed to the constructor.}

\item{data_format}{A character string indicating the format of the data.}

\item{dates}{An optional vector of dates to be included.}
}
\value{
An S4 object of class \code{assets_meta_xts} or \code{metrics_meta_xts},
depending on \code{type}.
}
\description{
This constructor automatically sets most slots for you based on an
input xts object and the desired type: "assets" or "metrics".
}
