% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_mmaf_parameters}
\alias{create_mmaf_parameters}
\title{Create MMAF (Micro Macro Allocation Framework) Parameters}
\usage{
create_mmaf_parameters(
  mmaf_method = "bottom_up",
  top_down_proxy_port_method = "ew",
  mmaf_group_col,
  micro_port_construction_method,
  macro_port_construction_method
)
}
\arguments{
\item{mmaf_method}{A character indicating the MMAF method to be used. Must be one of 'top_down' or 'bottom_up'.}

\item{top_down_proxy_port_method}{A character indicating the method to be used for constructing the top-down proxy portfolio.
Must be one of 'ew', 'rp', 'hrp', 'cs' or 'sw' if mmaf_method is 'top_down' and NULL if mmaf_method is 'bottom_up'.}

\item{mmaf_group_col}{A character string representing the MMAF group to which the assets belong. This is used to group assets when constructing micro and macro portfolios.
It must be a length 1 character.}

\item{micro_port_construction_method}{An object of class \code{micro_port_construction_method} representing the configuration for constructing micro portfolios.}

\item{macro_port_construction_method}{An object of class \code{macro_port_construction_method} representing the configuration for constructing macro portfolios.}
}
\value{
An S4 object of class \code{mmaf_parameters}.
}
\description{
Constructor function for creating an instance of \code{mmaf_parameters}.
}
