% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_mto_portfolio.R
\name{create_mto_portfolio}
\alias{create_mto_portfolio}
\title{Create a MTO portfolio}
\usage{
create_mto_portfolio(
  universe_m_d_ref,
  returns_upd_ref,
  covariance_matrix_sample_size,
  covariance_estimation_method,
  liquidity_constraint_policy = NULL,
  turnover_constraint_policy = NULL,
  concentration_constraint_policy = NULL,
  groups_m_d_ref = NULL,
  n_random_portfolios = 2000,
  rp_method = "sample",
  mto_port_objective = "IR",
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe with tickers, is_eligible and final_signal columns}

\item{covariance_matrix_sample_size}{Number of periods to subset returns_df sample when estimating the covariance_matrix. A high number will provide
higher degrees of freedom, but old returns might not reflect current risk due to parameter shift. A low number will tend to expose estimation
to dimensionality curse.}

\item{covariance_estimation_method}{One of SAM (Sample), EWMA, CC (Constant Correlation), PCA1, PCA2, Shrink_ID or Shrink_CC.
If NULL, blending_method can only be EW or IR}

\item{liquidity_constraint_policy}{Optional. A named list containing objects used to apply liquidity constraints. Possible elements of the list are:
\itemize{
\item \code{liquidity_floor_rule}: A character indicating the liquidity classification (e.g., micro_caps, small_caps) used to filter stocks. Stocks with less liquidity than specified in \code{liquidity_floor_rule} will be considered ineligible.
In the case of the \code{generate_box_constraints} function, \code{liquidity_constraint_policy} can also contain:
\item \code{liquidity_cap_rule} lists: One or many lists used to create upper bounds for weights based on a liquidity classification. Each list must contain:
\itemize{
\item \code{liquidity_classification}: A character indicating the classification for the cap.
\item \code{liquidity_cap}: A numeric value indicating the cap (upper bound) for stocks with that liquidity classification.
Many liquidity caps might be created, and in this case, each \code{liquidity_cap_rule} must be identified with a number (e.g., liquidity_cap_rule_1, liquidity_cap_rule_2, and so on).
}
}}

\item{turnover_constraint_policy}{A named list containing objects used to build buffer zones and apply turnover constraints.
\itemize{
\item Each element will constitute a \code{buffer_zone}, being a list with three elements:
\itemize{
\item \code{liquidity_classification} element: A liquidity classification (e.g., "micro_caps", "small_caps") for that buffer zone.
\item \code{top_assets_quantile_buffer}: A numeric value indicating a buffer value that relaxes \code{top_assets_quantile} for stocks with the specified liquidity classification.
\item \code{turnover_cap}: A numeric value specifying the turnover cap.
Stocks that are less liquid than specified for a buffer zone and have a signal higher than the respective buffer quantile will be considered eligible, even if they do not meet the \code{liquidity_floor_rule}.
}
}}

\item{groups_m_d_ref}{A data frame containing columns for id, tickers, dates, and group classification columns following a given classification method.
All tickers in the current stock universe must have a unique correspondence in the data frame.}

\item{mto_port_objective}{A character describing the objective to maximize in order to choose the best portfolio. One of "AR (active return)",
"TE (tracking error)" or " IR (information ratio)"}

\item{returns_d_ref}{A dataframe in which columns represent tickers present in universe_m_d_ref and row represent days
It should include all stocks in current_stock_universe and a dates column with at least covariance_matrix_sample_size days before current_date}

\item{current_date}{Current date}

\item{concentration_constraints_policy}{A named list containing up to four elements:
\itemize{
\item \code{benchmark}: A character vector describing the benchmark to be used to apply constraint.
Must have a correspondence in \code{benchmark_weights_m_d_ref}
\item \code{max_abs_active_individual_weight}: The maximum absolute individual active weights.
\item \code{group_classification}: A character vector describing the group classification to be used to apply group constraints.
Must have a correspondence in \code{groups_m_d_ref}
\item \code{max_abs_active_group_weight}: The maximum absolute group active weight used for creating group constraints in \code{generate_group_constraints}.
If a given group has no eligible stock, the one with the greatest signal will be automatically promoted.
Note that, in the context of \code{generate_group_constraints}, a \code{benchmark_weights_m_d_ref} data frame must also be supplied.
}}

\item{benchmark_weights_m_d_ref}{A data frame containing columns for id, tickers, dates, and current benchmark weights columns.
All tickers in the current stock universe must have a unique correspondence in this data frame.}
}
\description{
Create a MTO portfolio
}
