% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_portfolio_policies}
\alias{create_portfolio_policies}
\title{Create Portfolio Policies}
\usage{
create_portfolio_policies()
}
\description{
Constructs a \code{portfolio_policies} object.
portfolio <- create_portfolio_policies()
}
