% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_risk_parity_portfolio.R
\name{create_risk_parity_portfolio}
\alias{create_risk_parity_portfolio}
\title{Create a risk parity portfolio}
\usage{
create_risk_parity_portfolio(
  universe_m_d_ref,
  covariance_matrix,
  rp_method = "cyclical-spinu",
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe with tickers, is_eligible and final_signal columns generated by filter_stock_universe}

\item{covariance_matrix}{The covariance matrix of all eligible stocks in universe_m_d_ref.}

\item{rp_method}{The method used to calculate the risk parity weights. Default is "cyclical-spinu". See riskParityPortfolio::riskParityPortfolio for more options.}

\item{verbose}{If TRUE, will print messages to the console}
}
\description{
Create a risk parity portfolio
}
