% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_rp_parameters}
\alias{create_rp_parameters}
\title{Create RP (Risk Parity) Parameters}
\usage{
create_rp_parameters(rp_method = "cyclical-spinu")
}
\arguments{
\item{rp_method}{A character indicating the method to compute the risk-parity vanilla solution.
It is passed to \code{riskParityPortfolio::riskParityPortfolio()} function as \code{method_init}.
Default is \code{"cyclical-spinu"}.}
}
\value{
An S4 object of class \code{rp_parameters}.
}
\description{
Constructor function for creating an instance of \code{rp_parameters}.
}
