% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_se_benchmarks.R
\name{create_se_benchmarks}
\alias{create_se_benchmarks}
\title{Create Signal Engineering Benchmarks}
\usage{
create_se_benchmarks(signal_universe_m_d_ref, selected_signal_themes_m_d_ref)
}
\arguments{
\item{signal_universe_m_d_ref}{A data frame containing the signal universe. Must include columns:
\itemize{
\item \code{id}
\item \code{tickers}
\item \code{dates}
\item \code{pre_eligible_assets} (indicating which signals are statistically significant)
}}

\item{selected_signal_themes_m_d_ref}{An optional data frame with signal themes. Must include columns:
\itemize{
\item \code{tickers}
\item \code{theme} (the classification theme for each signal)
This parameter is mandatory if one wants to calculate theme-weighted benchmarks.
}}
}
\value{
A list with two data frames, each with benchmark weights following Signal-Selection and Signal-Blending methods. The columns of each data.frame include:
\itemize{
\item \code{id}
\item \code{tickers}
\item \code{dates}
\item \code{individual} (weight assigned to each signal)
\item \code{theme} (weight assigned to each signal given their theme, if \code{selected_signal_themes_m_d_ref} is provided)
}
}
\description{
Creates Signal-Selection and Signal-Blending benchmark weights:
\itemize{
\item{Signal-Selection Benchmark}: A benchmark that is built using the universe of all signals in \code{chosen_signals}.
It is used to evaluate the performance of the signal selection process.
\item{Signal-Blending Benchmark}: A benchmark that is built using only signals derived from signal selection process (those with pre_eligible_assets assigned as 1).
It is used to evaluate the performance of the signal blending process.
}
}
\details{
The function calculates weights for signals based on their statistical significance.
If \code{selected_signal_themes_m_d_ref} is provided, it also calculates theme-based weights and merges them
back into the main data frame.
}
