% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_tickers_catalog,raw_features_m_df,data.frame,data.frame-method}
\alias{create_tickers_catalog,raw_features_m_df,data.frame,data.frame-method}
\title{Create a tickers_catalog Object}
\usage{
\S4method{create_tickers_catalog}{raw_features_m_df,data.frame,data.frame}(
  raw_features_m_df,
  date_first_quote,
  date_last_quote,
  n_days_tolerance = 10
)
}
\arguments{
\item{raw_features_m_df}{A \code{meta_dataframe} object with \code{type == "generic"}, containing stock tickers and dates.}

\item{date_first_quote}{A \code{data.frame} with columns \code{tickers} (character) and \code{date_first_quote} (Date).}

\item{date_last_quote}{A \code{data.frame} with columns \code{tickers} (character) and \code{date_last_quote} (Date).}

\item{n_days_tolerance}{A \code{numeric} indicating the maximum tolerance for which \code{date_last_quote} can be smaller than
current_date without it making it a delisted stock.}
}
\description{
Constructs a \code{tickers_catalog} object by integrating stock metadata from multiple data sources.
Ensures data consistency, generates unique identifiers, and classifies stocks based on listing status.
}
