% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_transaction_costs_parameters}
\alias{create_transaction_costs_parameters}
\title{Create a New Transaction Cost Parameters Object}
\usage{
create_transaction_costs_parameters(
  direct_transaction_cost,
  strategy_aum,
  alpha,
  lambda
)
}
\arguments{
\item{direct_transaction_cost}{A numeric value for the direct transaction cost.}

\item{strategy_aum}{A numeric value for the strategy's assets under management.}

\item{alpha}{A numeric value for the alpha parameter.}

\item{lambda}{A numeric value or the string "dynamic" for the lambda parameter.}
}
\value{
An object of class \code{transaction_cost_parameters}.
}
\description{
This function constructs a new \code{transaction_cost_parameters} S4 object.
}
