% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_turnover_constraint_policy}
\alias{create_turnover_constraint_policy}
\title{Create Turnover Constraint Policy}
\usage{
create_turnover_constraint_policy(
  quantile_range_buffer,
  turnover_cap_rules = NULL
)
}
\arguments{
\item{quantile_range_buffer}{A numeric value indicating the increase in the quantile range for buffer zones.}

\item{turnover_cap_rules}{A named numeric vector indicating turnover cap rules.}
}
\value{
An S4 object of class \code{turnover_constraint_policy}.
}
\description{
Constructor for a \code{turnover_constraint_policy} object.
}
