% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instructions.R
\name{display_valid_custom_objectives}
\alias{display_valid_custom_objectives}
\title{Display Valid Custom Objectives}
\usage{
display_valid_custom_objectives()
}
\value{
Prints the valid heuristic SB metrics to the console, categorizing them as needed.
}
\description{
This function prints out the list of valid custom objectives, categorizing them based on the requirements
for the \code{sb_algorithm} slot in the \code{sb_backtest_config} class. The metrics are needed when the \code{sb_algorithm}
is set to either \code{sw} or \code{markowitz}, and some of them are specific to the \code{active_returns} setting or the \code{p_correction_method}.
}
\details{
Metrics containing either "act_", "info_ratio", or "track_err" are active performance metrics, which are used
when the \code{active_returns} slot of \code{sb_backtest_config} is TRUE. Metrics with "posterior" relate to a Bayesian
correction method when the \code{p_correction_method} slot is set to \code{bayesian}. Raw return metrics are used when
\code{active_returns} is FALSE. Frequentist CAPM metrics apply when \code{p_correction_method} is not \code{bayesian}.

Additionally, the error metrics 'squared_error', 'pseudo_huber_error', or 'absolute_error' are valid only when
\code{sb_algorithm} is either \code{xgb} or \code{nn}.
}
\examples{
display_valid_custom_objectives()

}
