% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dofuture.R
\name{\%dofuture\%}
\alias{\%dofuture\%}
\title{dofuture operator}
\usage{
foreach \%dofuture\% expr
}
\arguments{
\item{foreach}{object created by \code{foreach::foreach()} and \code{foreach::times()}}

\item{expr}{An R expression}
}
\value{
The value of the foreach call..
}
\description{
dofuture operator
}
\keyword{internal}
