% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_criteria.R
\name{eligibility_criteria}
\alias{eligibility_criteria}
\title{Display Eligibility Criteria with Colors}
\usage{
eligibility_criteria()
}
\value{
Prints the colored eligibility criteria to the console.
}
\description{
This function prints out the eligibility criteria for classifying the universe based on signals and other custom and user-defined rules, using colors to facilitate reading.
It uses cyan and magenta colors for headings and yellow to highlight important words.
}
\examples{
eligibility_criteria()

}
