% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-explain_prediction.R
\name{explain_prediction,sb_backtest_results,meta_dataframe,character,Date-method}
\alias{explain_prediction,sb_backtest_results,meta_dataframe,character,Date-method}
\title{Explain Prediction for sb_backtest_results}
\usage{
\S4method{explain_prediction}{sb_backtest_results,meta_dataframe,character,Date}(
  sb_backtest_results,
  features_m_df,
  selected_ticker,
  selected_date
)
}
\description{
Explain Prediction for sb_backtest_results
}
