% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-explain_prediction.R
\name{explain_prediction_inner}
\alias{explain_prediction_inner}
\title{inner function}
\usage{
explain_prediction_inner(
  sb_backtest_workflow,
  oos_sb_outputs_m_df,
  feature_importance_m_df,
  gsm_algorithm,
  features_m_df,
  selected_ticker = selected_ticker,
  selected_date = selected_date
)
}
\description{
inner function
}
