% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract_returns_m_xts.R
\name{extract_returns_m_xts}
\alias{extract_returns_m_xts}
\title{Extract Returns from port_backtest_cohort Object}
\usage{
extract_returns_m_xts(
  port_backtest_cohort,
  signals_m_df,
  benchmark_returns_m_xts,
  simplify_name = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{port_backtest_cohort}{An object of class \code{port_backtest_cohort}.}

\item{signals_m_df}{An object of class \code{signals_m_df} containing the signals used in the backtest.}

\item{benchmark_returns_m_xts}{An object of class \code{meta_xts} containing the benchmark returns used in the backtest.}

\item{simplify_name}{Logical. If \code{TRUE}, the column names will be simplified using the configuration names.}

\item{verbose}{Logical. If \code{TRUE}, messages will be printed about the simplification process.}
}
\value{
An \code{xts} object containing the extracted backtest returns.
}
\description{
This method extracts the \code{net_returns_m_xts} slot from the \code{port_backtest_cohort} object,
optionally simplifying the column names based on the configuration names of the cohort.
}
