% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_cia_adjust.R
\name{financial_cia_adjust}
\alias{financial_cia_adjust}
\title{Financial Cias Adjust Financials}
\usage{
financial_cia_adjust(
  characteristic_matrix,
  subsector_classification,
  characteristic_financial_cia_matrix,
  subsectors_to_adjust
)
}
\arguments{
\item{characteristic_matrix}{A matrix of financial characteristics, where rows represent observations and columns represent variables.}

\item{subsector_classification}{A matrix indicating the subsector classification of each observation. Must have the same dimensions as \code{characteristic_matrix}.}

\item{characteristic_financial_cia_matrix}{A matrix of financial characteristics specific to certain subsectors, where rows represent observations and columns represent variables. Must have the same dimensions as \code{characteristic_matrix}.}

\item{subsectors_to_adjust}{A vector of subsector labels for which the financial characteristics will be adjusted.}
}
\value{
A matrix of adjusted financial characteristics.
}
\description{
This function adjusts characteristics for specific financial subsectors for which that characteristic is not defined, replacing it
based on a provided matrix of data available for those subsectors.
}
\details{
This function is useful for when a specific variable is not defined for sectors in subsectors_to_adjust (eg. ebit/inv_cap). If that is the case, it will replace the value by the corresponding value in
characteristic_financial_cia_matrix (eg. roe)
}
\examples{
\dontrun{
# Create example matrices
characteristic_matrix <- matrix(c(10, 20, NA, 30, 40, 50), nrow = 3, ncol = 2)
subsector_classification <- matrix(c("A", "B", NA, "C", "A", "B"), nrow = 3, ncol = 2)
characteristic_financial_cia_matrix <- matrix(c(15, 25, 35, 45, 55, 65), nrow = 3, ncol = 2)
subsectors_to_adjust <- c("A", "B")

# Adjust financial characteristics
adjusted_matrix <- financial_cia_adjust(characteristic_matrix, subsector_classification, 
characteristic_financial_cia_matrix, subsectors_to_adjust)
}


}
\seealso{
\code{\link{fin_ratio}}
}
