% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_fill_panel.R
\name{financialcia_fill_panel}
\alias{financialcia_fill_panel}
\title{Fill Financial Cias Missing Data}
\usage{
financialcia_fill_panel(
  features_m_df,
  segment_column = "segment",
  segments_to_adjust,
  features_to_preserve = NULL
)
}
\arguments{
\item{features_m_df}{A meta dataframe or coercible object containing panel data with columns for "id", "tickers", "dates", "segment", and other features.}

\item{segment_column}{A character indicating which column in features_m_df contain economic segment classification to fill based on segments_to_adjust}

\item{segments_to_adjust}{A character vector specifying the segments (e.g., banks) for which missing data should be filled.}

\item{features_to_preserve}{A character vector specifying features that should not have missing data filled.}
}
\value{
A panel data matrix with missing data filled.
}
\description{
This function fills missing data in a panel data matrix, specifically for banks and other financial companies.
}
\details{
This function fills missing data in the panel data matrix for the specified segments. It calculates the mean for each feature across all companies (including non-banks) and fills missing values in bank segments with the calculated mean.
Features listed in \code{features_to_preserve} will not be filled even if they have missing values. Features that have missing values for all banks (but not for non-bank companies) are not filled, as it's assumed they require specific treatment.
}
\examples{
# Example usage
features_m_df <- data.frame(
  id = c(1, 2, 3, 4, 5),
  tickers = c("A", "B", "C", "D", "E"),
  dates = as.Date(c("2022-01-01", "2022-01-01", "2022-01-02", "2022-01-02", "2022-01-03")),
  segment = c("Bank", "Bank", "Other", "Bank", "Other"),
  feature1 = c(1, 2, NA, 4, 5),
  feature2 = c(NA, 6, NA, 8, NA)
)
segments_to_adjust <- c("Bank")
filled_data <- financialcia_fill_panel(features_m_df, segment_column = "segment", segments_to_adjust)

}
