% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fit_bayesian_model.R
\name{fit_bayesian_model}
\alias{fit_bayesian_model}
\title{Fit a Bayesian Model for a Given Theme}
\usage{
fit_bayesian_model(
  theme,
  signals_groups_m_d_ref,
  selected_signals_backtest_returns_upd_ref,
  selected_benchmark_returns_vector_upd_ref,
  elected_priors_list
)
}
\arguments{
\item{theme}{A character string specifying the theme for which the Bayesian model is to be fitted. This theme is used to select the relevant signals and priors.}

\item{groups}{A data frame that contains information about the signals and their associated themes. It should have at least two columns: \code{tickers} (the signals) and \code{theme} (the theme associated with each signal).}

\item{backtest_returns}{A data frame or matrix containing backtest returns data for the signals. The rows represent different time points, and the columns represent different signals.}

\item{priors}{A named list of priors for the Bayesian model. The list should contain prior distributions for each theme, as specified in the \code{brms} package format.}
}
\value{
An object of class \code{brmsfit} representing the fitted Bayesian regression model. This object contains the results of the Bayesian analysis, including estimates of the model parameters and diagnostics.
}
\description{
This function fits a Bayesian regression model to the backtest returns data for a specified theme using the \code{brms} package. It uses the priors defined for the theme and includes both fixed and random effects in the model.
}
