% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fit_frequentist_hierarchical_model.R
\name{fit_frequentist_hierarchical_model}
\alias{fit_frequentist_hierarchical_model}
\title{Fit a Frequentist Hierarchical Regression Model and Summarize Metrics}
\usage{
fit_frequentist_hierarchical_model(
  signal_universe_m_d_ref,
  selected_backtest_returns_corrected_positions_xts_upd_ref,
  selected_market_factor_proxy_xts_upd_ref,
  selected_backtest_returns_corrected_positions_m_upd_ref = NULL,
  selected_signal_themes_m_d_ref,
  model_spec_theme_level,
  lmer_optimizer,
  lmer_optimization_objective,
  hierarchical_p_value_method
)
}
\arguments{
\item{signal_universe_m_d_ref}{A data frame containing the signal universe. If provided, data in this object will be updated with hierarchical model metrics.}

\item{selected_backtest_returns_corrected_positions_m_upd_ref}{An already processed \code{selected_backtest_returns_corrected_positions_upd_ref}.
This data.frame is already in long format and contemplates both the \code{selected_market_factor_proxy_vector_upd_ref} and the
\code{selected_signal_themes_m_d_ref} theme data.}

\item{selected_signal_themes_m_d_ref}{A data frame containing metadata about signals. This data frame should include:
\itemize{
\item \code{tickers}: Signal identifiers matching those in \code{selected_backtest_returns_corrected_positions_upd_ref}.
\item \code{theme}: Group membership for each signal, defining the clusters for the Bayesian hierarchical model.
\item \code{dates}: Dates corresponding to the backtest data.
This input ensures proper alignment between signals and their associated themes.
}}

\item{model_spec_theme_level}{A character string specifying the desired model structure.
Options include:
\itemize{
\item \code{"random_intercept_fixed_slope"}: Includes random effects for the intercept at the theme level.
\item \code{"theme_specific_intercept_fixed_slope"}: Uses fixed intercepts for each theme.
\item \code{"theme_specific_intercept_theme_specific_slope"}: Includes fixed intercepts and slopes for each theme.
\item \code{"fixed_intercept_fixed_slope"}: Omits theme-level intercepts but includes random effects at the theme:signal level.
}}

\item{selected_backtest_returns_corrected_positions_upd_ref}{A data frame containing the backtest returns data for various signals.
\itemize{
\item The first column should include dates.
\item Remaining columns represent signals (e.g., tickers) and their respective active returns.
}}

\item{selected_market_factor_proxy_vector_upd_ref}{A numeric vector containing benchmark returns data. The vector will be recycled
to match the length of the backtest returns data.}

\item{verbose}{A logical indicating whether to display progress messages during the model fitting process. Defaults to \code{TRUE}.}
}
\value{
A named list containing:
\itemize{
\item \code{frequentist_model}: A \code{lmermod} object containing the fitted frequentist model.
\item \code{signal_universe_m_d_ref}: A data frame containing the updated \code{signal_universe_m_d_ref}.
}
}
\description{
This function fits a frequentist hierarchical regression model using the \code{lme4} package and computes summary
statistics for signal-level metrics. The model captures signal clustering within themes and provides detailed
metrics for use in performance analysis.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Preprocessing}: Converts the backtest returns data into long format, merges metadata about signals and themes,
and adds the market factor proxy.
\item \strong{Formula Specification}: Defines the model formula based on the chosen \code{model_spec_theme_level}.
\item \strong{Model Fitting}: Fits the frequentist model using the \code{lmer} function from the \code{lme4} package, incorporating the
prepared data.
\item \strong{Summarization}:
Updates the \code{signal_universe_m_d_ref} data frame with new metrics.
}

This hierarchical approach allows the model to capture both global (theme-level) and local (signal-level) effects.

'@section P-value calculation
The calculation of p-values is a controversial topic regarding linear mixed-effects models, which is the reason why
\code{lme4::lmer} does not provide p-values by default. The \code{afex} package provides a range of methods to calculate p-values for mixed models
}
