% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fit_sb_model.R
\name{fit_sb_model}
\alias{fit_sb_model}
\title{Fit SB Model}
\usage{
fit_sb_model(
  sb_algorithm,
  target_fwd_name,
  full_data_m_refit_clean = NULL,
  features_m_refit,
  target_m_refit,
  custom_objective_translated,
  huber_delta,
  quantile_tau,
  early_stop,
  keras_architecture_parameters,
  optimal_hyper = NULL,
  chosen_eval_metric_translated,
  most_recent_signal_universe_m_d_ref,
  selected_backtest_returns_corrected_positions_xts_upd_ref,
  selected_market_factor_proxy_xts_upd_ref,
  covariance_matrix_sample_size = 36,
  covariance_estimation_method = "sample",
  active_returns = TRUE,
  rp_method = "cyclical-spinu",
  n_random_ports = 2000,
  random_ports_method = "sample",
  opt_objective = "sharpe",
  concentration_constraint_policy,
  upper_quantile_winsorization = 0.95,
  lower_quantile_winsorization = 0.05,
  verbose
)
}
\description{
Fits a Statistical/Beta (SB) model based on the specified \code{sb_algorithm}, preparing and
training the model with given data, hyperparameters, and constraints. This function
acts as a dispatcher, calling different modeling workflows (OLS, GLMNET, Ranger, XGBoost,
Keras, heuristic portfolios, Risk Parity, or MVO) depending on \code{sb_algorithm}.
}
