% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{frequentist_alpha_test_strategy-class}
\alias{frequentist_alpha_test_strategy-class}
\title{frequentist_alpha_test_strategy Class}
\description{
A subclass of alpha_test_strategy for frequentist methods.
}
