% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generate_box_constraints.R
\name{generate_box_constraints}
\alias{generate_box_constraints}
\title{Create individual box constraints for MTO portfolios based on liquidity, concentration and turnover constraints}
\usage{
generate_box_constraints(
  universe_m_d_ref,
  liquidity_constraint_policy = NULL,
  turnover_constraint_policy = NULL,
  concentration_constraint_policy,
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe of assets with tickers, is_eligible, final_signal columns and also those generated by application of constraints policies in classify_investment_universe}

\item{liquidity_constraint_policy}{Optional. A named list containing objects used to apply liquidity constraints. Possible elements of the list are:
\itemize{
\item \code{liquidity_floor_rule}: A character indicating the liquidity classification (e.g., micro_caps, small_caps) used to filter stocks. Stocks with less liquidity than specified in \code{liquidity_floor_rule} will be considered ineligible.
In the case of the \code{generate_box_constraints} function, \code{liquidity_constraint_policy} can also contain:
\item \code{liquidity_cap_rule} lists: One or many lists used to create upper bounds for weights based on a liquidity classification. Each list must contain:
\itemize{
\item \code{liquidity_classification}: A character indicating the classification for the cap.
\item \code{liquidity_cap}: A numeric value indicating the cap (upper bound) for stocks with that liquidity classification.
Many liquidity caps might be created, and in this case, each \code{liquidity_cap_rule} must be identified with a number (e.g., liquidity_cap_rule_1, liquidity_cap_rule_2, and so on).
}
}}

\item{turnover_constraint_policy}{A named list containing objects used to build buffer zones and apply turnover constraints.
\itemize{
\item Each element will constitute a \code{buffer_zone}, being a list with three elements:
\itemize{
\item \code{liquidity_classification} element: A liquidity classification (e.g., "micro_caps", "small_caps") for that buffer zone.
\item \code{top_stock_quantile_buffer}: A numeric value indicating a buffer value that relaxes \code{top_stocks_quantile} for stocks with the specified liquidity classification.
\item \code{turnover_cap}: A numeric value specifying the turnover cap.
Stocks that are less liquid than specified for a buffer zone and have a signal higher than the respective buffer quantile will be considered eligible, even if they do not meet the \code{liquidity_floor_rule}.
}
}}

\item{concentration_constraints_policy}{A named list containing up to four elements:
\itemize{
\item \code{benchmark}: A character vector describing the benchmark to be used to apply constraint.
Must have a correspondence in \code{benchmark_weights_m_d_ref}
\item \code{max_abs_active_individual_weight}: The maximum absolute individual active weights.
\item \code{group_classification}: A character vector describing the sector classification to be used to apply sector constraints.
Must have a correspondence in \code{sector_m_d_ref}
\item \code{max_abs_active_sector_weight}: The maximum absolute sector active weight used for creating group constraints in \code{generate_sector_constraints}.
If a given sector has no eligible stock, the one with the greatest signal will be automatically promoted.
Note that, in the context of \code{generate_sector_constraints}, a \code{benchmark_weights_m_d_ref} data frame must also be supplied.
}}
}
\description{
Create individual box constraints for MTO portfolios based on liquidity, concentration and turnover constraints
}
