% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generate_group_constraints.R
\name{generate_group_constraints}
\alias{generate_group_constraints}
\title{Create group constraints for MTO portfolios}
\usage{
generate_group_constraints(
  universe_m_d_ref,
  concentration_constraint_policy,
  groups_m_d_ref,
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe of assets with tickers, is_eligible, final_signal columns and also those generated by application of constraints policies in filter_stock_universe}

\item{groups_m_d_ref}{A data frame containing columns for id, tickers, dates, and sector classification columns following a given classification method.
All tickers in the current stock universe must have a unique correspondence in the data frame.}

\item{concentration_constraints_policy}{A named list containing up to four elements:
\itemize{
\item \code{benchmark}: A character vector describing the benchmark to be used to apply constraint.
Must have a correspondence in \code{benchmark_weights_m_d_ref}
\item \code{max_abs_active_individual_weight}: The maximum absolute individual active weights.
\item \code{group_classification}: A character vector describing the sector classification to be used to apply sector constraints.
Must have a correspondence in \code{sector_m_d_ref}
\item \code{max_abs_active_sector_weight}: The maximum absolute sector active weight used for creating group constraints in \code{generate_sector_constraints}.
If a given sector has no eligible stock, the one with the greatest signal will be automatically promoted.
Note that, in the context of \code{generate_sector_constraints}, a \code{benchmark_weights_m_d_ref} data frame must also be supplied.
}}

\item{benchmark_weights_m_d_ref}{A data frame containing columns for id, tickers, dates, and current benchmark weights columns.
All tickers in the current stock universe must have a unique correspondence in this data frame.}
}
\description{
Create group constraints for MTO portfolios
}
