% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-generate_sector_constraints.R
\name{generate_sector_constraints}
\alias{generate_sector_constraints}
\title{Create sector constraints for MTO portfolios}
\usage{
generate_sector_constraints(
  current_stock_universe_df,
  sectors_constraint_policy,
  sectors_m_df
)
}
\arguments{
\item{current_stock_universe_df}{A dataframe of stocks with tickers, is_stock_eligible, final_signal columns and also those generated by application of constraints policies in filter_stock_universe}

\item{sectors_constraint_policy}{A named list with one or several elements, in which each element is a data frame with sector_classification, max and min sector weights.
Each element's name must correspond to the corresponding sector classification name, so that the user can implement sectors, subsectors, segments restrictions.
Each dataframe must have columns labeled as sector, max_weight and min_weight.}

\item{sectors_m_df}{A dataframe with id, tickers and dates with dummy sectors classifications to be used to fill NAs}
}
\description{
Create sector constraints for MTO portfolios
}
