% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_bayesian_fit}
\alias{get_bayesian_fit}
\title{Get bayesian_fit}
\usage{
get_bayesian_fit(object)
}
\arguments{
\item{object}{A \code{ss_backtest_results} object.}
}
\value{
A list of eligible signals.
}
\description{
Accessor function to retrieve bayesian_results
}
