% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_bayesian_model_parameters}
\alias{get_bayesian_model_parameters}
\alias{get_bayesian_model_parameters,bayesian_alpha_test_strategy-method}
\alias{get_bayesian_model_parameters,ss_backtest_config-method}
\title{Get Bayesian Model Parameters}
\usage{
get_bayesian_model_parameters(object)

\S4method{get_bayesian_model_parameters}{bayesian_alpha_test_strategy}(object)

\S4method{get_bayesian_model_parameters}{ss_backtest_config}(object)
}
\arguments{
\item{object}{An S4 object, typically \code{bayesian_alpha_test_strategy} or \code{ss_backtest_config}.}
}
\value{
An object of class \code{bayesian_model_parameters}.
}
\description{
Extracts the \code{bayesian_model_parameters} from an object (e.g. a
\code{bayesian_alpha_test_strategy} or an \code{ss_backtest_config} holding a Bayesian strategy).
}
\section{Functions}{
\itemize{
\item \code{get_bayesian_model_parameters(bayesian_alpha_test_strategy)}: Extract parameters from a \code{bayesian_alpha_test_strategy}.

\item \code{get_bayesian_model_parameters(ss_backtest_config)}: Extract parameters from an \code{ss_backtest_config}, if it has a Bayesian alpha test strategy.

}}
