% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_lambda.R
\name{get_best_lambda}
\alias{get_best_lambda}
\title{Get Best Lambda from Glmnet Model Based on Training/Test Sets}
\usage{
get_best_lambda(
  glmnet_fit,
  lambda_seq,
  features_validation_sample_clean,
  target_validation_sample,
  huber_delta,
  quantile_tau,
  chosen_eval_metric
)
}
\arguments{
\item{glmnet_fit}{A glmnet object obtained from fitting the model using glmnet.}

\item{lambda_seq}{Numeric vector of lambda values used in the glmnet model.}

\item{features_validation_sample_clean}{Matrix or data frame of validation set features.}

\item{target_validation_sample}{Vector of true values for the validation set.}

\item{huber_delta}{Huber loss parameter (optional, default is 1.345).}

\item{quantile_tau}{Quantile loss parameter (optional, default is 0.5).}

\item{chosen_eval_metric}{Character string specifying the evaluation metric to optimize (e.g., "mse", "mae", "quantile").}
}
\value{
The lambda value from lambda_seq that maximizes the chosen evaluation metric.
}
\description{
Given a glmnet model fit and a lambda sequence, returns the lambda (regularization parameter)
that yields the best performance metric on a validation sample.
}
