% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_concentration_constraint_policy}
\alias{get_concentration_constraint_policy}
\alias{get_concentration_constraint_policy,port_backtest_config-method}
\alias{get_concentration_constraint_policy,sb_backtest_config-method}
\title{Get the Concentration Constraint Policy}
\usage{
get_concentration_constraint_policy(object)

\S4method{get_concentration_constraint_policy}{port_backtest_config}(object)

\S4method{get_concentration_constraint_policy}{sb_backtest_config}(object)
}
\arguments{
\item{object}{An object of class \code{port_backtest_config} or \code{sb_backtest_config}.}
}
\value{
An S4 object of class \code{concentration_constraint_policy}.
}
\description{
Accessor method to extract the \code{concentration_constraint_policy} from an object.
}
\section{Functions}{
\itemize{
\item \code{get_concentration_constraint_policy(port_backtest_config)}: Extract the concentration policy from \code{port_backtest_config}.

\item \code{get_concentration_constraint_policy(sb_backtest_config)}: Extract the concentration policy from \code{sb_backtest_config},
which stores it inside \code{object@signal_port_parameters}.

}}
