% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_data}
\alias{get_data}
\alias{get_data,meta_dataframe-method}
\alias{get_data,meta_xts-method}
\title{Accessor for Data Slot}
\usage{
get_data(object)

\S4method{get_data}{meta_dataframe}(object)

\S4method{get_data}{meta_xts}(object)
}
\arguments{
\item{object}{An object of class \code{meta_dataframe} or \code{meta_xts}.}
}
\value{
A \code{data.frame} (for \code{meta_dataframe}) or \code{xts} object (for \code{meta_xts}) containing the data.
}
\description{
Retrieves the \code{data} slot from a \code{meta_dataframe} or \code{meta_xts} object.
}
