% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_dates}
\alias{get_dates}
\alias{get_dates,meta_dataframe-method}
\alias{get_dates,meta_xts-method}
\title{Accessor for Dates}
\usage{
get_dates(object, ...)

\S4method{get_dates}{meta_dataframe}(object)

\S4method{get_dates}{meta_xts}(object)
}
\arguments{
\item{object}{An object of class \code{meta_dataframe} or \code{meta_xts}.}
}
\value{
A sorted vector of dates.
}
\description{
Retrieves the unique dates from a \code{meta_dataframe} or \code{meta_xts} object.
}
