% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_eligible_signals}
\alias{get_eligible_signals}
\title{Get eligible signals}
\usage{
get_eligible_signals(object)
}
\arguments{
\item{object}{A \code{ss_backtest_results} object.}
}
\value{
A list of eligible signals.
}
\description{
Accessor function to retrieve eligible signals
}
