% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_hyperparameter_tuning_strategy}
\alias{get_hyperparameter_tuning_strategy}
\alias{get_hyperparameter_tuning_strategy,ml_experiment-method}
\title{Get Hyperparameter Tuning Strategy}
\usage{
get_hyperparameter_tuning_strategy(object)

\S4method{get_hyperparameter_tuning_strategy}{ml_experiment}(object)
}
\arguments{
\item{object}{An ml_experiment object.}
}
\value{
The \code{hyperparameter_tuning_strategy} slot of the \code{ml_experiment} object.
}
\description{
Accessor function to retrieve the hyperparameter tuning strategy from an ml_experiment object.
}
