% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_liquidity_floor_cutoffs}
\alias{get_liquidity_floor_cutoffs}
\title{Accessor for Liquidity Floor Cutoffs}
\usage{
get_liquidity_floor_cutoffs(portfolio_policies_obj)
}
\arguments{
\item{portfolio_policies_obj}{A \code{portfolio_policies} object.}
}
\value{
The liquidity floor cutoffs list.
}
\description{
Retrieves the liquidity floor cutoffs from a \code{portfolio_policies} object.
}
