% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_ml_backtest_config}
\alias{get_ml_backtest_config}
\alias{get_ml_backtest_config,ml_metabacktest_config-method}
\alias{get_ml_backtest_config,ml_backtest_results-method}
\title{Get ML Backtest Config Object}
\usage{
get_ml_backtest_config(object)

\S4method{get_ml_backtest_config}{ml_metabacktest_config}(object)

\S4method{get_ml_backtest_config}{ml_backtest_results}(object)
}
\arguments{
\item{object}{A \code{ml_metabacktest_config} or a \code{ml_backtest_results} object.}
}
\value{
The \code{ml_backtest_configs} slot of the \code{ml_metabacktest_config} object or a \code{ml_backtest_config} derived from a \code{ml_backtest_results} object.
}
\description{
Accessor function to retrieve the ml_backtest_config object from a ml_metabacktest_config object or a ml_backtest_results object
}
