% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_sb_backtest_config}
\alias{get_sb_backtest_config}
\alias{get_sb_backtest_config,sb_metabacktest_config-method}
\alias{get_sb_backtest_config,sb_backtest_results-method}
\title{Get ML Backtest Config Object}
\usage{
get_sb_backtest_config(object)

\S4method{get_sb_backtest_config}{sb_metabacktest_config}(object)

\S4method{get_sb_backtest_config}{sb_backtest_results}(object)
}
\arguments{
\item{object}{A \code{sb_metabacktest_config} or a \code{sb_backtest_results} object.}
}
\value{
The \code{sb_backtest_configs} slot of the \code{sb_metabacktest_config} object or a \code{sb_backtest_config} derived from a \code{sb_backtest_results} object.
}
\description{
Accessor function to retrieve the sb_backtest_config object from a sb_metabacktest_config object or a sb_backtest_results object
}
