% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_signal_selection_policy}
\alias{get_signal_selection_policy}
\title{Accessor for Signal Selection Policy}
\usage{
get_signal_selection_policy(portfolio_policies_obj)
}
\arguments{
\item{portfolio_policies_obj}{A \code{portfolio_policies} object.}
}
\value{
The signal selection policy list.
}
\description{
Retrieves the signal selection policy from a \code{portfolio_policies} object.
}
