% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_signal_universe}
\alias{get_signal_universe}
\title{Get signal_universe}
\usage{
get_signal_universe(object)
}
\arguments{
\item{object}{A \code{ss_backtest_results} object.}
}
\value{
A list of eligible signals.
}
\description{
Accessor function to retrieve signal universe
}
