% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_ss_backtest_config}
\alias{get_ss_backtest_config}
\alias{get_ss_backtest_config,sb_backtest_config-method}
\title{Get SS Backtest Config Object}
\usage{
get_ss_backtest_config(object)

\S4method{get_ss_backtest_config}{sb_backtest_config}(object)
}
\arguments{
\item{object}{A \code{ss_backtest_results} or a \code{sb_backtest_results} object.}
}
\value{
A \code{ss_backtest_config} derived from a \code{sb_backtest_config} \code{ss_backtest_results} or object.
}
\description{
Accessor function to retrieve the ss_backtest_config object from a ss_backtest_results object or a sb_backtest_results object
}
