% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_tickers}
\alias{get_tickers}
\alias{get_tickers,meta_dataframe-method}
\alias{get_tickers,meta_xts-method}
\title{Accessor for Tickers}
\usage{
get_tickers(object)

\S4method{get_tickers}{meta_dataframe}(object)

\S4method{get_tickers}{meta_xts}(object)
}
\arguments{
\item{object}{An object of class \code{meta_dataframe} or \code{meta_xts}.}
}
\value{
A character vector of unique tickers (column names in the case of \code{meta_xts}).
}
\description{
Retrieves the unique tickers from a \code{meta_dataframe} or \code{meta_xts} object.
}
