% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_tuning_strategy}
\alias{get_tuning_strategy}
\alias{get_tuning_strategy,sb_backtest_config-method}
\alias{get_tuning_strategy,sb_metabacktest_config-method}
\alias{get_tuning_strategy,sb_backtest_results-method}
\title{Get Hyperparameter Tuning Strategy}
\usage{
get_tuning_strategy(object)

\S4method{get_tuning_strategy}{sb_backtest_config}(object)

\S4method{get_tuning_strategy}{sb_metabacktest_config}(object)

\S4method{get_tuning_strategy}{sb_backtest_results}(object)
}
\arguments{
\item{object}{An sb_backtest_config object.}
}
\value{
The \code{tuning_strategy} slot of the \code{sb_backtest_config} object.
}
\description{
Accessor function to retrieve the hyperparameter tuning strategy from an sb_backtest_config object.
}
