% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_turnover_constraint_policy}
\alias{get_turnover_constraint_policy}
\title{Accessor for Turnover Constraint Policy}
\usage{
get_turnover_constraint_policy(portfolio_policies_obj)
}
\arguments{
\item{portfolio_policies_obj}{A \code{portfolio_policies} object.}
}
\value{
The turnover constraint policy list.
}
\description{
Retrieves the turnover constraint policy from a \code{portfolio_policies} object.
}
