% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_workflow}
\alias{get_workflow}
\alias{get_workflow,meta_dataframe-method}
\alias{get_workflow,meta_xts-method}
\title{Accessor for Workflow Slot}
\usage{
get_workflow(object)

\S4method{get_workflow}{meta_dataframe}(object)

\S4method{get_workflow}{meta_xts}(object)
}
\arguments{
\item{object}{An object of class \code{meta_dataframe} or \code{meta_xts}.}
}
\value{
A named list representing the workflow steps applied to the object.
}
\description{
Retrieves the \code{workflow} slot from a \code{meta_dataframe} or \code{meta_xts} object.
}
