% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-globals.R
\name{globalVariables_nse}
\alias{globalVariables_nse}
\title{Declare global variables used in NSE to avoid R CMD check notes}
\description{
Declare global variables used in NSE to avoid R CMD check notes
}
\keyword{internal}
