% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{grid_search_strategy-class}
\alias{grid_search_strategy-class}
\title{Grid Search Tuning Strategy}
\description{
A subclass of \code{tuning_strategy} that implements grid search.
}
\section{Slots}{

\describe{
\item{\code{tuning_method}}{The tuning method is set to 'grid_search'.}
}}

