% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{hyper_grid_domain-class}
\alias{hyper_grid_domain-class}
\title{Define the \code{hyper_grid_domain} S4 Class}
\description{
This class represents parameters for defining the hyperparameter domain based on which tuning will be performed.
It helps the user in correctly setting this object in the context of the \code{ml_walk_forward_validation} function.
}
\section{Slots}{

\describe{
\item{\code{tuning_method}}{A character string specifying the tuning method that will be applied (e.g., "grid_search", "random_search" or "bayesian_optimization").}

\item{\code{ml_algorithm}}{A character string specifying the machine-learning algorithm that will be used.}

\item{\code{hyperparameter_list}}{A list with the hyperparameters relevant to the specified machine learning algorithm.}
}}

