% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-hyper_tune.R
\name{hyper_tune}
\alias{hyper_tune}
\title{Perform Hyperparameter Tuning for Machine Learning Models}
\usage{
hyper_tune(
  tuning_method,
  ml_algorithm,
  target_fwd_name,
  full_data_training_sample_clean,
  features_validation_sample,
  target_validation_sample,
  eval_function,
  custom_objective_translated,
  chosen_eval_metric_translated,
  early_stop,
  chosen_eval_metric,
  huber_delta,
  quantile_tau,
  hyper_grid_domain_list,
  n_iter,
  init_points,
  k_iter,
  acq,
  keras_architecture_parameters,
  parallel,
  verbose
)
}
\arguments{
\item{tuning_method}{A character string specifying the tuning method. Possible values are \code{"random_search"}, \code{"grid_search"}, or \code{"bayesian_opt"}.}

\item{ml_algorithm}{A character string indicating the machine learning algorithm to be used.}

\item{target_fwd_name}{A character string specifying the name of the target variable in forward prediction.}

\item{full_data_training_sample_clean}{A data frame containing the clean training sample data.}

\item{features_validation_sample}{A data frame with feature values for validation.}

\item{target_validation_sample}{A data frame with target values for validation.}

\item{eval_function}{A function that computes the evaluation metric for given hyperparameters.}

\item{custom_objective_translated}{An optional custom objective function to be used in tuning.}

\item{chosen_eval_metric_translated}{The evaluation metric to be used internally for model performance assessment.}

\item{early_stop}{An integer specifying the number of iterations with no improvement before stopping early.}

\item{chosen_eval_metric}{A character string specifying the evaluation metric to be used.}

\item{huber_delta}{A numeric value for the Huber loss delta parameter, applicable if using Huber loss.}

\item{quantile_tau}{A numeric value for the quantile parameter, applicable if using quantile loss.}

\item{hyper_grid_domain_list}{A list specifying the domain of hyperparameters to search over.}

\item{n_iter}{An integer specifying the number of iterations for grid or random search or Bayesian optimization.}

\item{init_points}{An integer specifying the number of initial random points for Bayesian optimization.}

\item{k_iter}{An integer specifying the number of iterations for scoring function sampling in Bayesian optimization.}

\item{acq}{A character string specifying the acquisition function for Bayesian optimization.}

\item{keras_architecture_parameters}{A list of parameters specifying the architecture of the Keras model.}

\item{parallel}{A logical indicating whether to perform the search in parallel. Default is TRUE.}

\item{verbose}{A logical indicating whether to print detailed progress information. Default is FALSE.}
}
\value{
A list containing:
\item{chosen_eval_metric_validation_current_date}{A data frame of evaluation metrics for each set of hyperparameters tried.}
\item{optimal_hyper}{A named vector of the optimal hyperparameter values found.}
\item{validation_eval_metrics_hyper_choice_current_date}{A named vector of evaluation metrics corresponding to the optimal hyperparameter set.}
}
\description{
This function tunes hyperparameters of machine learning models using grid search, random search, or Bayesian optimization. It evaluates different hyperparameter combinations and returns the best-performing ones based on specified evaluation metrics.
}
