% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{hyperparameter_tuning_strategy-class}
\alias{hyperparameter_tuning_strategy-class}
\title{Base class for hyperparameter tuning strategies}
\description{
This class defines the common slots and structure for hyperparameter tuning strategies such as grid search, random search, and Bayesian optimization.
}
\section{Slots}{

\describe{
\item{\code{tuning_method}}{Character string indicating the hyperparameter tuning method ('grid_search', 'random_search', or 'bayesian_opt').}

\item{\code{ml_algorithm}}{A character string specifying the machine learning algorithm to be used ('glmnet', 'rf', 'xgb' or 'nn').}

\item{\code{validation_sample_size}}{Numeric value representing the size of the validation sample.}

\item{\code{split_method}}{Character string indicating the data splitting method ('expanding' or 'rolling').}

\item{\code{chosen_eval_metric}}{Character or NULL, specifying the evaluation metric to be optimized.}

\item{\code{hyper_grid_domain}}{An object of class hyper_grid_domain, representing the hyperparameter domain based on which tuning will pe performed.}

\item{\code{early_stop}}{Sets a halting criteria to prevent overfitting in xgb and nn.}
}}

