% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{hyperparameters}
\alias{hyperparameters}
\alias{hyperparameters,character-method}
\alias{hyperparameters,sb_backtest_config-method}
\title{Get Expected Hyperparameters for a Machine Learning Algorithm or Configuration}
\usage{
hyperparameters(object)

\S4method{hyperparameters}{character}(object)

\S4method{hyperparameters}{sb_backtest_config}(object)
}
\arguments{
\item{object}{An \code{sb_backtest_config} object.}
}
\value{
A character vector containing the names of the expected hyperparameters for the specified algorithm.

A character vector of expected hyperparameters.
If the algorithm is "ols" or not recognized, it returns an empty character vector.

A character vector of expected hyperparameters for the algorithm specified in the configuration.
If the algorithm is "ols" or not recognized, it returns an empty character vector.
}
\description{
The \code{hyperparameters} function returns a character vector of expected hyperparameters for a given machine learning algorithm or configuration.
}
\section{Methods (by class)}{
\itemize{
\item \code{hyperparameters(character)}: Get expected hyperparameters for a given machine learning algorithm.

\item \code{hyperparameters(sb_backtest_config)}: Get expected hyperparameters from an \code{sb_backtest_config} object.

}}
