% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_window.R
\name{idio_vol}
\alias{idio_vol}
\title{Calculate Idiosyncratic Volatility}
\usage{
idio_vol(ret_values, bench_ret_values, na.rm = TRUE)
}
\arguments{
\item{ret_values}{A numeric vector of stock returns.}

\item{bench_ret_values}{A numeric vector of benchmark returns.}

\item{na.rm}{A logical value indicating whether to remove NA values before computation (default is TRUE).}
}
\value{
A numeric value representing the idiosyncratic volatility. If the computed variance is negative,
\code{NA_real_} is returned.
}
\description{
This function computes the idiosyncratic volatility of a stock by fitting a linear regression model with
\code{ret_values} as the dependent variable and \code{bench_ret_values} as the independent variable.
Idiosyncratic volatility is defined as the square root of the difference between the variance of the stock returns
and the portion explained by the benchmark (i.e., \code{beta^2} times the variance of the benchmark returns).
}
\examples{
idio_vol(c(0.05, 0.06, 0.07), c(0.03, 0.04, 0.05))
}
