% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_sector_unavailable.R
\name{impute_sector_unavailable}
\alias{impute_sector_unavailable}
\title{Impute Missing Values for Entirely Unavailable Features by Sector}
\usage{
impute_sector_unavailable(
  meta_dataframe,
  sector_column,
  sectors_to_adjust,
  features_to_preserve,
  method = "mean",
  ...
)
}
\arguments{
\item{meta_dataframe}{A \code{meta_dataframe} object containing the target data.}

\item{sector_column}{A \code{character} specifying the column in \code{meta_dataframe} that contains sector classifications.}

\item{sectors_to_adjust}{A \verb{character vector} specifying which sectors should be imputed.}

\item{features_to_preserve}{A \verb{character vector} specifying which features should not be imputed.}

\item{method}{A \code{character} specifying the imputation method. Options:
\itemize{
\item "mean": Global mean
\item "median": Global median
\item "mode": Most frequent value
}}
}
\value{
A modified \code{meta_dataframe} with imputed values and an informative summary of imputation.
}
\description{
This function imputes missing values in a \code{meta_dataframe} for features that are completely missing
within specified sectors. If a feature has only NAs for all rows in the given sector, it will be
imputed using the specified method (mean, median, or mode) based on data from other sectors.
}
