% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_unavaiable_feature_fill.R
\name{industry_unavaiable_feature_fill}
\alias{industry_unavaiable_feature_fill}
\title{Fill Industry-Unavailable Feature}
\usage{
industry_unavaiable_feature_fill(
  features_df,
  unavaiable_feature,
  similar_features,
  industry_classification_column_name,
  selected_industries
)
}
\arguments{
\item{features_df}{A data frame containing the features matrix with columns for "id", "tickers" and "dates".}

\item{unavaiable_feature}{The name of the unavailable feature to be filled.}

\item{similar_features}{A character vector containing the names of similar features used for imputation.}

\item{industry_classification_column_name}{The name of the column in \code{features_df} containing industry classifications.}

\item{selected_industries}{A character vector containing the names of industries to which the imputation should be restricted.}
}
\value{
A data frame with missing values of \code{unavaiable_feature} filled with the mean of \code{similar_features} within specified industries.
}
\description{
This function fills missing values of an unavailable feature in a features matrix with the mean value of similar features.
An industry-unavaiable feature is a feature that is not defined for a given industry.
This is a rather common problem for banks, in which some accounting information are not defined.
For instance, one might replace FCF Yield by Dividend Yield.
}
\examples{
features_df <- data.frame(id = 1:5, 
                               tickers = c("AAPL", "GOOG", "JPM", "WFC", "FB"), 
                               dates = as.Date("2022-01-01") + 0:4,
                               industry = c("Tech", "Tech", "Banks", "Banks", "Tech"),
                               feature1 = c(NA, 2, NA, 4, 5),
                               feature2 = c(1, NA, NA, 3, NA))
industry_unavaiable_feature_fill(features_df, "feature1", "feature2", "industry", c("Finance"))
}
