% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/industry_weighted_mean.R
\name{industry_weighted_mean}
\alias{industry_weighted_mean}
\title{Industry Weighted Mean}
\usage{
industry_weighted_mean(
  characteristic_matrix,
  sector_classification,
  weighting_matrix,
  ew
)
}
\arguments{
\item{characteristic_matrix}{A matrix containing characteristics.}

\item{sector_classification}{A matrix containing sector classifications.}

\item{weighting_matrix}{A matrix containing weighting variables.}

\item{ew}{A binary indicator (0 or 1) indicating whether to use equal weights (ew = 1) or weighted means (ew = 0).}
}
\value{
A matrix of industry weighted means.
}
\description{
This function calculates the industry weighted mean of characteristics for each sector.
}
\details{
When elements in weighting matrix are missing, they are set to zero. When there is NA in sector classification, chacteristic value is ignored.
}
\examples{
# Example usage:
# industry_weighted_mean(characteristic_matrix, sector_classification, weighting_matrix, ew = 0)

}
