% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_coercible_to_fwd_meta_dataframe}
\alias{is_coercible_to_fwd_meta_dataframe}
\title{Check if an Object is Coercible to \code{fwd_meta_dataframe}}
\usage{
is_coercible_to_fwd_meta_dataframe(obj)
}
\arguments{
\item{obj}{The object to be checked.}
}
\value{
A logical value indicating whether the object is coercible to \code{fwd_meta_dataframe}.
Additionally, if the object is not coercible, a message will be displayed explaining why.
}
\description{
This function checks if an object can be coerced into a \code{fwd_meta_dataframe} class.
It ensures that the object meets the structural requirements of \code{fwd_meta_dataframe},
including correct column names and formats.
}
\examples{
df <- data.frame(
  id = c("A-2024-01-01", "B-2024-02-01"),
  tickers = c("A", "B"),
  dates = as.Date(c("2024-01-01", "2024-02-01")),
  value_01 = c(10, 20),
  value_02 = c(15, 25)
)
is_coercible_to_fwd_meta_dataframe(df)

}
