% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_coercible_to_meta_dataframe}
\alias{is_coercible_to_meta_dataframe}
\title{Check if an object is coercible to meta_dataframe}
\usage{
is_coercible_to_meta_dataframe(obj)
}
\arguments{
\item{obj}{An R object to be checked for coercibility.}
}
\value{
A logical value indicating whether the object can be coerced to \code{meta_dataframe}.
}
\description{
This function checks if an object can be converted to a \code{meta_dataframe} class
by verifying the required columns, data types, and other constraints. It provides
detailed messages explaining why the object is not coercible.
}
