% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_hyper_grid_domain}
\alias{is_hyper_grid_domain}
\title{Check if an object is of class hyper_grid_domain}
\usage{
is_hyper_grid_domain(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
Logical value: TRUE if the object is of class 'hyper_grid_domain', FALSE otherwise.
}
\description{
This function checks whether a given object is of class 'hyper_grid_domain'.
}
