% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_keras_architecture_parameters}
\alias{is_keras_architecture_parameters}
\title{Define the is keras_architecture_parameters function}
\usage{
is_keras_architecture_parameters(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether the object is of class \code{keras_architecture_parameters}.
}
\description{
Function to check if an object is of class \code{keras_architecture_parameters}.
}
