% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_meta_dataframe}
\alias{is_meta_dataframe}
\title{Check if an object is a meta_dataframe}
\usage{
is_meta_dataframe(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
TRUE if x is of class "meta_dataframe", FALSE otherwise.
}
\description{
Check if an object is a meta_dataframe
}
