% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_portfolio_policies}
\alias{is_portfolio_policies}
\title{Define the is portfolio_policies function}
\usage{
is_portfolio_policies(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether the object is of class \code{portfolio_policies}.
}
\description{
Function to check if an object is of class \code{portfolio_policies}.
}
