% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_tuning_strategy}
\alias{is_tuning_strategy}
\title{Define the is_hyperparameter_tuning_strategy function}
\usage{
is_tuning_strategy(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value indicating whether the object is of class \code{hyperparameter_tuning_strategy}.
}
\description{
Function to check if an object is of class \code{hyperparameter_tuning_strategy}.
}
