% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{lookup_catalog}
\alias{lookup_catalog}
\title{Lookup method for tickers_catalog}
\usage{
lookup_catalog(tickers_catalog, ...)
}
\arguments{
\item{tickers_catalog}{A tickers_catalog object.}

\item{tickers_to_lookup}{A character vector of tickers to filter.}

\item{perm_id_to_lookup}{A character vector of perm_ids to filter.}
}
\value{
A filtered data.frame.
}
\description{
Filters the catalog by provided tickers.
}
