% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_meta_dataframe.R
\name{mean_std}
\alias{mean_std}
\title{Calculate Mean-to-Standard Deviation Ratio (Signal-to-Noise Ratio)}
\usage{
mean_std(values, na.rm = TRUE)
}
\arguments{
\item{values}{A numeric vector.}

\item{na.rm}{A logical value indicating whether to remove NA values before computation (default is TRUE).}
}
\value{
A numeric value representing the mean-to-std ratio. If the standard deviation is zero,
\code{NA_real_} is returned.
}
\description{
This function computes the mean-to-standard deviation ratio of values in a numeric vector.
This ratio is also known as the inverse of the coefficient of variation or, in finance, as the Sharpe Ratio when the risk-free rate is zero.
}
\examples{
mean_std(c(1, 2, 3, 4, 5))
}
