% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_matrix_36m.R
\name{median_matrix_36m}
\alias{median_matrix_36m}
\title{Calculate element-wise 36m median matrix.
Useful for calculating the median of firm-level data from last 3 years.}
\usage{
median_matrix_36m(matrix1, matrix2, matrix3)
}
\arguments{
\item{matrix1}{A matrix with firm-level data}

\item{matrix2}{A matrix with firm-level data}

\item{matrix3}{A matrix with firm-level data}
}
\value{
A matrix in which inputs are element-wise medians of corresponding elements in each matrix.
The function automatically handles \code{NA} values by ignoring them during median calculation.
}
\description{
Calculate element-wise 36m median matrix.
Useful for calculating the median of firm-level data from last 3 years.
}
\details{
This function computes the element-wise median across three matrices of equal dimensions.
It is particularly useful for scenarios requiring aggregation of data over time or among different data sets.
By default, \code{NA} values are excluded from the median calculation. If all values in a position across matrices
are \code{NA}, the result will be \code{NA} for that position.
}
\examples{
median_matrix_36m(
  matrix(c(1,2,3,4), nrow=2, ncol=2),
  matrix(c(5,6,7,8), nrow=2, ncol=2),
  matrix(c(1,1,1,1), nrow=2, ncol=2))
}
