% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{meta_dataframe_accessors}
\alias{meta_dataframe_accessors}
\title{Accessor Methods for meta_dataframe}
\usage{
get_data(object)
}
\arguments{
\item{object}{An object of class \code{meta_dataframe}.}
}
\value{
The respective slot of the \code{meta_dataframe} object.
}
\description{
These methods are used to access components of a \code{meta_dataframe} object.
}
