% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabacktest.R
\name{metabacktest}
\alias{metabacktest}
\title{Backtest a porfolio based on signals derived from simple stock characteristics (simple factors) or
expected returns from ml model predictions}
\usage{
metabacktest(
  signals_m_df,
  ml_walk_forward_validation_results_list,
  dates_m_vector,
  rebalancing_months,
  n_months_until_first_rebalancing_date,
  selected_signal,
  chosen_characteristics,
  chosen_ml_algorithms,
  liquidity_m_df,
  benchmark_weights_m_df,
  volatility_m_df,
  target_m_df
)
}
\arguments{
\item{signals_m_df}{A matrix or data frame containing simple features, with columns: id, tickers, dates,
to be used for signal-portfolio construction}

\item{ml_walk_forward_validation_results_list}{Optional. A list of one or more objects of class ml_walk_forward_validation_results
whose oos_prediction_list is to be used as signals and chosen_eval_metric_validation can be used to produce ml ensembles}

\item{dates_m_vector}{A vector of dates corresponding to the data.}

\item{rebalancing_months}{Months (numeric) when portfolio should be rebalanced.}

\item{selected_signal}{One of ml_model, characteristics or meta_signal IS THIS REALLY NECESSARY????????}

\item{chosen_characteristics}{A vector of characteristics that indicates which cols in features_m_df should be used as signals}

\item{chosen_ml_algorithms}{A vector of ml_algorithms that indicate which ml model predictions of ml_walk_forward_validation_results_list
should be used as signals. If NULL, all objects of ml_walk_forward_validation_results wil be considered}

\item{liquidity_m_df}{A data frame (similar to \code{features_m_df}) containing columns for id, tickers, dates,
and one or more market liquidity measures (e.g., inflation-adjusted mean financial volume).}

\item{nmonths_until_first_rebalancing_date}{The amount of dates in dates_m_vector to be skipped before backtest starts}

\item{portfolio_construction_method}{The type of portfolio construction method for setting weights. Possibilities are heuristics (EW, CW, CS, SW),
Mean Tracking-Error (MTO) and Risk-Parity (RP)}

\item{covariance_estimation_method}{One of SAM (Sample), EWMA, CC (Constant Correlation), PCA1, PCA2, Shrink_ID or Shrink_CC.
If NULL, blending_method can only be EW or IR}
}
\description{
Backtest a porfolio based on signals derived from simple stock characteristics (simple factors) or
expected returns from ml model predictions
}
