% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{metrics_meta_xts-class}
\alias{metrics_meta_xts-class}
\title{An S4 subclass of meta_xts for metric time series (holes allowed).}
\description{
In addition to the parent slots, it has:
\itemize{
\item \code{metrics}: a character vector with names of metric columns.
\item \code{n_metrics}: a numeric value equal to the number of metric columns.
}
}
\section{Slots}{

\describe{
\item{\code{series}}{Character. Names of the columns (metrics).}

\item{\code{n_series}}{Numeric. Number of metric columns.}
}}

