% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_performance_m_df.R
\name{min_track_record}
\alias{min_track_record}
\title{Compute the Minimum Track Record Length}
\usage{
min_track_record(x)
}
\arguments{
\item{x}{A numeric vector of returns (in decimal form).}
}
\value{
A named list with:
\describe{
\item{min_trl}{Minimum track record length required.}
\item{is_significant}{Logical value: is the current Sharpe Ratio statistically significant?}
\item{num_of_extra_obs_needed}{Number of additional observations needed (0 if already significant).}
}
}
\description{
Calculates the minimum number of observations required for the observed Sharpe Ratio
to be statistically greater than a reference Sharpe Ratio at a given confidence level,
adjusting for skewness and kurtosis.
}
\details{
This implementation follows the methodology from Bailey and Lopez de Prado (2012),
and assumes the returns will maintain similar statistical properties out-of-sample.
}
