% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{ml_backtest_config-class}
\alias{ml_backtest_config-class}
\title{Machine Learning Experiment Class}
\description{
The ml_backtest_config class is designed to define an end-to-end machine learning experiment, including the hyperparameter tuning strategy, algorithm parameters, and other experiment-specific configurations.
}
\section{Slots}{

\describe{
\item{\code{target_fwd_name}}{Character string indicating the forward target variable's name. Should be in the format 'xxxxxx_ym'.}

\item{\code{ml_algorithm}}{Character string specifying the machine learning algorithm to be used ('glmnet', 'rf', 'xgb', 'nn').}

\item{\code{tuning_strategy}}{An object of class tuning_strategy, specifying the strategy for tuning hyperparameters.}

\item{\code{custom_objective}}{Character string specifying the custom objective function ('squared_error', 'pseudo_huber_error', 'absolute_error') or NULL.}

\item{\code{keras_architecture_parameters}}{List or NULL, providing parameters specific to keras-based neural networks.}

\item{\code{quantile_tau}}{Numeric value indicating the tau parameter used for quantile regression, between 0 and 1.}

\item{\code{huber_delta}}{Numeric value greater than 0, specifying the delta parameter for Huber loss function.}
}}

