% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{ml_metabacktest_config-class}
\alias{ml_metabacktest_config-class}
\title{ml_metabacktest_config Class}
\description{
The ml_metabacktest_config class is designed to store and manage a collection of ml_backtest_config objects.
}
\section{Slots}{

\describe{
\item{\code{meta_ml_backtest_config}}{A \code{ml_backtest_config} with the configuration for the meta learner}

\item{\code{base_ml_backtest_configs}}{A list of ml_backtest_config objects whose oos predictions will be fed to the meta learner.}
}}

