% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_walk_forward_validation.R
\name{ml_walk_forward_validation}
\alias{ml_walk_forward_validation}
\title{Perform out-of-sample testing for ML Algorithms with walk-forward time series validation}
\usage{
ml_walk_forward_validation(
  features_m_df,
  target_m_df,
  dates_m_vector,
  training_sample_size,
  target_fwd,
  target_fwd_name,
  validation_sample_size = 0,
  rebalancing_months,
  split_method = "expanding",
  ml_algorithm,
  custom_objective = "squared_error",
  chosen_eval_metric = NULL,
  huber_delta = 1,
  quantile_tau = 0.5,
  hyper_grid_domain_list = NULL,
  tuning_method = NULL,
  n_iter = NULL,
  k_iter = NULL,
  acq = "ucb",
  init_points = NULL,
  early_stop = NULL,
  keras_architecture_parameters = NULL,
  show_plots = TRUE,
  verbose = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{features_m_df}{A matrix or data frame containing features, with columns: id, tickers, dates.}

\item{target_m_df}{A matrix or data frame containing target variable(s), with corresponding dates.}

\item{dates_m_vector}{A vector of dates corresponding to the data.}

\item{training_sample_size}{Number of observations to include in each training sample.}

\item{target_fwd}{Number of periods ahead for the target variable.}

\item{target_fwd_name}{Name of the target variable in \code{target_m_df}.}

\item{validation_sample_size}{Number of observations to include in each validation sample.}

\item{rebalancing_months}{Months (numeric) when model should be rebalanced.}

\item{split_method}{Choice of split method (expanding or rolling)}

\item{ml_algorithm}{Choice of ml_algorithm: glmnet, rf, xgboost and neural networks.}

\item{custom_objective}{Custom objective (double differenciable loss function) to xgboost and nn algorithms
(current options are squared_error, absolute_error and (pseudo)-huber loss)}

\item{chosen_eval_metric}{Metric to optimize during tuning: "rss", "rmse", "cp", "mae", "mphe", "mpe", "mape", "hr" and "mb".}

\item{huber_delta}{#A single numeric value indicating the boundary that separates where loss function turns from quadratic to linear}

\item{quantile_tau}{#A single numeric value indicating target quantile when calculating quantile loss}

\item{hyper_grid_domain_list}{List specifying hyperparameter ranges for tuning.}

\item{tuning_method}{Method for hyperparameter tuning: "random_search", "grid_search", or "bayesian_opt".}

\item{n_iter}{Number of iterations for random search.}

\item{k_iter}{Integer that specifies the number of times to sample eval_function at each Epoch during bayesian optimization.
If running in parallel, set iters.k to multiple of number of cores. Must be lower and preferrably, multiple of n_ter}

\item{acq}{Acquisition function for Bayesian optimization: "ucb", "ei", or "poi".}

\item{init_points}{Number of initial random points for Bayesian optimization.}

\item{early_stop}{Sets a halting criteria to prevent overfitting in xgb and nn}

\item{keras_architecture_parameters}{Named list, containing units (a vector containing number of neurons in each layer),
n_layers (number of layers), activation (a character vector of activation function at each layer), nn_optimizer (Adam or RMSProp) and
batch_norm_option (a logical vector indicating whether normalization after each respective layer should be performed).}

\item{show_plots}{Logical, indicating whether to plot results (default is TRUE).}

\item{verbose}{Logical, indicating whether to print progress messages (default is TRUE).}

\item{parallel}{Logical, indicating whether to run hyperparameter tuning in parallel (default is TRUE)}
}
\value{
List with various outputs including model predictions, errors, and validation metrics.
}
\description{
This function performs walk-forward validation for time series data using
a range of ML models. It supports hyperparameter tuning via random search, grid search,
or Bayesian optimization. The function divides the data into training, validation,
and testing samples, and iteratively refits the model at specified rebalancing dates.
}
\details{
The function ensures all inputs are correctly formatted and performs checks to
validate the integrity of input data and parameters. It employs the glmnet package
for elastic net regularization, supporting both Lasso and Ridge regression.
}
\section{Running in Parallel}{


By default, tuning_method \%in\% c("random_search", "grid_search") utilizes furrr::future_pmap, which means they can run according to the built-in backends
from the future package. Therefore, if the user does not specify a different evaluation strategy with future::plan(),
tuning will be done sequentially by default (equivalent to future::plan(sequential)).In this case, however,
random number generator will be set to RNGkind("L'Ecuyer-CMRG"), instead of R default (RNGkind("Mersenne-Twister")), making results
not reproducible regarding using purrr:pmap(). In order to run using R's default random number generator, set parallel = FALSE.
Using a different evaluation strategy (eg., future::plan(multisession)) will tune hyperparameters asynchronously (in parallel).

For tuning_method = "bayesian_opt", the ParBayesianOptimization::bayesOpt function runs in parallel by using foreach::foreach with the \%dopar\% operator.
Therefore, in this case, the user can either: (i) use doFuture::registerDoFuture(), in order to use the \%dofuture\% foreach adapter
(actually, in this case, doFuture::withDoRNG is used to turn \%dopar\% to \%dorng\% in order to use parallel-safe RNG), which allows
usage of backends from future package or (ii) use parallel::makeCluster(), doParallel::registerDoParallel(), doParallel::clusterExport() and
doParallel::clusterEvalQ(), as exemplified by ParBayesianOptimization. If parallel = TRUE and neither strategy is being used,
code will result in error. Therefore, to run bayesian_opt synchronously, either use doFuture::registerDoFuture() with plan(sequential)
or set parallel = FALSE.

Keras has some limitations when working in parallel, especially when using bayesian optimization as tuning method.
}

\seealso{
\code{\link{glmnet}}, \code{\link{time_series_split}}
}
