% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{ml_wf_val_results_accessors}
\alias{ml_wf_val_results_accessors}
\title{Accessor Methods for ml_wf_val_results}
\usage{
get_oos_prediction_list(object)
}
\arguments{
\item{object}{An object of class \code{ml_wf_val_results}.}
}
\value{
The respective slot of the \code{ml_wf_val_results} object.
}
\description{
These methods are used to access various components of an \code{ml_wf_val_results} object.
}
