% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{mmaf_parameters-class}
\alias{mmaf_parameters-class}
\title{Define the \code{mmaf_parameters} S4 Class}
\description{
S4 class to represent a set of configurations for the MMAF portfolio construction method.
}
\section{Slots}{

\describe{
\item{\code{mmaf_method}}{A character indicating the MMAF method to be used. Must be one of 'top_down' or 'bottom_up'.}

\item{\code{top_down_proxy_port_method}}{A character indicating the method to be used for constructing the top-down proxy portfolio.
Must be one of 'ew', 'rp', 'hrp', 'cs' or 'sw' if mmaf_method is 'top_down' and NULL if mmaf_method is 'bottom_up'.}

\item{\code{mmaf_group_col}}{A character string representing the MMAF group to which the assets belong. This is used to group assets when constructing micro and macro portfolios.
It must be a length 1 character.}

\item{\code{micro_port_config}}{An object of class \code{mmaf_sub_port_config} representing the configuration for constructing micro portfolios.}

\item{\code{macro_port_config}}{An object of class \code{mmaf_sub_port_config} representing the configuration for constructing macro portfolios.}
}}

