% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dataframe.R
\docType{class}
\name{oos_sb_outputs_m_df-class}
\alias{oos_sb_outputs_m_df-class}
\title{Define the oos_sb_outputs_m_df S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to the output of a
signal blending backtest workflow.
}
