% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelize_data.R
\name{panelize_data}
\alias{panelize_data}
\title{Panelize Data}
\usage{
panelize_data(features_list, row_names, column_names, features_names)
}
\arguments{
\item{features_list}{A list of matrices, data frames or tibbles containing the features.}

\item{row_names}{A vector of row names for the panel data.}

\item{column_names}{A vector of column names for the panel data.}

\item{features_names}{A vector of names for each feature.}
}
\value{
A panel data matrix.
}
\description{
Convert a list of matrices or data frames into panel data format.
}
\details{
This function takes a list of matrices or data frames, each representing a set of features for a group of entities, and converts them into a panel data format. Each matrix or data frame in the list represents a time series of features for a set of entities. The resulting panel data matrix has one row for each combination of entity and time point, with columns representing different features.
}
\examples{
# Example usage
features_list <- list(matrix(1:9, nrow = 3), matrix(11:19, nrow = 3))
row_names <- c("A", "B", "C")
column_names <- c("X", "Y", "Z")
features_names <- c("Feature1", "Feature2")
panelize_data(features_list, row_names, column_names, features_names)



}
