% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,bayesian_alpha_test_strategy,ANY-method}
\alias{plot,bayesian_alpha_test_strategy,ANY-method}
\title{Plot Bayesian Alpha Test Strategy Priors}
\usage{
\S4method{plot}{bayesian_alpha_test_strategy,ANY}(x, y, ...)
}
\arguments{
\item{x}{A \code{bayesian_alpha_test_strategy} object.}

\item{y}{Ignored. Included for S3 compatibility.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Plots the prior distributions defined in a \code{bayesian_alpha_test_strategy} object.
}
