% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,bayesian_model_parameters,ANY-method}
\alias{plot,bayesian_model_parameters,ANY-method}
\title{Plot Bayesian Model Parameters Priors}
\usage{
\S4method{plot}{bayesian_model_parameters,ANY}(x, y, ...)
}
\arguments{
\item{x}{A \code{bayesian_model_parameters} object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Plots the distribution curves for each prior in a \code{bayesian_model_parameters} object, with a cyberpunk theme.
}
