% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,grid_search_strategy,missing-method}
\alias{plot,grid_search_strategy,missing-method}
\title{Plot Method for \code{grid_search_strategy}}
\usage{
\S4method{plot}{grid_search_strategy,missing}(x, y, palette = "cyberpunk")
}
\arguments{
\item{x}{An object of class \code{grid_search_strategy}.}

\item{y}{Unused. Included for consistency with the generic \code{plot} method.}

\item{palette}{Character. Color palette to use for the plot. Options include "cyberpunk" and "br". Default is "cyberpunk".}
}
\value{
A \code{ggplot} object visualizing the hyperparameter grid and possible limits.
}
\description{
Plot the values selected for each hyperparameter in \code{hyper_grid_domain} for grid search strategy.
}
