% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,meta_dataframe-method}
\alias{plot,meta_dataframe-method}
\title{Plot Method for Meta Dataframe}
\usage{
\S4method{plot}{meta_dataframe}(x, type = "distribution")
}
\arguments{
\item{x}{A \code{meta_dataframe} object containing the data to be plotted.}

\item{type}{A character string specifying the type of plot to create. Options are:
\itemize{
\item \strong{"distribution"}: Plots the distribution of values for numeric columns.
\item \strong{"date_range"}: Visualizes the frequency of data points over time.
\item \strong{"unique_tickers"}: Displays the number of unique tickers over time.
}}
}
\value{
A ggplot object representing the requested plot.
}
\description{
This method generates plots for a \code{meta_dataframe} object based on the specified plot type.
}
