% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,ml_backtest_config,missing-method}
\alias{plot,ml_backtest_config,missing-method}
\title{Plot Method for \code{ml_backtest_config}}
\usage{
\S4method{plot}{ml_backtest_config,missing}(x, y)
}
\arguments{
\item{x}{An object of class \code{ml_backtest_config}.}

\item{y}{Unused. Included for consistency with the generic \code{plot} method.}
}
\value{
A \code{ggplot} object visualizing the hyperparameter histograms with possible limits.
}
\description{
Calls the appropriate plot method for \code{tuning_strategy}.
}
