% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,ml_metabacktest_config,missing-method}
\alias{plot,ml_metabacktest_config,missing-method}
\title{Plot Method for \code{ml_metabacktest_config}}
\usage{
\S4method{plot}{ml_metabacktest_config,missing}(x, y)
}
\arguments{
\item{x}{An object of class \code{ml_metabacktest_config}.}

\item{y}{Unused. Included for consistency with the generic \code{plot} method.}
}
\value{
A combined \code{ggplot} object visualizing the hyperparameter distributions for all configurations.
}
\description{
Combines individual plots from \code{ml_backtest_config} objects in a grid layout.
}
