% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,ml_wf_val_results-method}
\alias{plot,ml_wf_val_results-method}
\title{Plot Machine Learning Walk-Forward Validation Results}
\usage{
\S4method{plot}{ml_wf_val_results}(x)
}
\arguments{
\item{x}{An object of class \code{ml_wf_val_results} containing the results of the walk-forward validation.}
}
\value{
The following plots:
\itemize{
\item \code{chosen_val_metric_over_time}: Plot of the chosen evaluation metric over time for the test data, including overall and yearly means.
\item \code{test_vs_val_chosen_eval_metric_over_time}: Plot comparing the chosen evaluation metric over time for both test and validation data.
\item \code{best_hyper_over_time}: Plot of the best hyperparameter values over time, with separate facets for each hyperparameter.
\item \code{hyper_vs_error}: Plot showing the performance of hyperparameter choices with respect to the chosen evaluation metric. The plot varies depending on the machine learning algorithm used.
\item \code{all_eval_metrics_over_time}: Plot of all evaluation metrics over time, including dashed lines for variable means and vertical lines for rebalancing dates.
}
}
\description{
This method generates various plots to visualize the performance of machine learning models using walk-forward validation metrics.
It creates plots comparing out-of-sample (OOS) testing metrics, validation metrics, and hyperparameter performance over time.
}
