% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,sb_metabacktest_config,missing-method}
\alias{plot,sb_metabacktest_config,missing-method}
\title{Plot Method for \code{sb_metabacktest_config}}
\usage{
\S4method{plot}{sb_metabacktest_config,missing}(x, y, plot_id = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{sb_metabacktest_config}.}

\item{y}{Unused. Included for consistency with the generic \code{plot} method.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A combined \code{ggplot} object visualizing the hyperparameter histograms for the selected configurations.
}
\description{
Allows the user to plot either the base learners or the meta learner configurations.
If \code{base_sb_backtest_results} is provided, it extracts the configurations using \code{get_sb_backtest_config}.
}
