% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,sb_model,missing-method}
\alias{plot,sb_model,missing-method}
\title{Plot Method for 'sb_model' Objects}
\usage{
\S4method{plot}{sb_model,missing}(x, type = NULL, palette = "cyberpunk", ...)
}
\arguments{
\item{x}{An object of class \code{sb_model}.}

\item{type}{Currently unused. Included for compatibility with other plot methods.}

\item{palette}{Character. Color palette to use for the plot. Options include "cyberpunk", "br" and "journal". Default is "cyberpunk". This will be passed to the underlying model's plot method if applicable.}

\item{...}{Additional arguments passed to the plot method of the underlying model.}
}
\description{
This method dispatches plotting to the underlying model stored in the \code{model} slot of a \code{sb_model} object.
}
