% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot_ml_walk_forward_validation.R
\name{plot_ml_walk_forward_validation}
\alias{plot_ml_walk_forward_validation}
\title{Plot Machine Learning Walk-Forward Validation Results}
\usage{
plot_ml_walk_forward_validation(
  oos_testing_eval_metrics,
  validation_eval_metrics_hyper_choice,
  hyper_choice_df,
  chosen_eval_metric,
  chosen_eval_metric_validation,
  ml_algorithm,
  rebalance_dates,
  show_plots = TRUE
)
}
\arguments{
\item{oos_testing_eval_metrics}{A data frame containing out-of-sample testing evaluation metrics. The data frame should have dates as row names and metric values as columns. The dates should be in "YYYY-MM-DD" format.}

\item{validation_eval_metrics_hyper_choice}{A data frame containing validation evaluation metrics and hyperparameter choices. The data frame should have dates as row names and metric values as columns. The dates should be in "YYYY-MM-DD" format.}

\item{hyper_choice_df}{A data frame containing hyperparameter choices and their corresponding evaluation metrics. The data frame should have dates as row names and hyperparameter values as columns.}

\item{chosen_eval_metric}{A string specifying the chosen evaluation metric for analysis. This metric must be present in both the OOS testing and validation data frames.}

\item{chosen_eval_metric_validation}{A list of data frames, where each data frame contains hyperparameter choices and evaluation metrics for validation. The list should have one data frame for each validation period.}

\item{ml_algorithm}{A string specifying the machine learning algorithm used. It determines how hyperparameter performance is plotted. Possible values include "glmnet", "rf", "xgb", and "nn".}

\item{rebalance_dates}{A vector of dates indicating the times when model rebalancing occurred. These dates are used to add vertical lines to some plots.}

\item{show_plots}{A logical value indicating whether to display the plots. Default is \code{TRUE}.}
}
\value{
A list containing the following plots:
\itemize{
\item \code{chosen_val_metric_over_time}: Plot of the chosen evaluation metric over time for the test data, including overall and yearly means.
\item \code{test_vs_val_chosen_eval_metric_over_time}: Plot comparing the chosen evaluation metric over time for both test and validation data.
\item \code{best_hyper_over_time}: Plot of the best hyperparameter values over time, with separate facets for each hyperparameter.
\item \code{hyper_vs_error}: Plot showing the performance of hyperparameter choices with respect to the chosen evaluation metric. The plot varies depending on the machine learning algorithm used.
\item \code{all_eval_metrics_over_time}: Plot of all evaluation metrics over time, including dashed lines for variable means and vertical lines for rebalancing dates.
}
}
\description{
This function generates various plots to visualize the performance of machine learning models using walk-forward validation metrics. It creates plots comparing out-of-sample (OOS) testing metrics, validation metrics, and hyperparameter performance over time.
}
