% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{portfolio_policies-class}
\alias{portfolio_policies-class}
\title{Define the \code{portfolio_policies} S4 Class}
\description{
S4 class to represent a set of portfolio policies, including liquidity, turnover, concentration and signal selection constraints.
}
\section{Slots}{

\describe{
\item{\code{liquidity_constraint_policy}}{A list that contains liquidity constraint policies.
It must include a character element named \code{liquidity_floor_rule} and one or more rules,
each represented as a list with \code{liquidity_classification} and \code{liquidity_cap}.}

\item{\code{signal_selection_policy}}{A list describing signal selection policy.}

\item{\code{turnover_constraint_policy}}{A list describing turnover constraint policy.}

\item{\code{concentration_constraint_policy}}{A list describing concentration constraint policy.}

\item{\code{liquidity_floor_cutoffs}}{A list for liquidity floor cutoffs.}
}}

