% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_methods.R
\name{predict,refit_ml_model-method}
\alias{predict,refit_ml_model-method}
\title{Predict Method for refit_ml_model Class}
\usage{
\S4method{predict}{refit_ml_model}(object, new_features_m_df)
}
\arguments{
\item{object}{An instance of the \code{refit_ml_model} class containing the
refitted model and its parameters.}

\item{new_features_m_df}{A data frame or an object of class \code{meta_dataframe}
containing new feature data for which predictions are to be made. The
data frame must be structured correctly and should not include the first
three columns, which are reserved for identifiers.}
}
\value{
A numeric vector of predictions for the new feature data.
}
\description{
This method generates predictions using a refitted machine learning model
based on the provided new feature data. It accommodates different machine
learning algorithms and applies the appropriate prediction logic.
}
\details{
The function first validates that \code{new_features_m_df} is coercible
to a \code{meta_dataframe}. It extracts the relevant data from the input and
uses the appropriate prediction method based on the specified machine
learning algorithm (e.g., OLS, GLMNET, RF, XGB, NN).
}
\examples{
# Assuming `refit_model` is an instance of `refit_ml_model`
# and `new_data` is a properly structured data frame:
predictions <- predict(refit_model, new_data)

}
