% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_methods.R
\name{predict,sb_backtest_results-method}
\alias{predict,sb_backtest_results-method}
\title{Predict Method for sb_backtest_results Class}
\usage{
\S4method{predict}{sb_backtest_results}(
  object,
  new_features_m_df,
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975
)
}
\arguments{
\item{object}{An instance of the \code{sb_backtest_results} class containing the
validated model, metadata, and best hyperparameters.}

\item{new_features_m_df}{A data frame or an object of class \code{meta_dataframe}
containing new feature data for which predictions are to be made. The
data frame must be structured correctly and should not include the first
three columns, which are reserved for identifiers.}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization}
}
\value{
A numeric vector of predictions for the new feature data.
}
\description{
This method generates predictions using a sb model that has been
validated through walk-forward validation. It uses the provided new feature
data and applies the appropriate prediction logic based on the underlying
model and its hyperparameters.
}
\details{
The function validates that \code{new_features_m_df} is coercible to a
\code{meta_dataframe}. It extracts the relevant data and uses the appropriate
prediction method based on the specified machine learning algorithm (e.g., OLS,
GLMNET, RF, XGB, NN). The method retrieves the refitted model and the best
hyperparameters for making predictions.
}
\examples{
# Assuming `ml_wf_model` is an instance of `sb_backtest_results`
# and `new_data` is a properly structured data frame:
predictions <- predict(ml_wf_model, new_data)

}
