% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_methods.R
\name{predict,sb_model-method}
\alias{predict,sb_model-method}
\title{Predict Method for sb_model Class and meta_dataframe}
\usage{
\S4method{predict}{sb_model}(
  object,
  new_features_m_df,
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975
)
}
\arguments{
\item{object}{An instance of the \code{sb_model} class containing the
refitted model and its parameters.}

\item{new_features_m_df}{A  \code{meta_dataframe} or a coercible data.frame
containing new feature data for which predictions are to be made.}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization}
}
\value{
A numeric vector of predictions for the new feature data.
}
\description{
This method generates predictions using a refitted sb model
based on the provided new feature data. It accommodates different machine
learning algorithms and applies the appropriate prediction logic.
}
\details{
The function first validates that \code{new_features_m_df} is coercible
to a \code{meta_dataframe}. It extracts the relevant data from the input and
uses the appropriate prediction method based on the specified signal blending algo
}
\examples{
# Assuming `refit_model` is an instance of `sb_model`
# and `new_data` is a properly structured data frame:
predictions <- predict(refit_model, new_data)

}
