% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ml_model.R
\name{predict_ml_model}
\alias{predict_ml_model}
\title{Predict with Machine Learning Models}
\usage{
predict_ml_model(
  ml_algorithm = NULL,
  refit_model = NULL,
  best_lam = NULL,
  ml_walk_forward_validation_results = NULL,
  new_features_m_d_ref
)
}
\arguments{
\item{ml_algorithm}{A character string specifying the machine learning algorithm used. Options are "ols", "glmnet", "rf", "xgb", and "nn".}

\item{refit_model}{The refitted model object to be used for making predictions. The type of model object depends on the algorithm specified.}

\item{best_lam}{A numeric value for the lambda parameter. Only used for "glmnet" models, where it specifies the lambda value for prediction.}

\item{ml_walk_forward_validation_results}{Results object from ml_walk_forward_validation}

\item{new_features_m_d_ref}{A data frame or matrix containing the new feature data for prediction. The first three columns are assumed to be non-predictive and are excluded from the predictions.}
}
\value{
A numeric vector of predictions.
}
\description{
This function generates predictions from a refitted machine learning model based on the specified algorithm and hyperparameters.
}
\details{
The function uses the appropriate prediction method based on the \code{ml_algorithm} specified. It assumes that:
\itemize{
\item For "ols", the model is a linear regression model from the \code{stats} package.
\item For "glmnet", the model is an elastic net model from the \code{glmnet} package. The \code{best_lam} parameter should be specified for lambda.
\item For "rf", the model is a random forest model from the \code{ranger} package. The new data is cleaned using \code{janitor::clean_names()}.
\item For "xgb", the model is an XGBoost model from the \code{xgboost} package.
\item For "nn", the model is a neural network model from the \code{keras} package.
}
The function excludes the first three columns of \code{new_features_m_d_ref} for prediction purposes.
}
