% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{prepare_tickers_catalog_slots}
\alias{prepare_tickers_catalog_slots}
\title{Prepare tickers_catalog slots}
\usage{
prepare_tickers_catalog_slots(tickers_catalog)
}
\description{
Inner helper to construct slots based on tickers_catalog. This allows for reusability between
create_tickers_catalog and update_tickers_catalog
}
