% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_performance_m_df.R
\name{prob_sharpe_ratio}
\alias{prob_sharpe_ratio}
\title{Compute the Probabilistic Sharpe Ratio}
\usage{
prob_sharpe_ratio(x)
}
\arguments{
\item{x}{A numeric vector of cleaned returns}
}
\value{
A single numeric value: the probabilistic Sharpe ratio, or NA if not computable.
}
\description{
Calculates the probability that the observed Sharpe Ratio exceeds a reference value,
adjusting for non-normality via skewness and kurtosis.
}
\details{
This implementation is based on the formula proposed in Lopez de Prado (2018),
and always incorporates skewness and kurtosis corrections. It does not use any external dependencies.
}
