% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{refit_ml_model-class}
\alias{refit_ml_model-class}
\title{Define the \code{refit_ml_model} S4 Class}
\description{
This class represents a refitted machine learning model. It encapsulates the algorithm used, hyperparameters,
feature data, target variable, and the fitted model object.
}
\section{Slots}{

\describe{
\item{\code{ml_algorithm}}{A character string specifying the machine learning algorithm used (e.g., "ols", "glmnet", "rf", "xgb", "nn").}

\item{\code{best_hyperparameters}}{The chosen hyperparameters relevant to the specified machine learning algorithm.}

\item{\code{model}}{The fitted model object, which varies based on the algorithm used.}
}}

\section{Methods}{

\describe{
\item{\code{refit()}}{Refits the model based on the specified algorithm and hyperparameters.}
\item{\code{predict(new_features)}}{Generates predictions using the fitted model on new feature data.}
}
}

