% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{refit_ml_model_accessors}
\alias{refit_ml_model_accessors}
\title{Accessor Methods for refit_ml_model}
\usage{
get_ml_algorithm(object)
}
\arguments{
\item{object}{An object of class \code{refit_ml_model}.}
}
\value{
The respective slot of the \code{refit_ml_model} object.
}
\description{
These methods are used to access components of a \code{refit_ml_model} object.
}
