% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_risk_contribution.R
\name{relative_risk_contribution}
\alias{relative_risk_contribution}
\title{Calculate Relative Risk Contribution}
\usage{
relative_risk_contribution(weights, covariance_matrix)
}
\arguments{
\item{weights}{A numeric vector of weights for each asset in the portfolio}

\item{covariance_matrix}{A numeric matrix representing the covariance matrix of the assets}
}
\value{
A numeric vector of the same length as \code{weights} with the relative risk contribution of each asset
}
\description{
Computes the relative risk contribution of each asset in a portfolio, given a vector of
weights and a covariance matrix
}
