% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{remove_sb_backtest_config}
\alias{remove_sb_backtest_config}
\title{Remove an sb_backtest_config by name from an sb_metabacktest_config}
\usage{
remove_sb_backtest_config(object, config_name)
}
\arguments{
\item{object}{An \code{sb_metabacktest_config} object.}

\item{config_name}{A character string specifying the name of the \code{sb_backtest_config} to remove.}
}
\value{
An updated \code{sb_metabacktest_config} object with the specified configuration removed.
}
\description{
Remove an sb_backtest_config by name from an sb_metabacktest_config
}
