% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{required_pkgs.step_impute_sector}
\alias{required_pkgs.step_impute_sector}
\title{Required Packages for step_impute_sector}
\usage{
required_pkgs.step_impute_sector(x, ...)
}
\arguments{
\item{x}{A \code{step_impute_sector} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A character vector of package names required for the step.
}
\description{
Declares the required packages for using the \code{step_impute_sector} step inside a recipe.
}
