% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_meta_dataframe.R
\name{res_mom}
\alias{res_mom}
\title{Calculate Residual Momentum Score}
\usage{
res_mom(ret_values, bench_ret_values, na.rm = TRUE)
}
\arguments{
\item{ret_values}{A numeric vector of returns for the stock.}

\item{bench_ret_values}{A numeric vector of benchmark returns corresponding to the same periods as \code{ret_values}.}

\item{na.rm}{A logical value indicating whether to remove NA values before computation (default is TRUE).}
}
\value{
A numeric value representing the residual momentum score. If the standard deviation of residuals is zero,
\code{NA_real_} is returned.
}
\description{
This function computes the residual momentum score by fitting a linear regression model with
\code{ret_values} as the dependent variable and \code{bench_ret_values} as the independent variable.
The score is defined as the sum of the regression residuals divided by their standard deviation.
}
\examples{
res_mom(c(0.05, 0.06, 0.07), c(0.03, 0.04, 0.05))
}
