% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{rp_parameters-class}
\alias{rp_parameters-class}
\title{Define the \code{rp_parameters} S4 Class}
\value{
An S4 object of class \code{port_backtest_config}.
}
\description{
S4 class to represent a set of configurations for risk-parity portfolios.
}
\section{Slots}{

\describe{
\item{\code{rp_method}}{A character indicating the method to compute the risk-parity vanilla solution. It is passed to riskParityPortfolio::riskParityPortfolio function as method_init.
Default is "cyclical-spinu"}
}}

