% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pp_backtest.R
\name{run_pp_backtest}
\alias{run_pp_backtest}
\title{Run Time-wise Preprocessing Backtest}
\usage{
run_pp_backtest(raw_features_m_df, config_obj, ...)
}
\arguments{
\item{raw_features_m_df}{A \code{raw_features_m_df} object.}

\item{pp_backtest_config}{A \code{pp_backtest_config} object that contains a recipe in its \code{recipe} slot.}

\item{verbose}{A logical indicating whether to print messages during the process. Default is \code{FALSE}.}
}
\value{
A time-wise preprocessed \code{meta_dataframe}.
}
\description{
This method performs a time-wise backtest preprocessing on a
\code{raw_features_m_df} object using the recipe stored in a
\code{pp_backtest_config} object. For each date, the recipe is prepped on the
subset of data available at the specific date. This ensures that no future
information is used during preprocessing.
}
\details{
Parallel processing is achieved using \code{furrr::future_map}; ensure that an appropriate
future plan is set (e.g., \code{future::plan(future::multisession)}).
}
\examples{
\dontrun{
  # Assume raw_obj is an instance of raw_features_m_df and config_obj is a
  # pp_backtest_config object (created via create_pp_backtest_config) that contains a recipe.
  preprocessed_df <- run_pp_backtest(raw_features_m_df = raw_obj, config_obj = config_obj)
}

}
