% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss_backtest.R
\name{run_ss_backtest_internal}
\alias{run_ss_backtest_internal}
\title{Run a Signal Selection Backtest}
\usage{
run_ss_backtest_internal(
  rebalancing_months,
  data_availability_cutoff = 60,
  split_method = "expanding",
  signals_m_df,
  chosen_signals,
  signal_positions,
  backtest_returns_df,
  benchmark_returns_df,
  market_factor_proxy = "IBOV",
  p_correction_method = "none",
  signal_significance_threshold = 0.05,
  heuristic_sb_metric = "IR",
  max_abs_active_group_weight,
  priors_m_df = NULL,
  priors_type = "all",
  signal_themes_m_df = NULL,
  user_priors_list = NULL,
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975,
  verbose = TRUE
)
}
\arguments{
\item{rebalancing_months}{Numeric months when signal selection should be implemented.}

\item{data_availability_cutoff}{The minimum number of non-NA observations required for a backtest to be considered.}

\item{split_method}{The method used for splitting the data, either "expanding" or "rolling" (default is "expanding").}

\item{signals_m_df}{A (meta) data frame with columns including "id", "tickers", "dates", and the selected signals.}

\item{chosen_signals}{A vector of user-defined characteristics to be considered.}

\item{signal_positions}{A named vector with the same length and names as \code{chosen_signals}, describing whether positions should be taken "long" or "short".}

\item{backtest_returns_df}{A data frame with a 'dates' column and remaining columns named according to signals in \code{signals_m_df}, containing historical backtested returns.}

\item{p_correction_method}{The method for p-value correction. Possible options are:
\itemize{
\item \strong{"none"}: No correction.
\item \strong{"bayesian"}: A hierarchical mixed-effects Bayesian linear model is fitted to the data using the \code{brms} package.

The user can also choose one of the following frequentist methods, which control the Family-Wise Error Rate (FWER) or the False Discovery Rate (FDR). FDR is less stringent than FWER.

For FWER, possible options are:
\itemize{
\item \strong{"bonferroni"}: Bonferroni correction.
\item \strong{"holm"}: Holm's (1979) method.
\item \strong{"hochberg"}: Hochberg's (1988) method.
\item \strong{"hommel"}: Hommel's (1988) method.
}

For FDR, possible options are:
\itemize{
\item \strong{"BH"} or \strong{"fdr"}: Benjamini-Hochberg (1995) procedure.
\item \strong{"BY"}: Benjamini-Yekutieli (2001) procedure.
}
}}

\item{signal_significance_threshold}{A decimal indicating the hypothesis testing zero-alpha null-hypothesis rejection criteria. If you want to select all \code{chosen_signals}, provide 1.}

\item{heuristic_sb_metric}{A signal in \code{signals_m_df} used for \code{sb_algorithms} "SW" and "MTO". It selects which signal will be elected in case of theme representativeness when no signal in a given theme was deemed significant.}

\item{max_abs_active_group_weight}{A decimal indicating the maximum absolute weight that a theme of signals can have in the final portfolio.}

\item{priors_m_df}{A (meta) data frame with columns including "id", "ticker", "dates", "theme" (used for clustering in the Bayesian hierarchical model),
and values for active_return, bench_return, alpha (mean and se), beta (mean and se), and sigma. Data should be exogenous, as it will be used to set priors for the hierarchical Bayesian model.}

\item{priors_type}{A flag indicating which priors should be set. Possible options are:
\itemize{
\item \strong{"all"}: Set priors for all parameters, including mean (\code{mu}), variance (\code{tau}), and correlation based on \code{priors_m_df} data.
\item \strong{"mean"}: Set priors only for mean (\code{mu}).
\item \strong{"uninformative"}: Set uninformative priors for all parameters.
\item \strong{"user"}: Set priors defined by the user.
}}

\item{signal_themes_m_df}{A (meta) data frame with "id", "tickers" ("signals"), and "dates" columns, including all signals in \code{signals_m_df}, and a "theme" column providing group membership for each signal.}

\item{user_priors_list}{A list with user-defined priors for the hierarchical Bayesian model, used when \code{priors_type = "user"}.
Each element of the list represents a theme and should contain priors set with brms::set_prior (class \code{brmsprior}). For instance, setting priors only for
alpha and beta global parameters can be done with:
c(brms::set_prior("normal(0,1)", class = "Intercept"), brms::set_prior("normal(0,1)", class = "b", coef = "bench_return"))}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization.}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization.}

\item{verbose}{A boolean indicating whether to print messages.}

\item{selected_benchmark_returns_df}{A data frame with a 'dates' column and a column with benchmark returns, named accordingly.}
}
\value{
A list containing eligible signals and signal universes throughout time.
}
\description{
This function performs iterative signal selection based on frequentist or Bayesian methods, which are applied to portfolio returns backtests to identify which signals can be considered significant in stock-level return prediction.
}
\details{
To determine whether a signal matters in cross-sectional predictability, the literature typically runs regressions of signal portfolios against a benchmark factor model (e.g., CAPM), computing alphas and corresponding t-stats. Due to the large number of signals identified in the literature (the "factor zoo"), methods to control for multiple testing are often advocated.

The function performs the following operations:
\itemize{
\item Extracts the user-defined chosen signals from \code{signals_m_df} and subsets the data frame to include only these signals.
\item Checks for consistency between the length of \code{chosen_signals} and \code{signal_positions} and ensures that all chosen signals have corresponding positions.
\item Adjusts the signal positions based on whether they are "short" by multiplying their values by -1.
\item Updates column names in the data frame to reflect the corrected positions of the signals.
\item Validates that all adjusted signals have corresponding columns in \code{backtest_returns_df}.
\item Calculates time-series metrics from backtested returns, including mean active return, IR, alpha, t-stat, beta, Treynor ratio, and p-value. Signals without adequate data (less than \code{data_availability_cutoff} periods) are removed.
\item Adjusts p-values based on the method selected by the user.
\item Defines which signals are eligible for blending.
}
}
