% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{sb_metabacktest_results-class}
\alias{sb_metabacktest_results-class}
\title{sb_metabacktest_results Class}
\description{
An S4 class designed to store and manage a collection of \code{sb_backtest_results} objects,
along with consolidated and time series evaluation metrics for machine learning models.
}
\section{Slots}{

\describe{
\item{\code{sb_backtest_results}}{A list of \code{sb_backtest_results} objects.}

\item{\code{consolidated_oos_testing_metrics}}{A data frame containing consolidated out-of-sample testing evaluation metrics for each algorithm.}

\item{\code{mean_validation_metrics}}{A data frame containing the mean validation metrics for each algorithm.}

\item{\code{time_series_oos_testing_metrics}}{A list of data frames for each evaluation metric over time (out-of-sample testing).}

\item{\code{time_series_validation_metrics}}{A list of data frames for each evaluation metric over time (validation).}
}}

