% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{sb_model-class}
\alias{sb_model-class}
\title{Define the \code{sb_model} S4 Class}
\description{
This class represents a (re)fitted sb model. It encapsulates the algorithm used, hyperparameters, custom objective,
feature data, target variable, and the fitted model object.
}
\section{Slots}{

\describe{
\item{\code{sb_algorithm}}{A character string specifying the machine learning algorithm used (e.g., "ols", "glmnet", "rf", "xgb", "nn", "ew", "rp" or "sw").}

\item{\code{best_hyperparameters}}{The chosen hyperparameters relevant to the specified machine learning algorithm. Applicable only for machine-learning algorithms.}

\item{\code{model}}{The fitted model object, which varies based on the algorithm used.}
}}

\section{Methods}{

\describe{
\item{\code{refit()}}{Refits the model based on the specified algorithm and hyperparameters.}
\item{\code{predict(new_features)}}{Generates predictions using the fitted model on new feature data.}
}
}

