% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_by_conditions.R
\name{screen_by_conditions,meta_dataframe-method}
\alias{screen_by_conditions,meta_dataframe-method}
\title{Method for screening a meta_dataframe based on conditions}
\usage{
\S4method{screen_by_conditions}{meta_dataframe}(meta_dataframe, ...)
}
\arguments{
\item{meta_dataframe}{A \code{meta_dataframe} object.}

\item{...}{A list of conditions to filter the data. Those will be passed to \code{dplyr::filter()}.}
}
\value{
A new \code{meta_dataframe} object with screened data.
}
\description{
Method for screening a meta_dataframe based on conditions
}
