% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_by_conditions.R
\name{screen_by_conditions}
\alias{screen_by_conditions}
\title{Generic function for screening a meta_dataframe based on conditions}
\usage{
screen_by_conditions(meta_dataframe, ...)
}
\arguments{
\item{meta_dataframe}{A \code{meta_dataframe} object.}

\item{...}{A list of conditions to filter the data. Those will be passed to \code{dplyr::filter()}.}
}
\description{
This is a generic function for screening \code{meta_dataframe} objects based on conditions.
}
