% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_by_liquidity.R
\name{screen_by_liquidity,meta_dataframe,meta_dataframe,data.frame,character,logical-method}
\alias{screen_by_liquidity,meta_dataframe,meta_dataframe,data.frame,character,logical-method}
\title{Method for screening a meta_dataframe based on liquidity}
\usage{
\S4method{screen_by_liquidity}{meta_dataframe,meta_dataframe,data.frame,character,logical}(
  meta_dataframe,
  liquidity_m_df,
  liquidity_floor_cutoffs,
  liquidity_floor_rule,
  verbose = TRUE
)
}
\arguments{
\item{meta_dataframe}{A \code{meta_dataframe} object.}

\item{liquidity_m_df}{A \code{liquidity_m_df} meta_dataframe containing one or more market liquidity measures (e.g., inflation-adjusted mean financial volume).
All ids in meta_dataframe must have a unique correspondence to this object.}

\item{liquidity_floor_cutoffs}{A data.frame containing cutoff values for liquidity metrics specified in \code{liquidity_m_df}.
The names should match the metrics and values should be the minimum acceptable values (adjust for inflation)
Stocks that have all metrics higher than defined in a \code{liquidity_floor_cutoffs} element will receive a liquidity classification at least equal to it.
Elements should be: "micro_caps", "small_caps", "mid_caps", "large_caps" and "mega_caps"
Classification should be in ascending order (from lest liquid to most liquid) for all metrics.
If set in decimals, values will be interpreted as quantiles and classification will be set according to quantiles
The first column is named "liquidity_classification"
It has at most 5 rows and no duplicates or NAs.}

\item{liquidity_floor_rule}{Optional. Character string specifying the liquidity classification to apply the liquidity floor rule (eg. "nano_caps", "micro_caps", "small_caps", "mid_caps", "large_caps", "mega_caps").}

\item{verbose}{Logical. If TRUE, prints processing messages.}
}
\value{
A new \code{meta_dataframe} object with screened data.
}
\description{
Method for screening a meta_dataframe based on liquidity
}
