% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_rolling.R
\name{sd_rolling}
\alias{sd_rolling}
\title{Row-wise Standard Deviation}
\usage{
sd_rolling(main_matrix, complementary_matrix)
}
\arguments{
\item{main_matrix}{A matrix, data.frame, or tibble of characteristics.}

\item{complementary_matrix}{A matrix, data.frame, or tibble containing lagged information necessary
to calculate SD for the first columns in main_matrix. The number of columns
of this matrix specifies how many previous observations should be considered
in the SD calculation, determining the width of the rolling windows used.}
}
\value{
A numeric matrix of row-wise SD. Non-unique values are ignored to limit the impact of repeated
information, quite common when dealing with accounting information in monthly observations. By
default, NAs are ignored.
}
\description{
Calculates the row-wise Standard Deviation (SD) of a matrix, considering lagged realizations.
}
\examples{
# Create a complete matrix with lagged information
main_matrix <- matrix(c(5, 3, 7, 8), nrow = 2, ncol = 2)
complementary_matrix <- matrix(c(1, 2, 6, 4), nrow = 2, ncol = 2)

# Calculate SD using a rolling window of 3 (1 + 2 additional columns)
sd_rolling(main_matrix, complementary_matrix)
}
