% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_custom_moments.R
\name{set_custom_moments}
\alias{set_custom_moments}
\title{Sets portfolio custom moments}
\usage{
set_custom_moments(
  R,
  portfolio,
  user_moments = NULL,
  user_method = c(returns, input, two_moment)
)
}
\arguments{
\item{R}{asset returns as in PortfoloAnalytics}

\item{portfolio}{a portfolio object as in PortfolioAnalytics}

\item{user_moments}{a list of four matices containing user-specified values for the first four return moments}

\item{user_method}{user-specified method for computing moment matrices. Defaults to calculation used by PortfolioAnalytics "sample" method,
which uses PerformanceAnalytics functions to computer the higher-order moments}
}
\description{
Sets portfolio custom moments
}
