% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_final_signal.R
\name{set_final_signal}
\alias{set_final_signal}
\title{Compute and Set Final Signal}
\usage{
set_final_signal(
  selected_signals_corrected_positions_m_d_ref,
  eligible_signals,
  signal_weights,
  ml_walk_forward_validation_results,
  upper_quantile_winsorization = 0.975,
  lower_quantile_winsorization = 0.025
)
}
\arguments{
\item{signal_weights}{A vector of weights for each signal.}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization when creating signals}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization when creating signals}

\item{selected_signals_corrected_positions_m_upd_ref}{A data frame containing selected signals with corrected positions,
including date information. This data frame is updated with the final signal values after computation.}

\item{signals_positions}{A data frame that specifies the positions (long or short) of the signals.}

\item{signal_blending_method}{A character string specifying the method used to blend signals. Possible values are
"ML" for machine learning or any other value for non-ML methods (EW, SW, RP or MTO).}

\item{target_m_upd_ref}{A data frame containing the target variables used for machine learning model fitting,
including date information.}
}
\description{
This function calculates and assigns a final signal value to a set of financial signals based on a specified
blending method. The blending method can either be a non-machine-learning approach or a machine learning
approach. The function handles signal weighting, quantile winsorization, and, if applicable, fits and uses
a machine learning model to generate predictions.
}
