% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_priors.R
\name{set_priors}
\alias{set_priors}
\title{Set Priors for Signals Based on Informative Data}
\usage{
set_priors(selected_priors_informative_data_m_upd_ref, set_priors_on = "all")
}
\arguments{
\item{selected_priors_informative_data_m_upd_ref}{A data frame containing historical data with columns: \code{dates}, \code{theme}, \code{alpha}, \code{beta}, and \code{sigma}. This data is used to estimate priors for each theme.}

\item{set_priors_on}{A character indicating to which parameters should priors be defined. If \code{mean}, priors will be set on alpha and beta.
If \code{all}, priors will also be set on variance (\code{tau_u1}, \code{tau_u2}, \code{tau_u_sigma}) and correlation parameters.
If \code{none}, brms default uninformative priors will be used}
}
\value{
A list, where each component contains \code{brms::set_prior} objects for each theme.
}
\description{
This function sets prior distributions for model parameters based on historical data, choosing appropriate priors for the intercept, beta coefficient, and optionally for variance components and correlations.
Hierarchical bayesian models can be built based on such priors. Signals are groups and are defined theme-wise.
The priors are determined using a Bayesian Information Criterion (BIC) approach to select the best-fitting distribution to reference historical data
}
\details{
The function performs the following steps:
\itemize{
\item Extracts the most recent dates and the unique themes from the data.
\item Computes summary statistics for \code{alpha}, \code{beta}, and \code{sigma} for each theme.
\item Chooses appropriate priors for \code{alpha}, \code{beta}, and optionally for \code{sigma} and variance components based on the chosen distributions.
\item Returns a list of \code{brms::set_prior} objects for each theme.
}
}
