% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe_matrix.R
\name{sharpe_matrix}
\alias{sharpe_matrix}
\title{Sharpe Ratio Matrix}
\usage{
sharpe_matrix(initial_list)
}
\arguments{
\item{initial_list}{A list of matrices, data frames, or tibbles with the same dimensions.}
}
\value{
A matrix containing the Sharpe ratio calculated for each element of the input matrices.
}
\description{
This function calculates the Sharpe ratio matrix from a list of matrices or data frames.
}
\details{
The Sharpe ratio is calculated as the mean divided by the standard deviation for each element of the input matrices.
}
\examples{
# Create a list of matrices
mat1 <- matrix(rnorm(9), nrow = 3)
mat2 <- matrix(rnorm(9), nrow = 3)
initial_list <- list(mat1, mat2)
# Calculate the Sharpe ratio matrix
sharpe_matrix(initial_list)

}
\references{
William F. Sharpe. "The Sharpe Ratio". The Journal of Portfolio Management 21 (1994): 49–58.
}
