% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,alpha_test_strategy-method}
\alias{show,alpha_test_strategy-method}
\title{Show Alpha Test Strategy}
\usage{
\S4method{show}{alpha_test_strategy}(object)
}
\arguments{
\item{object}{An \code{alpha_test_strategy} object to be displayed.}
}
\description{
Prints the contents of an \code{alpha_test_strategy} object, detailing the various parameters and their configurations.
}
