% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,bayesian_alpha_test_strategy-method}
\alias{show,bayesian_alpha_test_strategy-method}
\title{Show Bayesian Alpha Test Strategy}
\usage{
\S4method{show}{bayesian_alpha_test_strategy}(object)
}
\arguments{
\item{object}{A \code{bayesian_alpha_test_strategy} object to be displayed.}
}
\description{
Prints the contents of a \code{bayesian_alpha_test_strategy} object, detailing the various parameters and their configurations.
}
