% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,bayesian_model_parameters-method}
\alias{show,bayesian_model_parameters-method}
\title{Show Bayesian Model Parameters}
\usage{
\S4method{show}{bayesian_model_parameters}(object)
}
\arguments{
\item{object}{A \code{bayesian_model_parameters} object to be displayed.}
}
\description{
Prints the contents of a \code{bayesian_model_parameters} object, detailing the various parameters and configurations.
}
