% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,bayesian_opt_strategy-method}
\alias{show,bayesian_opt_strategy-method}
\title{Show Method for \code{bayesian_opt_strategy}}
\usage{
\S4method{show}{bayesian_opt_strategy}(object)
}
\arguments{
\item{object}{An object of class \code{bayesian_opt_strategy}.}
}
\value{
Printed information about the object.
}
\description{
Custom show method for displaying information about objects of class \code{bayesian_opt_strategy}.
This method will display the tuning method, machine learning algorithm, validation sample size,
and details specific to Bayesian optimization such as \code{n_iter}, acquisition function (\code{acq}),
initial points, and hyperparameter bounds.
}
