% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,concentration_constraint_policy-method}
\alias{show,concentration_constraint_policy-method}
\title{Show Concentration Constraint Policy}
\usage{
\S4method{show}{concentration_constraint_policy}(object)
}
\arguments{
\item{object}{A \code{concentration_constraint_policy} object.}
}
\description{
Prints the contents of a \code{concentration_constraint_policy} object,
detailing the benchmark, the maximum absolute active weight for individual assets,
and the maximum absolute active group weights.
}
