% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,frequentist_alpha_test_strategy-method}
\alias{show,frequentist_alpha_test_strategy-method}
\title{Show Frequentist Alpha Test Strategy}
\usage{
\S4method{show}{frequentist_alpha_test_strategy}(object)
}
\arguments{
\item{object}{A \code{frequentist_alpha_test_strategy} object to be displayed.}
}
\description{
Prints the contents of a \code{frequentist_alpha_test_strategy} object: a header, followed by the
inherited \code{alpha_test_strategy} display (main information, model structure, and \code{lmer_control}).
}
