% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,grid_search_strategy-method}
\alias{show,grid_search_strategy-method}
\title{Show Method for \code{grid_search_strategy}}
\usage{
\S4method{show}{grid_search_strategy}(object)
}
\arguments{
\item{object}{An object of class \code{grid_search_strategy}.}
}
\value{
Printed information about the object.
}
\description{
Custom show method for displaying information about objects of class \code{grid_search_strategy}.
This method will display the tuning method, machine learning algorithm, validation sample size,
and details about the hyperparameter grid.
}
