% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,groups_m_df-method}
\alias{show,groups_m_df-method}
\title{Show Method for meta_dataframe Class}
\usage{
\S4method{show}{groups_m_df}(object)
}
\arguments{
\item{object}{An instance of the \code{meta_dataframe} class.}
}
\value{
The method returns the object invisibly.
}
\description{
This method displays a summary of the \code{meta_dataframe} object, including
sb_backtest_workflow information, number of signals, unique dates, unique tickers,
total observations, and the first few rows of the data.
}
