% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,hrp_parameters-method}
\alias{show,hrp_parameters-method}
\title{Show HRP Parameters}
\usage{
\S4method{show}{hrp_parameters}(object)
}
\arguments{
\item{object}{A \code{hrp_parameters} object.}

\item{hide_title}{Logical indicating whether to hide the title when displaying the parameters. Default is FALSE.}
}
\description{
Displays the hierarchical risk parity configuration contained in a \code{hrp_parameters} object.
}
