% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,hyper_grid_domain-method}
\alias{show,hyper_grid_domain-method}
\title{Print method for hyper_grid_domain}
\usage{
\S4method{show}{hyper_grid_domain}(object)
}
\arguments{
\item{object}{An object of class hyper_grid_domain.}
}
\description{
This method prints the contents of a \code{hyper_grid_domain} object in a user-friendly format.
}
