% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,keras_architecture_parameters-method}
\alias{show,keras_architecture_parameters-method}
\title{Print keras_architecture_parameters}
\usage{
\S4method{show}{keras_architecture_parameters}(object)
}
\arguments{
\item{object}{An object of class \code{keras_architecture_parameters}.}
}
\description{
Method to print an object of class \code{keras_architecture_parameters}.
}
