% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,liquidity_constraint_policy-method}
\alias{show,liquidity_constraint_policy-method}
\title{Show Liquidity Constraint Policy}
\usage{
\S4method{show}{liquidity_constraint_policy}(object)
}
\arguments{
\item{object}{A \code{liquidity_constraint_policy} object.}
}
\description{
Prints the contents of a \code{liquidity_constraint_policy} object,
including the liquidity floor rule and liquidity cap rules.
}
