% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,meta_dataframe-method}
\alias{show,meta_dataframe-method}
\title{Show Method for meta_dataframe Class}
\usage{
\S4method{show}{meta_dataframe}(object)
}
\arguments{
\item{object}{An instance of the \code{meta_dataframe} class.}
}
\value{
The method returns the object invisibly.
}
\description{
This method displays a summary of the \code{meta_dataframe} object, including its name and current date, the
signal (or target) columns and their count, the unique dates and tickers, the total number of observations,
and the first few rows of the data.
}
