% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ml_backtest_config-method}
\alias{show,ml_backtest_config-method}
\title{Show ML Backtest Config}
\usage{
\S4method{show}{ml_backtest_config}(object)
}
\arguments{
\item{object}{An \code{ml_backtest_config} object to be displayed.}
}
\description{
Prints the contents of an \code{ml_backtest_config} object, detailing the various parameters and their configurations.
}
