% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ml_experiment-method}
\alias{show,ml_experiment-method}
\title{Show ML Experiment}
\usage{
\S4method{show}{ml_experiment}(object)
}
\arguments{
\item{object}{An \code{ml_experiment} object to be displayed.}
}
\description{
Prints the contents of an \code{ml_experiment} object, detailing the various parameters and their configurations.
}
