% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ml_metabacktest_config-method}
\alias{show,ml_metabacktest_config-method}
\title{Show Method for ml_metabacktest_config Class}
\usage{
\S4method{show}{ml_metabacktest_config}(object)
}
\arguments{
\item{object}{An \code{ml_metabacktest_config} object.}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effect of displaying information.
}
\description{
Displays detailed information about each configuration in the \code{ml_metabacktest_config} object.
}
\examples{
# Assuming you have an ml_metabacktest_config object named meta_config
show(meta_config)

}
