% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ml_metabacktest_results-method}
\alias{show,ml_metabacktest_results-method}
\title{Show Method for ml_metabacktest_results Class}
\usage{
\S4method{show}{ml_metabacktest_results}(object)
}
\arguments{
\item{object}{An object of class \code{ml_metabacktest_results}.}
}
\description{
Displays the contents of an \code{ml_metabacktest_results} object,
including consolidated and time series evaluation metrics.
}
