% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ml_wf_val_results-method}
\alias{show,ml_wf_val_results-method}
\title{Show Method for ml_wf_val_results Class}
\usage{
\S4method{show}{ml_wf_val_results}(object)
}
\arguments{
\item{object}{An instance of the \code{ml_wf_val_results} class.}
}
\value{
The method returns the object invisibly.
}
\description{
This method displays a detailed summary of the \code{ml_wf_val_results} object,
including metadata on the machine learning workflow validation results,
algorithm details, sample sizes, stock information, features, tuning,
Keras architecture parameters, performance, and the original call.
}
