% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,mmaf_parameters-method}
\alias{show,mmaf_parameters-method}
\title{Show MMAF Parameters}
\usage{
\S4method{show}{mmaf_parameters}(object)
}
\arguments{
\item{object}{A \code{mmaf_parameters} object.}
}
\description{
Displays the Micro-Macro Allocation Framework configuration contained in a \code{mmaf_parameters} object.
}
