% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,port_backtest_cohort-method}
\alias{show,port_backtest_cohort-method}
\title{Show Method for port_backtest_cohort Class}
\usage{
\S4method{show}{port_backtest_cohort}(object)
}
\arguments{
\item{object}{An object of class \code{port_backtest_cohort}.}
}
\description{
Displays a detailed summary of a \code{port_backtest_cohort} object.
It focuses on the configuration settings contained in the common
port_backtest_workflow slot (e.g., selected_benchmark, port_construction_method,
object names, dates, and more).
}
