% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,port_backtest_config-method}
\alias{show,port_backtest_config-method}
\title{Show Port Backtest Config}
\usage{
\S4method{show}{port_backtest_config}(object)
}
\arguments{
\item{object}{A \code{port_backtest_config} object to be displayed.}
}
\description{
Prints the contents of a \code{port_backtest_config} object, detailing its various
parameters and configurations for stock-level portfolio backtesting.
}
