% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,portfolio_policies-method}
\alias{show,portfolio_policies-method}
\title{Show Portfolio Policies}
\usage{
\S4method{show}{portfolio_policies}(object)
}
\arguments{
\item{object}{A \code{portfolio_policies} object to be displayed.}
}
\description{
Prints the contents of a \code{portfolio_policies} object, detailing
the various policies and their configurations.
}
