% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,random_search_strategy-method}
\alias{show,random_search_strategy-method}
\title{Show Method for \code{random_search_strategy}}
\usage{
\S4method{show}{random_search_strategy}(object)
}
\arguments{
\item{object}{An object of class \code{random_search_strategy}.}
}
\value{
Printed information about the object.
}
\description{
Custom show method for displaying information about objects of class \code{random_search_strategy}.
This method will display the tuning method, machine learning algorithm, validation sample size,
and the number of iterations (\code{n_iter}) along with hyperparameter distributions.
}
\examples{
# Create a random_search_strategy object
random_search_obj <- create_hyperparameter_tuning_strategy(
  tuning_method = "random_search",
  ml_algorithm = "rf",
  validation_sample_size = 1000,
  n_iter = 20
)
show(random_search_obj)
}
