% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,refit_ml_model-method}
\alias{show,refit_ml_model-method}
\title{Show Method for refit_ml_model Class}
\usage{
\S4method{show}{refit_ml_model}(object)
}
\arguments{
\item{object}{An instance of the \code{refit_ml_model} class.}
}
\value{
The method returns the object invisibly.
}
\description{
This method provides a summary of the \code{refit_ml_model} object, including
the machine learning algorithm used, best hyperparameters, custom objective,
Huber delta, Keras architecture parameters, and the model structure.
}
