% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,sb_metabacktest_config-method}
\alias{show,sb_metabacktest_config-method}
\title{Show Method for sb_metabacktest_config Class}
\usage{
\S4method{show}{sb_metabacktest_config}(object)
}
\arguments{
\item{object}{An \code{sb_metabacktest_config} object.}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effect of displaying information.
}
\description{
Displays detailed information about each configuration in the \code{sb_metabacktest_config} object.
}
