% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,sb_metabacktest_results-method}
\alias{show,sb_metabacktest_results-method}
\title{Show Method for sb_metabacktest_results Class}
\usage{
\S4method{show}{sb_metabacktest_results}(object)
}
\arguments{
\item{object}{An object of class \code{sb_metabacktest_results}.}
}
\description{
Displays the contents of an \code{sb_metabacktest_results} object,
including consolidated and time series evaluation metrics.
}
