% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,sb_model-method}
\alias{show,sb_model-method}
\title{Show Method for sb_model Class}
\usage{
\S4method{show}{sb_model}(object)
}
\arguments{
\item{object}{An instance of the \code{sb_model} class.}
}
\value{
The method returns the object invisibly.
}
\description{
This method provides a summary of the \code{sb_model} object, including
the machine learning algorithm used, best hyperparameters, custom objective,
Huber delta, Keras architecture parameters, and the model structure.
}
