% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,signal_port_parameters-method}
\alias{show,signal_port_parameters-method}
\title{Show Signal Portfolio Parameters}
\usage{
\S4method{show}{signal_port_parameters}(object)
}
\arguments{
\item{object}{A \code{signal_port_parameters} object to be displayed.}
}
\description{
Prints the contents of a \code{signal_port_parameters} object, detailing
the various parameters (covariance estimation, MVO, RP, and concentration constraints).
}
