% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,signal_universe_m_df-method}
\alias{show,signal_universe_m_df-method}
\title{Show Method for signal_universe_m_df Class}
\usage{
\S4method{show}{signal_universe_m_df}(object)
}
\arguments{
\item{object}{An instance of the \code{signal_universe_m_df} class.}
}
\value{
Returns the object invisibly.
}
\description{
This method extends the parent \code{meta_dataframe} show method by displaying
additional elements from \code{ss_backtest_workflow} and \code{sb_backtest_workflow}.
It focuses on the key fields you specified:
\itemize{
\item \strong{ss_backtest_workflow}: active_returns, model_structure, market_factor_proxy,
backtest_type, p_correction_method, theme_level_intercept (can be NULL),
theme_level_slope (can be NULL), signals_object_name, signal_themes_object_name,
priors_object_name, backtest_returns_object_name, rebalancing_months
\item \strong{sb_backtest_workflow}: sb_algorithm, custom_objective, backtest_type,
keras_architecture_parameters, tuning_method, chosen_eval_metric, huber_delta, quantile_tau
\item Note: \code{sb_backtest_workflow} can be \code{NULL}.
}
}
