% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ss_backtest_config-method}
\alias{show,ss_backtest_config-method}
\title{Show Signal Selection Backtest Config}
\usage{
\S4method{show}{ss_backtest_config}(object)
}
\arguments{
\item{object}{An \code{ss_backtest_config} object to be displayed.}
}
\description{
Prints the contents of an \code{ss_backtest_config} object: config name; backtest parameters
(initial sample size, rebalancing months, active returns, split method, chosen signals and positions); and,
if set, the \code{alpha_test_strategy} (delegated to its own \code{show} method).
}
