% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ss_backtest_results-method}
\alias{show,ss_backtest_results-method}
\title{Show Method for ss_backtest_results Class}
\usage{
\S4method{show}{ss_backtest_results}(object)
}
\arguments{
\item{object}{An instance of the \code{ss_backtest_results} class.}
}
\value{
The method returns the object invisibly.
}
\description{
This method displays a detailed summary of the \code{ss_backtest_results} object,
including metadata on the signal selection backtest results,
configuration details, date information, signals information,
p-value correction methods, Bayesian model parameters (if applicable),
performance metrics, and the original call.
}
