% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,tickers_catalog-method}
\alias{show,tickers_catalog-method}
\title{Print method for tickers_catalog}
\usage{
\S4method{show}{tickers_catalog}(object)
}
\description{
Displays key information about a tickers_catalog object.
}
