% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,transaction_costs_parameters-method}
\alias{show,transaction_costs_parameters-method}
\title{Show Transaction Cost Parameters}
\usage{
\S4method{show}{transaction_costs_parameters}(object)
}
\arguments{
\item{object}{A \code{transaction_costs_parameters} object.}
}
\description{
Prints the contents of a \code{transaction_costs_parameters} object,
including direct transaction cost, strategy AUM, alpha, and lambda.
}
