% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,turnover_constraint_policy-method}
\alias{show,turnover_constraint_policy-method}
\title{Show Turnover Constraint Policy}
\usage{
\S4method{show}{turnover_constraint_policy}(object)
}
\arguments{
\item{object}{A \code{turnover_constraint_policy} object.}
}
\description{
Prints the contents of a \code{turnover_constraint_policy} object,
including the quantile range buffer and the turnover cap rules.
}
