% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{signal_port_parameters-class}
\alias{signal_port_parameters-class}
\title{Signal Portfolio Parameters}
\description{
Class to encapsulate parameters for constructing signal portfolios (portfolio-blending). Only needed when
sb_algorithm is 'rp' or 'mvo'.
}
\section{Slots}{

\describe{
\item{\code{cov_est_method}}{An object of class \code{cov_est_method} representing the covariance estimation method and relevant parameters. Current methods are: 'sample', 'ewma', 'cc' (constant correlation),
'pca1', 'pca2', 'shrink_id' (shrinkage to identity matrix), 'shrink_cc' (shrinkage to constant correlation). This is only relevant for 'rp' and 'mvo'.}

\item{\code{mvo_parameters}}{An object of class \code{mvo_parameters} representing the parameters for mean-variance optimization. This is only relevant for 'mvo'.}

\item{\code{rp_parameters}}{An object of class \code{rp_parameters} representing the parameters for risk parity. This is only relevant for 'rp'.}

\item{\code{hrp_parameters}}{An object of class \code{hrp_parameters} representing the parameters for hierarchical risk parity. This is only relevant for 'hrp'.}

\item{\code{mmaf_parameters}}{An object of class \code{mmaf_parameters} representing the parameters for the MMAF portfolio construction method. This is only relevant for 'mmaf'.}

\item{\code{concentration_constraint_policy}}{The policy to handle concentration constraints.
It contains up to to three elements:
\itemize{
\item \code{benchmark}: A character vector describing the benchmark to be used to apply constraint.
For signal portfolios, possible options are theme_ss or theme_sb.
For stock portfolios, there must be a correspondence in \code{benchmark_weights_m_df}
\item \code{max_abs_active_individual_weight}: The maximum absolute individual active weights.
\item \code{max_abs_active_group_weight}: The maximum absolute theme active weight used for creating group constraints.
}}
}}

