% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-signal_transform.R
\name{signal_transform}
\alias{signal_transform}
\title{Signal Transformation Function}
\usage{
signal_transform(
  vector,
  lower_quantile_winsorization = 0.05,
  upper_quantile_winsorization = 0.95
)
}
\arguments{
\item{vector}{A numeric vector to be transformed.}

\item{lower_quantile_winsorization}{A numeric value between 0 and 1 specifying the quantile threshold for lower winsorization.}

\item{upper_quantile_winsorization}{A numeric value between 0 and 1 specifying the quantile threshold for upper winsorization.}
}
\value{
A numeric vector of the same length as \code{vector} with transformed values.
}
\description{
This function performs a sequence of transformations on a numeric vector:
\itemize{
\item Winsorizes the values based on specified quantiles.
\item Computes the z-score of the winsorized values.
\item Transforms the z-scores into a new vector based on their sign.
}
}
\details{
The function first applies winsorization to the \code{vector} based on the provided quantile thresholds.
Values exceeding the upper quantile are replaced with the upper quantile value, and values below the lower quantile are replaced with the lower quantile value.
The function then computes the z-scores of the winsorized values.
Finally, it transforms these z-scores:
\itemize{
\item Positive z-scores are adjusted to \eqn{1 + Z}
\item Negative z-scores are transformed to \eqn{\frac{1}{1 - Z}}
\item A z-score of zero is transformed to 1#'
}
}
\examples{
vector <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
upper_quantile_winsorization <- 0.9
lower_quantile_winsorization <- 0.1
transformed_vector <- signal_transform(vector, upper_quantile_winsorization, lower_quantile_winsorization)
print(transformed_vector)

}
