% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_window.R
\name{skew}
\alias{skew}
\title{Calculate Skewness of a Numeric Vector}
\usage{
skew(values, na.rm = TRUE)
}
\arguments{
\item{values}{A numeric vector whose skewness is to be calculated.}

\item{na.rm}{A logical value indicating whether NA values should be removed before computation (default is TRUE).}
}
\value{
A numeric value representing the skewness of the input vector. If the standard deviation is zero, \code{NA_real_} is returned.
}
\description{
This function computes the skewness (the third standardized moment) of a numeric vector.
It handles missing values according to the \code{na.rm} parameter and returns \code{NA} if the standard deviation is zero.
}
\examples{
skew(c(1, 2, 3, 4, 5))
}
