% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_rolling.R
\name{skew_rolling}
\alias{skew_rolling}
\title{Row-wise Skewness Rolling Calculation of a Matrix}
\usage{
skew_rolling(main_matrix, complementary_matrix)
}
\arguments{
\item{main_matrix}{A matrix, data frame, or tibble whose skewness needs to be calculated.}

\item{complementary_matrix}{A matrix, data frame, or tibble containing lagged observations of the main matrix. This matrix is used to extend the data available for calculating skewness at the beginning of the main matrix. The number of columns in the complementary matrix defines the width of the rolling window used for calculating skewness.}
}
\value{
A numeric matrix with the same dimensions as main_matrix where each entry is the skewness calculated from a rolling window of observations that includes both the lagged and current data. By default, NAs are ignored.
}
\description{
Calculates the row-wise skewness of elements in a matrix using a rolling window, which incorporates both lagged and current observations to assess the asymmetry of data distribution at each point.
}
\details{
The function appends the complementary matrix (containing lagged observations) to the main matrix to create a complete matrix used for skewness calculation. The skewness for each element in the main matrix is calculated using a rolling window that spans across the corresponding columns in the complete matrix, starting from the lagged values and extending to include the current value.
}
\examples{
# Create a main matrix and a complementary matrix
main_matrix <- matrix(c(5, 3, 7, 8), nrow = 2, ncol = 2)
complementary_matrix <- matrix(c(1, 2, 6, 4), nrow = 2, ncol = 2)

# Calculate skewness using a rolling window
skew_rolling(main_matrix, complementary_matrix)

}
