% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{ss_backtest_results_accessors}
\alias{ss_backtest_results_accessors}
\title{Accessor Methods for ss_backtest_results}
\usage{
\S4method{get_final_model}{ss_backtest_results}(object)
}
\arguments{
\item{object}{An object of class \code{ss_backtest_results}.}
}
\value{
The respective slot of the \code{ss_backtest_results} object.
}
\description{
These methods are used to access various components of an \code{ss_backtest_results} object.
}
