% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{step_impute_sector_new}
\alias{step_impute_sector_new}
\title{Constructor for step_impute_sector}
\usage{
step_impute_sector_new(
  terms,
  sector,
  method,
  role,
  trained,
  impute_values,
  skip,
  id
)
}
\description{
Constructor for step_impute_sector
}
