% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_matrix_ignore_na_matrix2.R
\name{sum_matrix_ignore_na_matrix2}
\alias{sum_matrix_ignore_na_matrix2}
\title{Element-wise Matrix Sum with Special NA Handling}
\usage{
sum_matrix_ignore_na_matrix2(matrix1, matrix2_ignoreifna)
}
\arguments{
\item{matrix1}{A matrix, data frame, or tibble. Values in this matrix are not ignored during addition.}

\item{matrix2_ignoreifna}{A matrix, data frame, or tibble. with the same dimensions as matrix1. Values in this matrix are ignored during addition if there are NAs in corresponding element in matrix1.}
}
\value{
A numeric matrix containing the element-wise sum of matrix1 and matrix2_ignoreifna, with values in matrix2_ignoreifna ignored if corresponding element in matrix1 is NA.
}
\description{
Computes the element-wise sum of two matrices while ignoring values in the second matrix if corresponding element in matrix1 is NA.
}
\details{
This function takes two matrices of the same dimensions and returns another matrix where each element is the sum of the corresponding elements in the input matrices. Unlike traditional matrix addition, this function treats values in the second matrix differently depending on corresponding element in matrix1: if an element is NA in first matrix, corresponding element in the resulting matrix is NA, regardless of the value in the second matrix.
}
\examples{
# Example matrices
matrix1 <- matrix(1:4, nrow = 2)
matrix2_ignoreifna <- matrix(c(5,NA,7,8), nrow = 2)

# Compute the sum
result <- sum_matrix_ignore_na_matrix2(matrix1, matrix2_ignoreifna)
result

matrix1 <- matrix(c(1,2,NA,4), nrow = 2)
matrix2_ignoreifna <- matrix(c(5,6,7,8), nrow = 2)

# Compute the sum
result <- sum_matrix_ignore_na_matrix2(matrix1, matrix2_ignoreifna)
result

}
