% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_window.R
\name{sum_top_n}
\alias{sum_top_n}
\title{Sum of the top n maximum values}
\usage{
sum_top_n(values, top_n = 1L, na.rm = TRUE)
}
\arguments{
\item{values}{Numeric vector.}

\item{top_n}{Integer >= 1.}

\item{na.rm}{Logical.}
}
\value{
Numeric scalar.
}
\description{
Sum of the top n maximum values
}
