% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_performance.R
\name{summarize_performance}
\alias{summarize_performance}
\title{Summarize Performance Metrics of Backtests}
\usage{
summarize_performance(
  selected_backtest_returns_corrected_positions_m_xts_upd_ref,
  selected_market_factor_proxy_m_xts_upd_ref,
  model_structure,
  model_spec_theme_level,
  lmer_control,
  selected_signal_themes_m_d_ref,
  custom_signal_universe_metrics_m_upd_ref = NULL,
  active_returns = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{selected_backtest_returns_corrected_positions_m_xts_upd_ref}{A xts containing the backtest returns data for various signals.}

\item{selected_market_factor_proxy_m_xts_upd_ref}{A xts containing benchmark returns data.}

\item{model_spec_theme_level}{A character string specifying the desired Bayesian model structure.
Options include:
\itemize{
\item \code{"random_intercept_fixed_slope"}: Includes random effects for the intercept at the theme level.
\item \code{"theme_specific_intercept_fixed_slope"}: Uses fixed intercepts for each theme.
\item \code{"theme_specific_intercept_theme_specific_slope"}: Includes fixed intercepts and slopes for each theme.
\item \code{"fixed_intercept_fixed_slope"}: Omits theme-level intercepts but includes random effects at the theme:signal level.
}}

\item{selected_signal_themes_m_d_ref}{A data frame containing metadata about signals. This data frame should include:
\itemize{
\item \code{tickers}: Signal identifiers matching those in \code{selected_backtest_returns_corrected_positions_m_xts_upd_ref}.
\item \code{theme}: Group membership for each signal, defining the clusters for the Bayesian hierarchical model.
\item \code{dates}: Dates corresponding to the backtest data.
This input ensures proper alignment between signals and their associated themes.
}}

\item{active_returns}{If TRUE, calculate ative returns before applying performance functions.}
}
\description{
Summarize Performance Metrics of Backtests
}
