% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summarize_posterior_draws.R
\name{summarize_posterior_draws}
\alias{summarize_posterior_draws}
\title{Summarize Posterior Draws for Signal Universe}
\usage{
summarize_posterior_draws(
  signal_universe_m_d_ref,
  posteriors_draws_list,
  signals_groups_m_d_ref,
  selected_benchmark_returns_upd_ref_vector
)
}
\arguments{
\item{signal_universe_m_d_ref}{A data frame containing the signal universe data with tickers and additional columns that will be updated.}

\item{posteriors_draws}{A named list where each element is a data frame containing posterior draws for a specific theme.}

\item{groups}{A data frame that maps signals groups to themes}
}
\value{
The \code{signal_universe_m_d_ref} data frame is updated in place with posterior summary statistics.
}
\description{
This function computes various posterior summary statistics for a given set of signals, themes, and posterior draws.
It updates the \code{signal_universe_m_d_ref} data frame with posterior statistics including alphas, betas, sigmas, and other metrics.
}
\details{
The function performs the following operations for each theme:
\itemize{
\item Computes and updates the posterior overall alpha and individual alpha for each signal.
\item Computes and updates the probability of the (positive) direction for posterior alphas.
\item Computes and updates the posterior overall beta and individual beta for each signal.
\item Computes and updates the posterior sigma for each signal.
\item Computes and updates posterior metrics such as active return, tracking error, and information ratio (IR).
\item Computes and updates additional performance metrics like Appraisal Ratio (AP) and Treynor ratio.
}
}
