% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-summarize_posterior_draws.R
\name{summarize_posteriors_draws}
\alias{summarize_posteriors_draws}
\title{Summarize Posterior Draws for Signal Universe}
\usage{
summarize_posteriors_draws(
  brm_model,
  signal_universe_m_d_ref,
  signal_themes_m_d_ref,
  model_spec_theme_level
)
}
\arguments{
\item{brm_model}{A bayesian model fit with \code{brms::brm}.}

\item{signal_universe_m_d_ref}{A dataframe with tickers, is_eligible and final_signal columns}

\item{signal_themes_m_d_ref}{A (meta) data frame with id, tickers ("signals") and dates column contemplating all signals in \code{signals_m_df} and a "theme" column providing group membership for each signal, which is needed
for defining clusters in bayesian hierarchical model. It should contain data only for current date.}

\item{model_spec_theme_level}{A character string specifying the desired Bayesian model structure.
Options include:
\itemize{
\item \code{"random_intercept"}: Includes random effects for the intercept at the theme level.
\item \code{"fixed_intercepts"}: Uses fixed intercepts for each theme.
\item \code{"fixed_intercepts_and_slopes"}: Includes fixed intercepts and slopes for each theme.
\item \code{"none"}: Omits theme-level intercepts but includes random effects at the theme:signal level.
}}
}
\value{
The \code{signal_universe_m_d_ref} data frame is updated in place with posterior summary statistics.
}
\description{
This function computes various posterior summary statistics for a given set of signals, themes, and posterior draws.
It updates the \code{signal_universe_m_d_ref} data frame with posterior statistics including alphas, betas, sigmas, and other metrics.
}
\details{
The function performs the following operations for each theme:
\itemize{
\item Computes and updates the posterior overall alpha and individual alpha for each signal.
\item Computes and updates the probability of the (positive) direction for posterior alphas.
\item Computes and updates the posterior overall beta and individual beta for each signal.
\item Computes and updates the posterior sigma for each signal.
\item Computes and updates posterior metrics such as active return, tracking error, and information ratio (IR).
\item Computes and updates additional performance metrics like Appraisal Ratio (AP) and Treynor ratio.
}
}
