% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,ml_backtest_results-method}
\alias{summary,ml_backtest_results-method}
\title{Summary Method for ml_backtest_results Class}
\usage{
\S4method{summary}{ml_backtest_results}(object)
}
\arguments{
\item{object}{An instance of the \code{ml_backtest_results} class.}
}
\value{
The method prints a summary to the console, including:
\itemize{
\item Main information about the ML algorithm, final model, and objective function.
\item Summary statistics (mean, median, SD, min, max) for out-of-sample predictions, errors, actual values, and evaluation metrics.
\item Chosen evaluation metric validation summaries.
\item Best hyperparameters and validation evaluation metrics hyper choice.
}
}
\description{
This method provides a comprehensive summary of the \code{ml_backtest_results}
object, including statistics for out-of-sample predictions, errors,
actual values, evaluation metrics, and hyperparameters.
}
