% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,ml_metabacktest_results-method}
\alias{summary,ml_metabacktest_results-method}
\title{Summary Method for ml_metabacktest_results Class}
\usage{
\S4method{summary}{ml_metabacktest_results}(object, summary_id = NULL)
}
\arguments{
\item{object}{An object of class \code{ml_metabacktest_results}.}

\item{summary_id}{A character string or numeric value specifying which table to display.
\itemize{
\item By name: Options are:
\itemize{
\item \code{"Time Series OOS Testing Metrics Summary"}
\item \code{"Time Series Validation Metrics Summary"}
}
\item By number: Provide a number corresponding to the table (as listed when \code{summary_id} is \code{NULL}).
If \code{NULL} (default), the method lists available summaries.
}}
}
\value{
Invisibly returns the input \code{object}.
}
\description{
Provides a detailed summary of an \code{ml_metabacktest_results} object.
Users can select which summary table to display by specifying the \code{summary_id} parameter,
either by name or by number.
The summary includes interactive tables styled using the \code{DT} package.
}
