% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,port_backtest_results-method}
\alias{summary,port_backtest_results-method}
\title{Summary Method for port_backtest_results Class}
\usage{
\S4method{summary}{port_backtest_results}(object, summary_id = NULL)
}
\arguments{
\item{object}{An object of class \code{port_backtest_results}.}

\item{summary_id}{A character string or numeric value specifying which table to display.}
}
\value{
Invisibly returns the input \code{object}.
}
\description{
Provides a detailed summary of \code{port_backtest_results} object.
Users can select which summary table to display by specifying the \code{summary_id} parameter.
The summary includes interactive tables styled using the \code{DT} package.
}
