% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,sb_backtest_results-method}
\alias{summary,sb_backtest_results-method}
\title{Summary Method for sb_backtest_results Class}
\usage{
\S4method{summary}{sb_backtest_results}(object, summary_id = NULL)
}
\arguments{
\item{object}{An object of class \code{sb_backtest_results}.}

\item{summary_id}{A character string or numeric value specifying which table to display.
\itemize{
\item By name: Options are:
\itemize{
\item \code{"OOS_SB_Outputs_Summary"} (Delegates to \code{meta_dataframe} summary)
\item \code{"OOS_Testing_Eval_Metrics"}
\item \code{"Chosen_Eval_Metric_Validation"}
}
\item By number: Provide a number corresponding to the table (as listed when \code{summary_id} is \code{NULL}).
If \code{NULL} (default), the method lists available tables.
}}
}
\value{
Invisibly returns the input \code{object}.
}
\description{
Provides a detailed summary of an \code{sb_backtest_results} object.
Users can select which summary table to display by specifying the \code{summary_id} parameter,
either by name or by number.
The summary includes interactive tables styled using the \code{DT} package and leverages
the summary method of the \code{meta_dataframe} class for the \code{oos_sb_outputs_m_df} component.
}
