% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,sb_metabacktest_results-method}
\alias{summary,sb_metabacktest_results-method}
\title{Summary Method for sb_metabacktest_results Class}
\usage{
\S4method{summary}{sb_metabacktest_results}(object, summary_id = NULL, which_backtest_results = NULL)
}
\arguments{
\item{object}{An object of class \code{sb_metabacktest_results}.}

\item{summary_id}{A character string or numeric value specifying which table to display.
\itemize{
\item By name: Options are:
\itemize{
\item \code{"Consolidated_OOS_Testing_Metrics"}
\item \code{"Mean_Validation_Metrics"}
\item \code{"Time_Series_OOS_Testing_Metrics"}
\item \code{"Time_Series_Validation_Metrics"}
\item \code{"Base_Learners_OOS_Predictions"}
}
\item By number: Provide a number corresponding to the table (as listed when \code{summary_id} is \code{NULL}).
If \code{NULL} (default), the method lists available tables.
}}

\item{which_backtest_results}{A character string or numeric value specifying which backtest results to display.}
}
\value{
Invisibly returns the input \code{object}.
}
\description{
Provides a detailed summary of an \code{sb_metabacktest_results} object.
Users can select which summary table to display by specifying the \code{summary_id} parameter,
either by name or by number.
The summary includes interactive tables styled using the \code{DT} package.
}
