% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,tickers_catalog-method}
\alias{summary,tickers_catalog-method}
\title{Summary Method for tickers_catalog Class}
\usage{
\S4method{summary}{tickers_catalog}(object, summary_id = NULL)
}
\arguments{
\item{object}{An instance of the \code{tickers_catalog} class.}

\item{summary_id}{A character string or numeric value specifying which summary to display.
\itemize{
\item By name: Options are:
\itemize{
\item "Catalog Overview"
}
\item By number: Provide a number corresponding to the summary.
If \code{NULL} (default), the method lists available summaries.
}}
}
\value{
Invisibly returns the input \code{object}.
}
\description{
Provides a detailed summary of a \code{tickers_catalog} object, allowing users to view
key statistics and filter the catalog slot interactively.
}
