% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_meta_dataframe.R
\name{sur}
\alias{sur}
\title{Calculate Standardized Unexpected Realization (SUR)}
\usage{
sur(final_value, past_values, na.rm = TRUE)
}
\arguments{
\item{na.rm}{A logical value indicating whether NA values should be removed before computation (default is TRUE).}

\item{values}{A numeric vector representing a series of observations.}
}
\value{
A numeric value representing the SUR. If the standard deviation is zero, \code{NA_real_} is returned.
}
\description{
This function computes the Standardized Unexpected Realization (SUR) for a numeric vector.
SUR is defined as the deviation of the most recent (final) value from the mean of the vector,
divided by the standard deviation. This metric quantifies how extreme the last observation is compared to the historical distribution.
}
\examples{
sur(c(1, 2, 3, 4, 5))
}
