% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dataframe.R
\docType{class}
\name{target_m_df-class}
\alias{target_m_df-class}
\title{Define the target S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to a target meta_dataframe.
}
