% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{tidy.step_winsorize}
\alias{tidy.step_winsorize}
\title{Tidy method for step_winsorize}
\usage{
\method{tidy}{step_winsorize}(x, ...)
}
\arguments{
\item{x}{A \code{step_winsorize} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Tidy method for step_winsorize
}
