% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{transaction_costs_parameters-class}
\alias{transaction_costs_parameters-class}
\title{Transaction Cost Parameters S4 Class}
\description{
This S4 class stores transaction cost parameters based on the BARRA model.
}
\section{Slots}{

\describe{
\item{\code{direct_transaction_cost}}{A numeric value representing the direct transaction cost (ie, brokerage fees). Should be in percentage (0.07 = 0.07\%) .}

\item{\code{strategy_aum}}{A numeric value representing the strategy's assets under management (AUM). Should be in same units as main_liquidity_metrics}

\item{\code{alpha}}{A numeric value representing the alpha parameter.}

\item{\code{lambda}}{A numeric value or the string "dynamic" representing the lambda parameter.}
}}

