% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{transactions_log-class}
\alias{transactions_log-class}
\title{Define the transactions_log S4 Class}
\description{
This class enforces that the underlying data is adherent to the output of a
port_backtest_workflow
}
