% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{transactions_log_m_df-class}
\alias{transactions_log_m_df-class}
\title{Define the transactions_log_m_df S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to the output of a
port_backtest_workflow
}
