% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-translate_metrics.R
\name{translate_metrics}
\alias{translate_metrics}
\title{Translate metrics based on machine learning algorithm and chosen evaluation metric.}
\usage{
translate_metrics(
  ml_algorithm,
  chosen_eval_metric,
  custom_objective,
  early_stop,
  huber_delta,
  verbose
)
}
\arguments{
\item{ml_algorithm}{Character string specifying the machine learning algorithm.}

\item{chosen_eval_metric}{Character string specifying the chosen evaluation metric.}

\item{custom_objective}{Character string specifying the custom objective function.}

\item{early_stop}{Numeric or null, indicating whether early stopping is enabled.}

\item{huber_delta}{Numeric specifying the delta parameter for Huber loss if applicable.}

\item{verbose}{Logical indicating whether to display verbose messages.}
}
\value{
A list containing the adapted chosen evaluation metric and custom objective function.
}
\description{
This function adapts the chosen evaluation metric based on the provided parameters.
}
\details{
\itemize{
\item If \code{chosen_eval_metric} is \code{NULL}, it selects an appropriate metric based on \code{custom_objective}.
\item Provides commentary if the selected evaluation metric is not supported for early stopping.
\item Adjusts \code{custom_objective} based on the specified \code{ml_algorithm}.
}
}
\examples{
translate_metrics("xgb", NULL, "squared_error", early_stop = FALSE, verbose = TRUE)

}
