% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-undummy_sector_features.R
\name{undummy_sector_features}
\alias{undummy_sector_features}
\title{Un-dummy sector features}
\usage{
undummy_sector_features(sectors_m_df)
}
\arguments{
\item{sectors_m_df}{A dataframe with id, tickers and dates with dummy sectors classifications to be used to fill NAs}
}
\description{
This functions takes dummy sectors columns and melt them into the single original column that created the dummies
}
