% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{update_meta_dataframe}
\alias{update_meta_dataframe}
\title{Update Meta Dataframe}
\usage{
update_meta_dataframe(old_features_m_df, new_features_m_df)
}
\arguments{
\item{old_features_m_df}{A features meta_dataframe object with previous data}

\item{new_features_m_df}{A features meta dataframe object with new data}
}
\value{
Updated meta_dataframe object
}
\description{
Update Meta Dataframe
}
