% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss_backtest.R
\name{update_ss_backtest}
\alias{update_ss_backtest}
\alias{update_ss_backtest,meta_dataframe,missing,port_backtest_cohort,meta_xts,meta_dataframe,ss_backtest_results-method}
\alias{update_ss_backtest,meta_dataframe,meta_xts,missing,meta_xts,meta_dataframe,ss_backtest_results-method}
\title{Update Signal Selection Backtest
The \code{update_ss_backtest} function will take an existing \code{port_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.}
\usage{
update_ss_backtest(
  signals_m_df,
  updated_backtest_returns_m_xts,
  updated_port_backtest_cohort,
  benchmark_returns_m_xts,
  signal_themes_m_df,
  old_results,
  ...
)

\S4method{update_ss_backtest}{meta_dataframe,missing,port_backtest_cohort,meta_xts,meta_dataframe,ss_backtest_results}(
  signals_m_df,
  updated_port_backtest_cohort,
  benchmark_returns_m_xts,
  signal_themes_m_df,
  old_results,
  priors_m_df = NULL,
  custom_signal_universe_metrics_m_df = NULL,
  verbose = TRUE,
  parallel = TRUE
)

\S4method{update_ss_backtest}{meta_dataframe,meta_xts,missing,meta_xts,meta_dataframe,ss_backtest_results}(
  signals_m_df,
  updated_backtest_returns_m_xts,
  benchmark_returns_m_xts,
  signal_themes_m_df,
  old_results,
  priors_m_df = NULL,
  custom_signal_universe_metrics_m_df = NULL,
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{signals_m_df}{A meta_dataframe containing the signal features. It must include at least the columns \code{id}, \code{tickers}, and \code{dates}.}

\item{updated_backtest_returns_m_xts}{An up-to-date xts containing historical backtested returns named according to signals in \code{signals_m_df},}

\item{updated_port_backtest_cohort}{An up-to-date \code{port_backtest_cohort} object containing historical backtested returns named according to signals in \code{signals_m_df},}

\item{benchmark_returns_m_xts}{A xts with benchmark returns, named accordingly.}

\item{signal_themes_m_df}{A (meta) data frame with "id", "tickers" ("signals"), and "dates" columns, including all signals in \code{signals_m_df}, and a "theme" column providing group membership for each signal.}

\item{...}{Additional arguments (not used in this method).}

\item{priors_m_df}{A (meta) data frame with columns including "id", "ticker", "dates", "theme" (used for clustering in the Bayesian hierarchical model),
and values for active_return, bench_return, alpha (mean and se), beta (mean and se), and sigma. Data should be exogenous, as it will be used to set priors for the hierarchical Bayesian model.}

\item{verbose}{A boolean indicating whether to print messages.}

\item{parallel}{A boolean indicating whether to run the backtest in parallel.}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization.}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization.}
}
\value{
An object of class \code{ss_backtest_results} containing the portfolio backtest results.
}
\description{
Update Signal Selection Backtest
The \code{update_ss_backtest} function will take an existing \code{port_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.
}
\section{Functions}{
\itemize{
\item \code{update_ss_backtest(
  signals_m_df = meta_dataframe,
  updated_backtest_returns_m_xts = missing,
  updated_port_backtest_cohort = port_backtest_cohort,
  benchmark_returns_m_xts = meta_xts,
  signal_themes_m_df = meta_dataframe,
  old_results = ss_backtest_results
)}: Updates a signal selection backtest based on a \code{ss_backtest_results} object and a \code{port_backtest_cohort}.

This method extracts the parameters from the \code{results} object (of class \code{ss_backtest_results}), modifies initial_sample_size, performs the
new backtest and then binds results to the old results.

\item \code{update_ss_backtest(
  signals_m_df = meta_dataframe,
  updated_backtest_returns_m_xts = meta_xts,
  updated_port_backtest_cohort = missing,
  benchmark_returns_m_xts = meta_xts,
  signal_themes_m_df = meta_dataframe,
  old_results = ss_backtest_results
)}: Updates a signal selection backtest based on a \code{ss_backtest_results} object and a \code{backtest_returns_m_xts}.

This method extracts the parameters from the \code{results} object (of class \code{ss_backtest_results}), modifies initial_sample_size, performs the
new backtest and then binds results to the old results.

}}
