% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_exp_ret_score_tilt}
\alias{validate_exp_ret_score_tilt}
\title{Valida exp_ret_score_tilt}
\usage{
validate_exp_ret_score_tilt(exp_ret_score_tilt, exp_ret_score_tilt_eta)
}
\arguments{
\item{exp_ret_score_tilt}{A character indicating when to apply the expected return score tilt. Possible options are 'none', 'inner' or 'final'.
'none' means no tilt is applied, 'inner' means the tilt is applied within the risk parity algorithm and 'final' means the tilt is applied after the risk parity solution is computed.}

\item{exp_ret_score_tilt_eta}{A numeric value representing the intensity of the expected return score tilt. Must be a positive value. Only needed when exp_ret_score_tilt is 'inner' or 'final'.}
}
\value{
Invisibly returns TRUE if the parameters are valid.
}
\description{
Internal function to validate the exp_ret_score_tilt and exp_ret_score_tilt_eta parameters
that belong to rp and hrp port constructions methods.
}
