% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_liquidity_floor_cutoffs}
\alias{validate_liquidity_floor_cutoffs}
\title{Validate Liquidity Floor Cutoffs}
\usage{
validate_liquidity_floor_cutoffs(
  liquidity_floor_cutoffs,
  main_liquidity_metric = NULL
)
}
\arguments{
\item{liquidity_floor_cutoffs}{A data.frame containing liquidity floor cutoffs.}

\item{main_liquidity_metric}{Optional character string specifying the column (other than
\code{"liquidity_classification"}) that should be used to check ascending order.}
}
\value{
Invisibly returns \code{TRUE} if the data frame passes all validations.
}
\description{
Internal function to validate a liquidity_floor_cutoffs data frame.
}
\details{
This function checks that the liquidity_floor_cutoffs data frame meets the following requirements:
\itemize{
\item It is a data.frame with at least two columns.
\item The first column is named \code{"liquidity_classification"}.
\item It has at most 5 rows.
\item The values in the \code{liquidity_classification} column are among
\code{"micro_caps"}, \code{"small_caps"}, \code{"mid_caps"}, \code{"large_caps"}, or \code{"mega_caps"}
and contain no duplicates.
\item All other columns are numeric and contain no missing values.
\item If \code{main_liquidity_metric} is provided, then it must be one of the non-classification
column names, the data frame must be arranged in ascending order by that metric, and the
ranking (order) of the values across all liquidity metric columns must be identical.
\item The liquidity metric values must not be normalized; that is, it is an error if all values in each
liquidity metric column (all but the first) are between -1 and 1.
}
}
\keyword{internal}
