% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_window.R
\name{validate_returns_bench}
\alias{validate_returns_bench}
\title{Validator function for returns and benchmark returns}
\usage{
validate_returns_bench(ret_values, bench_ret_values)
}
\arguments{
\item{ret_values}{Numeric vector (signal returns).}

\item{bench_ret_values}{Numeric vector (benchmark returns).}
}
\value{
NA_real_ if ret_values are all NAs or contain Inf/-Inf, else throws error if bench_ret_values contain NA/Inf or lengths differ.
}
\description{
Validator function for returns and benchmark returns
}
