% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_slsaf_parameters}
\alias{validate_slsaf_parameters}
\title{Validate SLSAF Parameters}
\usage{
validate_slsaf_parameters(object)
}
\arguments{
\item{object}{An object of class \code{slsaf_parameters}.}
}
\value{
\code{TRUE} invisibly if the configuration is valid. Stops with an informative
message otherwise.
}
\description{
Validates an \code{slsaf_parameters} object, the configuration of the Simulated Long-Short
Allocation Framework.
}
