% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_turnover_constraint_policy}
\alias{validate_turnover_constraint_policy}
\title{Validate Turnover Constraint Policy}
\usage{
validate_turnover_constraint_policy(turnover_constraint_policy)
}
\arguments{
\item{turnover_constraint_policy}{A list with the elements described above.}
}
\value{
Invisibly returns TRUE if the policy is valid.
}
\description{
Internal function to validate the turnover constraint policy.
}
\details{
The policy is provided as a list that must contain the following elements:
\describe{
\item{quantile_range_buffer}{A numeric value between 0 and 1 that defines the buffer for turnover quantiles.}
\item{turnover_cap_rules}{A named numeric vector whose names correspond to valid liquidity categories
("micro_caps", "small_caps", "mid_caps", "large_caps", "mega_caps"). Each value must be between 0 and 1.
Additionally, for any two categories, if one category is less liquid than another then its cap must not exceed
that of the more liquid category.}
}
}
\keyword{internal}
