% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{weights_m_df-class}
\alias{weights_m_df-class}
\title{Define the weights_m_df S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to a weights_meta_dataframe
}
