% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize_panel_data.R
\name{winsorize_panel_data}
\alias{winsorize_panel_data}
\title{Winsorize Panel Data}
\usage{
winsorize_panel_data(
  features_m_df,
  dates_vector,
  probs,
  Infs_to_preserve = NULL
)
}
\arguments{
\item{features_m_df}{A meta_dataframe object or an object coercible to data frame containing panel data with columns for "id", "tickers" and "dates".}

\item{dates_vector}{A vector of dates for which winsorization is performed.}

\item{probs}{A vector of probabilities specifying the quantiles for winsorization (e.g., c(0.99, 0.01)).}

\item{Infs_to_preserve}{A vector of column names for features where infinite values should be replaced by max/min and not by NA.}
}
\value{
A winsorized panel data matrix.
}
\description{
Winsorizes a panel data matrix, replacing extreme values with specified quantiles.
}
\details{
This function winsorizes a panel data matrix, replacing extreme values with specified quantiles. For each specified date in the dates_vector, it calculates the quantiles for each feature and replaces values above the upper quantile with the upper quantile and values below the lower quantile with the lower quantile. Infinite values are treated differently depending on whether they should be preserved or not.
}
\examples{
# Example usage
features_m_df <- data.frame(
  id = c(1, 2, 3, 4),
  tickers = c("A", "B", "C", "D"),
  dates = as.Date(c("2022-01-01", "2022-01-01", "2022-01-02", "2022-01-02")),
  feature1 = c(1, 2, 3, 4),
  feature2 = c(5, 6, 7, 8)
)
dates_vector <- as.factor(c("2022-01-01", "2022-01-02"))
probs <- c(0.05, 0.95)
winsorized_data <- winsorize_panel_data(features_m_df, dates_vector, probs)

}
