% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_alpha_test_strategy}
\alias{add_alpha_test_strategy}
\alias{add_alpha_test_strategy,ss_backtest_config,alpha_test_strategy-method}
\alias{add_alpha_test_strategy,ss_backtest_config,missing-method}
\title{Add Alpha Test Strategy to ss_backtest_config}
\usage{
add_alpha_test_strategy(object, alpha_test_strategy, ...)

\S4method{add_alpha_test_strategy}{ss_backtest_config,alpha_test_strategy}(object, alpha_test_strategy)

\S4method{add_alpha_test_strategy}{ss_backtest_config,missing}(
  object,
  signal_significance_threshold = 0.05,
  p_correction_method = "none",
  market_factor_proxy,
  model_structure = "partial_pooled",
  theme_level_intercept = NULL,
  theme_level_slope = NULL,
  enable_theme_representativeness = TRUE,
  bayesian_model_parameters = NULL,
  lmer_control = NULL
)
}
\arguments{
\item{object}{An \code{ss_backtest_config} object.}

\item{alpha_test_strategy}{An \code{alpha_test_strategy} object to be added.}
}
\value{
The updated \code{ss_backtest_config} object with the specified \code{alpha_test_strategy} added.
}
\description{
This method allows you to add an \code{alpha_test_strategy} object to an \code{ss_backtest_config} object.
}
