% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_bayesian_model_parameters}
\alias{add_bayesian_model_parameters}
\title{Add Bayesian Model Parameters}
\usage{
add_bayesian_model_parameters(object, ...)
}
\arguments{
\item{object}{An object of class \code{bayesian_alpha_test_strategy}.}

\item{model_spec_theme_level}{A character string specifying the desired Bayesian model structure.
Options: \code{"random_intercept"}, \code{"fixed_intercepts"}, \code{"fixed_intercepts_and_slopes"}, \code{"none"}.}

\item{user_priors}{An optional object of class \code{brmsprior}.}

\item{prior_derivation_control}{An optional list containing \code{half_t_df}, \code{lmer_optimizer} and/or \code{lmer_optimization_objective}}

\item{brms_control}{An optional list of parameters for \code{brms::brm}.}
}
\value{
An updated \code{bayesian_alpha_test_strategy} object with the \code{bayesian_model_parameters} slot populated.
}
\description{
Adds an object of class \code{bayesian_model_parameters} to a \code{bayesian_alpha_test_strategy} object.
}
