% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_brms_prior}
\alias{add_brms_prior}
\alias{add_brms_prior,bayesian_alpha_test_strategy-method}
\alias{add_brms_prior,ss_backtest_config-method}
\title{Add Prior to Bayesian Alpha Test Strategy}
\usage{
add_brms_prior(object, ...)

\S4method{add_brms_prior}{bayesian_alpha_test_strategy}(
  object,
  distribution_choice,
  pars,
  class,
  coef = NULL,
  group = NULL
)

\S4method{add_brms_prior}{ss_backtest_config}(
  object,
  distribution_choice,
  pars,
  class,
  coef = NULL,
  group = NULL
)
}
\arguments{
\item{object}{An object of class \code{bayesian_alpha_test_strategy}.}

\item{distribution_choice}{A vector of strings representing the distribution of the prior. See details for valid options.}

\item{pars}{A list of named numeric vectors. The names should match the expected parameter names for the chosen distribution.}

\item{class}{A vector of character strings representing the class of the prior. Should be one of 'Intercept', 'b', 'sd', 'sigma', or 'cor'.}

\item{coef}{A vector of character strings representing the coefficient name. Only applicable when \code{class} is 'b' or 'sd'.}

\item{group}{A vector of character strings representing the group name. Only applicable when \code{class} is 'b' or 'sd'.}
}
\value{
An updated \code{bayesian_alpha_test_strategy} object with the added prior.
}
\description{
Adds a prior to a \code{bayesian_alpha_test_strategy} object based on provided parameters.
}
