% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_cov_est_method}
\alias{add_cov_est_method}
\alias{add_cov_est_method,sb_backtest_config,cov_est_method-method}
\alias{add_cov_est_method,sb_backtest_config,missing-method}
\alias{add_cov_est_method,port_backtest_config,cov_est_method-method}
\alias{add_cov_est_method,port_backtest_config,missing-method}
\title{Add a cov_est_method to a \code{sb_backtest_config} or \code{port_backtest_config} object.

This function allows either directly add a pre-existing \code{cov_est_method} object or create one dynamically by passing additional arguments.
When \code{cov_est_method} is not provided, a new one will be created using the values for \code{cov_estimation_method}, \code{cov_matrix_sample_size}, \code{active_returns}, passed via the \code{...} argument.}
\usage{
add_cov_est_method(object, cov_est_method, ...)

\S4method{add_cov_est_method}{sb_backtest_config,cov_est_method}(object, cov_est_method, ...)

\S4method{add_cov_est_method}{sb_backtest_config,missing}(
  object,
  cov_est_method,
  cov_estimation_method = "sample",
  cov_matrix_sample_size = 36,
  active_returns = TRUE,
  ...
)

\S4method{add_cov_est_method}{port_backtest_config,cov_est_method}(object, cov_est_method, ...)

\S4method{add_cov_est_method}{port_backtest_config,missing}(
  object,
  cov_est_method,
  cov_estimation_method = "sample",
  cov_matrix_sample_size = 252,
  active_returns = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{port_backtest_config}.}

\item{cov_est_method}{An existing object of class \code{cov_est_method}.}

\item{...}{Additional arguments used to create a new \code{cov_est_method} when \code{cov_est_method} is missing. These arguments must include:
\itemize{
\item \strong{cov_estimation_method}: A character string representing the covariance estimation method. Must be one of 'sample', 'ewma', 'cc', 'pca1', 'pca2', 'shrink_id' or 'shrink_cc'.
\item \strong{cov_matrix_sample_size}: Number of periods to subset return sample when estimating the covariance matrix. A high number will provide
higher degrees of freedom, but old returns might not reflect current risk due to parameter shift. A low number will tend to expose estimation
to dimensionality curse.
\item \strong{active_returns}: logical. If TRUE, the covariance matrix will be estimated using active returns. If FALSE, the covariance matrix will be estimated using raw returns.
}}
}
\value{
An updated object of class \code{sb_backtest_config} with the \code{cov_est_method} added.
}
\description{
Add a cov_est_method to a \code{sb_backtest_config} or \code{port_backtest_config} object.

This function allows either directly add a pre-existing \code{cov_est_method} object or create one dynamically by passing additional arguments.
When \code{cov_est_method} is not provided, a new one will be created using the values for \code{cov_estimation_method}, \code{cov_matrix_sample_size}, \code{active_returns}, passed via the \code{...} argument.
}
\section{Functions}{
\itemize{
\item \code{add_cov_est_method(
  object = sb_backtest_config,
  cov_est_method = cov_est_method
)}: Add existing \code{cov_est_method} object to a \code{sb_backtest_config} object.

This method allows to add an already existing \code{cov_est_method} object to an \code{sb_backtest_config}.

\item \code{add_cov_est_method(object = sb_backtest_config, cov_est_method = missing)}: Create a \code{cov_est_method} object to a \code{sb_backtest_config} object.

This method allows to dynamically create a \code{cov_est_method} object and add to \code{sb_backtest_config}.

\item \code{add_cov_est_method(
  object = port_backtest_config,
  cov_est_method = cov_est_method
)}: Add existing \code{cov_est_method} object to a \code{port_backtest_config} object.

This method allows to add an already existing \code{cov_est_method} object to an \code{port_backtest_config}.

\item \code{add_cov_est_method(object = port_backtest_config, cov_est_method = missing)}: Create a \code{cov_est_method} object to a \code{port_backtest_config} object.

This method allows to dynamically create a \code{cov_est_method} object and add to \code{port_backtest_config}.

}}
