% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_hyperparameter}
\alias{add_hyperparameter}
\alias{add_hyperparameter,hyper_grid_domain-method}
\alias{add_hyperparameter,grid_search_strategy-method}
\alias{add_hyperparameter,random_search_strategy-method}
\alias{add_hyperparameter,bayesian_opt_strategy-method}
\alias{add_hyperparameter,sb_backtest_config-method}
\title{Add a Hyperparameter to a \code{hyper_grid_domain}, whether inside a \code{sb_backtest_config}, a \code{tuning_strategy} or on its own.}
\usage{
add_hyperparameter(object, hyperparameter, ...)

\S4method{add_hyperparameter}{hyper_grid_domain}(object, new_hyperparameter_list)

\S4method{add_hyperparameter}{grid_search_strategy}(object, hyperparameter, grid, ...)

\S4method{add_hyperparameter}{random_search_strategy}(object, hyperparameter, distribution_choice, pars, ...)

\S4method{add_hyperparameter}{bayesian_opt_strategy}(object, hyperparameter, bounds, ...)

\S4method{add_hyperparameter}{sb_backtest_config}(
  object,
  hyperparameter,
  grid = NULL,
  distribution_choice = NULL,
  pars = NULL,
  bounds = NULL
)
}
\arguments{
\item{object}{A \code{tuning_strategy} or a \code{hyper_grid_domain} object.}

\item{hyperparameter}{A vector of characters indicating the name of the hyperparameter to be added.}

\item{new_hyperparameter_list}{Used to pass new_hyperparameters_list after specific method at tuning_strategy level.}

\item{grid}{A numeric vector or list of numeric vectors for grid search values (only used for grid_search).}

\item{distribution_choice}{A character vector indicating the distribution to sample from (only used for random_search).}

\item{pars}{A numeric named vector or list of numeric named vectors specifying parameter values (only used for random_search).}

\item{bounds}{A vector of length 2 indicating minimum and maximum bounds for each hyperparameter (only used for bayesian_opt).}
}
\value{
A \code{hyper_grid_domain} object with the updated hyperparameters.
}
\description{
This generic function adds a new hyperparameter to an existing \code{hyper_grid_domain} object. The function is overloaded to handle different types of hyperparameters for different machine learning algorithms.
}
\section{Functions}{
\itemize{
\item \code{add_hyperparameter(hyper_grid_domain)}: Add hyperparameter to \code{hyper_grid_domain} object

\item \code{add_hyperparameter(grid_search_strategy)}: Add hyperparameter to \code{grid_search_strategy} object

\item \code{add_hyperparameter(random_search_strategy)}: Add hyperparameter to \code{random_search_strategy} object

\item \code{add_hyperparameter(bayesian_opt_strategy)}: Add hyperparameter to \code{bayesian_opt_strategy} object

\item \code{add_hyperparameter(sb_backtest_config)}: Add Hyperparameter to \code{sb_backtest_config} object

}}
