% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_hyperparameter}
\alias{add_hyperparameter}
\title{Add a Hyperparameter to a \code{hyper_grid_domain}}
\usage{
add_hyperparameter(hyper_grid_domain, new_hyperparameters)
}
\arguments{
\item{hyper_grid_domain}{A \code{hyper_grid_domain} object to which the hyperparameter will be added.}

\item{new_hyperparameters}{A named list representing the hyperparameter to be added.}
}
\value{
A new \code{hyper_grid_domain} object with the updated hyperparameters.
}
\description{
This method adds a new hyperparameter to an existing \code{hyper_grid_domain} object.
}
\examples{
# Create an initial hyper_grid_domain object for grid-search
hyper_grid <- create_hyper_grid_domain(
  tuning_method = "grid_search",
  ml_algorithm = "glmnet"
)

#Add hyperparameter for grid_search
hyper_grid <- add_hyperparameter(
hyper_grid_domain = hyper_grid,
new_hyperparameters = list(alpha = c(0.2, 0.5)) #Hyperparameters may be added one by one or all at once (eg.: list(alpha = c(0.2, 0.5), lambda.min.ratio = c(0.5, 0.2, 0.3)))
)

random_search and xgb algorithm
hyper_grid <- create_hyper_grid_domain(
  tuning_method = "random_search",
  ml_algorithm = "xgb"
)

hyper_grid <- add_hyperparameter(
hyper_grid_domain = hyper_grid,
new_hyperparameters =
list(min_child_weight = list(distribution_choice = "constant", value = 3),
     max_depth = list(distribution_choice = "uniform", pars = c(min = 1L, max = 2L)),
     subsample = list(distribution_choice = "uniform", pars = c(min = 0.25, max = 0.50)),
     colsample_bytree = list(distribution_choice = "uniform", pars = c(min = 0.25, max = 0.50)),
     eta = list(distribution_choice = "uniform", pars = c(min = 0.1, max = 0.2)),
     alpha = list(distribution_choice = "uniform", pars = c(min = 2, max = 5)),
     gamma = list(distribution_choice = "constant", value = 0),
     nrounds = list(distribution_choice = "uniform", pars = c(min = 200L, max = 500L)))
)

bayesian_opt and rf algorithm
#' hyper_grid <- create_hyper_grid_domain(
  tuning_method = "bayesian_opt",
  ml_algorithm = "rf"
)


hyper_grid <- add_hyperparameter(
hyper_grid_domain = hyper_grid,
new_hyperparameters =
list(mtry = c(0,1),
     num.trees = c(100L, 1000L),
     max.depth = c(2L, 8L),
     min.bucket = c(1, 5))
)


}
