% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_liquidity_constraint_policy}
\alias{add_liquidity_constraint_policy}
\title{Add Liquidity Constraint Policy}
\usage{
add_liquidity_constraint_policy(
  port_backtest_config_obj,
  liquidity_floor_rule = NULL,
  liquidity_cap_rules = NULL
)
}
\arguments{
\item{port_backtest_config_obj}{A \code{port_backtest_config} object to which the liquidity constraint policy will be added.}

\item{liquidity_floor_rule}{A character string representing the liquidity floor rule (optional).}

\item{liquidity_cap_rules}{A named numeric vector where names represent liquidity classifications and values represent liquidity caps (optional).}
}
\value{
The updated \code{port_backtest_config} object with the added liquidity constraint policy.
}
\description{
Adds a liquidity constraint policy to a \code{port_backtest_config} object.
}
\examples{
portfolio <- create_port_backtest_config()
portfolio <- add_liquidity_constraint_policy(portfolio, liquidity_floor_rule = "micro_caps")
portfolio <- add_liquidity_constraint_policy(portfolio, liquidity_cap_rules = c(micro_caps = 0.01))
portfolio <- add_liquidity_constraint_policy(portfolio, liquidity_floor_rule = "small_caps") # This should update the liquidity_floor_rule

}
