% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_liquidity_floor_cutoffs}
\alias{add_liquidity_floor_cutoffs}
\title{Add Liquidity Floor Cutoffs}
\usage{
add_liquidity_floor_cutoffs(portfolio_policies_obj, liquidity_metric, cutoffs)
}
\arguments{
\item{portfolio_policies_obj}{An S4 object of class \code{portfolio_policies}.}

\item{liquidity_metric}{A character string representing the liquidity metric.}

\item{cutoffs}{A numeric vector of length 5 representing the cutoff values for
micro_caps, small_caps, mid_caps, large_caps, and mega_caps.}
}
\value{
An updated \code{portfolio_policies} object.
}
\description{
Add liquidity floor cutoffs to the portfolio_policies object.
}
