% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_sb_backtest_config}
\alias{add_sb_backtest_config}
\title{Add one or more sb_backtest_config objects to an sb_metabacktest_config}
\usage{
add_sb_backtest_config(object, ...)
}
\arguments{
\item{object}{An \code{sb_metabacktest_config} object.}

\item{...}{One or more \code{sb_backtest_config} objects to add.}
}
\value{
An updated \code{sb_metabacktest_config} object with added configurations.
}
\description{
Add one or more sb_backtest_config objects to an sb_metabacktest_config
}
