% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_tuning_strategy}
\alias{add_tuning_strategy}
\alias{add_tuning_strategy,sb_backtest_config,tuning_strategy-method}
\alias{add_tuning_strategy,sb_backtest_config,missing-method}
\title{Add a \code{tuning_strategy} to an existing \code{sb_backtest_config}.}
\usage{
add_tuning_strategy(object, tuning_strategy, ...)

\S4method{add_tuning_strategy}{sb_backtest_config,tuning_strategy}(object, tuning_strategy)

\S4method{add_tuning_strategy}{sb_backtest_config,missing}(
  object,
  tuning_strategy = NULL,
  tuning_method,
  validation_sample_size,
  chosen_eval_metric = NULL,
  hyper_grid_domain = NULL,
  early_stop = NULL,
  n_iter = NULL,
  acq = "ucb",
  init_points = NULL,
  k_iter = NULL
)
}
\arguments{
\item{object}{A \code{sb_backtest_config} object.}

\item{tuning_strategy}{\code{NULL}, indicating that a new \code{tuning_strategy} should be created.}

\item{...}{Additional arguments required to create a new \code{tuning_strategy}, only needed when tuning_strategy is \code{NULL}.}

\item{tuning_method}{Character string indicating the hyperparameter tuning method. Must be one of 'grid_search', 'random_search', or 'bayesian_opt'.}

\item{validation_sample_size}{Numeric value representing the size of the validation sample.}

\item{chosen_eval_metric}{Character or \code{NULL}, specifying the evaluation metric to be optimized.}

\item{early_stop}{Optional, stopping criteria for early termination. Can be of any type.}

\item{n_iter}{Numeric, number of iterations for 'random_search' or 'bayesian_opt'.}

\item{acq}{Character string specifying the acquisition function for Bayesian optimization (for 'bayesian_opt' only).}

\item{init_points}{Numeric, number of initial random points for Bayesian optimization (for 'bayesian_opt' only).}

\item{k_iter}{Numeric, number of samples to evaluate during Bayesian optimization (for 'bayesian_opt' only).}
}
\value{
An updated \code{sb_backtest_config} object with the specified or newly created \code{tuning_strategy}.

The updated \code{sb_backtest_config} object with the provided \code{tuning_strategy}.

An updated \code{sb_backtest_config} object with a newly created \code{grid_search_strategy}, \code{random_search_strategy}, or \code{bayesian_opt_strategy}, depending on the selected \code{tuning_method}.
}
\description{
This generic function adds an existing \code{tuning_strategy} object or creates a new one if none is provided
(i.e., when tuning_strategy is \code{NULL}).
}
\section{Functions}{
\itemize{
\item \code{add_tuning_strategy(
  object = sb_backtest_config,
  tuning_strategy = tuning_strategy
)}: Add an existing \code{tuning_strategy} to the \code{sb_backtest_config}.

This method adds a pre-existing \code{tuning_strategy} to the \code{sb_backtest_config}. It replaces any existing tuning strategy in the experiment.

\item \code{add_tuning_strategy(object = sb_backtest_config, tuning_strategy = missing)}: Create and add a new \code{tuning_strategy} to the \code{sb_backtest_config}.

This method is used when \code{tuning_strategy} is \code{NULL}. It creates a new tuning strategy based on the provided parameters and adds it to the \code{sb_backtest_config}.

}}
