% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{as.list,liquidity_constraint_policy-method}
\alias{as.list,liquidity_constraint_policy-method}
\title{Convert Liquidity Constraint Policy to List}
\usage{
\S4method{as.list}{liquidity_constraint_policy}(x)
}
\arguments{
\item{x}{An S4 object of class \code{liquidity_constraint_policy}.}
}
\value{
A named list containing:
\itemize{
\item \code{liquidity_floor_rule}
\item \code{liquidity_cap_rules}
}
}
\description{
Convert Liquidity Constraint Policy to List
}
