% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{as.list,turnover_constraint_policy-method}
\alias{as.list,turnover_constraint_policy-method}
\title{as.list Method for turnover_constraint_policy S4 Class}
\usage{
\S4method{as.list}{turnover_constraint_policy}(x, ...)
}
\arguments{
\item{x}{A turnover_constraint_policy S4 object.}

\item{...}{Additional arguments (unused).}
}
\value{
A list with elements \code{quantile_range_buffer} and \code{turnover_cap_rules}.
}
\description{
Converts a turnover_constraint_policy S4 object to a list with elements
\code{quantile_range_buffer} and \code{turnover_cap_rules}.
}
