% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{bake.step_winsorize}
\alias{bake.step_winsorize}
\title{Apply winsorization during baking}
\usage{
\method{bake}{step_winsorize}(object, new_data, ...)
}
\arguments{
\item{object}{A trained \code{step_winsorize} object.}

\item{new_data}{A data frame to which the winsorization will be applied.}
}
\description{
Apply winsorization during baking
}
