% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-bayesian_adjustment.R
\name{bayesian_adjustment}
\alias{bayesian_adjustment}
\title{Bayesian Adjustment Function}
\usage{
bayesian_adjustment(
  selected_signals_backtest_returns_upd_ref,
  selected_benchmark_returns_vector_upd_ref,
  signal_universe_m_d_ref,
  selected_priors_informative_data_m_upd_ref,
  priors_type,
  user_priors_list,
  signals_groups_m_d_ref,
  verbose = TRUE
)
}
\arguments{
\item{selected_signals_backtest_returns_upd_ref}{A data frame or matrix containing the backtest returns for the selected signals. This data is used for model fitting.}

\item{selected_priors_informative_data_m_upd_ref}{A data frame or matrix containing informative prior data. This data is used to set priors if \code{priors_type} is not \code{"none"}.}

\item{priors_type}{A character string indicating if and how priors should be set. If \code{"none"}, no priors are set. Otherwise, the function \code{set_priors} is used to determine the priors.}

\item{signals_groups_m_d_ref}{A data frame containing groups of signals with associated metadata, such as themes. This information is used to fit the Bayesian hierarchical model.}

\item{verbose}{A logical value indicating whether to print progress messages. If TRUE, progress messages are printed; otherwise, they are suppressed. Default is TRUE.}

\item{selected_benchmark_returns_upd_ref}{A data frame or matrix containing the updated benchmark returns. The second column is used for model fitting and posterior analysis.}
}
\value{
A data frame or matrix (\code{signal_universe_m_d_ref}) updated with posterior statistics based on the Bayesian model fitting.
}
\description{
Performs Bayesian p-value adjustment based on priors, fits a Bayesian hierarchical model to signals, and summarizes posterior draws. Optionally provides progress updates if verbose is set to TRUE.
}
\details{
This function performs the following steps:
\itemize{
\item If \code{priors_type} is not \code{"uninformative"}, it sets priors based on the provided informative data using the \code{set_priors} function.
\item It fits a Bayesian hierarchical model to each theme in \code{signals_groups_m_d_ref} using parallel processing. Progress messages are printed if \code{verbose} is TRUE.
\item It extracts posterior draws from the fitted models and summarizes them using the \code{summarize_posterior_draws} function.
}
}
