% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{bayesian_opt_strategy-class}
\alias{bayesian_opt_strategy-class}
\title{Bayesian Opt Tuning Strategy}
\arguments{
\item{init_points}{Number of initial random points for Bayesian optimization.}

\item{k_iter}{Integer that specifies the number of times to sample eval_function at each Epoch during Bayesian optimization.
If the intention is running in parallel, set k_iter to a multiple of the number of cores. Must be lower and preferably a multiple of n_iter.}
}
\description{
A subclass of \code{hyperparameter_tuning_strategy} that implements bayesian optimization.
}
\section{Slots}{

\describe{
\item{\code{tuning_method}}{The tuning method is set to 'bayesian_opt'.}

\item{\code{n_iter}}{For bayesian_opt, it should be the number of times the ml algorithm will be evaluated after initialization.}

\item{\code{acq}}{Acquisition function for Bayesian optimization: "ucb", "ei", or "poi".}
}}

