% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_window.R
\name{cagr}
\alias{cagr}
\title{Calculate Compound Annual Growth Rate (CAGR)}
\usage{
cagr(begin, final, period)
}
\arguments{
\item{begin}{A numeric value representing the initial value.}

\item{final}{A numeric value representing the final value.}

\item{period}{A numeric value representing the number of periods over which the growth is calculated.}
}
\value{
A numeric value representing the CAGR. If either \code{begin} or \code{final} is \code{NA},
or if the period is less than or equal to zero, appropriate errors or \code{NA} are returned.
}
\description{
This function computes the Compound Annual Growth Rate (CAGR) given an initial value, a final value,
and the number of periods. It handles various scenarios, including cases with negative values or missing data.
}
\examples{
cagr(100, 200, 5)
}
