% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_inputs_meta_sb_backtest.R
\name{check_inputs_meta_sb_backtest}
\alias{check_inputs_meta_sb_backtest}
\title{Validate Meta Signal Selection and Data Inputs}
\usage{
check_inputs_meta_sb_backtest(
  config,
  features_m_df,
  target_m_df,
  base_sb_backtest_results_list,
  base_signal_themes_m_df,
  base_custom_signal_weights_m_df,
  base_custom_signal_universe_metrics_m_df,
  meta_signal_themes_m_df,
  meta_custom_signal_weights_m_df,
  meta_custom_signal_universe_metrics_m_df,
  base_backtest_returns_m_xts,
  base_benchmark_returns_m_xts,
  meta_backtest_returns_m_xts,
  meta_benchmark_returns_m_xts,
  verbose = TRUE
)
}
\arguments{
\item{config}{An object containing the meta backtest configuration.}

\item{features_m_df}{A dataframe containing features}

\item{target_m_df}{A dataframe containing targets}

\item{base_sb_backtest_results_list}{A list of base signal blend backtest results.}

\item{base_signal_themes_m_df}{Optional base signal themes meta dataframe.}

\item{base_custom_signal_weights_m_df}{Optional base custom signal weights meta dataframe.}

\item{base_custom_signal_universe_metrics_m_df}{Optional base custom signal universe metrics meta dataframe.}

\item{meta_signal_themes_m_df}{Optional meta signal themes meta dataframe.}

\item{meta_custom_signal_weights_m_df}{Optional meta custom signal weights meta dataframe.}

\item{meta_custom_signal_universe_metrics_m_df}{Optional meta custom signal universe metrics meta dataframe.}

\item{base_backtest_returns_m_xts}{Optional xts object for base backtest returns.}

\item{base_benchmark_returns_m_xts}{Optional xts object for base benchmark returns.}

\item{meta_backtest_returns_m_xts}{Optional xts object for meta backtest returns.}

\item{meta_benchmark_returns_m_xts}{Optional xts object for meta benchmark returns.}

\item{verbose}{A boolean indicating whether to print detailed messages.}
}
\value{
None. Stops execution if validation checks fail.
}
\description{
This function performs a series of validation checks on the meta-level
signal selection, features passthrough, meta dataframes, and xts objects.
}
