% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_inputs_sb_backtest.R
\name{check_inputs_sb_backtest}
\alias{check_inputs_sb_backtest}
\title{Perform validation checks on inputs for SB workflow}
\usage{
check_inputs_sb_backtest(
  features_m_df,
  target_m_df,
  training_sample_size,
  target_fwd_name,
  validation_sample_size,
  rebalancing_months,
  split_method,
  signal_universe_m_df,
  backtest_returns_m_xts,
  benchmark_returns_m_xts,
  cov_matrix_benchmark,
  cov_matrix_sample_size,
  cov_estimation_method,
  active_returns,
  signal_themes_m_df,
  rp_method,
  n_random_ports,
  random_ports_method,
  opt_objective,
  concentration_constraint_policy,
  custom_signal_weights_m_df,
  sb_algorithm,
  gsm_algorithm,
  custom_objective,
  chosen_eval_metric,
  huber_delta,
  quantile_tau,
  hyper_grid_domain_list,
  tuning_method,
  n_iter,
  k_iter,
  acq,
  init_points,
  early_stop,
  keras_architecture_parameters,
  parallel,
  verbose = TRUE,
  .test_seed
)
}
\arguments{
\item{features_m_df}{A data frame or matrix containing features data.}

\item{target_m_df}{A data frame or matrix containing target variable data.}

\item{training_sample_size}{Numeric, size of the training sample.}

\item{target_fwd_name}{Character, name of the forward looking target.}

\item{validation_sample_size}{Numeric, size of the validation sample.}

\item{rebalancing_months}{Numeric, number of months for rebalancing.}

\item{split_method}{Character, method of data splitting (currently only "expanding" is supported).}

\item{backtest_returns_m_xts}{A xts containing historical backtested returns named according to signals in \code{signals_m_df},}

\item{benchmark_returns_m_xts}{A xts with benchmark returns, named accordingly.}

\item{sb_algorithm}{Character, choice of signal blending algorithm ("ols", "glmnet", "rf", "xgb", "nn", "ew", "sw", "rp", "mvo").}

\item{custom_objective}{Character, custom objective function for loss.}

\item{chosen_eval_metric}{Character, chosen evaluation metric ("rmse", "mae", "cp", "rss", "mphe", "mpe", "hr", "mape").}

\item{huber_delta}{Numeric, delta parameter for Huber loss (for "pseudo_huber_error" custom objective).}

\item{quantile_tau}{Numeric, tau parameter for quantile loss (for "quantile_error" custom objective).}

\item{hyper_grid_domain_list}{List, domain list of hyperparameters for tuning.}

\item{tuning_method}{Character, method of hyperparameter tuning ("random_search", "grid_search", "bayesian_opt").}

\item{n_iter}{number of iterations for tuning.}

\item{k_iter}{number of iterations for k-fold cross-validation.}

\item{acq}{Character, acquisition function for Bayesian optimization ("ucb", "ei", "poi").}

\item{init_points}{Numeric, number of initial points for Bayesian optimization.}

\item{early_stop}{Numeric, number of epochs for early stopping (for "xgb" and "nn" algorithms).}

\item{keras_architecture_parameters}{List, containing units (numeric), n_layers (numeric between 1 and 5), activation_function and nn_optimizer ("Adam" or "RMSProp")}

\item{parallel}{Logical, whether to use parallel computation.}

\item{verbose}{Logical, whether to print verbose output.}
}
\value{
NULL. This function is used for validation and does not return a value; it stops on errors.
}
\description{
This function validates and checks various inputs required for a signal blending workflow.
}
\details{
This function performs comprehensive validation checks on various inputs required for a signal blending workflow.
It validates data formats, correctness of hyperparameters, consistency of dates, and other specific requirements for
different signal blending algorithms.
}
\references{
For more information on signal blending algorithms and their usage, refer to appropriate documentation.
}
\keyword{internal}
\keyword{machine-learning}
\keyword{validation}
