% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_inputs_ss_backtest.R
\name{check_inputs_ss_backtest}
\alias{check_inputs_ss_backtest}
\title{Perform validation checks on inputs for Signal Selection (SS) Workflow}
\usage{
check_inputs_ss_backtest(
  initial_sample_size,
  rebalancing_months,
  data_availability_cutoff,
  split_method,
  signals_m_df,
  chosen_signals,
  signal_positions,
  backtest_returns_df,
  benchmark_returns_df,
  market_factor_proxy,
  p_correction_method,
  signal_significance_threshold,
  enable_theme_representativeness,
  priors_m_df,
  user_priors,
  model_spec_theme_level,
  brms_control,
  prior_derivation_control,
  signal_themes_m_df,
  lower_quantile_winsorization,
  upper_quantile_winsorization
)
}
\arguments{
\item{initial_sample_size}{A numeric indicating the minimum number of observations required to begin the backtest.}

\item{rebalancing_months}{Months (numeric) when signal selection should be implemented.}

\item{data_availability_cutoff}{The minimum number of non-NA observations required for a backtest to be considered.}

\item{signals_m_df}{A (meta) data frame with columns including "id", "tickers", "dates", and the selected signals.}

\item{chosen_signals}{A vector with user-defined characteristics to be considered.}

\item{signal_positions}{A named vector with same length and names as chosen_signals describing whether positions should be taken "long" or "short".}

\item{backtest_returns_df}{A data frame with a 'dates' column and remaining columns named according to signals in signals_m_df, containing historical backtested returns.}

\item{p_correction_method}{The method for p-value correction. Possible options are:
\itemize{
\item{"none"}: No correction.
\item{"bayesian"}: When bayesian is set, a hierarchical mixed-effects bayesian linear model is fitted to the data, using the \code{brms} package,
which is an interface to the \code{Stan} probabilistic programming language.
The user can also choose one of the following frequentist methods, which will control Family-Wise Error Rate (FWER) or the False Discovery Rate (FDR).
FDR is less stringent than FWER.
For FWER, possible options are:
\item{"bonferroni"}: Bonferroni correction, which is dominated by Holm's method.
\item{"holm"}: Holm's (1979) method.
\item{"hochberg"}: Hochberg's (1988) method, valid when hypothesis tests are independent or non-negatively associated. Less powerful than Hommel's (1988) method, but
faster to compute.
\item{"hommel"}: Hommel's (1988) method, also valid when hypothesis tests are independent or non-negatively associated, but is more powerful than Hochberg (1988).
For FDR, possible options are:
\item{"BH" or "fdr"}: Benjamini-Hochberg (1995) procedure.
\item{"BY"}: Benjamini-Yekutieli (2001) procedure.
}}

\item{signal_significance_threshold}{A decimal indicating the hypothesis testing zero-alpha null-hypothesis rejection criteria. If one wants to select all chosen_signals,
provide 1. In any case, a signal being selected demands a significant CAPM alpha.}

\item{enable_theme_representativeness}{If TRUE, in case a given theme in \code{signal_themes_m_df} does not have any eligible signal, the signal
with highest alpha t-stat will be elected.}

\item{priors_m_df}{A (meta) data frame with columns including "id", "characteristic/signal", "dates", "theme" (used for clustering in hierarchical bayesian model)
and values for alpha (mean and se), beta (mean and se) and sigma, which are used to build priors.}

\item{model_spec_theme_level}{A character string specifying the desired Bayesian model structure.
Options include:
\itemize{
\item \code{"random_intercept"}: Includes random effects for the intercept at the theme level.
\item \code{"fixed_intercepts"}: Uses fixed intercepts for each theme.
\item \code{"fixed_intercepts_and_slopes"}: Includes fixed intercepts and slopes for each theme.
\item \code{"none"}: Omits theme-level intercepts but includes random effects at the theme:signal level.
}}

\item{prior_derivation_control}{A list of additional parameters to be passed to the \code{lme4::lmer} function:
\itemize{
\item{half_t_df} A numeric indicating the degrees of freedom in the half-t distribution to be applied in sd parameters.

\item{lmer_optimizer} A character string specifying the optimizer to be used in the \code{lme4::lmer} function.
It will be passed to lme4::lmerControl, which will be used in the \code{lme4::lmer} function.
Options include: 'nloptwrap', 'bobyqa', 'Nelder_Mead' or 'nlminbwrap'

\item{lmer_optimization_objective} A character string indicating whether estimates should be chosen to optimize the 'REML' criterion or the 'likelihood'.
}}

\item{signal_themes_m_df}{A (meta) data frame with id, tickers ("signals") and dates column contemplating all signals in \code{signals_m_df} and a "theme" column providing group membership for each signal, which is needed
for defining clusters in bayesian hierarchical model.}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization.}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization.}

\item{selected_benchmark_returns_df}{A data frame with a 'dates' column and a column with benchmark returns, named accordingly.}

\item{parallel}{Logical.
Indicates whether to enable parallel computation using the \code{future} package. Only avaialable for bayesian model. Default is \code{TRUE}.}
}
\description{
This function validates and checks various inputs required for signal selection in the context of run_ss_backtest function.
}
