% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_inputs_ss_backtest.R
\name{check_inputs_ss_backtest}
\alias{check_inputs_ss_backtest}
\title{Perform validation checks on inputs for Signal Selection (SS) Workflow}
\usage{
check_inputs_ss_backtest(
  initial_sample_size,
  rebalancing_months,
  data_availability_cutoff,
  split_method,
  signals_m_df,
  chosen_signals_and_positions,
  backtest_returns_xts,
  benchmark_returns_xts,
  market_factor_proxy,
  p_correction_method,
  signal_significance_threshold,
  enable_theme_representativeness,
  model_structure,
  theme_level_intercept,
  theme_level_slope,
  lmer_control,
  priors_m_df,
  user_priors,
  brms_control,
  prior_derivation_control,
  signal_themes_m_df,
  lower_quantile_winsorization,
  upper_quantile_winsorization
)
}
\arguments{
\item{initial_sample_size}{A numeric indicating the minimum number of observations required to begin the backtest.}

\item{rebalancing_months}{Months (numeric) when signal selection should be implemented.}

\item{data_availability_cutoff}{The minimum number of non-NA observations required for a backtest to be considered.}

\item{signals_m_df}{A (meta) data frame with columns including "id", "tickers", "dates", and the selected signals.}

\item{chosen_signals_and_positions}{A named vector indicating signals and their corresponding positions (long or short).
For example, chosen_signals_and_positions = c(book_yield = "long", vol_36m = "short").}

\item{backtest_returns_xts}{A xts containing historical backtested returns named according to signals in \code{signals_m_df},}

\item{benchmark_returns_xts}{A xts with benchmark returns, named accordingly.}

\item{p_correction_method}{The method for p-value correction. Possible options are:
\itemize{
\item{"none"}: No correction.
\item{"bayesian"}: When bayesian is set, a hierarchical mixed-effects bayesian linear model is fitted to the data, using the \code{brms} package,
which is an interface to the \code{Stan} probabilistic programming language.
The user can also choose one of the following frequentist methods, which will control Family-Wise Error Rate (FWER) or the False Discovery Rate (FDR).
FDR is less stringent than FWER.
For FWER, possible options are:
\item{"bonferroni"}: Bonferroni correction, which is dominated by Holm's method.
\item{"holm"}: Holm's (1979) method.
\item{"hochberg"}: Hochberg's (1988) method, valid when hypothesis tests are independent or non-negatively associated. Less powerful than Hommel's (1988) method, but
faster to compute.
\item{"hommel"}: Hommel's (1988) method, also valid when hypothesis tests are independent or non-negatively associated, but is more powerful than Hochberg (1988).
For FDR, possible options are:
\item{"BH" or "fdr"}: Benjamini-Hochberg (1995) procedure.
\item{"BY"}: Benjamini-Yekutieli (2001) procedure.
}}

\item{signal_significance_threshold}{A decimal indicating the hypothesis testing negative-alpha null-hypothesis rejection criteria. If one wants to select all signals,
provide 1. In any case, a signal being selected demands a significant CAPM alpha.}

\item{enable_theme_representativeness}{If TRUE, in case a given theme in \code{signal_themes_m_df} does not have any eligible signal, the signal
with highest alpha t-stat will be elected.}

\item{model_structure}{A character indicating whether the model should be estimated using a ´partial_pooled´ or ´no_pooled´ approach.}

\item{theme_level_intercept}{A character specifying the specification of effects of the intercept at the theme level.}

\item{theme_level_slope}{A character specifying the specification of effects of the slope at the theme level.}

\item{lmer_control}{Other additional parameters to be passed to \code{lme4::lmer} function.
\itemize{
#' \item{lmer_optimizer} A character string specifying the optimizer to be used in the
It will be passed to lme4::lmerControl, which will be used in the \code{lme4::lmer} function.
Options include: 'nloptwrap', 'bobyqa', 'Nelder_Mead' or 'nlminbwrap'

\item{lmer_optimization_objective} A character string indicating whether estimates should be chosen to optimize the 'REML' criterion or the 'likelihood'.
}}

\item{priors_m_df}{A (meta) data frame with columns including "id", "characteristic/signal", "dates", "theme" (used for clustering in hierarchical bayesian model)
and values for alpha (mean and se), beta (mean and se) and sigma, which are used to build priors.}

\item{prior_derivation_control}{A list of additional parameters to be passed to the \code{lme4::lmer} function:
\itemize{
\item{half_t_df} A numeric indicating the degrees of freedom in the half-t distribution to be applied in sd parameters.
}}

\item{signal_themes_m_df}{A (meta) data frame with id, tickers ("signals") and dates column contemplating all signals in \code{signals_m_df} and a "theme" column providing group membership for each signal, which is needed
for defining clusters in bayesian hierarchical model.}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization.}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization.}

\item{parallel}{Logical.
Indicates whether to enable parallel computation using the \code{future} package. Only avaialable for bayesian model. Default is \code{TRUE}.}
}
\description{
This function validates and checks various inputs required for signal selection in the context of run_ss_backtest function.
}
