% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score.R
\name{compute_score}
\alias{compute_score}
\title{Compute Score Based on Conditions}
\usage{
compute_score(features_m_df, conditions, feature_name, ...)
}
\arguments{
\item{features_m_df}{A \code{meta_dataframe} object.}

\item{conditions}{A named list of functions, where each function takes a column vector and returns \code{TRUE} or \code{FALSE}.}

\item{feature_name}{The name of the feature to be created.}

\item{min_non_na}{The minimum number of non-NA values required to compute the score.}
}
\value{
A \code{meta_dataframe} object with an additional \code{score} column indicating how many conditions are met per row.
}
\description{
This method computes a score by evaluating multiple user-defined conditions on a \code{meta_dataframe}.
Each condition corresponds to a column in the data and applies a function row-wise. If all conditions
for a row are met, it contributes \code{1}, otherwise \code{0}.
}
