% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{consolidate_generic_meta_dataframes}
\alias{consolidate_generic_meta_dataframes}
\title{Consolidate meta and base m_df}
\usage{
consolidate_generic_meta_dataframes(meta_generic_m_df, base_generic_m_df, type)
}
\description{
Consolidate meta and base m_df
}
