% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sb_backtest.R
\name{convert_oos_list_to_m_df}
\alias{convert_oos_list_to_m_df}
\title{Convert OOS Lists to Meta Data Frame}
\usage{
convert_oos_list_to_m_df(oos_obj_list)
}
\arguments{
\item{oos_obj_list}{A list of out-of-sample objects, where each object is a named numeric vector.}
}
\description{
Convert OOS Lists to Meta Data Frame
}
