% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-convert_oos_predictions_lists_to_m_df.R
\name{convert_oos_list_to_m_df}
\alias{convert_oos_list_to_m_df}
\title{Convert OOS Lists to Meta Data Frame}
\usage{
convert_oos_list_to_m_df(oos_obj_list)
}
\description{
Convert OOS Lists to Meta Data Frame
}
