% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_hyper_grid_domain}
\alias{create_hyper_grid_domain}
\title{Create a \code{hyper_grid_domain} Object}
\usage{
create_hyper_grid_domain(
  hyperparameter = NULL,
  tuning_method,
  ml_algorithm,
  grid = NULL,
  distribution_choice = NULL,
  pars = NULL,
  bounds = NULL
)
}
\arguments{
\item{hyperparameter}{A named list representing the hyperparameter to be added. Should be as follows:
\itemize{
\item \strong{glmnet}: alpha, lambda.min.ratio
\item \strong{rf}: mtry, num.trees, max.depth, min.bucket
\item \strong{xgb}: min_child_weight, max_depth, subsample, colsample_bytree, eta, gamma, nrounds
\item \strong{nn}: regularizer_l1, regularizer_l2, droprate, lr, size_of_batch, number_of_epochs
}}

\item{tuning_method}{Character string indicating the hyperparameter tuning method. Must be one of 'grid_search', 'random_search', or 'bayesian_opt'.}

\item{grid}{A numeric vector or list of numeric vectors for grid search values (only used for grid_search).}

\item{distribution_choice}{A character vector indicating the distribution to sample from (only used for random_search).}

\item{pars}{A numeric named vector or list of numeric named vectors specifying parameter values (only used for random_search).}

\item{bounds}{A vector of length 2 indicating minimum and maximum bounds for each hyperparameter (only used for bayesian_opt).}
}
\value{
An instance of the \code{hyper_grid_domain} S4 class.
}
\description{
This function creates an instance of the \code{hyper_grid_domain} S4 class.
It allows users to specify relevant hyperparameters.
}
\examples{
# Creating a hyper_grid_domain object for a xgb model
hyper_grid_xgb_random <- create_hyper_grid_domain(
 tuning_method = "random_search",
 ml_algorithm = "xgb",
 hyperparameter = c("min_child_weight", "max_depth", "subsample", "colsample_bytree",
                    "eta", "alpha", "gamma", "nrounds"),
 distribution_choice = c("constant", "uniform", "uniform", "uniform",
                         "uniform", "uniform", "constant", "uniform"),
 pars = list(3, c(min = 1L, max = 2L), c(min = 0.25, max = 0.50),
           c(min = 0.25, max = 0.50), c(min = 0.1, max = 0.2),
           c(min = 2, max = 5), 0, c(min = 200L, max = 500L))
)




}
