% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_keras_architecture}
\alias{create_keras_architecture}
\title{Create Keras Architecture}
\usage{
create_keras_architecture(
  nn_optimizer,
  units = NULL,
  activation = NULL,
  batch_norm_option = NULL
)
}
\arguments{
\item{nn_optimizer}{A character string specifying the optimizer to use (e.g., "adam").}

\item{units}{A numeric value for the number of units in the new layer.}

\item{activation}{A character string specifying the activation function for the new layer (e.g., "relu").}

\item{batch_norm_option}{A character string indicating whether to apply batch normalization for the new layer (e.g., "yes").}
}
\value{
An object of class \code{keras_architecture_parameters}.
}
\description{
Constructor for creating an instance of \code{keras_architecture_parameters}.
}
