% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_keras_architecture}
\alias{create_keras_architecture}
\title{Create Keras Architecture}
\usage{
create_keras_architecture(nn_optimizer, units, activation, batch_norm_option)
}
\arguments{
\item{nn_optimizer}{A character string specifying the optimizer to use (e.g., "adam").}
}
\value{
An object of class \code{keras_architecture_parameters}.
}
\description{
Constructor for creating an instance of \code{keras_architecture_parameters}.
}
