% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_meta_xts}
\alias{create_meta_xts}
\title{Create a meta_xts, assets_meta_xts or metrics_meta_xts object.}
\usage{
create_meta_xts(
  data,
  type = c("returns", "metrics"),
  asset_type = "not_identified",
  meta_xts_name = "not_identified",
  metric_name = NULL,
  workflow = NULL,
  source = NULL
)
}
\arguments{
\item{data}{An xts object containing your time series data.}

\item{type}{Character. Either \code{"assets"} (no holes allowed) or
\code{"metrics"} (holes allowed). Defaults to \code{c("assets","metrics")}
which means you must pick one.}

\item{meta_xts_name}{A character string. Defaults to \code{"default_name"}.}

\item{workflow}{An ANY object for workflow. Defaults to \code{NULL}.}

\item{source}{A character vector indicating data origin for each column.
If \code{NULL}, defaults to \code{"not_identified"} repeated for each column.}
}
\value{
An S4 object of class \code{assets_meta_xts} or \code{metrics_meta_xts},
depending on \code{type}.
}
\description{
This constructor automatically sets most slots for you based on an
input xts object and the desired type: "assets" or "metrics".
}
\examples{
\dontrun{
library(xts)
# Simple example data
dates <- seq(as.Date("2020-01-01"), as.Date("2020-01-05"), by = "day")
x <- matrix(rnorm(5 * 2), ncol = 2)
colnames(x) <- c("AssetA", "AssetB")
my_xts <- xts::xts(x, order.by = dates)

# Create an assets_meta_xts
obj_assets <- create_meta_xts(
  data = my_xts, type = "port",
  meta_xts_name = "My Assets Data"
)
validObject(obj_assets)

# Create a metrics_meta_xts
obj_metrics <- create_meta_xts(
  data = my_xts, type = "metrics",
  meta_xts_name = "My Metrics Data"
)
validObject(obj_metrics)
}

}
