% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_ml_backtest_config}
\alias{create_ml_backtest_config}
\title{Create ml_backtest_config Object}
\usage{
create_ml_backtest_config(
  target_fwd_name,
  ml_algorithm = "ols",
  tuning_strategy = NULL,
  custom_objective = "squared_error",
  keras_architecture_parameters = NULL,
  quantile_tau = 0.5,
  huber_delta = 1
)
}
\arguments{
\item{target_fwd_name}{Character string indicating the target variable's forward name.}

\item{ml_algorithm}{Character string specifying the machine learning algorithm to be used ('glmnet', 'rf', 'xgb', 'nn').}

\item{tuning_strategy}{An object of class tuning_strategy, specifying the strategy for tuning hyperparameters.}

\item{custom_objective}{Character string specifying the custom objective function ('squared_error', 'pseudo_huber_error', 'absolute_error') or NULL.}

\item{keras_architecture_parameters}{List or NULL, providing parameters specific to keras-based neural networks.}

\item{quantile_tau}{Numeric value indicating the tau parameter used for quantile regression, between 0 and 1.}

\item{huber_delta}{Numeric value greater than 0, specifying the delta parameter for Huber loss function.}
}
\value{
An ml_backtest_config object.
}
\description{
Constructs an ml_backtest_config object.
}
