% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_mvo_parameters}
\alias{create_mvo_parameters}
\title{Create MVO Parameters}
\usage{
create_mvo_parameters(
  opt_method = "random",
  random_ports_method = "sample",
  n_random_ports = 1000,
  opt_objective = "sharpe"
)
}
\arguments{
\item{opt_method}{A character indicating the optimization method.
The only current available method is 'random'. In this case, \code{n_random_ports} are
generated under the constraints defined in the \code{mvo_parameters} object and the one that
optimizes the \code{opt_objective} will be selected.}

\item{random_ports_method}{A character string representing the method that will be
passed to \code{PortfolioAnalytics::random_portfolios} to generate random portfolios.
Options are 'sample', 'simplex' or 'grid'.}

\item{n_random_ports}{Number of random portfolios to generate. Only needed when
\code{opt_method} is 'random'.}

\item{opt_objective}{A character indicating the optimization objective. Possible options
are 'return', 'risk' or 'sharpe'.}
}
\value{
An S4 object of class \code{mvo_parameters}.
}
\description{
Constructor function for creating an instance of \code{mvo_parameters}.
}
