% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_port_backtest_config}
\alias{create_port_backtest_config}
\title{Create port_backtest_config Object}
\usage{
create_port_backtest_config(
  chosen_score_metric_and_position = NULL,
  eligibility_quantile_range = c(0.9, 1),
  min_eligible_assets_fallback = NULL,
  chosen_scaler = NULL,
  scaler_shrinkage = NULL,
  use_raw_for_eligibility = NULL,
  enable_group_representativeness = NULL,
  selected_benchmark = NULL,
  initial_buffer_period,
  rebalancing_months,
  cov_est_method = NULL,
  port_construction_method = "ew",
  mvo_parameters = NULL,
  rp_parameters = NULL,
  hrp_parameters = NULL,
  mmaf_parameters = NULL,
  main_liquidity_metric,
  liquidity_floor_cutoffs = NULL,
  liquidity_constraint_policy = NULL,
  turnover_constraint_policy = NULL,
  concentration_constraint_policy = NULL,
  transaction_costs_parameters = NULL,
  config_name = "not_identified"
)
}
\arguments{
\item{chosen_score_metric_and_position}{An object (or named vector) specifying the expected return score metric and its associated position. Required if \code{sb_backtest_results} is not provided.}

\item{eligibility_quantile_range}{A numeric vector of length 2 (e.g., c(0.9, 1.0)) specifying the quantile range used to determine eligible assets.}

\item{min_eligible_assets_fallback}{A numeric value indicating the minimum number of eligible assets to include in the portfolio.}

\item{chosen_scaler}{An object of class \code{scaler} specifying the scaling variable to be applied to the scores.}

\item{scaler_shrinkage}{A numeric value between 0 and 1 indicating the shrinkage intensity for the scaler.}

\item{use_raw_for_eligibility}{A logical value indicating whether to use raw scores for determining eligibility.}

\item{enable_group_representativeness}{Logical; if TRUE, ensures at least one asset in all groups in groups_m_d_ref}

\item{selected_benchmark}{A character string indicating the benchmark to use for benchmark-relative backtests.}

\item{initial_buffer_period}{A numeric value indicating the number of initial dates to skip before starting the backtest.}

\item{rebalancing_months}{A numeric vector (e.g., c(3,6,9,12)) indicating the months when the portfolio should be rebalanced.}

\item{cov_est_method}{A \code{cov_est_method} object specifying the covariance estimation method and its parameters.
If not provided, a default using method "sample", sample size 36, active returns = TRUE, and the provided selected_benchmark is created.}

\item{port_construction_method}{A character string representing the portfolio construction method.
Must be one of "ew", "sw", "cw", "cs", "rp", or "mvo".}

\item{mvo_parameters}{An object of class \code{mvo_parameters} for mean-variance optimization. Only required if \code{port_construction_method} is "mvo".
If missing and port_construction_method is "mvo", a default is created.}

\item{rp_parameters}{An object of class \code{rp_parameters} for risk parity portfolios. Only required if \code{port_construction_method} is "rp".
If missing and port_construction_method is "rp", a default is created.}

\item{hrp_parameters}{An object of class \code{hrp_parameters} for hierarchical risk parity portfolios. Only required if \code{port_construction_method} is "hrp".}

\item{mmaf_parameters}{An object of class \code{mmaf_parameters} for micro-macro allocation framework portfolios. Only required if \code{port_construction_method} is "mmaf".}

\item{main_liquidity_metric}{A character string indicating which liquidity metric (i.e. column in liquidity_m_df) to use.}

\item{liquidity_floor_cutoffs}{An object (e.g., a data frame) containing liquidity cutoff values.}

\item{liquidity_constraint_policy}{An object of class \code{liquidity_constraint_policy} (optional).}

\item{turnover_constraint_policy}{An object of class \code{turnover_constraint_policy} (optional).}

\item{concentration_constraint_policy}{An object of class \code{concentration_constraint_policy} (optional).}

\item{transaction_costs_parameters}{An object specifying transaction cost parameters (optional).}

\item{config_name}{A character string representing the name of the configuration.}
}
\value{
An object of class \code{port_backtest_config}.
}
\description{
Constructs a \code{port_backtest_config} object containing all necessary parameters for backtesting stock-level portfolios.
}
