% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_risk_parity_portfolio.R
\name{create_risk_parity_portfolio}
\alias{create_risk_parity_portfolio}
\title{Create a risk parity portfolio}
\usage{
create_risk_parity_portfolio(
  universe_m_d_ref,
  covariance_matrix,
  rp_method = "cyclical-spinu",
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe with tickers, is_eligible and final_signal columns generated by filter_stock_universe}

\item{covariance_matrix}{The covariance matrix of all eligible stocks in universe_m_d_ref.}

\item{returns_xts_upd_ref}{A dataframe in which columns represent tickers present in universe_m_d_ref and row represent periods
It should include all stocks in current_stock_universe and a dates column with at least cov_matrix_sample_size periods before current_date}
}
\description{
Create a risk parity portfolio
}
