% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_risk_parity_portfolio.R
\name{create_risk_parity_portfolio}
\alias{create_risk_parity_portfolio}
\title{Create a risk parity portfolio}
\usage{
create_risk_parity_portfolio(
  universe_m_d_ref,
  returns_upd_ref,
  covariance_matrix_sample_size,
  covariance_estimation_method,
  groups_m_d_ref,
  verbose = TRUE
)
}
\arguments{
\item{universe_m_d_ref}{A dataframe with tickers, is_eligible and final_signal columns generated by filter_stock_universe}

\item{covariance_matrix_sample_size}{Number of periods to subset returns_d_ref sample when estimating the covariance_matrix. A high number will provide
higher degrees of freedom, but old returns might not reflect current risk due to parameter shift. A low number will tend to expose estimation
to dimensionality curse.}

\item{covariance_estimation_method}{One of SAM (Sample), EWMA, CC (Constant Correlation), PCA1, PCA2, Shrink_ID or Shrink_CC.
If NULL, blending_method can only be EW or IR}

\item{groups_m_d_ref}{A dataframe with id, tickers, dates and group classification columns to be used to fill NAs in the return sample}

\item{returns_d_ref}{A dataframe in which columns represent tickers present in universe_m_d_ref and row represent periods
It should include all stocks in current_stock_universe and a dates column with at least covariance_matrix_sample_size periods before current_date}
}
\description{
Create a risk parity portfolio
}
