% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_sb_backtest_config}
\alias{create_sb_backtest_config}
\title{Create sb_backtest_config Object}
\usage{
create_sb_backtest_config(
  sb_algorithm = "ols",
  target_fwd_name,
  tuning_strategy = NULL,
  training_sample_size,
  rebalancing_months,
  split_method = "expanding",
  chosen_signals_and_positions = NULL,
  custom_objective = NULL,
  keras_architecture_parameters = NULL,
  signal_port_parameters = NULL,
  quantile_tau = 0.5,
  huber_delta = 1,
  config_name = "not_identified"
)
}
\arguments{
\item{sb_algorithm}{Character string specifying the machine learning algorithm to be used ('glmnet', 'rf', 'xgb', 'nn', 'sw', 'ew', etc.).}

\item{target_fwd_name}{Name of the target variable in \code{target_m_df}.}

\item{tuning_strategy}{An object of class tuning_strategy, specifying the strategy for tuning hyperparameters.}

\item{training_sample_size}{Number of observations to include in each training sample.}

\item{rebalancing_months}{Months (numeric) when model should be rebalanced (refit).}

\item{split_method}{Character string indicating the data splitting method ('expanding' or 'rolling').}

\item{chosen_signals_and_positions}{A vector of chosen signals and positions (long, short)}

\item{custom_objective}{Character string specifying the custom objective function ('squared_error', 'pseudo_huber_error', 'absolute_error') or NULL.}

\item{keras_architecture_parameters}{An object of class \code{keras_architecture_parameters} providing parameters specific to keras-based neural networks.}

\item{signal_port_parameters}{An object of class \code{signal_port_parameters}, specifying the parameters for constructing signal portfolios (portfolio-blending).}

\item{quantile_tau}{Numeric value indicating the tau parameter used for quantile regression, between 0 and 1.}

\item{huber_delta}{Numeric value greater than 0, specifying the delta parameter for Huber loss function.}

\item{config_name}{Name of the backtest configuration.}
}
\value{
An sb_backtest_config object.
}
\description{
Constructs an sb_backtest_config object.
}
