% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_sb_backtest_config}
\alias{create_sb_backtest_config}
\title{Create sb_backtest_config Object}
\usage{
create_sb_backtest_config(
  sb_algorithm = "ols",
  tuning_strategy = NULL,
  training_sample_size,
  rebalancing_months,
  split_method = "expanding",
  ss_backtest_config = NULL,
  ss_backtest_results = NULL,
  custom_objective = "squared_error",
  keras_architecture_parameters = NULL,
  quantile_tau = 0.5,
  huber_delta = 1,
  config_name = "not_identified"
)
}
\arguments{
\item{sb_algorithm}{Character string specifying the machine learning algorithm to be used ('glmnet', 'rf', 'xgb', 'nn').}

\item{tuning_strategy}{An object of class tuning_strategy, specifying the strategy for tuning hyperparameters.}

\item{training_sample_size}{Number of observations to include in each training sample.}

\item{rebalancing_months}{Months (numeric) when model should be rebalanced (refit).}

\item{split_method}{Character string indicating the data splitting method ('expanding' or 'rolling').}

\item{ss_backtest_config}{An object of class \code{ss_backtest_config}, specifying the single strategy backtest configuration.}

\item{ss_backtest_results}{An object of class \code{ss_backtest_results}, containing the results of the single strategy backtest.}

\item{custom_objective}{Character string specifying the custom objective function ('squared_error', 'pseudo_huber_error', 'absolute_error') or NULL.}

\item{keras_architecture_parameters}{List or NULL, providing parameters specific to keras-based neural networks.}

\item{quantile_tau}{Numeric value indicating the tau parameter used for quantile regression, between 0 and 1.}

\item{huber_delta}{Numeric value greater than 0, specifying the delta parameter for Huber loss function.}

\item{config_name}{Name of the backtest configuration.}
}
\value{
An sb_backtest_config object.
}
\description{
Constructs an sb_backtest_config object.
}
