% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_sb_metabacktest_config}
\alias{create_sb_metabacktest_config}
\alias{create_sb_metabacktest_config,sb_backtest_config,character-method}
\title{Create SB Meta Backtest Configuration}
\usage{
create_sb_metabacktest_config(
  meta_sb_backtest_config,
  features_passthrough,
  ...
)

\S4method{create_sb_metabacktest_config}{sb_backtest_config,character}(
  meta_sb_backtest_config,
  features_passthrough = "none",
  config_name = "not_identified",
  normalize_base_predictions = TRUE,
  winsorize_base_predictions = TRUE,
  ...
)
}
\arguments{
\item{meta_sb_backtest_config}{A \code{sb_backtest_config} with the configuration for the meta learner.}

\item{features_passthrough}{A character vector of features to pass through to the meta-learner.}

\item{...}{Additional arguments (not used).}

\item{config_name}{Name of the backtest configuration.}

\item{base_sb_backtest_configs}{A list of \code{sb_backtest_config} objects with \code{tuning_strategy} set to \code{NULL}.}

\item{base_sb_backtest_results}{A list of \code{sb_backtest_results} objects.}

\item{base_ss_backtest_configs}{A list of \code{ss_backtest_config} objects (optional).}

\item{base_ss_backtest_results}{A list of \code{ss_backtest_results} objects (optional)}
}
\value{
A \code{sb_metabacktest_config} object containing all viable combinations of configs.

An \code{sb_metabacktest_config} object containing the provided sb_backtest objects.
}
\description{
The \code{create_sb_metabacktest_config} function creates an \code{sb_metabacktest_config} object by combining a \code{sb_backtest_config} for the meta-learner
and a group of \code{sb_backtest_config} objects for the base learners.
Those can be passed with our without a \code{ss_backtest_config} set. In this latter case, by passing
a list of \code{ss_backtest_config} objects, the function will generate all possible combinations of configurations for running multiple backtests,
possibly in parallel.
Alternatively, the user can provide a list of \code{sb_backtest_results} directly to be used by the function.
}
\section{Functions}{
\itemize{
\item \code{create_sb_metabacktest_config(
  meta_sb_backtest_config = sb_backtest_config,
  features_passthrough = character
)}: Create meta config from ss_backtest_results

}}
