% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_sb_metabacktest_config}
\alias{create_sb_metabacktest_config}
\alias{create_sb_metabacktest_config,sb_backtest_config,list-method}
\title{Create SB Meta Backtest Configuration}
\usage{
create_sb_metabacktest_config(
  meta_sb_backtest_config,
  base_sb_backtest_results,
  ...
)

\S4method{create_sb_metabacktest_config}{sb_backtest_config,list}(
  meta_sb_backtest_config,
  base_sb_backtest_results,
  config_name = "not_identified",
  features_passthrough = "none",
  normalize_base_predictions = TRUE,
  winsorize_base_predictions = TRUE,
  ...
)
}
\arguments{
\item{meta_sb_backtest_config}{A \code{sb_backtest_config} with the configuration for the meta learner.}

\item{base_sb_backtest_results}{A list of \code{sb_backtest_results} objects.}

\item{...}{Additional arguments (not used).}

\item{config_name}{Name of the backtest configuration.}

\item{base_sb_backtest_configs}{A list of \code{sb_backtest_config} objects with \code{tuning_strategy} set to \code{NULL}.}

\item{base_ss_backtest_configs}{A list of \code{ss_backtest_config} objects (optional).}

\item{base_ss_backtest_results}{A list of \code{ss_backtest_results} objects (optional)}
}
\value{
A \code{sb_metabacktest_config} object containing all viable combinations of configs.

An \code{sb_metabacktest_config} object containing the provided sb_backtest objects.
}
\description{
The \code{create_sb_metabacktest_config} function creates an \code{sb_metabacktest_config} object by combining a \code{sb_backtest_config} for the meta-learner
and a group of \code{sb_backtest_config} objects for the base learners.
Those can be passed with our without a \code{ss_backtest_config} set. In this latter case, by passing
a list of \code{ss_backtest_config} objects, the function will generate all possible combinations of configurations for running multiple backtests,
possibly in parallel.
Alternatively, the user can provide a list of \code{sb_backtest_results} directly to be used by the function.
}
\section{Functions}{
\itemize{
\item \code{create_sb_metabacktest_config(
  meta_sb_backtest_config = sb_backtest_config,
  base_sb_backtest_results = list
)}: Create meta config from ss_backtest_results

}}
\examples{
# Example usage:
# Assuming you have sb_backtest_config objects sb_config1, sb_config2
# and ss_backtest_config objects ss_config1, ss_config2

# First method: Combine sb_configs and ss_configs
meta_config <- create_sb_metabacktest_config(
  meta_sb_backtest_config = sb_backtest_config,
  base_sb_backtest_configs = list(sb_config1, sb_config2),
  ss_backtest_configs = list(ss_config1, ss_config2)
)

# Second method: Configs already have ss_backtest_config set
meta_config <- create_sb_metabacktest_config(
  meta_sb_backtest_config = sb_backtest_config,
  base_sb_backtest_configs = list(config1, config2)
)

}
\seealso{
\code{\link{sb_backtest_config}}, \code{\link{ss_backtest_config}}, \code{\link{sb_metabacktest_config}}
}
