% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_sb_metabacktest_results}
\alias{create_sb_metabacktest_results}
\alias{create_sb_metabacktest_results,sb_backtest_results,list,meta_dataframe,sb_metabacktest_config-method}
\alias{create_sb_metabacktest_results,list-method}
\title{Create an sb_metabacktest_results Object}
\usage{
create_sb_metabacktest_results(
  meta_sb_backtest_results,
  base_sb_backtest_results_list,
  oos_predictions_m_df,
  sb_metabacktest_config,
  ...
)

\S4method{create_sb_metabacktest_results}{sb_backtest_results,list,meta_dataframe,sb_metabacktest_config}(
  meta_sb_backtest_results,
  base_sb_backtest_results_list,
  oos_predictions_m_df,
  sb_metabacktest_config
)
}
\arguments{
\item{meta_sb_backtest_results}{A \code{sb_backtest_results} object for the meta learner}

\item{base_sb_backtest_results_list}{A named list of \code{sb_backtest_results} objects for the base learners.}
}
\value{
An object of class \code{sb_metabacktest_results}.
}
\description{
Constructs an \code{sb_metabacktest_results} object from a list of \code{sb_backtest_results} objects for base learners a single
\code{sb_backtest_results} object for the meta learner.
It computes consolidated and time series evaluation metrics for machine learning backtests.
}
