% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_ss_backtest_config}
\alias{create_ss_backtest_config}
\title{Create an ss_backtest_config Object}
\usage{
create_ss_backtest_config(
  data_availability_cutoff,
  initial_sample_size,
  rebalancing_months,
  active_returns = TRUE,
  split_method = "expanding",
  alpha_test_strategy = NULL,
  config_name = "not_identified",
  chosen_signals_and_positions = "all"
)
}
\arguments{
\item{data_availability_cutoff}{A numeric indicating the minimum number of non-NA observations required for a backtest.}

\item{initial_sample_size}{A numeric indicating the minimum number of observations required to begin the backtest.}

\item{rebalancing_months}{A numeric indicating the number of months for rebalancing.}

\item{active_returns}{Logical, whether to calculate active returns when calculating performance metrics, except for CAPM (default is TRUE).}

\item{split_method}{A character string specifying the splitting method, either "expanding" (default) or "rolling".}

\item{alpha_test_strategy}{An \code{alpha_test_strategy} object defining the alpha test configuration.}

\item{config_name}{A character string naming the configuration.}

\item{enable_theme_representativeness}{Logical, whether to enable theme representativeness (default is TRUE).}
}
\value{
An object of class \code{ss_backtest_config}.
}
\description{
This function constructs an object of class \code{ss_backtest_config}, ensuring the proper initialization
and validation of its slots.
}
\examples{
# Example usage:
config <- create_ss_backtest_config(
  data_availability_cutoff = 100,
  initial_sample_size = 200,
  rebalancing_months = 6,
  alpha_test_strategy = alpha_test_strategy_obj,
  config_name = "ExampleConfig"
)
}
