% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_target_m_df}
\alias{create_target_m_df}
\alias{create_target_m_df,meta_dataframe,meta_xts,meta_dataframe-method}
\title{Create a target_m_df from a meta_dataframe}
\usage{
create_target_m_df(
  daily_returns_m_df,
  daily_bench_returns_m_xts,
  features_m_df,
  ...
)

\S4method{create_target_m_df}{meta_dataframe,meta_xts,meta_dataframe}(
  daily_returns_m_df,
  daily_bench_returns_m_xts,
  features_m_df,
  past_ret_column,
  selected_bench,
  fwd_horizon,
  active_returns,
  parallel = TRUE
)
}
\arguments{
\item{daily_returns_m_df}{A \code{meta_dataframe} object containing daily stock returns data.}

\item{daily_bench_returns_m_xts}{A \code{meta_xts} object containing daily benchmark returns.}

\item{features_m_df}{A \code{meta_dataframe} object containing features data.}

\item{...}{Additional arguments to be passed to the method.}

\item{past_ret_column}{Character, indicating the column in \code{daily_returns_m_df} that contains returns.}

\item{selected_bench}{Character, indicating the column in \code{benchmark_returns_m_xts} to use as the benchmark.}

\item{fwd_horizon}{Integer, the number of months to compute forward returns}

\item{active_returns}{Logical, indicating whether to calculate active returns.}

\item{parallel}{Logical, indicating whether to use parallel processing.}
}
\value{
A \code{target_m_df} object containing future return metrics.
}
\description{
Create a target_m_df from a meta_dataframe
}
\details{
All ids from features_m_df are used to create the target_m_df.
The target_m_df is created by calculating the forward returns for each id in features_m_df.
The forward returns are calculated by shifting the returns in daily_returns_m_df by fwd_horizon months.
The benchmark returns are calculated by shifting the returns in daily_bench_returns_m_xts by fwd_horizon months.
The active returns are calculated by subtracting the benchmark returns from the forward returns.
}
