% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{create_tuning_strategy}
\alias{create_tuning_strategy}
\title{Hyperparameter Tuning Strategy Constructor}
\usage{
create_tuning_strategy(
  tuning_method,
  validation_sample_size,
  chosen_eval_metric,
  hyper_grid_domain = NULL,
  early_stop = NULL,
  n_iter = NULL,
  acq = NULL,
  init_points = NULL,
  k_iter = NULL
)
}
\arguments{
\item{tuning_method}{Character string indicating the hyperparameter tuning method. Must be one of 'grid_search', 'random_search', or 'bayesian_opt'.}

\item{validation_sample_size}{Numeric value representing the size of the validation sample.}

\item{chosen_eval_metric}{Character or NULL, specifying the evaluation metric to be optimized.}

\item{hyper_grid_domain}{An object of class \code{hyper_grid_domain}, representing the hyperparameter search space. If NULL, an empty hyper_grid_domain is created.}

\item{early_stop}{ANY, optional argument for halting criteria.}

\item{n_iter}{Numeric, number of iterations for 'random_search' or 'bayesian_opt'.}

\item{acq}{Character string for the acquisition function (for 'bayesian_opt' only).}

\item{init_points}{Numeric, number of initial random points for Bayesian optimization (for 'bayesian_opt' only).}

\item{k_iter}{Numeric, number of samples to evaluate during Bayesian optimization (for 'bayesian_opt' only).}
}
\value{
An object of class \code{grid_search_strategy}, \code{random_search_strategy}, or \code{bayesian_opt_strategy}, depending on the selected \code{tuning_method}.
}
\description{
A constructor function to create a tuning_strategy object, based on the specified tuning method.
}
