% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-explain_prediction.R
\name{decompose_feature_importance}
\alias{decompose_feature_importance}
\title{inner helper function to decompose feature importance}
\usage{
decompose_feature_importance(
  most_recent_meta_date,
  meta_feature_importance,
  base_identifiers,
  base_feature_importance_filtered
)
}
\description{
inner helper function to decompose feature importance
}
