% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-define_signal_eligibility.R
\name{define_signal_eligibility}
\alias{define_signal_eligibility}
\title{Define Signal Eligibility}
\usage{
define_signal_eligibility(
  selected_signals_backtest_returns_upd_ref,
  selected_benchmark_returns_upd_ref,
  signal_selection_policy,
  signals_groups_m_d_ref,
  selected_priors_informative_data_m_upd_ref,
  upper_quantile_winsorization,
  lower_quantile_winsorization
)
}
\arguments{
\item{selected_signals_backtest_returns_upd_ref}{A data frame containing backtest returns for various signals. The first column should be identifiers for the signals, and the subsequent columns should contain the returns data.}

\item{selected_benchmark_returns_upd_ref}{A data frame containing benchmark returns data. The first column should be identifiers for the benchmarks, and the subsequent columns should contain the returns data.}

\item{signal_selection_policy}{A list containing the signal selection policy. This list should include:
\describe{
\item{\code{data_availability_cutoff}}{The minimum number of non-NA observations required for a backtest to be considered.}
\item{\code{p_correction_method}}{The method for p-value correction ("none", holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "bayesian").}
\item{\code{benchmark}}{The benchmark to use for performance comparisons.}
\item{\code{chosen_prior}}{Dataset used to build prior for Bayesian adjustments.}
\item{\code{priors_type}}{A flag indicating whether priors should be set ("none" if not).}
\item{\code{chosen_sb_metric}}{The signal to use for final selection after adjustments.}
\item{\code{sb_benchmark_weighting}}{The benchmark weighting for the concentration constraint policy.}
\item{\code{max_abs_active_group_weight}}{The maximum absolute weight for any group in the concentration constraint policy.}
}}

\item{signals_groups_d_ref}{An optional data frame that maps signals to groups for Bayesian modeling and concentration constraints. It should contain a column named \code{theme} that categorizes the signals.}

\item{priors_m_d_ref}{An optional list of prior distributions for Bayesian adjustments. The list should contain prior data frames or objects indexed by names corresponding to \code{chosen_prior}.}
}
\value{
A data frame containing the updated signal universe with computed performance metrics, adjusted p-values, and final signal classifications. The columns include:
\describe{
\item{\code{tickers}}{The identifiers for the signals.}
\item{\code{mean_active_return}}{The mean active return of each signal.}
\item{\code{tracking_error}}{The tracking error of each signal.}
\item{\code{IR}}{The information ratio of each signal.}
\item{\code{alpha}}{The OLS CAPM alpha of each signal.}
\item{\code{AP}}{The t-statistic of the alpha.}
\item{\code{beta}}{The beta of each signal.}
\item{\code{treynor}}{The Treynor ratio of each signal.}
\item{\code{p_value}}{The p-value of the alpha.}
\item{\code{adjusted_p_value}}{The p-value adjusted for multiple comparisons, if applicable.}
\item{\code{final_signal}}{The final signal classification after applying transformations and adjustments.}
}
}
\description{
This function evaluates the eligibility of signals for inclusion in the investment universe based on various performance metrics and statistical adjustments. It performs initial data checks, computes performance metrics, applies statistical adjustments (both frequentist and Bayesian), and classifies signals according to the specified selection policy.
}
