% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-consolidate_meta_objects.R
\name{derive_adapted_custom_signal_universe_m_df}
\alias{derive_adapted_custom_signal_universe_m_df}
\title{Derive adapted_custom_signal_universe_metrics_m_df}
\usage{
derive_adapted_custom_signal_universe_m_df(
  meta_custom_objective,
  base_sb_backtest_results_list,
  meta_custom_signal_universe_metrics_m_df,
  base_custom_signal_universe_metrics_m_df
)
}
\description{
Derive adapted_custom_signal_universe_metrics_m_df
}
