% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fit_keras_model.R
\name{fit_keras_model}
\alias{fit_keras_model}
\title{Fit a Keras Neural Network Model}
\usage{
fit_keras_model(
  regularizer_l1,
  regularizer_l2,
  droprate,
  lr,
  number_of_epochs,
  size_of_batch,
  keras_architecture_parameters,
  early_stop = NULL,
  custom_objective_translated,
  huber_delta,
  features_matrix_train_clean,
  target_vector_train,
  verbose,
  ...
)
}
\arguments{
\item{regularizer_l1}{Numeric. L1 regularization parameter.}

\item{regularizer_l2}{Numeric. L2 regularization parameter.}

\item{droprate}{Numeric. Dropout rate.}

\item{lr}{Numeric. Learning rate.}

\item{number_of_epochs}{Integer. Maximum number of training epochs.}

\item{size_of_batch}{Integer. Batch size for training.}

\item{keras_architecture_parameters}{List, containing n_layers, units, activation, nn_optimizer and batch_norm_option}

\item{early_stop}{Integer or NULL. Number of epochs with no improvement to stop early, or NULL for no early stopping.}

\item{custom_objective_translated}{Custom objective in keras format}

\item{huber_delta}{Numeric. Delta parameter for Huber loss function.}

\item{features_matrix_train_clean}{Matrix. Training features matrix.}

\item{target_vector_train}{Vector. Training target vector.}

\item{verbose}{Integer. Verbosity level during training.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
A list containing:
\item{model_nn}{The Keras model object.}
\item{fit_nn}{The fitted Keras model object.}
}
\description{
This function fits a Keras neural network model based on the specified hyperparameters,
architecture choices, training configurations, and loss function parameters.
}
