% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_alpha_test_strategy}
\alias{get_alpha_test_strategy}
\title{Get alpha test strategy}
\usage{
get_alpha_test_strategy(object)
}
\arguments{
\item{object}{A \code{ss_backtest_config} object.}
}
\value{
A \code{alpha_test_strategy} S4 class.
}
\description{
Accessor function to retrieve alpha_test_strategy
}
