% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_concentration_constraint_policy}
\alias{get_concentration_constraint_policy}
\title{Accessor for Concentration Constraint Policy}
\usage{
get_concentration_constraint_policy(portfolio_policies_obj)
}
\arguments{
\item{portfolio_policies_obj}{A \code{portfolio_policies} object.}
}
\value{
The concentration constraint policy list.
}
\description{
Retrieves the concentration constraint policy from a \code{portfolio_policies} object.
}
