% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_liquidity_floor_cutoffs}
\alias{get_liquidity_floor_cutoffs}
\alias{get_liquidity_floor_cutoffs,port_backtest_config-method}
\title{Accessor for Liquidity Floor Cutoffs}
\usage{
get_liquidity_floor_cutoffs(port_backtest_config_obj)

\S4method{get_liquidity_floor_cutoffs}{port_backtest_config}(port_backtest_config_obj)
}
\arguments{
\item{port_backtest_config_obj}{A \code{port_backtest_config} object.}
}
\value{
The liquidity floor cutoffs list.
}
\description{
Retrieves the liquidity floor cutoffs from a \code{port_backtest_config} object.
}
