% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_priors_from_informative_data.R
\name{get_priors_from_informative_data}
\alias{get_priors_from_informative_data}
\title{Get Priors for Signals Based on Informative Data}
\usage{
get_priors_from_informative_data(priors_m_upd_ref, priors_type)
}
\arguments{
\item{priors_m_upd_ref}{A (meta) data frame with columns including "id", "characteristic/signal", "dates", "theme" (used for clustering in hierarchical bayesian model)
and values for alpha (mean and se), beta (mean and se) and sigma, which are used to build priors. It should contain data only for current date.}
}
\value{
A list, where each component contains \code{brms::set_prior} objects for each theme.
}
\description{
This function sets prior distributions for model parameters based on historical data, choosing appropriate priors for the intercept, beta coefficient, and optionally for variance components and correlations.
Hierarchical bayesian models can be built based on such priors. Signals are groups and are defined theme-wise.
The priors are determined using a Bayesian Information Criterion (BIC) approach to select the best-fitting distribution to reference historical data
}
\details{
The function performs the following steps:
\itemize{
\item Extracts the most recent dates and the unique themes from the data.
\item Computes summary statistics for \code{alpha}, \code{beta}, and \code{sigma} for each theme.
\item Chooses appropriate priors for \code{alpha}, \code{beta}, and optionally for \code{sigma} and variance components based on the chosen distributions.
\item Returns a list of \code{brms::set_prior} objects for each theme.
}
}
