% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_signal_selection_policy}
\alias{get_signal_selection_policy}
\title{Accessor for Signal Selection Policy}
\usage{
get_signal_selection_policy(port_backtest_config_obj)
}
\arguments{
\item{port_backtest_config_obj}{A \code{port_backtest_config} object.}
}
\value{
The signal selection policy list.
}
\description{
Retrieves the signal selection policy from a \code{port_backtest_config} object.
}
