% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{get_tuning_strategy}
\alias{get_tuning_strategy}
\alias{get_tuning_strategy,ml_backtest_config-method}
\title{Get Hyperparameter Tuning Strategy}
\usage{
get_tuning_strategy(object)

\S4method{get_tuning_strategy}{ml_backtest_config}(object)
}
\arguments{
\item{object}{An ml_backtest_config object.}
}
\value{
The \code{tuning_strategy} slot of the \code{ml_backtest_config} object.
}
\description{
Accessor function to retrieve the hyperparameter tuning strategy from an ml_backtest_config object.
}
