% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{keras_architecture_parameters-class}
\alias{keras_architecture_parameters-class}
\title{Keras Architecture Parameters}
\description{
Class to encapsulate parameters for constructing a Keras neural network architecture.
}
\section{Slots}{

\describe{
\item{\code{units}}{A numeric vector specifying the number of units (neurons) for each layer.}

\item{\code{n_layers}}{A numeric value representing the total number of layers in the model.}

\item{\code{activation}}{A character vector containing the activation functions for each layer.}

\item{\code{nn_optimizer}}{A character string indicating the optimization algorithm used (length = 1).}

\item{\code{batch_norm_option}}{A character vector specifying whether to apply batch normalization for each layer.}
}}

