% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{map_recipe_timewise}
\alias{map_recipe_timewise}
\alias{map_recipe_timewise,meta_dataframe,recipe-method}
\title{Map a Recipe to Sequential Dates}
\usage{
map_recipe_timewise(
  meta_dataframe,
  recipe,
  verbose = TRUE,
  parallel = TRUE,
  type = "signals"
)

\S4method{map_recipe_timewise}{meta_dataframe,recipe}(
  meta_dataframe,
  recipe,
  verbose = TRUE,
  parallel = TRUE,
  type = "signals"
)
}
\arguments{
\item{meta_dataframe}{A \code{meta_dataframe} object.}

\item{recipe}{A \code{recipe} object that contains a recipe in its \code{recipe} slot.
In special, make sure that:
\itemize{
\item The required meta columns (id, tickers, dates) are assigned the role \code{"id_vars"}.
\item All columns present in the \code{meta_dataframe} object have an assigned role in the recipe.
\item Target variables are preproessed separetely
}}

\item{verbose}{A logical indicating whether to print messages during the process. Default is \code{FALSE}.}
}
\value{
A time-wise preprocessed \code{meta_dataframe}.
}
\description{
This method maps recipes in a time-wise manner to a
\code{meta_dataframe} object using the \code{recipe} supplied.
For each date, the recipe is prepped and baked on the
subset of data available at the specific date. This ensures that no future
information is used during preprocessing.
}
\details{
Parallel processing is achieved using \code{furrr::future_map}; ensure that an appropriate
future plan is set (e.g., \code{future::plan(future::multisession)}).
}
