% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{meta_dataframe-class}
\alias{meta_dataframe-class}
\title{Define the \code{meta_dataframe} S4 Class}
\description{
This class represents a sb_backtest_workflow-enhanced data frame. It extends the functionality
of a standard data frame by including additional sb_backtest_workflow slots. The class is designed
to ensure that the input data frame adheres to specific structural requirements, including
unique identifiers, valid date formats, and unique column names.
}
\details{
The \code{meta_dataframe} class ensures that the data frame is structured correctly with the required columns,
and includes sb_backtest_workflow about the data. The \code{signals} slot holds the names of columns representing various signals.
The \code{unique_dates}, \code{unique_tickers}, and \code{n_obs} slots store the sb_backtest_workflow related to the number of unique dates,
tickers, and total observations respectively.
}
\section{Slots}{

\describe{
\item{\code{data}}{A \code{data.frame} containing the actual data.}

\item{\code{workflow}}{A \code{list} for storing sb_backtest_workflow about the data manipulation workflow.}

\item{\code{signals}}{A \code{character} vector containing the names of columns that represent signals.}

\item{\code{unique_dates}}{A \code{numeric} value representing the count of unique dates in the data.}

\item{\code{unique_tickers}}{A \code{numeric} value representing the count of unique tickers in the data.}

\item{\code{n_obs}}{A \code{numeric} value representing the total number of observations in the data.}
}}

\examples{
# Define a sample data frame
df <- data.frame(
  id = c("A-2024-01-01", "B-2024-02-01"),
  tickers = c("A", "B"),
  dates = as.Date(c("2024-01-01", "2024-02-01")),
  value = c(10, 20)
)

# Create a meta_dataframe object
meta_df <- create_meta_dataframe(df)

# Print the meta_dataframe object
print(meta_df)

}
