% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{ml_metabacktest_config-class}
\alias{ml_metabacktest_config-class}
\title{ml_metabacktest_config Class}
\description{
The ml_metabacktest_config class is designed to store and manage a collection of ml_backtest_config objects.
}
\section{Slots}{

\describe{
\item{\code{configs}}{A list of ml_backtest_config objects.}
}}

