% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{mmaf_sub_port_config-class}
\alias{mmaf_sub_port_config-class}
\title{MMAF Sub Portfolio Configuration}
\description{
An S4 class to represent the configuration of micro or macro portfolios
in the MMAF portfolio construction method.
}
\section{Slots}{

\describe{
\item{\code{port_construction_method}}{A character string indicating the method used for constructing micro portfolios.}

\item{\code{mvo_parameters}}{An object of class \code{mvo_parameters} representing the parameters for mean-variance optimization. This is only relevant for 'mvo'.}

\item{\code{rp_parameters}}{An object of class \code{rp_parameters} representing the parameters for risk parity. This is only relevant for 'rp'.}

\item{\code{hrp_parameters}}{An object of class \code{hrp_parameters} representing the parameters for hierarchical risk parity. This is only relevant for 'hrp'.}
}}

