% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,bayesian_opt_strategy,missing-method}
\alias{plot,bayesian_opt_strategy,missing-method}
\title{Plot Method for \code{bayesian_opt_strategy}}
\usage{
\S4method{plot}{bayesian_opt_strategy,missing}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{bayesian_opt_strategy}.}

\item{y}{Unused. Included for consistency with the generic \code{plot} method.}

\item{...}{Additional arguments passed to the plotting method (currently unused).}
}
\value{
A \code{ggplot} object visualizing the bounds.
}
\description{
Plot the bounds for each hyperparameter in \code{bayesian_opt_strategy}.
}
