% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,port_backtest_cohort,ANY-method}
\alias{plot,port_backtest_cohort,ANY-method}
\title{Plot Method for port_backtest_cohort Class}
\usage{
\S4method{plot}{port_backtest_cohort,ANY}(x, plot_id = NULL, vertical_lines = NULL)
}
\arguments{
\item{x}{An object of class \code{port_backtest_cohort}.}

\item{plot_id}{A character string or numeric value specifying which plot to display.}
}
\value{
Invisibly returns the input object.
}
\description{
Generates various plots to visualize metrics from the \code{port_backtest_cohort} object.
Users can select which plot to display by specifying the \code{plot_id} parameter.
}
