% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,port_backtest_results,ANY-method}
\alias{plot,port_backtest_results,ANY-method}
\title{Plot Method for port_backtest_results Class}
\usage{
\S4method{plot}{port_backtest_results,ANY}(
  x,
  plot_id = NULL,
  vertical_lines = NULL,
  palette = "cyberpunk",
  group_col = NULL,
  facet_by_year = NULL
)
}
\arguments{
\item{x}{An object of class \code{port_backtest_results}.}

\item{plot_id}{A character string or numeric value specifying which plot to display. If \code{NULL}, the user will be prompted.}

\item{vertical_lines}{Optional. A vector of \code{Date} objects indicating vertical lines to display in time-series plots (e.g., rebalance dates). If \code{NULL}, the user will be prompted.}

\item{palette}{A character string specifying the color palette to use for the plots. Options are "cyberpunk" (default) and "br". If an invalid option is provided, it will default to "cyberpunk".}

\item{group_col}{Optional. A character string specifying the column name in the backtest results to use for grouping in certain plots. If \code{NULL}, the user will be prompted to select from available columns.}

\item{facet_by_year}{Logical. If \code{TRUE}, time-series plots will be faceted by year. Default is \code{FALSE}.}
}
\value{
Invisibly returns the input object.
}
\description{
Generates various plots to visualize metrics from the \code{port_backtest_results} object.
Users can select which plot to display by specifying the \code{plot_id} parameter.
}
