% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,random_search_strategy,missing-method}
\alias{plot,random_search_strategy,missing-method}
\title{Plot Method for \code{random_search_strategy}}
\usage{
\S4method{plot}{random_search_strategy,missing}(x, y)
}
\arguments{
\item{x}{An object of class \code{random_search_strategy}.}

\item{y}{Unused. Included for consistency with the generic \code{plot} method.}
}
\value{
A \code{ggplot} object visualizing the hyperparameter distributions with possible limits.
}
\description{
Generate a sample for each hyperparameter and plot the distribution for random search.
}
