% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,sb_backtest_results,ANY-method}
\alias{plot,sb_backtest_results,ANY-method}
\title{Plot Signal Blending Walk-Forward Validation Results}
\usage{
\S4method{plot}{sb_backtest_results,ANY}(x, plot_id = NULL, features_m_df = NULL)
}
\arguments{
\item{x}{An object of class \code{sb_backtest_results} containing the results of the walk-forward validation.}

\item{plot_id}{A character string or numeric value specifying which plot to display.
\itemize{
\item By name: Options are:
\itemize{
\item \code{"Chosen Evaluation Metric Over Time"}
\item \code{"Test vs Validation Chosen Evaluation Metric Over Time"}
\item \code{"Best Hyperparameters Over Time"}
\item \code{"Hyperparameters vs Error"}
\item \code{"All Evaluation Metrics Over Time"}
\item \code{"Consolidated OOS Testing Metrics"}
\item \code{"Average Validation Metrics"}
\item \code{"Signal Portfolio"}
}
\item By number: Provide a number corresponding to the plot (as listed when \code{plot_id} is \code{NULL}).
If \code{NULL} (default), the method lists available plots.
}}

\item{features_m_df}{A \code{meta_dataframe} containing features used in the backtest. Required for plots like \code{"Explain Prediction"}.}
}
\value{
Invisibly returns the input \code{x}.
}
\description{
This method generates various plots to visualize the performance of machine learning models using walk-forward validation metrics.
Users can select which plot to display by specifying the \code{plot_id} parameter,
either by name or by number.
}
