% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,sb_model,missing-method}
\alias{plot,sb_model,missing-method}
\title{Plot Method for 'sb_model' Objects}
\usage{
\S4method{plot}{sb_model,missing}(x, type = NULL, ...)
}
\description{
This method generates plots for a \code{sb_model} object, depending on the \code{model} slot.
}
