% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot,ss_backtest_config,ANY-method}
\alias{plot,ss_backtest_config,ANY-method}
\title{Plot Priors from ss_backtest_config}
\usage{
\S4method{plot}{ss_backtest_config,ANY}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{ss_backtest_config} with a \code{bayesian_alpha_test_strategy} alpha test strategy.}

\item{y}{Not used. Included for S4 method compatibility.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Plots the distribution curves for each prior in the \code{bayesian_model_parameters} inside a \code{ss_backtest_config} object.
}
