% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{prep.step_impute_sector}
\alias{prep.step_impute_sector}
\title{Prepare method for step_impute_sector}
\usage{
\method{prep}{step_impute_sector}(x, training, info = NULL, ...)
}
\description{
Prepare method for step_impute_sector
}
