% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{prep.step_winsorize}
\alias{prep.step_winsorize}
\title{Prepare method for step_winsorize}
\usage{
prep.step_winsorize(x, training, info = NULL, ...)
}
\description{
Prepare method for step_winsorize
}
