% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pp_backtest.R
\name{print.step_impute_sector}
\alias{print.step_impute_sector}
\title{Print method for step_impute_sector}
\usage{
\method{print}{step_impute_sector}(x, width = max(20, options()$width - 30), ...)
}
\description{
Print method for step_impute_sector
}
