% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pp_backtest.R
\name{print.step_winsorize}
\alias{print.step_winsorize}
\title{Print method for step_winsorize}
\usage{
\method{print}{step_winsorize}(x, width = max(20, options()$width - 30), ...)
}
\description{
Print method for step_winsorize
}
