% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dataframe.R
\docType{class}
\name{priors_m_df-class}
\alias{priors_m_df-class}
\title{Define the priors_m_df S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to a priors meta_dataframe.
}
