% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{random_search_strategy-class}
\alias{random_search_strategy-class}
\title{Random Search Tuning Strategy}
\description{
A subclass of \code{tuning_strategy} that implements random search.
}
\section{Slots}{

\describe{
\item{\code{tuning_method}}{The tuning method is set to 'random_search'.}

\item{\code{n_iter}}{For random_search, it should be the number of random draws for each hyperparameter to which a distribution has been assigned.
Random samples of n_iter size will be generated for each hyperparameter and their unique values will be exhaustively combined.
Therefore, for n_iter = 5 and 2 hyperparameters, the ml algorithm validation error should be generally evaluated 5² = 25 times.
In case a constant vector is passed, the n_iter argument is not applied to this hyperparameter.}
}}

