% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{raw_features_m_df-class}
\alias{raw_features_m_df-class}
\title{Define the raw_features_m_df S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is a product of using create_meta_dataframe to list.
}
