% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tickers_catalog.R
\name{read_tickers_catalog}
\alias{read_tickers_catalog}
\title{Apply tickers_catalog Transformations}
\usage{
read_tickers_catalog(data, tickers_catalog, ...)
}
\arguments{
\item{data}{An object of class \code{raw_features_m_df} or \code{returns_meta_xts}.}

\item{tickers_catalog}{A \code{tickers_catalog} object containing reference data including
ticker status, trading range, and permanent IDs.}

\item{...}{Additional arguments passed to methods. Supports \code{verbose = TRUE/FALSE}.}
}
\value{
An object of the same class as \code{data}, modified according to the catalog rules.
}
\description{
Applies ticker ID mapping and filtering based on a \code{tickers_catalog} object. Can handle either
a \code{raw_features_m_df} or a \code{returns_meta_xts} object. This ensures consistency in ticker
identifiers and filters out data outside trading ranges or for untraded tickers.
}
