% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ml_backtest.R
\name{run_ml_backtest}
\alias{run_ml_backtest}
\alias{run_ml_backtest,meta_dataframe,meta_dataframe,ml_backtest_config,numeric,numeric,logical,logical-method}
\alias{run_ml_backtest,meta_dataframe,meta_dataframe,ml_metabacktest_config,numeric,numeric,logical,logical-method}
\title{Run Machine Learning Backtest}
\usage{
run_ml_backtest(
  features_m_df,
  target_m_df,
  config,
  training_sample_size,
  rebalancing_months,
  verbose,
  parallel,
  ...
)

\S4method{run_ml_backtest}{meta_dataframe,meta_dataframe,ml_backtest_config,numeric,numeric,logical,logical}(
  features_m_df,
  target_m_df,
  config,
  training_sample_size,
  rebalancing_months,
  verbose = TRUE,
  parallel = TRUE
)

\S4method{run_ml_backtest}{meta_dataframe,meta_dataframe,ml_metabacktest_config,numeric,numeric,logical,logical}(
  features_m_df,
  target_m_df,
  config,
  training_sample_size,
  rebalancing_months,
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{features_m_df}{A meta_dataframe containing features.}

\item{target_m_df}{A meta_dataframe containing target variable(s), with corresponding dates. Columns should follow the format XXXX_number_m, where
XXXX is the name of the target variable, number is the amount of forward periods and m indicates periods are measured in months.}

\item{config}{A \code{ml_metabacktest_config} object containing multiple configurations.}

\item{training_sample_size}{Number of observations to include in each training sample.}

\item{rebalancing_months}{Months (numeric) when model should be rebalanced (refit).}

\item{verbose}{Logical vector, indicating whether to print progress messages (default is TRUE).}

\item{parallel}{Logical vector, indicating whether to run  hyperparameter tuning in parallel (default is TRUE).}

\item{...}{Additional arguments (not used in this method).}
}
\value{
An object of class \code{ml_backtest_results} or \code{ml_metabacktest_results} with various outputs including model predictions, errors, and validation metrics.

A \code{ml_metabacktest_results} objects, one for each configuration.
}
\description{
The \code{run_ml_backtest} function performs out-of-sample testing for a range of machine learning algorithms using walk-forward time series validation.
It supports many ml algorithms and can handle different configurations for objective functions and tuning strategies through S4 objects.
The function divides the data into training, validation, and testing samples, and iteratively refits the model at specified rebalancing dates.
}
\details{
The function ensures all inputs are correctly formatted and performs checks to
validate the integrity of input data and parameters. It employs the glmnet package
for elastic net regularization, supporting both Lasso and Ridge regression.
}
\section{Methods (by class)}{
\itemize{
\item \code{run_ml_backtest(
  features_m_df = meta_dataframe,
  target_m_df = meta_dataframe,
  config = ml_backtest_config,
  training_sample_size = numeric,
  rebalancing_months = numeric,
  verbose = logical,
  parallel = logical
)}: Runs a backtest with a single configuration.

\item \code{run_ml_backtest(
  features_m_df = meta_dataframe,
  target_m_df = meta_dataframe,
  config = ml_metabacktest_config,
  training_sample_size = numeric,
  rebalancing_months = numeric,
  verbose = logical,
  parallel = logical
)}: Runs backtests with multiple configurations.

}}
\section{Running in Parallel}{

By default, tuning_method \%in\% c("random_search", "grid_search") utilizes furrr::future_pmap, which means they can run according to the built-in backends
from the future package. Therefore, if the user does not specify a different evaluation strategy with future::plan(),
tuning will be done sequentially by default (equivalent to future::plan(sequential)). In this case, however,
random number generator will be set to RNGkind("L'Ecuyer-CMRG"), instead of R default (RNGkind("Mersenne-Twister")), making results
not reproducible regarding using purrr:pmap(). In order to run using R's default random number generator, set parallel = FALSE.
Using a different evaluation strategy (e.g., future::plan(multisession)) will tune hyperparameters asynchronously (in parallel).

For tuning_method = "bayesian_opt", the ParBayesianOptimization::bayesOpt function runs in parallel by using foreach::foreach with the \%dopar\% operator.
Therefore, in this case, the user can either: (i) use doFuture::registerDoFuture(), in order to use the \%dofuture\% foreach adapter
(actually, in this case, doFuture::withDoRNG is used to turn \%dopar\% into \%dorng\% in order to use parallel-safe RNG), which allows
usage of backends from the future package or (ii) use parallel::makeCluster(), doParallel::registerDoParallel(), doParallel::clusterExport() and
doParallel::clusterEvalQ(), as exemplified by ParBayesianOptimization. If parallel = TRUE and neither strategy is being used,
code will result in error. Therefore, to run bayesian_opt synchronously, either use doFuture::registerDoFuture() with plan(sequential)
or set parallel = FALSE.

Keras has some limitations when working in parallel, especially when using bayesian optimization as tuning method.
}

\examples{
# Assuming features_m_df and target_m_df are meta_dataframe objects
# and config is an `ml_backtest_config` object
result <- run_ml_backtest(features_m_df, target_m_df, config, training_sample_size = 30, rebalacing_months = c(6,12), verbose = TRUE, parallel = TRUE)
# Assuming features_m_df and target_m_df are meta_dataframe objects
# and meta_config is an ml_metabacktest_config object
results_list <- run_ml_backtest(features_m_df, target_m_df, training_sample_size = 30, rebalacing_months = c(6,12), meta_config)
}
\seealso{
\code{\link{glmnet}}, \code{\link{ranger}}, \code{\link{xgboost}}, \code{\link{keras}}, \code{\link{time_series_split},
\code{\link{ml_backtest_config}}, \code{\link{ml_metabacktest_config}}
}
