% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_port_backtest.R
\name{run_port_backtest}
\alias{run_port_backtest}
\alias{run_port_backtest,meta_dataframe,meta_dataframe,meta_dataframe,meta_dataframe,port_backtest_config-method}
\title{Run Portfolio Backtest}
\usage{
run_port_backtest(
  signals_m_df,
  fwd_return_m_df,
  liquidity_m_df,
  volatility_m_df,
  config,
  ...
)

\S4method{run_port_backtest}{meta_dataframe,meta_dataframe,meta_dataframe,meta_dataframe,port_backtest_config}(
  signals_m_df,
  fwd_return_m_df,
  liquidity_m_df,
  volatility_m_df,
  config,
  target_m_df = NULL,
  port_backtest_cohort = NULL,
  backtest_returns_m_xts = NULL,
  benchmark_returns_m_xts = NULL,
  signal_themes_m_df = NULL,
  priors_m_df = NULL,
  custom_signal_weights_m_df = NULL,
  custom_signal_universe_metrics_m_df = NULL,
  gsm_algorithm = "ols",
  .test_seed = NULL,
  stock_groups_m_df = NULL,
  benchmark_weights_m_df = NULL,
  daily_stock_returns_m_xts = NULL,
  daily_bench_returns_m_xts = NULL,
  custom_stock_weights_m_df = NULL,
  custom_stock_metrics_m_df = NULL,
  user_defined_OR_rules_m_df = NULL,
  user_defined_AND_rules_m_df = NULL,
  winsorization_probs = c(0.025, 0.975),
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{signals_m_df}{A meta_dataframe containing the signal features. It must include at least the columns \code{id}, \code{tickers}, and \code{dates}.}

\item{fwd_return_m_df}{A meta_dataframe containing forward returns.}

\item{liquidity_m_df}{A meta_dataframe containing liquidity metrics.}

\item{volatility_m_df}{A meta_dataframe containing volatility metrics.}

\item{config}{An object of class \code{port_backtest_config} specifying the portfolio backtest configuration.}

\item{...}{Additional arguments (if needed).}

\item{parallel}{Logical; if \code{TRUE}, executes parts of the backtest in parallel (default is \code{TRUE}).}
}
\value{
An object of class \code{port_backtest_results} containing the portfolio backtest results.
}
\description{
The \code{run_port_backtest} function serves as a wrapper for the internal function
\code{run_port_backtest_internal}. It extracts the required parameters from a
\code{port_backtest_config} object, applies any additional pre‐processing to the inputs,
and then calls the internal function to perform the portfolio backtest.
}
\section{Methods (by class)}{
\itemize{
\item \code{run_port_backtest(
  signals_m_df = meta_dataframe,
  fwd_return_m_df = meta_dataframe,
  liquidity_m_df = meta_dataframe,
  volatility_m_df = meta_dataframe,
  config = port_backtest_config
)}: Runs a portfolio backtest using a \code{port_backtest_config} object.

This method extracts the parameters from the \code{config} object (of class \code{port_backtest_config}) and then calls
the internal function \code{run_port_backtest_internal} to perform the backtest.

}}
