% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_port_backtest.R
\name{run_port_backtest_internal}
\alias{run_port_backtest_internal}
\title{Backtest a porfolio based on signals derived from simple stock characteristics (simple factors) or
expected returns from ml model predictions}
\usage{
run_port_backtest_internal(
  signals_m_df,
  oos_predictions_m_df = NULL,
  exp_ret_score_metric = NULL,
  rebalancing_months,
  initial_sample_size,
  port_construction_method = "EW",
  stock_selection_quantile_range = c(0.9, 1),
  rp_method = "cyclical-spinu",
  n_random_ports = 2000,
  random_ports_method = "sample",
  opt_objective = "sharpe",
  opt_method = "random",
  cov_estimation_method = "sample",
  cov_matrix_sample_size = 252,
  active_returns = FALSE,
  cov_matrix_benchmark = NULL,
  returns_m_xts = NULL,
  benchmark_returns_m_xts = NULL,
  liquidity_constraint_policy,
  turnover_constraint_policy,
  concentration_constraint_policy,
  liquidity_m_df = NULL,
  liquidity_floor_cutoffs,
  main_liquidity_metric,
  stocks_groups_m_df = NULL,
  benchmark_weights_m_df = NULL,
  volatility_m_df = NULL,
  target_m_df,
  transaction_costs_list = NULL,
  custom_stock_universe_metrics_m_df = NULL,
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975,
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{signals_m_df}{A matrix or data frame containing simple features, with columns: id, tickers, dates,
to be used for signal-portfolio construction}

\item{oos_predictions_m_df}{A meta_dataframe object containing predictions from sb models, result of a run_sb_backtest call.}

\item{rebalancing_months}{Months (numeric) when portfolio should be rebalanced.}

\item{liquidity_m_df}{A data frame (similar to \code{features_m_df}) containing columns for id, tickers, dates,
and one or more market liquidity measures (e.g., inflation-adjusted mean financial volume).}

\item{nmonths_until_first_rebalancing_date}{The amount of dates in dates_m_vector to be skipped before backtest starts}

\item{selected_signal}{One of ml_model, characteristics or meta_signal IS THIS REALLY NECESSARY????????}

\item{chosen_characteristics}{A vector of characteristics that indicates which cols in features_m_df should be used as signals}

\item{chosen_ml_algorithms}{A vector of ml_algorithms that indicate which ml model predictions of ml_walk_forward_validation_results_list
should be used as signals. If NULL, all objects of ml_walk_forward_validation_results wil be considered}

\item{portfolio_construction_method}{The type of portfolio construction method for setting weights. Possibilities are heuristics (EW, CW, CS, SW),
Mean Tracking-Error (MTO) and Risk-Parity (RP)}

\item{covariance_estimation_method}{One of SAM (Sample), EWMA, CC (Constant Correlation), PCA1, PCA2, Shrink_ID or Shrink_CC.
If NULL, blending_method can only be EW or IR}

\item{custom_target_name}{A character indicating a custom target to be used in the backtest. Default is to use fwd_return_1m in \code{target_m_df}. If
a different variable is provided, the function will first seek that variable in columns of \code{target_m_df} and then in \code{signals_m_df}, in that order. If a variable
in target_m_df is chosen, the function will derive its prediction horizon based on naming convention and then name the time series accordingly. Otherwise,
ir will refer to current date.}
}
\description{
Backtest a porfolio based on signals derived from simple stock characteristics (simple factors) or
expected returns from ml model predictions
}
