% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sb_backtest.R
\name{run_sb_backtest}
\alias{run_sb_backtest}
\alias{run_sb_backtest,meta_dataframe,meta_dataframe,character,sb_backtest_config-method}
\alias{run_sb_backtest,meta_dataframe,meta_dataframe,character,sb_metabacktest_config-method}
\alias{run_sb_backtest_internal}
\title{Run Signal Blending Backtest}
\usage{
run_sb_backtest(features_m_df, target_m_df, target_fwd_name, config, ...)

\S4method{run_sb_backtest}{meta_dataframe,meta_dataframe,character,sb_backtest_config}(
  features_m_df,
  target_m_df,
  target_fwd_name,
  config,
  verbose = TRUE,
  parallel = TRUE
)

\S4method{run_sb_backtest}{meta_dataframe,meta_dataframe,character,sb_metabacktest_config}(
  features_m_df,
  target_m_df,
  target_fwd_name,
  config,
  verbose = TRUE,
  parallel = TRUE,
  winsorize_predictions = TRUE,
  winsorization_probs = c(0.025, 0.975),
  normalize_predictions = TRUE,
  features_passthrough = "none"
)

run_sb_backtest_internal(
  features_m_df,
  target_m_df,
  training_sample_size,
  target_fwd_name,
  validation_sample_size = 0,
  rebalancing_months,
  split_method = "expanding",
  sb_algorithm = "ols",
  custom_objective = "squared_error",
  chosen_eval_metric = NULL,
  huber_delta = 1,
  quantile_tau = 0.5,
  hyper_grid_domain_list = NULL,
  tuning_method = NULL,
  n_iter = NULL,
  k_iter = NULL,
  acq = "ucb",
  init_points = NULL,
  early_stop = NULL,
  keras_architecture_parameters = NULL,
  verbose = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{features_m_df}{A meta dataframe or data frame containing features, with columns: id, tickers, dates.}

\item{target_m_df}{A meta dataframe or data frame containing target variable(s), with corresponding dates. Columns should follow the format XXXX_number_m, where
XXXX is the name of the target variable, number is the amount of forward periods and m indicates periods are measured in months.}

\item{target_fwd_name}{Name of the target variable in \code{target_m_df}.}

\item{config}{A \code{sb_metabacktest_config} object containing a meta learner configuration and multiple individual configurations.}

\item{...}{Additional arguments (not used in this method).}

\item{verbose}{Logical, indicating whether to print progress messages (default is TRUE).}

\item{parallel}{Logical, indicating whether to run hyperparameter tuning in parallel (default is TRUE).}

\item{winsorize_predictions}{Logical; if \code{TRUE}, winsorizes the base learners' predictions before passing them to the meta learner. Default is \code{TRUE}.}

\item{winsorization_probs}{Numeric vector of length 2 specifying the lower and upper quantiles for winsorization. Default is \code{c(0.025, 0.975)}.}

\item{normalize_predictions}{Logical; if \code{TRUE}, normalizes the base learners' predictions before passing them to the meta learner. Default is \code{TRUE}.}

\item{features_passthrough}{A character vector indicating which features from \code{features_m_df} are to be passed through to the meta learner. Should correspond to column names in \code{features_m_df}.
Alternatively, if \code{'all'}, all features are passed through. If \code{'none'}, no features are passed through. Default is \code{'none'}.}

\item{training_sample_size}{Number of observations to include in each training sample.}

\item{validation_sample_size}{Number of observations to include in each validation sample. If provided a decimal, it will be considered as a percentage of the training sample size.}

\item{rebalancing_months}{Months (numeric) when model should be rebalanced.}

\item{split_method}{Choice of split method (expanding or rolling).}

\item{sb_algorithm}{Choice of sb_algorithm: ols (Ordinary Least Squares), glmnet (Elastic Net), rf (Random Forest), xgb (eXtreme Gradient Boosting), and nn (Keras Neural Networks).}

\item{custom_objective}{Custom objective (double differentiable loss function) for xgboost and nn algorithms.
(current options are squared_error, absolute_error and (pseudo)-huber loss)}

\item{chosen_eval_metric}{Metric to optimize during tuning: "rss", "rmse", "cp", "mae", "mphe", "mpe", "mape", "hr", and "mb".}

\item{huber_delta}{A single numeric value indicating the boundary that separates where the loss function turns from quadratic to linear.}

\item{quantile_tau}{A single numeric value indicating target quantile when calculating quantile loss.}

\item{tuning_method}{Method for hyperparameter tuning: "random_search", "grid_search", or "bayesian_opt".}

\item{n_iter}{Number of iterations.
For grid_search, the value does not make difference, as the number of times the ml algorithm validation error will be evaluated equals the exhaustive combination of unique hyperparameters values provided.
For random_search, it should be the number of random draws for each hyperparameter. Random samples of n_iter size will be generated for each hyperparameter and their unique values will be exhaustively combined.
Therefore, for n_iter = 5 and 2 hyperparameters, the ml algorithm validation error should be generally evaluated 5² = 25 times.
For bayesian_opt, it should be the number of times the ml algorithm will be evaluated after initialization.}

\item{k_iter}{Integer that specifies the number of times to sample eval_function at each Epoch during Bayesian optimization.
If running in parallel, set iters.k to a multiple of the number of cores. Must be lower and preferably a multiple of n_iter.}

\item{acq}{Acquisition function for Bayesian optimization: "ucb", "ei", or "poi".}

\item{init_points}{Number of initial random points for Bayesian optimization.}

\item{early_stop}{Sets a halting criteria to prevent overfitting in xgb and nn.}

\item{keras_architecture_parameters}{A named list containing parameters for configuring the Keras neural network architecture. It includes:
\itemize{
\item \strong{units}: A numeric vector specifying the number of neurons in each layer.
\item \strong{n_layers}: An integer indicating the total number of layers in the neural network.
\item \strong{activation}: A character vector listing the activation functions for each layer (e.g., "relu", "sigmoid", "tanh").
\item \strong{nn_optimizer}: A character string specifying the optimizer used for training the model (options: "Adam" or "RMSProp").
\item \strong{batch_norm_option}: A logical vector indicating whether batch normalization should be applied after each respective layer (TRUE or FALSE).
}}

\item{hyper_grid_domain}{A named list containing hyperparameter definitions. The structure of this list depends on the specified tuning method:
\itemize{
\item \strong{For grid search:} Must be a list of named vectors:
\item \strong{For random search:} Must be a list of named lists, where each named list contains:
\itemize{
\item \code{distribution_choice}: A character string specifying the distribution (one of "normal", "uniform", "lognormal", "constant").
\item \code{pars}: A named numeric vector of parameters corresponding to the chosen distribution.
\item \code{value}: A numeric value (only present if \code{distribution_choice} is "constant").
}
\item \strong{For Bayesian optimization:} Must be a list of named numeric vectors, each of length 2, representing the boundaries for the hyperparameters.
}}
}
\value{
An object of class \code{sb_backtest_results} or \code{sb_metabacktest_results} with various outputs including model predictions, errors, and validation metrics.

An \code{sb_metabacktest_results} object containing the results of the meta-backtest.

An object of class sb_backtest_results with various outputs including model predictions, errors, and validation metrics.
}
\description{
The \code{run_sb_backtest} function performs out-of-sample testing for a range of signal-blending algorithms (including machine-learning),
using walk-forward time series validation.
It supports many algorithms and can handle different configurations for objective functions and tuning strategies through S4 objects.
The function divides the data into training, validation, and testing samples, and iteratively refits the model at specified rebalancing dates.

Runs machine learning backtests for base learners and a meta learner using the provided configurations.
Users can provide their own base learner results or have the function compute them internally.
}
\details{
The function ensures all inputs are correctly formatted and performs checks to validate the integrity of input data and parameters. It employs the glmnet package
for elastic net regularization, supporting both Lasso and Ridge regression.

The function ensures all inputs are correctly formatted and performs checks to
validate the integrity of input data and parameters. It employs the glmnet package
for elastic net regularization, supporting both Lasso and Ridge regression.
}
\section{Methods (by class)}{
\itemize{
\item \code{run_sb_backtest(
  features_m_df = meta_dataframe,
  target_m_df = meta_dataframe,
  target_fwd_name = character,
  config = sb_backtest_config
)}: Runs a backtest with a single configuration.

\item \code{run_sb_backtest(
  features_m_df = meta_dataframe,
  target_m_df = meta_dataframe,
  target_fwd_name = character,
  config = sb_metabacktest_config
)}: Run Machine Learning Meta-Backtest

}}
\section{Functions}{
\itemize{
\item \code{run_sb_backtest_internal()}: Run ML Backtest
Perform out-of-sample testing for ML Algorithms with walk-forward time series validation

This function performs walk-forward validation for time series data using
a range of ML models. It supports hyperparameter tuning via random search, grid search,
or Bayesian optimization. The function divides the data into training, validation,
and testing samples, and iteratively refits the model at specified rebalancing dates.

}}
\section{Running in Parallel}{

The function supports parallel execution for runing multiple ml backtests and for hyperparameter tuning, both using the future package.

The method for \code{sb_metabacktest_config} is basically a wrapper for \code{sb_backtest_config} method, which is called iteratively for each configuration,
and possibly run in parallel.

By default, the method for \code{sb_metabacktest_config} and, individually, the method for \code{sb_backtest_config} when tuning_method \%in\% c("random_search", "grid_search"),
utilizes furrr::future_pmap, which means they can run according to the built-in backends from the future package. Therefore, if the user does not specify a different evaluation strategy with future::plan(),
tuning will be done sequentially by default (equivalent to future::plan(sequential)). In this case, however,
random number generator will be set to RNGkind("L'Ecuyer-CMRG"), instead of R default (RNGkind("Mersenne-Twister")), making results
not reproducible regarding using purrr:pmap(). In order to run using R's default random number generator, set parallel = FALSE.
Using a different evaluation strategy (e.g., future::plan(multisession)) will tune hyperparameters asynchronously (in parallel).

For tuning_method = "bayesian_opt", under the application in \code{sb_backtest_config} method, the ParBayesianOptimization::bayesOpt function runs in parallel by using foreach::foreach with the \%dopar\% operator.
Therefore, in this case, the user can either: (i) use doFuture::registerDoFuture(), in order to use the \%dofuture\% foreach adapter
(actually, in this case, doFuture::withDoRNG is used to turn \%dopar\% into \%dorng\% in order to use parallel-safe RNG), which allows
usage of backends from the future package or (ii) use parallel::makeCluster(), doParallel::registerDoParallel(), doParallel::clusterExport() and
doParallel::clusterEvalQ(), as exemplified by ParBayesianOptimization. If parallel = TRUE and neither strategy is being used,
code will result in error. Therefore, to run bayesian_opt synchronously, either use doFuture::registerDoFuture() with plan(sequential)
or set parallel = FALSE.

Keras has some limitations when working in parallel, especially when using bayesian optimization as tuning method.


By default, tuning_method \%in\% c("random_search", "grid_search") utilizes furrr::future_pmap, which means they can run according to the built-in backends
from the future package. Therefore, if the user does not specify a different evaluation strategy with future::plan(),
tuning will be done sequentially by default (equivalent to future::plan(sequential)). In this case, however,
random number generator will be set to RNGkind("L'Ecuyer-CMRG"), instead of R default (RNGkind("Mersenne-Twister")), making results
not reproducible regarding using purrr:pmap(). In order to run using R's default random number generator, set parallel = FALSE.
Using a different evaluation strategy (e.g., future::plan(multisession)) will tune hyperparameters asynchronously (in parallel).

For tuning_method = "bayesian_opt", the ParBayesianOptimization::bayesOpt function runs in parallel by using foreach::foreach with the \%dopar\% operator.
Therefore, in this case, the user can either: (i) use doFuture::registerDoFuture(), in order to use the \%dofuture\% foreach adapter
(actually, in this case, doFuture::withDoRNG is used to turn \%dopar\% into \%dorng\% in order to use parallel-safe RNG), which allows
usage of backends from the future package or (ii) use parallel::makeCluster(), doParallel::registerDoParallel(), doParallel::clusterExport() and
doParallel::clusterEvalQ(), as exemplified by ParBayesianOptimization. If parallel = TRUE and neither strategy is being used,
code will result in error. Therefore, to run bayesian_opt synchronously, either use doFuture::registerDoFuture() with plan(sequential)
or set parallel = FALSE.

Keras has some limitations when working in parallel, especially when using bayesian optimization as tuning method.
}

\examples{
# Assuming features_m_df and target_m_df are meta_dataframe objects
# and config is an `sb_backtest_config` object
result <- run_sb_backtest(features_m_df, target_m_df, config, training_sample_size = 30, rebalacing_months = c(6,12), verbose = TRUE, parallel = TRUE)
# Assuming features_m_df and target_m_df are meta_dataframe objects
# and meta_config is an sb_metabacktest_config object
results_list <- run_sb_backtest(features_m_df, target_m_df, training_sample_size = 30, rebalacing_months = c(6,12), meta_config)
# Example of creating hyper_grid_domain_list for grid search
hyper_grid <- list(
   alpha = c(0.2, 0.5),
   lambda.min.ratio = c(0.1, 0.5, 0.9)
   )

# Example of creating hyper_grid_domain_list for random search
hyper_grid <- list(
  alpha = list(distribution_choice = "uniform", pars = c(min = 0, max = 1), value = NULL),
  lambda.min.ratio = list(distribution_choice = "uniform", pars = c(min = 0, max = 0.9), value = NULL)
)

# Example of creating hyper_grid_domain_list for bayesian optimization
hyper_grid <- list(
   alpha = c(0.2, 0.9),
   lambda.min.ratio = c(0.1, 0.9)
   )

# Example of creating a 3 layers keras_architecture_parameters for neural networks.
keras_architecture_parameters = list(units = c(32,16,8), n_layers = 3, activation = c("relu", "relu", "relu"),
                                     nn_optimizer = "Adam", batch_norm_option = c(TRUE,TRUE,TRUE)
)
}
\seealso{
\code{\link{glmnet}}, \code{\link{ranger}}, \code{\link{xgboost}}, \code{\link{keras}}, \code{\link{time_series_split},
\code{\link{sb_backtest_config}}, \code{\link{sb_metabacktest_config}}

\code{\link{glmnet}}, \code{\link{ranger}}, \code{\link{xgboost}}, \code{\link{keras}}, \code{\link{time_series_split}}
}
