% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss_backtest.R
\name{run_ss_backtest}
\alias{run_ss_backtest}
\alias{run_ss_backtest,ss_backtest_config,meta_dataframe,missing,port_backtest_cohort,meta_xts,meta_dataframe-method}
\alias{run_ss_backtest,ss_backtest_config,meta_dataframe,meta_xts,missing,meta_xts,meta_dataframe-method}
\alias{run_ss_backtest_internal}
\title{Run a Signal Selection Backtest}
\usage{
run_ss_backtest(
  config,
  signals_m_df,
  backtest_returns_m_xts,
  port_backtest_cohort,
  benchmark_returns_m_xts,
  signal_themes_m_df,
  ...
)

\S4method{run_ss_backtest}{ss_backtest_config,meta_dataframe,missing,port_backtest_cohort,meta_xts,meta_dataframe}(
  config,
  signals_m_df,
  port_backtest_cohort,
  benchmark_returns_m_xts,
  signal_themes_m_df,
  priors_m_df = NULL,
  custom_signal_universe_metrics_m_df = NULL,
  verbose = TRUE,
  parallel = TRUE,
  winsorization_probs = c(0.025, 0.975)
)

\S4method{run_ss_backtest}{ss_backtest_config,meta_dataframe,meta_xts,missing,meta_xts,meta_dataframe}(
  config,
  signals_m_df,
  backtest_returns_m_xts,
  benchmark_returns_m_xts,
  signal_themes_m_df,
  priors_m_df = NULL,
  custom_signal_universe_metrics_m_df = NULL,
  verbose = TRUE,
  parallel = TRUE,
  winsorization_probs = c(0.025, 0.975)
)

run_ss_backtest_internal(
  initial_sample_size,
  rebalancing_months,
  split_method = "expanding",
  signals_m_df,
  chosen_signals_and_positions,
  custom_signal_universe_metrics_m_df = NULL,
  backtest_returns_m_xts,
  benchmark_returns_m_xts,
  market_factor_proxy = "IBOV",
  p_correction_method = "none",
  signal_significance_threshold = 0.05,
  enable_theme_representativeness = TRUE,
  model_structure = "no_pooled",
  theme_level_intercept = NULL,
  theme_level_slope = NULL,
  lmer_control = list(lmer_optimizer = "nloptwrap", lmer_optimization_objective = "REML",
    hierarchical_p_value_method = "Satterthwaite"),
  active_returns = TRUE,
  priors_m_df = NULL,
  user_priors = NULL,
  brms_control = list(iter = 2000, chains = 4, thin = 1, seed = NA, adapt_delta = 0.8),
  prior_derivation_control = list(half_t_df = 30),
  signal_themes_m_df,
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975,
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{config}{An object of class \code{ss_backtest_config} specifying the backtest configuration.}

\item{signals_m_df}{A (meta) data frame with columns including "id", "tickers", "dates", and the selected signals.}

\item{backtest_returns_m_xts}{A xts containing historical backtested returns named according to signals in \code{signals_m_df},}

\item{benchmark_returns_m_xts}{A xts with benchmark returns, named accordingly.}

\item{signal_themes_m_df}{A (meta) data frame with "id", "tickers" ("signals"), and "dates" columns, including all signals in \code{signals_m_df}, and a "theme" column providing group membership for each signal.}

\item{...}{Additional arguments (not used in this method).}

\item{priors_m_df}{A (meta) data frame with columns including "id", "ticker", "dates", "theme" (used for clustering in the Bayesian hierarchical model),
and values for active_return, bench_return, alpha (mean and se), beta (mean and se), and sigma. Data should be exogenous, as it will be used to set priors for the hierarchical Bayesian model.}

\item{verbose}{A boolean indicating whether to print messages.}

\item{parallel}{A boolean indicating whether to use parallel processing.}

\item{initial_sample_size}{A numeric indicating the minimum number of observations required to begin the backtest.}

\item{rebalancing_months}{Numeric months when signal selection should be implemented.}

\item{split_method}{The method used for splitting the data, either "expanding" or "rolling" (default is "expanding").}

\item{chosen_signals_and_positions}{A named vector indicating signals and their corresponding positions (long or short).
For example, chosen_signals_and_positions = c(book_yield = "long", vol_36m = "short").}

\item{p_correction_method}{The method for p-value correction. Possible options are:
\itemize{
\item \strong{"none"}: No correction.
\item \strong{"bayesian"}: A hierarchical mixed-effects Bayesian linear model is fitted to the data using the \code{brms} package.

The user can also choose one of the following frequentist methods, which control the Family-Wise Error Rate (FWER) or the False Discovery Rate (FDR). FDR is less stringent than FWER.

For FWER, possible options are:
\itemize{
\item \strong{"bonferroni"}: Bonferroni correction.
\item \strong{"holm"}: Holm's (1979) method.
\item \strong{"hochberg"}: Hochberg's (1988) method.
\item \strong{"hommel"}: Hommel's (1988) method.
}

For FDR, possible options are:
\itemize{
\item \strong{"BH"} or \strong{"fdr"}: Benjamini-Hochberg (1995) procedure.
\item \strong{"BY"}: Benjamini-Yekutieli (2001) procedure.
}
}}

\item{signal_significance_threshold}{A decimal indicating the hypothesis testing negative-alpha null-hypothesis rejection criteria. If you want to select all \code{chosen_signals}, provide 1.}

\item{enable_theme_representativeness}{If TRUE, in case a given theme in \code{signal_themes_m_df} does not have any eligible signal, the signal
with highest alpha t-stat will be elected.}

\item{model_structure}{A character indicating whether the model should be estimated using a ´partial_pooled´ or ´no_pooled´ approach.}

\item{theme_level_intercept}{A character specifying the specification of effects of the intercept at the theme level.}

\item{theme_level_slope}{A character specifying the specification of effects of the slope at the theme level.}

\item{lmer_control}{Other additional parameters to be passed to \code{lme4::lmer} function.
\itemize{
\item{lmer_optimizer} A character string specifying the optimizer to be used in the
It will be passed to lme4::lmerControl, which will be used in the \code{lme4::lmer} function.
Options include: 'nloptwrap', 'bobyqa', 'Nelder_Mead' or 'nlminbwrap'

\item{lmer_optimization_objective}: A character string indicating whether estimates should be chosen to optimize the 'REML' criterion or the 'likelihood'.

\item{hierarchical_p_value_method}: One of "Satterthwaite", "Kenward-Roger" and "lme4". Default is "Satterthwaite".
}}

\item{active_returns}{A character string indicating whether performance metrics should be calculated based on active returns or raw returns. If TRUE,
backtest_returns_m_xts will be adjusted by subtracting the selected market factor proxy in benchmark_returns_m_xts. This does not
impact calculation of the CAPM model, whether it is frequentist or bayesian, pooled or partial pooled.}

\item{user_priors}{An object with user-defined priors for the hierarchical Bayesian model, used when \code{priors_type = "user"}.
Each element of the list represents a theme and should contain priors set with brms::set_prior (class \code{brmsprior}). For instance, setting priors only for
alpha and beta global parameters can be done with:
c(brms::set_prior("normal(0,1)", class = "Intercept"), brms::set_prior("normal(0,1)", class = "b", coef = "bench_return"))}

\item{prior_derivation_control}{A list of additional parameters to be passed to the \code{lme4::lmer} function:
\itemize{
\item{half_t_df} A numeric indicating the degrees of freedom in the half-t distribution to be applied in sd parameters.
}}

\item{lower_quantile_winsorization}{Numeric value for lower winsorization.}

\item{upper_quantile_winsorization}{Numeric value for upper winsorization.}

\item{The}{minimum number of non-NA observations required for a backtest to be considered.}
}
\value{
A list containing eligible signals and signal universes throughout time.
}
\description{
This function performs iterative signal selection based on frequentist or Bayesian methods, which are applied to portfolio returns backtests to identify which signals can be considered significant in stock-level return prediction.

Runs signal selection backtest given bayesian or frequentist approaches for p-value correction. This acts as a wrapper for backtest_returns_m_xts

Runs signal selection backtest given bayesian or frequentist approaches for p-value correction. This acts as a wrapper for run_ss_backtest_internal
}
\details{
To determine whether a signal matters in cross-sectional predictability, the literature typically runs regressions of signal portfolios against a benchmark factor model (e.g., CAPM),
computing alphas and corresponding t-stats. Due to the large number of signals identified in the literature (the "factor zoo"), methods to control for multiple testing are often advocated.

The function performs the following operations:
\itemize{
\item Extracts the user-defined chosen signals from \code{signals_m_df} and subsets the data frame to include only these signals.
\item Checks for consistency in \code{chosen_signals_and_positions}.
\item Adjusts the signal positions based on whether they are "short" by multiplying their values by -1.
\item Updates column names in the data frame to reflect the corrected positions of the signals.
\item Validates that all adjusted signals have corresponding columns in \code{backtest_returns_m_xts}.
\item Calculates time-series metrics from backtested returns, including mean active return, IR, alpha, t-stat, beta, Treynor ratio, and p-value. Signals without adequate data (less than `` periods) are removed.
\item Adjusts p-values based on the method selected by the user.
\item Defines which signals are eligible for blending.
}
}
\section{Functions}{
\itemize{
\item \code{run_ss_backtest(
  config = ss_backtest_config,
  signals_m_df = meta_dataframe,
  backtest_returns_m_xts = missing,
  port_backtest_cohort = port_backtest_cohort,
  benchmark_returns_m_xts = meta_xts,
  signal_themes_m_df = meta_dataframe
)}: Run Signal Selection Backtest

\item \code{run_ss_backtest(
  config = ss_backtest_config,
  signals_m_df = meta_dataframe,
  backtest_returns_m_xts = meta_xts,
  port_backtest_cohort = missing,
  benchmark_returns_m_xts = meta_xts,
  signal_themes_m_df = meta_dataframe
)}: Run Signal Selection Backtest

\item \code{run_ss_backtest_internal()}: Run Signal Selection Backtest
Run a Signal Selection Backtest

This function performs iterative signal selection based on frequentist or Bayesian methods, which are applied to portfolio returns backtests to identify which signals can be considered significant in stock-level return prediction.

To determine whether a signal matters in cross-sectional predictability, the literature typically runs regressions of signal portfolios against a benchmark factor model (e.g., CAPM), computing alphas and corresponding t-stats. Due to the large number of signals identified in the literature (the "factor zoo"), methods to control for multiple testing are often advocated.

}}
