% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{sb_backtest_results-class}
\alias{sb_backtest_results-class}
\title{S4 Class for Time Series Walk-Forward Validation Results of Machine-Learning Models}
\value{
An S4 object of class \code{sb_backtest_results} containing all the specified results and sb_backtest_workflow.
}
\description{
This S4 class encapsulates the results and parameters from performing walk-forward
validation on time series data using machine learning algorithms. It includes
information about the model, data, tuning process, and performance metrics.
}
\section{Slots}{

\describe{
\item{\code{oos_sb_outputs_m_df}}{A meta dataframe containing out-of-sample predictions, target and errors, all indexed by testing dates.}

\item{\code{oos_testing_eval_metrics}}{A list of evaluation metrics for the out-of-sample testing samples.}

\item{\code{final_model}}{The final refitted machine learning model with best hyperparameters found after tuning. Possibly a object of sb_model S4 class.}

\item{\code{chosen_eval_metric_validation}}{A list of data.frames with the chosen evaluation metric calculated for the hyperparameter grid.}

\item{\code{best_hyperparameters_xts}}{A data frame containing the best hyperparameters selected during tuning for each rebalancing period.}

\item{\code{validation_eval_metrics_hyper_choice}}{All evaluation metrics calculated for the set of best hyperparameters.}

\item{\code{sb_backtest_workflow}}{A list containing sb_backtest_workflow about the walk-forward validation process.}
}}

