% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{sb_metabacktest_config-class}
\alias{sb_metabacktest_config-class}
\title{sb_metabacktest_config Class}
\description{
The sb_metabacktest_config class is designed to store and manage a collection of sb_backtest_config objects.
}
\section{Slots}{

\describe{
\item{\code{meta_sb_backtest_config}}{A \code{sb_backtest_config} with the configuration for the meta learner}

\item{\code{base_sb_backtest_configs}}{A list of \code{sb_backtest_config} objects whose oos predictions will be fed to the meta learner.}

\item{\code{base_sb_backtest_results}}{A list of \code{sb_backtest_result} objects whose oos predictions will be fed to the meta learner.}

\item{\code{config_name}}{A character string with the name of the configuration}
}}

