% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{sb_model_accessors}
\alias{sb_model_accessors}
\title{Accessor Methods for sb_model}
\usage{
get_sb_algorithm(object)
}
\arguments{
\item{object}{An object of class \code{sb_model}.}
}
\value{
The respective slot of the \code{sb_model} object.
}
\description{
These methods are used to access components of a \code{sb_model} object.
}
