% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_by_feature.R
\name{screen_by_feature}
\alias{screen_by_feature}
\title{Generic function for selecting features from a meta_dataframe}
\usage{
screen_by_feature(meta_dataframe, ...)
}
\arguments{
\item{meta_dataframe}{A \code{meta_dataframe} object.}

\item{...}{A list of feature names or tidyselect helpers to select columns.}
}
\description{
This is a generic function for selecting columns (features) from a \code{meta_dataframe} object.
}
