% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_eval_function.R
\name{set_eval_function}
\alias{set_eval_function}
\title{Helper function to run hyperparameter tuning}
\usage{
set_eval_function(ml_algorithm, tuning_method)
}
\arguments{
\item{ml_algorithm}{Choice of ml algorithm}

\item{tuning_method}{Choice of tuning method}
}
\value{
A eval_function to apply hyperparameter tuning
}
\description{
Helper function to run hyperparameter tuning
}
