% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-set_portfolio_weights.R
\name{set_portfolio_weights}
\alias{set_portfolio_weights}
\title{Set Portfolio Weights}
\usage{
set_portfolio_weights(
  universe_m_d_ref,
  port_construction_method,
  liquidity_m_d_ref = NULL,
  cap_weighting_metric = NULL,
  groups_m_d_ref = NULL,
  returns_xts_upd_ref = NULL,
  selected_benchmark_xts_upd_ref = NULL,
  active_returns = if (is.null(selected_benchmark_xts_upd_ref)) FALSE else TRUE,
  cov_estimation_method = "sample",
  cov_matrix_sample_size = if (is.null(returns_xts_upd_ref)) NULL else
    nrow(returns_xts_upd_ref),
  liquidity_constraint_policy = NULL,
  turnover_constraint_policy = NULL,
  concentration_constraint_policy = NULL,
  n_random_ports = 2000,
  random_ports_method = "sample",
  opt_objective = "sharpe",
  opt_method = "random",
  rp_method = "cyclical-spinu",
  lower_quantile_winsorization = 0.025,
  upper_quantile_winsorization = 0.975
)
}
\arguments{
\item{universe_m_d_ref}{A data frame containing the current universe of signals, including their associated metrics. The structure of this data frame depends on the portfolio construction method used.}

\item{port_construction_method}{A character string indicating the method to use for portfolio construction. Supported methods are:
\describe{
\item{\code{EW}}{Equal-Weighted Portfolio}
\item{\code{SW}}{Signal-Weighted Portfolio}
\item{\code{CW}}{Cap-Weighted Portfolio}
\item{\code{CS}}{Cap-Scaled Portfolio}
\item{\code{RP}}{Risk-Parity Portfolio}
\item{\code{MTO}}{Mean-Tracking Error Optimization}
}}

\item{liquidity_m_d_ref}{An optional data frame or matrix containing liquidity metrics for the signals, used for capitalization weighting and scaling. Defaults to \code{NULL}.}

\item{cap_weighting_metric}{An optional character string specifying the metric to use for capitalization weighting or scaling. Defaults to \code{NULL}.}

\item{groups_m_d_ref}{An optional data frame used for group constraints and covariance matrix estimation. Should include group information if used. Defaults to \code{NULL}.}

\item{cov_estimation_method}{An optional character string specifying the method for estimating the covariance matrix. Defaults to \code{NULL}.}

\item{liquidity_constraint_policy}{An optional list specifying the policy for liquidity constraints. Defaults to \code{NULL}.}

\item{turnover_constraint_policy}{An optional list specifying the policy for turnover constraints. Defaults to \code{NULL}.}

\item{n_random_ports}{An optional numeric value indicating the number of random portfolios to generate for optimization methods. Defaults to \code{NULL}.}

\item{random_ports_method}{An optional character string specifying the method for risk parity optimization. Defaults to \code{NULL}.}

\item{opt_objective}{Objective of mean-tracking error optimization. Defaults to \code{NULL}.}

\item{lower_quantile_winsorization}{An optional numeric value for lower quantile winsorization when handling cap weighting and scaling. Defaults to \code{NULL}.}

\item{upper_quantile_winsorization}{An optional numeric value for upper quantile winsorization when handling cap weighting and scaling. Defaults to \code{NULL}.}

\item{benchmark_weights_m_d_ref}{An optional data frame or matrix specifying the benchmark weights for setting sector constraints. Defaults to \code{NULL}.}
}
\value{
A data frame or object (depending on the portfolio construction method) with the updated portfolio weights assigned based on the specified method and constraints.
}
\description{
This function assigns weights to a portfolio based on the specified portfolio construction method. It supports various methods including equal weighting, signal weighting, capitalization weighting, capitalization scaling, risk parity, and mean-tracking error optimization. The function also accommodates additional constraints and policies related to liquidity, turnover, and benchmark weights.
}
