% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,port_backtest_results-method}
\alias{show,port_backtest_results-method}
\title{Show Port Backtest Results}
\usage{
\S4method{show}{port_backtest_results}(object)
}
\arguments{
\item{object}{An instance of the \code{port_backtest_results} class.}
}
\value{
The object is returned invisibly.
}
\description{
Displays a detailed summary of the \code{port_backtest_results} object,
including workflow metadata, portfolio construction details, date and stock information,
performance metrics (if available), and the final stock portfolio.
}
