% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,rp_parameters-method}
\alias{show,rp_parameters-method}
\title{Show Risk-Parity Parameters}
\usage{
\S4method{show}{rp_parameters}(object)
}
\arguments{
\item{object}{A \code{rp_parameters} object.}

\item{hide_title}{Logical indicating whether to hide the title when displaying the parameters. Default is FALSE.}
}
\description{
Displays the risk-parity configuration contained in a \code{rp_parameters} object.
}
