% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ss_backtest_config-method}
\alias{show,ss_backtest_config-method}
\title{Show Signal Selection Backtest Config}
\usage{
\S4method{show}{ss_backtest_config}(object)
}
\arguments{
\item{object}{An \code{ss_backtest_config} object to be displayed.}
}
\description{
Prints the contents of an \code{ss_backtest_config} object, detailing the various parameters and their configurations.
}
