% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{show,ss_backtest_results-method}
\alias{show,ss_backtest_results-method}
\title{Show Method for ss_backtest_results Class}
\usage{
\S4method{show}{ss_backtest_results}(object)
}
\arguments{
\item{object}{An instance of the \code{ss_backtest_results} class.}
}
\value{
The method returns the object invisibly.
}
\description{
This method displays a detailed summary of the most recent batch in the \code{ss_backtest_results} object's
\code{ss_backtest_workflow}, including: backtest configuration (config name, alpha test strategy parameters,
lmer control), p-value correction method (with Bayesian model parameters, if applicable), date/rebalancing
information, signals information, signal themes and priors information (if applicable), winsorization
parameters, and execution performance (elapsed time and timestamps).
}
