% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{signals_m_df-class}
\alias{signals_m_df-class}
\title{Define the signals_m_df S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to a signals meta_dataframe.

This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to a priros meta_dataframe.
}
