% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{ss_backtest_config-class}
\alias{ss_backtest_config-class}
\title{ss_backtest_config Class}
\description{
The ss_backtest_config class is designed to define an end-to-end signal selection experiment based on
backtest returns of associated strategies. The class includes parameters for manipulating the backtest returns object and
conducting hypothesis tests regarding CAPM alpha under a multiple testing framework, with frequentist and bayesian approaches. In the
latter, a hierarhical model is fit, with informative priors set according to an exogeneous dataset or by the user, or
default uninformative priors.
}
\section{Slots}{

\describe{
\item{\code{initial_sample_size}}{A numeric indicating the minimum number of observations required to begin the backtest.}

\item{\code{split_method}}{The method used for splitting the data, either "expanding" or "rolling" (default is "expanding").}

\item{\code{alpha_test_strategy}}{An \code{alpha_test_strategy} object with the configuration for the alpha test.}

\item{\code{chosen_signals_and_positions}}{A character indicating to which signals ss_backtest should be applied and their positions (long and short).
For example, chosen_signals_and_positions = c(book_yield = "long", vol_36m = "short").}
}}

