% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{step_impute_sector}
\alias{step_impute_sector}
\title{Prepare Step: step_impute_sector}
\usage{
step_impute_sector(
  recipe,
  ...,
  sector,
  method = "mean",
  role = NA,
  trained = FALSE,
  impute_values = NULL,
  skip = FALSE,
  id = recipes::rand_id("impute_sector")
)
}
\arguments{
\item{...}{Additional arguments passed to methods (not used).}

\item{x}{A \code{step_impute_sector} object.}

\item{training}{A data frame of training data that will be used to estimate the imputation values.}

\item{info}{An optional \code{term_info} object. Not used here but required by the S3 method signature.}
}
\value{
A \code{step_impute_sector} object with computed \code{impute_values}.
}
\description{
Prepares the \code{step_impute_sector} by computing mean or median values by sector for imputation.
This function is called automatically during \code{prep()} on a recipe.
}
