% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{step_impute_sector}
\alias{step_impute_sector}
\title{Prepare Step: step_impute_sector}
\usage{
step_impute_sector(
  recipe,
  ...,
  sector,
  method = "mean",
  role = NA,
  trained = FALSE,
  impute_values = NULL,
  skip = FALSE,
  id = recipes::rand_id("impute_sector")
)
}
\arguments{
\item{recipe}{A recipe object.}

\item{...}{Additional arguments passed to methods (not used).}

\item{sector}{A character string specifying the name of the sector variable.}

\item{method}{A character string specifying the imputation method. Either "mean" or "median".}

\item{role}{The role to assign to the new variables created by this step. Default is \code{NA}.}

\item{trained}{A logical indicating if the step has been trained. Default is \code{FALSE}.}

\item{impute_values}{A list of computed imputation values for each variable.}

\item{skip}{A logical indicating if the step should be skipped during baking. Default is \code{FALSE}.}

\item{id}{A character string for the step ID. Default is a random ID.}
}
\value{
A \code{step_impute_sector} object with computed \code{impute_values}.
}
\description{
Prepares the \code{step_impute_sector} by computing mean or median values by sector for imputation.
This function is called automatically during \code{prep()} on a recipe.
}
