% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pp_backtest.R
\name{step_winsorize_new}
\alias{step_winsorize_new}
\title{Constructor for step_winsorize}
\usage{
step_winsorize_new(terms, role, trained, probs, winsor_limits, skip, id)
}
\description{
Constructor for step_winsorize
}
