% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dataframe.R
\docType{class}
\name{stock_universe_m_df-class}
\alias{stock_universe_m_df-class}
\title{Define the signal_universe_meta_dataframe S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to the output of a signal selection backtest workflow.
}
\section{Slots}{

\describe{
\item{\code{universe_name}}{A \code{character} string describing the universe name.}

\item{\code{port_backtest_workflow}}{A \code{list} storing the ss_backtest_workflow that generated the signal_universe_meta_dataframe object.}
}}

