% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,meta_dataframe-method}
\alias{summary,meta_dataframe-method}
\title{Summary Method for meta_dataframe Class}
\usage{
\S4method{summary}{meta_dataframe}(object)
}
\arguments{
\item{object}{An instance of the \code{meta_dataframe} class.}
}
\value{
The method prints a summary to the console, including:
\itemize{
\item Number of observations and columns.
\item Data types of each column.
\item Missing values count per column.
\item Descriptive statistics for numeric columns.
\item Counts of unique values for categorical columns.
\item Summary statistics for tickers and date range.
\item Overall data frame summary.
}
}
\description{
This method provides a detailed summary of the \code{meta_dataframe} object,
including basic statistics of the data frame, information about missing
values, descriptive statistics for numeric columns, and counts for
categorical columns.
}
