% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary,meta_dataframe-method}
\alias{summary,meta_dataframe-method}
\title{Summary Method for meta_dataframe Class}
\usage{
\S4method{summary}{meta_dataframe}(object, summary_id = NULL)
}
\arguments{
\item{object}{An instance of the \code{meta_dataframe} class.}

\item{summary_id}{A character string or numeric value specifying which summary to display.
\itemize{
\item By name: Options are:
\itemize{
\item \code{"Numeric Summary Table"}
\item \code{"Tickers Frequency Table"}
\item \code{"Categorical Variables Plots"}
}
\item By number: Provide a number corresponding to the summary (as listed when \code{summary_id} is \code{NULL}).
If \code{NULL} (default), the method lists available summaries.
}}
}
\value{
Invisibly returns the input \code{object}.
}
\description{
Provides a detailed summary of a \code{meta_dataframe} object.
Users can select which summary to display by specifying the \code{summary_id} parameter,
either by name or by number.
The summary includes tables and plots styled accordingly.
}
