% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{target_m_df-class}
\alias{target_m_df-class}
\title{Define the tickers S4 Class}
\description{
This class inherits from \code{meta_dataframe} and enforces that the underlying data is adherent to a tickers meta_dataframe.
}
