% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{tickers_catalog-class}
\alias{tickers_catalog-class}
\title{tickers_catalog Class}
\description{
An S4 class to store stock metadata, including listing and delisting dates,
a unique identifier (\code{perm_id}), and classification flags for private and delisted stocks.
}
\section{Slots}{

\describe{
\item{\code{tickers}}{A character vector of stock tickers.}

\item{\code{perm_id}}{A character vector of unique stock identifiers, combining tickers and inception date.}

\item{\code{date_first_quote}}{A Date vector representing the first trading date for each stock.}

\item{\code{date_last_quote}}{A Date vector representing the last trading date for each stock.}

\item{\code{untraded}}{A character vector of untraded stocks (both dates are NA).}

\item{\code{delisted}}{A character vector of delisted stocks (date_last_quote < current_date and public).}

\item{\code{listed}}{A character vector of listed stocks (date_last_quote >= current_date).}

\item{\code{current_date}}{A Date object representing the most recent available date in the dataset.}

\item{\code{n_days_tolerance}}{A numeric value representing the number of days to consider a stock as delisted.}
}}

\examples{
showClass("tickers_catalog")

}
