% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_recipe_timewise.R
\name{tidy.step_impute_sector}
\alias{tidy.step_impute_sector}
\title{Tidy method for step_impute_sector}
\usage{
\method{tidy}{step_impute_sector}(x, ...)
}
\description{
Tidy method for step_impute_sector
}
