% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_split.R
\name{time_series_split}
\alias{time_series_split}
\title{Time Series Split}
\usage{
time_series_split(
  current_date,
  features_m_df,
  target_m_df,
  dates_m_vector,
  training_sample_size,
  validation_sample_size = 0,
  target_fwd,
  target_fwd_name,
  split_method = "expanding"
)
}
\arguments{
\item{current_date}{A single date in the format '\%Y-\%m-\%d'.}

\item{features_m_df}{A matrix or data frame containing features with columns 'id', 'tickers', and 'dates'.}

\item{target_m_df}{A matrix or data frame containing target variables.}

\item{dates_m_vector}{A vector of dates in the format '\%Y-\%m-\%d'.}

\item{training_sample_size}{Number of observations to include in the training sample.}

\item{validation_sample_size}{Number of observations to include in the validation sample (default is 0).}

\item{target_fwd}{The number of periods forward to forecast.}

\item{target_fwd_name}{The name of the target variable.}

\item{split_method}{The method used for splitting the data, either 'expanding' or 'rolling' (default is 'expanding').}
}
\value{
A list containing the training, validation, and refit samples.
}
\description{
Split time series data into training, validation, and refitting samples, in order to perform a walk-forward expanding or rolling time series validation.
}
