% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sb_backtest.R
\name{update_port_backtest}
\alias{update_port_backtest}
\title{Update Signal Blending Backtest
The \code{update_sb_backtest} function will take an existing \code{sb_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.}
\usage{
update_port_backtest(
  signals_m_df,
  fwd_return_m_df,
  liquidity_m_df,
  volatility_m_df,
  old_results,
  ...
)
}
\arguments{
\item{...}{Additional arguments (if needed).}

\item{features_m_df}{A meta_dataframe containing features.}

\item{target_m_df}{A meta_dataframe containing target variable(s), with corresponding dates. Columns should follow the format XXXX_number_m, where
XXXX is the name of the target variable, number is the amount of forward periods and m indicates periods are measured in months.}

\item{results}{An object of class \code{port_backtest_results} specifying the portfolio backtest results to be updated.}

\item{parallel}{Logical; if \code{TRUE}, executes parts of the backtest in parallel (default is \code{TRUE}).}
}
\value{
An object of class \code{port_backtest_results} containing the portfolio backtest results.
}
\description{
Update Signal Blending Backtest
The \code{update_sb_backtest} function will take an existing \code{sb_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.
}
