% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_port_backtest.R
\name{update_port_backtest}
\alias{update_port_backtest}
\alias{update_port_backtest,meta_dataframe,meta_dataframe,meta_dataframe,meta_dataframe,numeric,port_backtest_results,port_backtest_config-method}
\title{Update Portfolio Backtest
The \code{update_port_backtest} function will take an existing \code{port_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.}
\usage{
update_port_backtest(
  signals_m_df,
  fwd_return_m_df,
  liquidity_m_df,
  volatility_m_df,
  n_update,
  old_results,
  new_config,
  ...
)

\S4method{update_port_backtest}{meta_dataframe,meta_dataframe,meta_dataframe,meta_dataframe,numeric,port_backtest_results,port_backtest_config}(
  signals_m_df,
  fwd_return_m_df,
  liquidity_m_df,
  volatility_m_df,
  n_update,
  old_results,
  new_config,
  stock_groups_m_df = NULL,
  benchmark_weights_m_df = NULL,
  daily_stock_returns_m_xts = NULL,
  daily_bench_returns_m_xts = NULL,
  custom_stock_weights_m_df = NULL,
  custom_stock_metrics_m_df = NULL,
  user_defined_OR_rules_m_df = NULL,
  user_defined_AND_rules_m_df = NULL,
  winsorization_probs = c(0.025, 0.975),
  verbose = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{signals_m_df}{A meta_dataframe containing the signal features. It must include at least the columns \code{id}, \code{tickers}, and \code{dates}.}

\item{fwd_return_m_df}{A meta_dataframe containing forward returns.}

\item{liquidity_m_df}{A meta_dataframe containing liquidity metrics.}

\item{volatility_m_df}{A meta_dataframe containing volatility metrics.}

\item{n_update}{The expected number of periods to update}

\item{...}{Additional arguments (if needed).}

\item{parallel}{Logical; if \code{TRUE}, executes parts of the backtest in parallel (default is \code{TRUE}).}

\item{results}{An object of class \code{port_backtest_results} specifying the portfolio backtest results to be updated.}
}
\value{
An object of class \code{port_backtest_results} containing the portfolio backtest results.
}
\description{
Update Portfolio Backtest
The \code{update_port_backtest} function will take an existing \code{port_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.
}
\section{Methods (by class)}{
\itemize{
\item \code{update_port_backtest(
  signals_m_df = meta_dataframe,
  fwd_return_m_df = meta_dataframe,
  liquidity_m_df = meta_dataframe,
  volatility_m_df = meta_dataframe,
  n_update = numeric,
  old_results = port_backtest_results,
  new_config = port_backtest_config
)}: Updates a portfolio backtest using based on a \code{port_backtest_results} object.

This method extracts the parameters from the \code{results} object (of class \code{port_backtest_results}), modifies initial_buffer_period, performs the
new backtest and then binds results#'

}}
