% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_port_backtest.R
\name{update_port_backtest}
\alias{update_port_backtest}
\alias{update_port_backtest,meta_dataframe,meta_dataframe,meta_dataframe,meta_dataframe,port_backtest_results-method}
\title{Update Portfolio Backtest
The \code{update_port_backtest} function will take an existing \code{port_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.}
\usage{
update_port_backtest(
  signals_m_df,
  fwd_return_m_df,
  liquidity_m_df,
  volatility_m_df,
  old_results,
  ...
)

\S4method{update_port_backtest}{meta_dataframe,meta_dataframe,meta_dataframe,meta_dataframe,port_backtest_results}(
  signals_m_df,
  fwd_return_m_df,
  liquidity_m_df,
  volatility_m_df,
  old_results,
  updated_sb_backtest_results = NULL,
  stock_groups_m_df = NULL,
  benchmark_weights_m_df = NULL,
  daily_stock_returns_m_xts = NULL,
  daily_bench_returns_m_xts = NULL,
  benchmark_returns_m_xts = NULL,
  custom_stock_weights_m_df = NULL,
  custom_stock_metrics_m_df = NULL,
  user_defined_OR_rules_m_df = NULL,
  user_defined_AND_rules_m_df = NULL,
  verbose = TRUE,
  parallel = TRUE,
  .test_seed = NULL
)
}
\arguments{
\item{signals_m_df}{A meta_dataframe containing the signal features. It must include at least the columns \code{id}, \code{tickers}, and \code{dates}.}

\item{fwd_return_m_df}{A meta_dataframe containing forward returns.}

\item{liquidity_m_df}{A meta_dataframe containing liquidity metrics.}

\item{volatility_m_df}{A meta_dataframe containing volatility metrics.}

\item{old_results}{An object of class \code{port_backtest_results} specifying the portfolio backtest results to be updated.}

\item{...}{Additional arguments (if needed).}

\item{updated_sb_backtest_results}{An optional object of class \code{sb_backtest_results} or \code{sb_metabacktest_results} used to update the backtest using style-based results.}

\item{stock_groups_m_df}{A \code{meta_dataframe} containing stock group classifications, if applicable.}

\item{benchmark_weights_m_df}{A \code{meta_dataframe} with benchmark weights.}

\item{daily_stock_returns_m_xts}{An object of class \code{meta_xts} with daily stock returns, used in covariance estimation.}

\item{daily_bench_returns_m_xts}{An object of class \code{meta_xts} with daily benchmark returns, used in covariance estimation.}

\item{benchmark_returns_m_xts}{An object of class \code{meta_xts} containing benchmark returns over the rebalancing periods.}

\item{custom_stock_weights_m_df}{A \code{meta_dataframe} with user-defined stock weights for portfolio construction.}

\item{custom_stock_metrics_m_df}{A \code{meta_dataframe} with user-defined metrics to be used in portfolio rules.}

\item{user_defined_OR_rules_m_df}{A \code{meta_dataframe} specifying OR-based portfolio inclusion rules.}

\item{user_defined_AND_rules_m_df}{A \code{meta_dataframe} specifying AND-based portfolio inclusion rules.}

\item{verbose}{Logical; if \code{TRUE}, prints progress messages (default is \code{TRUE}).}

\item{parallel}{Logical; if \code{TRUE}, executes parts of the backtest in parallel (default is \code{TRUE}).}

\item{.test_seed}{Optional seed for reproducibility when testing.}
}
\value{
An object of class \code{port_backtest_results} containing the portfolio backtest results.
}
\description{
Update Portfolio Backtest
The \code{update_port_backtest} function will take an existing \code{port_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.
}
\section{Methods (by class)}{
\itemize{
\item \code{update_port_backtest(
  signals_m_df = meta_dataframe,
  fwd_return_m_df = meta_dataframe,
  liquidity_m_df = meta_dataframe,
  volatility_m_df = meta_dataframe,
  old_results = port_backtest_results
)}: Updates a portfolio backtest using based on a \code{port_backtest_results} object.

This method extracts the parameters from the \code{results} object (of class \code{port_backtest_results}), modifies initial_buffer_period, performs the
new backtest and then binds results

}}
