% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sb_backtest.R
\name{update_sb_backtest}
\alias{update_sb_backtest}
\alias{update_sb_backtest,meta_dataframe,meta_dataframe,sb_backtest_results-method}
\alias{update_sb_backtest,meta_dataframe,meta_dataframe,sb_metabacktest_results-method}
\title{Update Signal Blending Backtest}
\usage{
update_sb_backtest(features_m_df, target_m_df, old_results, ...)

\S4method{update_sb_backtest}{meta_dataframe,meta_dataframe,sb_backtest_results}(
  features_m_df,
  target_m_df,
  old_results,
  updated_ss_backtest_results = NULL,
  updated_port_backtest_cohort = NULL,
  updated_backtest_returns_m_xts = NULL,
  benchmark_returns_m_xts = NULL,
  signal_themes_m_df = NULL,
  custom_signal_weights_m_df = NULL,
  custom_signal_universe_metrics_m_df = NULL,
  verbose = TRUE,
  parallel = TRUE,
  .test_seed = NULL
)

\S4method{update_sb_backtest}{meta_dataframe,meta_dataframe,sb_metabacktest_results}(
  features_m_df,
  target_m_df,
  old_results,
  updated_base_sb_backtest_results,
  updated_base_port_backtest_cohort = NULL,
  updated_base_backtest_returns_m_xts = NULL,
  base_benchmark_returns_m_xts = NULL,
  base_signal_themes_m_df = NULL,
  base_priors_m_df = NULL,
  base_custom_signal_weights_m_df = NULL,
  base_custom_signal_universe_metrics_m_df = NULL,
  updated_meta_port_backtest_cohort = NULL,
  updated_meta_backtest_returns_m_xts = NULL,
  meta_benchmark_returns_m_xts = NULL,
  meta_signal_themes_m_df = NULL,
  meta_priors_m_df = NULL,
  meta_custom_signal_weights_m_df = NULL,
  meta_custom_signal_universe_metrics_m_df = NULL,
  verbose = TRUE,
  parallel = TRUE,
  .test_seed = NULL
)
}
\arguments{
\item{features_m_df}{A meta_dataframe containing features.}

\item{target_m_df}{A meta_dataframe containing target variable(s), with corresponding dates. Columns should follow the format XXXX_number_m, where
XXXX is the name of the target variable, number is the amount of forward periods and m indicates periods are measured in months.}

\item{old_results}{An object of class \code{sb_backtest_results} specifying the sb backtest results to be updated.}

\item{...}{Additional arguments (if needed).}

\item{updated_ss_backtest_results}{An optional object of class \code{ss_backtest_results}, used when the SB model depends on updated SS results.}

\item{updated_port_backtest_cohort}{An optional object of class \code{port_backtest_cohort} used to derive updated return and benchmark data.}

\item{updated_backtest_returns_m_xts}{A \code{meta_xts} object containing updated signal-based backtest returns.}

\item{benchmark_returns_m_xts}{A \code{meta_xts} object with benchmark returns over the backtest period.}

\item{signal_themes_m_df}{A \code{meta_dataframe} class object containing signal theme information for risk parity and MVO algorithms.}

\item{custom_signal_weights_m_df}{A \code{meta_dataframe} with user-defined signal weights.}

\item{custom_signal_universe_metrics_m_df}{A \code{meta_dataframe} with custom signal-level metrics for filtering or selection.}

\item{parallel}{Logical; if \code{TRUE}, executes parts of the backtest in parallel (default is \code{TRUE}).}

\item{updated_base_sb_backtest_results}{A list of \code{sb_backtest_results} used for updating the base learners in the SB meta-backtest.}

\item{updated_base_port_backtest_cohort}{An optional \code{port_backtest_cohort} object used for the base learners.}

\item{updated_base_backtest_returns_m_xts}{A \code{meta_xts} with signal-based returns for the base learners.}

\item{base_benchmark_returns_m_xts}{A \code{meta_xts} object with benchmark returns used for the base learners.}

\item{base_signal_themes_m_df}{A \code{meta_dataframe} with signal themes used in the base learners.}

\item{base_priors_m_df}{A \code{meta_dataframe} containing prior beliefs or constraints used for Bayesian learners in the base layer.}

\item{base_custom_signal_weights_m_df}{A \code{meta_dataframe} of custom signal weights for base learners.}

\item{base_custom_signal_universe_metrics_m_df}{A \code{meta_dataframe} of custom signal metrics for base learners.}

\item{updated_meta_port_backtest_cohort}{An optional \code{port_backtest_cohort} used in the meta learner layer.}

\item{updated_meta_backtest_returns_m_xts}{A \code{meta_xts} with returns for the meta learner.}

\item{meta_benchmark_returns_m_xts}{A \code{meta_xts} with benchmark returns for the meta learner.}

\item{meta_signal_themes_m_df}{A \code{meta_dataframe} with signal themes used in the meta learner.}

\item{meta_priors_m_df}{A \code{meta_dataframe} with priors used by the}
}
\value{
An object of class \code{sb_backtest_results} containing the sb backtest results.
}
\description{
The \code{update_sb_backtest} function will take an existing \code{sb_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.
}
\section{Methods (by class)}{
\itemize{
\item \code{update_sb_backtest(
  features_m_df = meta_dataframe,
  target_m_df = meta_dataframe,
  old_results = sb_backtest_results
)}: Updates a sb backtest using based on a \code{sb_backtest_results} object.

This method extracts the parameters from the \code{results} object (of class \code{sb_backtest_results}), modifies training_sample_size, performs the
new backtest and then binds results

\item \code{update_sb_backtest(
  features_m_df = meta_dataframe,
  target_m_df = meta_dataframe,
  old_results = sb_metabacktest_results
)}: Updates a sb backtest using based on a \code{sb_metabacktest_results} object.

This method extracts the parameters from the \code{results} object (of class \code{sb_metabacktest_results}), modifies training_sample_size, performs the
new backtest and then binds results

}}
