% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sb_backtest.R
\name{update_sb_backtest}
\alias{update_sb_backtest}
\alias{update_sb_backtest,meta_dataframe,meta_dataframe,sb_backtest_results-method}
\title{Update Signal Blending Backtest}
\usage{
update_sb_backtest(features_m_df, target_m_df, old_results, ...)

\S4method{update_sb_backtest}{meta_dataframe,meta_dataframe,sb_backtest_results}(
  features_m_df,
  target_m_df,
  old_results,
  updated_ss_backtest_results = NULL,
  backtest_returns_m_xts = NULL,
  benchmark_returns_m_xts = NULL,
  signal_themes_m_df = NULL,
  custom_signal_weights_m_df = NULL,
  custom_signal_universe_metrics_m_df = NULL,
  winsorization_probs = c(0.025, 0.975),
  gsm_algorithm = "ols",
  verbose = TRUE,
  parallel = TRUE,
  .test_seed = NULL
)
}
\arguments{
\item{features_m_df}{A meta_dataframe containing features.}

\item{target_m_df}{A meta_dataframe containing target variable(s), with corresponding dates. Columns should follow the format XXXX_number_m, where
XXXX is the name of the target variable, number is the amount of forward periods and m indicates periods are measured in months.}

\item{old_results}{An object of class \code{sb_backtest_results} specifying the sb backtest results to be updated.}

\item{...}{Additional arguments (if needed).}

\item{parallel}{Logical; if \code{TRUE}, executes parts of the backtest in parallel (default is \code{TRUE}).}
}
\value{
An object of class \code{sb_backtest_results} containing the sb backtest results.
}
\description{
The \code{update_sb_backtest} function will take an existing \code{sb_backtest_results} object and update it with
new dates. This function is useful when you want to add new dates to an existing backtest without having to re-run the entire backtest.
}
\section{Methods (by class)}{
\itemize{
\item \code{update_sb_backtest(
  features_m_df = meta_dataframe,
  target_m_df = meta_dataframe,
  old_results = sb_backtest_results
)}: Updates a sb backtest using based on a \code{sb_backtest_results} object.

This method extracts the parameters from the \code{results} object (of class \code{sb_backtest_results}), modifies training_sample_size, performs the
new backtest and then binds results

}}
