% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_transaction_costs_parameters}
\alias{validate_transaction_costs_parameters}
\title{Validate Transaction Cost Parameters}
\usage{
validate_transaction_costs_parameters(transaction_costs_parameters)
}
\arguments{
\item{transaction_costs_parameters}{A list containing transaction cost parameters.}
}
\value{
TRUE if all validations pass; otherwise, the function stops with an error.
}
\description{
This function validates a list of transaction cost parameters based on the BARRA model.
It checks that the list contains the names: "direct_transaction_cost", "strategy_aum",
"alpha", and "lambda". Additionally, it ensures that:
\itemize{
\item \code{direct_transaction_cost} is a single numeric value and positive.
\item \code{strategy_aum} is a single numeric value and positive.
\item \code{alpha} is a single numeric value and positive.
\item \code{lambda} is a single value that is either numeric or exactly the string "dynamic".
}
}
